### 2023-07-02 #596 @gilbertocamara
SITS: Satellite Image Time Series Analysis for Earth Observation Data Cubes (submit R package for review)
Submitting Author Name: Gilberto Camara
Submitting Author Github Handle: @gilbertocamara
Other Package Authors Github handles:  @rolfsimoes, @OldLipe, @M3nin0, @lorenalves
Repository:  https://github.com/e-sensing/sits
Version submitted: 1.4.2
Submission type: Standard
Reviewers: @mdsumner, @loreabad6, @nowosad, @edzer, @alexgleith
Version accepted: TBD
---
```
Package: sits
Version: 1.5.3
Title: Satellite Image Time Series Analysis for Earth Observation Data Cubes
Authors@R: c(person('Rolf', 'Simoes', role = c('aut'), email = 'rolfsimoes@gmail.com'),
             person('Gilberto', 'Camara', role = c('aut', 'cre', 'ths'), email = 'gilberto.camara.inpe@gmail.com'),
             person('Felipe', 'Souza', role = c('aut'), email = 'felipe.carvalho@inpe.br'),
             person('Felipe', 'Carlos', role = c('aut'), email = "efelipecarlos@gmail.com"),
             person('Lorena', 'Santos', role = c('aut'), email = 'lorena.santos@inpe.br')
             )
Maintainer: Gilberto Camara <gilberto.camara.inpe@gmail.com>
Description: An end-to-end toolkit for land use and land cover classification
    using big Earth observation data. Builds satellite image data cubes from cloud collections.
    Supports visualization methods for images and time series and
    smoothing filters for dealing with noisy time series.
    Includes functions for quality assessment of training samples using self-organized maps and
    to reduce training samples imbalance. Provides machine learning algorithms including support vector machines,
    random forests, extreme gradient boosting, multi-layer perceptrons,
    temporal convolution neural networks, and temporal attention encoders.
    Performs efficient classification of big Earth observation data cubes and includes
    functions for post-classification smoothing based on Bayesian inference.
    Enables best practices for estimating area and assessing accuracy of land change.
    Minimum recommended requirements: 16 GB RAM and 4 CPU dual-core.
    yaml (>= 2.3.0),
    grDevices,
    graphics,
    leafgl,
    leaflet (>= 2.2.2),
    luz (>= 0.4.0),
    parallel,
    purrr (>= 1.0.2),
    randomForest,
    Rcpp (>= 1.0.13),
    rstac (>= 1.0.1),
    sf (>= 1.0-19),
    slider (>= 0.2.0),
    terra (>= 1.8-5),
    tibble (>= 3.1),
    tidyr (>= 1.3.0),
    tmap (>= 4.0),
    torch (>= 0.14.0),
    units,
    utils
    aws.s3,
    caret,
    cli,
    cols4all (>= 0.8.0),
    covr,
    dendextend,
    dtwclust,
    DiagrammeR,
    digest,
    e1071,
    exactextractr,
    FNN,
    gdalcubes (>= 0.7.0),
    geojsonsf,
    httr2 (>= 1.1.0),
    kohonen (>= 3.0.11),
    mgcv,
    nnet,
    openxlsx,
    proxy,
    randomForestExplainer,
    RColorBrewer,
    RcppArmadillo (>= 0.12),
    scales,
    spdep,
    stars,
    supercells (>= 1.0.0),
    tools,
    xgboost
Config/testthat/parallel: false
Config/testthat/start-first: cube, raster, regularize, data, ml
    Rcpp,
    RcppArmadillo
Collate:
    'api_accessors.R'
    'api_accuracy.R'
    'api_apply.R'
    'api_band.R'
    'api_bayts.R'
    'api_bbox.R'
    'api_block.R'
    'api_check.R'
    'api_chunks.R'
    'api_classify.R'
    'api_clean.R'
    'api_cluster.R'
    'api_colors.R'
    'api_combine_predictions.R'
    'api_comp.R'
    'api_conf.R'
    'api_crop.R'
    'api_csv.R'
    'api_cube.R'
    'api_data.R'
    'api_debug.R'
    'api_detect_change.R'
    'api_download.R'
    'api_dtw.R'
    'api_environment.R'
    'api_factory.R'
    'api_file_info.R'
    'api_file.R'
    'api_gdal.R'
    'api_gdalcubes.R'
    'api_grid.R'
    'api_jobs.R'
    'api_kohonen.R'
    'api_label_class.R'
    'api_mask.R'
    'api_merge.R'
    'api_mixture_model.R'
    'api_ml_model.R'
    'api_mosaic.R'
    'api_opensearch.R'
    'api_parallel.R'
    'api_patterns.R'
    'api_period.R'
    'api_plot_time_series.R'
    'api_plot_raster.R'
    'api_plot_vector.R'
    'api_point.R'
    'api_predictors.R'
    'api_preconditions.R'
    'api_raster.R'
    'api_raster_sub_image.R'
    'api_reclassify.R'
    'api_reduce.R'
    'api_regularize.R'
    'api_request.R'
    'api_request_httr2.R'
    'api_roi.R'
    'api_samples.R'
    'api_segments.R'
    'api_select.R'
    'api_sf.R'
    'api_shp.R'
    'api_signal.R'
    'api_smooth.R'
    'api_smote.R'
    'api_som.R'
    'api_source.R'
    'api_source_aws.R'
    'api_source_bdc.R'
    'api_source_cdse.R'
    'api_source_deafrica.R'
    'api_source_deaustralia.R'
    'api_source_hls.R'
    'api_source_local.R'
    'api_source_mpc.R'
    'api_source_sdc.R'
    'api_source_stac.R'
    'api_source_terrascope.R'
    'api_source_usgs.R'
    'api_space_time_operations.R'
    'api_stac.R'
    'api_stats.R'
    'api_summary.R'
    'api_texture.R'
    'api_tibble.R'
    'api_tile.R'
    'api_timeline.R'
    'api_tmap.R'
    'api_torch.R'
    'api_torch_psetae.R'
    'api_ts.R'
    'api_tuning.R'
    'api_uncertainty.R'
    'api_utils.R'
    'api_validate.R'
    'api_values.R'
    'api_variance.R'
    'api_vector.R'
    'api_vector_info.R'
    'api_view.R'
    'RcppExports.R'
    'data.R'
    'sits-package.R'
    'sits_add_base_cube.R'
    'sits_apply.R'
    'sits_accuracy.R'
    'sits_bands.R'
    'sits_bayts.R'
    'sits_bbox.R'
    'sits_classify.R'
    'sits_colors.R'
    'sits_combine_predictions.R'
    'sits_config.R'
    'sits_csv.R'
    'sits_cube.R'
    'sits_cube_copy.R'
    'sits_cube_local.R'
    'sits_clean.R'
    'sits_cluster.R'
    'sits_detect_change.R'
    'sits_detect_change_method.R'
    'sits_dtw.R'
    'sits_factory.R'
    'sits_filters.R'
    'sits_geo_dist.R'
    'sits_get_data.R'
    'sits_get_class.R'
    'sits_get_probs.R'
    'sits_histogram.R'
    'sits_imputation.R'
    'sits_labels.R'
    'sits_label_classification.R'
    'sits_lighttae.R'
    'sits_machine_learning.R'
    'sits_merge.R'
    'sits_mixture_model.R'
    'sits_mlp.R'
    'sits_mosaic.R'
    'sits_model_export.R'
    'sits_patterns.R'
    'sits_plot.R'
    'sits_predictors.R'
    'sits_reclassify.R'
    'sits_reduce.R'
    'sits_reduce_imbalance.R'
    'sits_regularize.R'
    'sits_sample_functions.R'
    'sits_segmentation.R'
    'sits_select.R'
    'sits_sf.R'
    'sits_smooth.R'
    'sits_som.R'
    'sits_stars.R'
    'sits_summary.R'
    'sits_tae.R'
    'sits_tempcnn.R'
    'sits_terra.R'
    'sits_texture.R'
    'sits_timeline.R'
    'sits_train.R'
    'sits_tuning.R'
    'sits_utils.R'
    'sits_uncertainty.R'
    'sits_validate.R'
    'sits_view.R'
    'sits_variance.R'
    'sits_xlsx.R'
    'zzz.R'
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under:
	- [X ] geospatial data

`sits` is a package for satellite image time series analysis that works with big Earth observation data sets.

-   Who is the target audience, and what are the scientific applications of this package?
Remote sensing and environmental experts wanting to classify remote sensing images for applications such as deforestation detection, agricultural and land use/land cover mapping, biodiversity conservation, and land degradation monitoring comprise the target audience.

There are currently no other open-source software packages that have the same capabilities.

(a) Vignettes: Instead of preparing vignettes, the authors have written an online book that describes the contents of the package in detail. The book is available at the URL https://e-sensing.github.io/sitsbook/

Important notes:

(1) To run the tests, examples, and code coverage, please make
Ensure that the following environment variables are set in the R session.
Sys.setenv("SITS_RUN_TESTS" = "YES")
Sys.setenv("SITS_RUN_EXAMPLES" = "YES")
`sits` is a fairly large package, and the tests take a long time to run, since they access cloud services. We must manually enable testing for this reason.

(2) Please review version 1.5.3, not yet on CRAN, which is available in the "dev" branch in the GitHub repository.

- [x] contains a vignette with examples of its essential functions and uses (see above).

The package is already on CRAN.

- [ x] Do you wish to submit an Applications Article about your package to [Methods in Ecology and Evolution](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? If so:
