### 2023-07-03 12e818a
remove NA values from probs cube
