### 2023-07-03 9a9ff83
optimize clean function
