+++ R/api_classify.R
@@ -80,6 +80,8 @@
         # Should bbox of resulting tile be updated?
         update_bbox <- nrow(chunks) != nchunks
     }
+    # Compute fractions probability
+    probs_fractions <- 1/length(.ml_labels(ml_model))
     # Process jobs in parallel
     block_files <- .jobs_map_parallel_chr(chunks, function(chunk) {
         # Job block
