+++ NAMESPACE
@@ -75,6 +75,7 @@ S3method(.raster_ymax,terra)
 S3method(.raster_ymin,terra)
 S3method(.raster_yres,terra)
 S3method(.slice_dfr,numeric)
+S3method(.source_collection_access_test,"mpc_cube_sentinel-1-grd")
 S3method(.source_collection_access_test,mpc_cube)
 S3method(.source_collection_access_test,stac_cube)
 S3method(.source_collection_access_test,usgs_cube)
