### 2023-07-05 464ad0b
collect more than one sample per segment
