### 2023-07-05 b163355
fix error in documentation of sits_segment
