### 2023-07-06 0cd7a40
fix bug in sits_cube for many classified cubes
