### 2023-07-07 #596 @maelle
Thank you for your submission @gilbertocamara! As well as your careful response to the automatic checks.  We agree with all your responses above.

However, as this a package that implements statistical and ML methods of geospatial data, rather than just the “accessing, manipulating, converting” and converting in our scope, it falls under under our newer [statistical peer review program](https://stats-devguide.ropensci.org/), which has its own [time series](https://stats-devguide.ropensci.org/standards.html#standards-time-series) and [geospatial standards](https://stats-devguide.ropensci.org/standards.html#standards-spatial.  Submission requirements are different for this as authors need to document their standards compliance with our code annotation system.

A note - SITS is a package that is very large in scope and code base, as exemplified by the fact that it has a whole book for its documentation. As such, we anticipate that it will be challenging to find reviewers and we will need to give them considerably longer than usual to review the code base and documentation in full. Most of our submissions are not as large or mature at the point of review and up for significant API or architecture changes in response to review. For something in an earlier stage we would likely have suggested breaking functionality up into smaller, more focused packages. Nonetheless, we are up for the challenge if you are up for the higher statistical submission requirements and potential changes.

One last note regarding check results: As of now we can't set any environment variables when running the checks automatically, so they'd have to be set on your side, maybe using the [withr::local_envvar()](https://withr.r-lib.org/reference/with_envvar.html) function or similar.

Thanks again! We're happy to answer further questions.
