### 2023-07-07 #596 @gilbertocamara
Dear @maelle, many thanks for your response. Please see my comments below:
```
Nonetheless, we are up for the challenge if you are up for the higher statistical submission requirements and potential changes.
```
Good! Looking at the specific requirements for ROpenSci statistical packages, the `sits` package meets most of them, such as G.2 (related to data input), G.3 (algorithms), G.4 (output data). We will have to review the `sits` package carefully as for requirements G.1 (documentation), G.5 (testing) and those for machine learning.

At a first glance, `sits` complies with requirements SP (spatial software) and TS (time series) and UL (unsupervised learning) . Since these requirements are very detailed, we will carefully review them  to ensure compliance. We believe we meet the PD reqs (probability distr).

```
One last note regarding check results: As of now we can't set any environment variables when running the checks automatically, so they'd have to be set on your side, maybe using the [withr::local_envvar()](https://withr.r-lib.org/reference/with_envvar.html) function or similar.
```
Allow me to propose an alternative: please consider that the information provided in "codecov.io" to be sufficient to assert that `sits` meets the code coverage requirements of ROpenSci.  If you accept this proposal, it will save us both time and work.
```
https://app.codecov.io/github/e-sensing/sits/tree/dev/R
```
We will work on improving `sits` so that it meets the specifications  for ROpenSci statistical packages. We will report back to you when we have a new version that fully meets such specs.

Thanks,
Gilberto
