### 2023-07-07 #596 @maelle
Thank you! :tada: Here's a direct link to the author guide for stat submission: https://stats-devguide.ropensci.org/pkgdev.html
