### 2023-07-07 2df149d
fix sits_join_segments to account for classification of many points inside a segment
