+++ R/sits_cube_copy.R
@@ -26,7 +26,7 @@
 #'     # Creating a sits cube from BDC
 #'     bdc_cube <- sits_cube(
 #'         source = "BDC",
-#'         collection = "CB4-16D-2",
+#'         collection = "CBERS-WFI-16D",
 #'         tiles = c("007004", "007005"),
 #'         bands = c("B15", "CLOUD"),
 #'         start_date = "2018-01-01",
