+++ R/sits_segment2.R
@@ -1,0 +1,383 @@
+sits_classify.raster_cube <- function(cube,
+                                      roi = NULL,
+                                      start_date = NULL,
+                                      end_date = NULL,
+                                      memsize = 8,
+                                      multicores = 2,
+                                      output_dir,
+                                      version = "v1",
+                                      verbose = FALSE,
+                                      progress = TRUE) {
+    # preconditions
+    .check_is_raster_cube(cube)
+    .check_is_regular(cube)
+    .check_memsize(memsize)
+    .check_output_dir(output_dir)
+    .check_version(version)
+
+    # Spatial filter
+    if (.has(roi)) {
+        roi <- .roi_as_sf(roi)
+        cube <- .cube_filter_spatial(cube = cube, roi = roi)
+    }
+    # Temporal filter
+    if (.has(start_date) || .has(end_date)) {
+        cube <- .cube_filter_interval(
+            cube = cube, start_date = start_date, end_date = end_date
+        )
+    }
+    # Check memory and multicores
+    # Get block size
+    block <- .raster_file_blocksize(.raster_open_rast(.tile_path(cube)))
+    # Check minimum memory needed to process one block
+    job_memsize <- .jobs_memsize(
+        job_size = .block_size(block = block, overlap = 0),
+        npaths = length(.tile_paths(cube)),
+        nbytes = 8,
+        proc_bloat = .conf("processing_bloat")
+    )
+    # Update multicores parameter
+    multicores <- .jobs_max_multicores(
+        job_memsize = job_memsize, memsize = memsize, multicores = multicores
+    )
+    # Update block parameter
+    block <- .jobs_optimal_block(
+        job_memsize = job_memsize, block = block,
+        image_size = .tile_size(.tile(cube)), memsize = memsize,
+        multicores = multicores
+    )
+    # Prepare parallel processing
+    .parallel_start(
+        workers = multicores, log = verbose,
+        output_dir = output_dir
+    )
+    on.exit(.parallel_stop(), add = TRUE)
+    # Show block information
+    if (verbose) {
+        start_time <- Sys.time()
+        message(
+            "Using blocks of size (", .nrows(block),
+            " x ", .ncols(block), ")"
+        )
+    }
+    # Classification
+    # Process each tile sequentially
+    segs_cube <- .cube_foreach_tile(cube, function(tile) {
+        # Segment the data
+        segs_tile <- .segment_tile(
+            tile = tile,
+            band = "segment",
+            seg_fn = seg_fn,
+            block = block,
+            roi = roi,
+            output_dir = output_dir,
+            version = version,
+            verbose = verbose,
+            progress = progress
+        )
+        return(segs_tile)
+    })
+    return(segs_cube)
+}
+
+
+.segment_tile <- function(tile,
+                          band,
+                          seg_fn,
+                          block,
+                          roi,
+                          output_dir,
+                          version,
+                          verbose,
+                          progress) {
+    # Output file
+    out_file <- .file_derived_name(
+        tile = tile, band = band, version = version, output_dir = output_dir
+    )
+    # Resume feature
+    if (file.exists(out_file)) {
+        if (.check_messages()) {
+            message("Recovery: tile '", tile[["tile"]], "' already exists.")
+            message(
+                "(If you want to produce a new image, please ",
+                "change 'output_dir' or 'version' parameters)"
+            )
+        }
[…]
