### 2023-07-08 cc72697
Fix errors in examples file
