+++ R/sits_view.R
@@ -33,7 +33,7 @@
 #' if (sits_run_examples()) {
 #'     # view samples
 #'     sits_view(cerrado_2classes)
-#'     . # create a local data cube
+#'     # create a local data cube
 #'     data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
 #'     modis_cube <- sits_cube(
 #'         source = "BDC",
@@ -46,7 +46,7 @@
 #'     )
 #'     # train a model
 #'     rf_model <- sits_train(samples_modis_ndvi, sits_rfor())
-#'     . # classify the cube
+#'     # classify the cube
 #'     modis_probs <- sits_classify(
 #'         data = modis_cube,
 #'         ml_model = rf_model,
@@ -59,9 +59,9 @@
 #'         modis_probs,
 #'         output_dir = tempdir()
 #'     )
-#'     . # view the classified map
+#'     # view the classified map
 #'     sits_view(modis_label)
-#'     . # view the classified map with the B/W image
+#'     # view the classified map with the B/W image
 #'     sits_view(modis_cube,
 #'         band = "NDVI",
 #'         class_cube = modis_label,
@@ -87,7 +87,7 @@
 #'         tile = "012010",
 #'         bands = "NDVI",
 #'         date = sits_timeline(modis_cube)[1],
-#'         seg_fn = sits_slic(step = 20)
+#'         seg_fn = sits_supercells(step = 20)
 #'     )
 #'     # view image and segments
 #'     sits_view(
