+++ man/sits_tuning.Rd
@@ -64,10 +64,10 @@ if (sits_run_examples()) {
         ml_method = sits_tempcnn(),
         params = sits_tuning_hparams(
             optimizer = choice(
-                torchopt::optim_adamw
+                torch::optim_adamw
             ),
             opt_hparams = list(
-                lr = beta(0.3, 5)
+                lr = loguniform(10^-2, 10^-4)
             )
         ),
         trials = 4,
