### 2023-07-08 #596 @gilbertocamara
Dear @maelle, thanks for the clarification. As part of the pre-submission process of the package "sits" to the ROpenSci statistical review, as recommended by you, I tried to run autotest_package . I made sure that all the examples and tests run OK. However, when running autotest_package, I am finding an error, which I reported in
```
https://github.com/ropensci-review-tools/autotest/issues/81
```
