### 2023-07-10 #596 @maelle
@gilbertocamara thanks for the report. I installed the dependencies of the dev branch, and I'm working with the dev branch itself. When running the examples I got

```
Error: No BDC_ACCESS_KEY defined in environment
```

Is there a comprehensive list of credentials needed to run the test? If there's such a list and one can create free accounts easily, I can get the credentials. If not, could these examples (and tests, if it happens in tests) be skipped? For examples, see the `@examplesIf` tag https://roxygen2.r-lib.org/articles/rd.html?q=examplesIf#examples

A good place to add setup info would be the installation instructions, and/or the file `.github/CONTRIBUTING.md`.

Thank you!
