### 2023-07-10 5056b56
update on documentation of apis and improved examples
