+++ R/api_gdal.R
@@ -5,7 +5,10 @@
     "INT4U" = "UInt32", "INT4S" = "Int32", "FLT4S" = "Float32",
     "FLT8S" = "Float64"
 )
-
+#' @title Get GDAL parameters
+#' @noRd
+#' @param params   Params used to describe GDAL file
+#' @returns        Cleaned GDAL parameters
 .gdal_params <- function(params) {
     # Check if parameters are named
     if (!all(.has_name(params))) {
@@ -23,7 +26,12 @@
         }
     }, names(params), unname(params), USE.NAMES = FALSE))
 }
-
+#' @title Format GDAL parameters
+#' @noRd
+#' @param asset  File to be accessed (with path)
+#' @param sf_roi Region of interest (sf object)
+#' @param res    Spatial resolution
+#' @returns      Formatted GDAL parameters
 .gdal_format_params <- function(asset, sf_roi, res) {
     gdal_params <- list()
     if (.has(res)) {
@@ -42,7 +50,11 @@
     gdal_params[["-a_nodata"]] <- .miss_value(band_conf)
     return(gdal_params)
 }
-
+#' @title Format GDAL block parameters for data access
+#' @noRd
+#' @param asset  File to be accessed (with path)
+#' @param sf_roi Region of interest (sf object)
+#' @returns      Formatted GDAL block parameters for data access
 .gdal_as_srcwin <- function(asset, sf_roi) {
     block <- .raster_sub_image(tile = asset, sf_roi = sf_roi)
     list(
@@ -52,21 +64,36 @@
         ysize = block[["nrows"]]
     )
 }
-
+#' @title Run gdal_translate
+#' @noRd
+#' @param file        File to be created (with path)
+#' @param base_file   File to be copied from (with path)
+#' @param param       GDAL parameters
+#' @param quiet       TRUE/FALSE
+#' @returns           Called for side effects
 .gdal_translate <- function(file, base_file, params, quiet) {
     sf::gdal_utils(
         util = "translate", source = base_file[[1]], destination = file[[1]],
         options = .gdal_params(params), quiet = quiet
     )
 }
-
+#' @title Run gdal_warp
+#' @noRd
+#' @param file        File to be created (with path)
+#' @param base_files  Files to be copied from (with path)
+#' @param param       GDAL parameters
+#' @param quiet       TRUE/FALSE
+#' @returns           Called for side effects
 .gdal_warp <- function(file, base_files, params, quiet) {
     sf::gdal_utils(
         util = "warp", source = base_files, destination = file[[1]],
         options = .gdal_params(params), quiet = quiet
     )
 }
-
+#' @title Run gdal_addo
+#' @noRd
+#' @param base_file   Base file to be processed
+#' @returns           Called for side effects
 .gdal_addo <- function(base_file) {
     conf_cog <- .conf("gdal_presets", "cog")
     suppressMessages(
@@ -78,7 +105,15 @@
         )
     )
 }
-
+#' @title Run gdal_translate from a block to a file
+#' @noRd
+#' @param block        Block with
+#' @param bbox         Bounding box for file
+#' @param file         Files to be written to (with path)
+#' @param nlayers      Number of layers in GDAL file
+#' @param miss_value   Missing value
+#' @param data_type.   GDAL data type
+#' @returns            Name of file that was written to
 .gdal_template_block <- function(block, bbox, file, nlayers, miss_value,
                                  data_type) {
     # Get first file
@@ -124,7 +159,12 @@
     # Return file
     file
 }
-
+#' @title Merge files into a single file
+#' @noRd
+#' @param file         Files to be written to (with path)
+#' @param base_files   Files to be copied from (with path)
+#' @param multicores   Number of cores to be used in parallel
[…]
