+++ R/api_apply.R
@@ -33,6 +33,23 @@
     data[[col]] <- x[[col]]
     return(data)
 }
+#' @title Apply an expression to block of a set of input bands
+#' @name .apply_feature
+#' @keywords internal
+#' @noRd
+#' @author Rolf Simoes, \email{rolf.simoes@@inpe.br}
+#'
+#' @param  feature         Subset of a data cube containing the input bands
+#'                         used in the expression
+#' @param  block           Individual block that will be processed
+#' @param  window_size     Size of the neighbourhood (if required)
+#' @param  expr            Expression to be applied
+#' @param  out_band        Output band
+#' @param  in_bands        Input bands
+#' @param  overlap         Overlap between tiles (if required)
+#' @param  output_dir      Directory where image will be save
+#'
+#' @return                 A feature compose by a combination of tile and band.
 .apply_feature <- function(feature, block, window_size, expr,
                            out_band, in_bands, overlap, output_dir) {
     # Output file
@@ -65,7 +82,8 @@
         block <- .block(chunk)
         # Block file name for each fraction
         block_files <- .file_block_name(
-            pattern = .file_pattern(out_file), block = block,
+            pattern = .file_pattern(out_file),
+            block = block,
             output_dir = output_dir
         )
         # Resume processing in case of failure
@@ -148,13 +166,13 @@
 }
 
 #' @title Apply an expression across all bands
-#'
 #' @name .apply_across
 #' @keywords internal
 #' @noRd
 #'
 #' @param data  Tile name.
-#'
+#' @param fn    Function to be applied
+#' @param ...   Further parameters for the function
 #' @return      A sits tibble with all processed bands.
 #'
 .apply_across <- function(data, fn, ...) {
@@ -172,14 +190,12 @@
     return(result)
 }
 #' @title Captures a band expression
-#'
 #' @name .apply_capture_expression
 #' @keywords internal
 #' @noRd
 #'
-#' @param tile_name  Tile name.
-#'
-#' @return           Named list with one expression
+#' @param ...  Expression to be applied
+#' @return     Named list with one expression
 #'
 .apply_capture_expression <- function(...) {
     # Capture dots as a list of quoted expressions
@@ -198,16 +214,13 @@
     return(list_expr)
 }
 #' @title Finds out all existing bands in an expression
-#'
 #' @name .apply_input_bands
 #' @keywords internal
 #' @noRd
 #'
-#' @param tile       Data cube tile.
-#' @param expr       Band expression.
-#'
-#' @return           List of combination among tiles, bands, and dates
-#'                   that are missing from the cube.
+#' @param cube       Data cube.
+#' @param expr       Band combination expression.
+#' @return           List of input bands required to run the expression
 #'
 .apply_input_bands <- function(cube, expr) {
     # Get all required bands in expression
@@ -252,6 +265,14 @@
         character()
     }
 }
+#' @title Kernel function for window operations in spatial neighbourhoods
+#' @name .kern_functions
+#' @noRd
+#' @param windows size of local window
+#' @param img_nrow  image size in rows
+#' @param img_ncol  image size in cols
+#' @return operations on local kernels
+#'
 .kern_functions <- function(window_size, img_nrow, img_ncol) {
     # Pre-conditions
     .check_window_size(window_size, max = min(img_nrow, img_ncol) - 1)
