+++ R/sits_mixture_model.R
@@ -196,7 +196,8 @@ sits_mixture_model.raster_cube <- function(data, endmembers, ...,
     )
     # Update block parameter
     block <- .jobs_optimal_block(
-        job_memsize = job_memsize, block = block,
+        job_memsize = job_memsize,
+        block = block,
         image_size = .tile_size(.tile(data)), memsize = memsize,
         multicores = multicores
     )
