### 2023-07-10 #596 @gilbertocamara
Dear @maelle, many thanks. One of the problems with the examples is that they assume the user has obtained a key to access the Brazil Data Cube (BDC). I have fixed the problem and removed this restriction. I ran R CMD CHECK with these changes, and now everything should work for you. Please use the new "dev" version that I have just uploaded.
