### 2023-07-11 5c85ea9
Removed examples in sits_run_examples and sits_run_tests
