+++ R/api_tile.R
@@ -26,9 +26,10 @@ NULL
     cube[1, ]
 }
 
-#  Tile field accessors
-#  These functions enable access components of tiles. A
-#
+#' @title Get source cloud provider for a tile
+#' @noRd
+#' @param tile A tile.
+#' @return Source cloud provider
 .tile_source <- function(tile) {
     UseMethod(".tile_source", tile)
 }
@@ -37,18 +38,22 @@ NULL
     tile <- .tile(tile)
     .as_chr(tile[["source"]])
 }
-#
+#' @title Get image collection for a tile
+#' @noRd
+#' @param tile A tile.
+#' @return Image collection
 .tile_collection <- function(tile) {
     UseMethod(".tile_collection", tile)
 }
-
 #' @export
 .tile_collection.raster_cube <- function(tile) {
     tile <- .tile(tile)
     .as_chr(tile[["collection"]])
 }
-
-#
+#' @title Get/Set tile name
+#' @noRd
+#' @param tile A tile.
+#' @return Name of the tile
 .tile_name <- function(tile) {
     UseMethod(".tile_name", tile)
 }
@@ -66,7 +71,10 @@ NULL
     tile <- .tile(tile)
     .as_chr(tile[["tile"]])
 }
-#
+#' @title Get tile number of columns
+#' @noRd
+#' @param tile A tile.
+#' @return Number of columns
 .tile_ncols <- function(tile) {
     UseMethod(".tile_ncols", tile)
 }
@@ -75,7 +83,10 @@ NULL
     tile <- .tile(tile)
     .ncols(.fi(tile))
 }
-#
+#' @title Get tile number of rows
+#' @noRd
+#' @param tile A tile.
+#' @return Number of rows
 .tile_nrows <- function(tile) {
     UseMethod(".tile_nrows", tile)
 }
@@ -84,7 +95,10 @@ NULL
     tile <- .tile(tile)
     .nrows(.fi(tile))
 }
-#
+#' @title Get tile size
+#' @noRd
+#' @param tile A tile.
+#' @return Size (list of nrows x ncols)
 .tile_size <- function(tile) {
     UseMethod(".tile_size", tile)
 }
@@ -92,7 +106,10 @@ NULL
 .tile_size.raster_cube <- function(tile) {
     list(ncols = .tile_ncols(tile), nrows = .tile_nrows(tile))
 }
-#
+#' @title Get X resolution
+#' @noRd
+#' @param tile A tile.
+#' @return x resolution
 .tile_xres <- function(tile) {
     UseMethod(".tile_xres", tile)
 }
@@ -101,7 +118,10 @@ NULL
     tile <- .tile(tile)
     .xres(.fi(tile))
 }
-#
+#' @title Get Y resolution
+#' @noRd
+#' @param tile A tile.
+#' @return y resolution
 .tile_yres <- function(tile) {
     UseMethod(".tile_yres", tile)
 }
@@ -110,7 +130,10 @@ NULL
     tile <- .tile(tile)
     .yres(.fi(tile))
 }
-#
+#' @title Get/Set labels
+#' @noRd
+#' @param tile A tile.
+#' @return vector of labels
 .tile_labels <- function(tile) {
     UseMethod(".tile_labels", tile)
 }
@@ -167,8 +190,7 @@ NULL
 #' @keywords internal
 #' @noRd
 #' @param tile A tile.
-#'
-#' @return date
+#' @return file ID
 .tile_fid <- function(tile) {
     UseMethod(".tile_fid", tile)
 }
@@ -182,7 +204,6 @@ NULL
 #' @keywords internal
 #' @noRd
 #' @param tile A tile.
-#'
 #' @return a timeline
[…]
