### 2023-07-12 082fe20
implement S2 tiling system generator
