+++ R/api_mgrs.R
@@ -1,0 +1,627 @@
+# The following code is a translation of parts of the proj4js/mgrs repository
+# (https://github.com/proj4js/mgrs) which is licensed under the MIT License.
+#
+# Copyright (c) 2012, Mike Adair, Richard Greenwood, Didier Richard,
+# Stephen Irons, Olivier Terral, Calvin Metcalf
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+
+# UTM zones are grouped, and assigned to one of a group of 6 sets.
+NUM_100K_SETS <- 6
+
+# The column letters (for easting) of the lower left value, per set.
+SET_ORIGIN_COLUMN_LETTERS <- 'AJSAJS'
+
+# The row letters (for northing) of the lower left value, per set.
+SET_ORIGIN_ROW_LETTERS <- 'AFAFAF'
+
+A <- 65  # A
+I <- 73  # I
+O <- 79  # O
+V <- 86  # V
+Z <- 90  # Z
+
+# First eccentricity squared
+ECC_SQUARED <- 0.00669438
+
+# Scale factor along the central meridian
+SCALE_FACTOR <- 0.9996
+
+# Semimajor axis (half the width of the earth) in meters
+SEMI_MAJOR_AXIS <- 6378137
+
+# The easting of the central meridian of each UTM zone
+EASTING_OFFSET <- 500000
+
+# The northing of the equator for southern hemisphere locations (in UTM)
+NORTHING_OFFSET <- 10000000
+
+# UTM zone width in degrees
+UTM_ZONE_WIDTH <- 6
+
+# Half the width of a UTM zone in degrees
+HALF_UTM_ZONE_WIDTH <- UTM_ZONE_WIDTH / 2
+
+# Convert lat/lon to MGRS.
+#
+# ll: Array with longitude and latitude on a WGS84 ellipsoid.
+# accuracy: Accuracy in digits (5 for 1 m, 4 for 10 m, 3 for 100 m, 2 for 1 km, 1 for 10 km or 0 for 100 km).
+#   Optional, default is 5.
+# Returns: The MGRS string for the given location and accuracy.
+forward <- function(ll, accuracy = 5) {
+    if (!is.numeric(accuracy)) {
+        accuracy <- 5  # default accuracy 1m
+    }
+
+    if (!is.vector(ll) || length(ll) != 2 || any(sapply(ll, function(x) !is.numeric(x)))) {
+        stop("forward did not receive a numeric array")
+    }
+
[…]
