### 2023-07-15 b8e6304
update segment code
