+++ R/sits_segment2.R
@@ -1,19 +1,18 @@
-sits_classify.raster_cube <- function(cube,
-                                      roi = NULL,
-                                      start_date = NULL,
-                                      end_date = NULL,
-                                      memsize = 8,
-                                      multicores = 2,
-                                      output_dir,
-                                      version = "v1",
-                                      verbose = FALSE,
-                                      progress = TRUE) {
-    # preconditions
+sits_segment <- function(cube,
+                         seg_fn = sits_supercells_temp(),
+                         roi = NULL,
+                         start_date = NULL,
+                         end_date = NULL,
+                         memsize = 8,
+                         multicores = 2,
+                         output_dir,
+                         version = "v1",
+                         progress = TRUE) {
+    # Preconditions
     .check_is_raster_cube(cube)
     .check_is_regular(cube)
     .check_memsize(memsize)
     .check_output_dir(output_dir)
-    .check_version(version)
 
     # Spatial filter
     if (.has(roi)) {
@@ -26,6 +25,9 @@ sits_classify.raster_cube <- function(cube,
             cube = cube, start_date = start_date, end_date = end_date
         )
     }
+    start_date <- .default(start_date, .cube_start_date(cube))
+    end_date <- .default(end_date, .cube_end_date(cube))
+
     # Check memory and multicores
     # Get block size
     block <- .raster_file_blocksize(.raster_open_rast(.tile_path(cube)))
@@ -34,7 +36,7 @@ sits_classify.raster_cube <- function(cube,
         job_size = .block_size(block = block, overlap = 0),
         npaths = length(.tile_paths(cube)),
         nbytes = 8,
-        proc_bloat = .conf("processing_bloat")
+        proc_bloat = .conf("processing_bloat_seg")
     )
     # Update multicores parameter
     multicores <- .jobs_max_multicores(
@@ -47,32 +49,20 @@ sits_classify.raster_cube <- function(cube,
         multicores = multicores
     )
     # Prepare parallel processing
-    .parallel_start(
-        workers = multicores, log = verbose,
-        output_dir = output_dir
-    )
+    .parallel_start(workers = multicores, output_dir = output_dir)
     on.exit(.parallel_stop(), add = TRUE)
-    # Show block information
-    if (verbose) {
-        start_time <- Sys.time()
-        message(
-            "Using blocks of size (", .nrows(block),
-            " x ", .ncols(block), ")"
-        )
-    }
-    # Classification
+    # Segmentation
     # Process each tile sequentially
     segs_cube <- .cube_foreach_tile(cube, function(tile) {
         # Segment the data
         segs_tile <- .segment_tile(
             tile = tile,
-            band = "segment",
             seg_fn = seg_fn,
+            band = "segment",
             block = block,
             roi = roi,
             output_dir = output_dir,
             version = version,
-            verbose = verbose,
             progress = progress
         )
         return(segs_tile)
@@ -80,42 +70,38 @@ sits_classify.raster_cube <- function(cube,
     return(segs_cube)
 }
 
-
 .segment_tile <- function(tile,
-                          band,
                           seg_fn,
+                          band,
                           block,
                           roi,
                           output_dir,
                           version,
-                          verbose,
                           progress) {
[…]
