+++ NAMESPACE
@@ -145,6 +145,7 @@ S3method(.tile_timeline,raster_cube)
 S3method(.tile_within,raster_cube)
 S3method(.tile_xres,raster_cube)
 S3method(.tile_yres,raster_cube)
+S3method(.vector_check_package,sf)
 S3method(plot,class_cube)
 S3method(plot,geo_distances)
 S3method(plot,patterns)
