### 2023-07-16 #596 @gilbertocamara
Dear @maelle @mpadge

Please, could you explain what appears to be an unexpected behaviour of `autotest`?

Today, I ran `autotest` twice on version 1.4.2 (dev) of the `sits` package. The first response had 16 issues (please see the RDS file in https://www.dropbox.com/s/cu4lpm9vjcgxhmw/autotest_1.rds?dl=0). From what I could understand from the  `autotest` output, it complains about the expected return values of R functions that are called for side-effects.

I tried to fix some of these problems by considering the recommendations of the tidyverse design guide. In Section 26 ("Side-effect functions should return invisibly"), the guide states: "If a function is called primarily for its side-effects, it should invisibly return a useful output. If there’s no obvious output, return the first argument". See more at https://design.tidyverse.org/out-invisible.html.

I am assuming that `autotest` follows the same guidelines. Thus, I included invisible return values in all `sits` functions that are called for side-effects. Then, I ran `autotest` again. To my surprise, it flagged 48 issues. Please see the second `autotest` output at https://www.dropbox.com/s/43zin4ithlectbz/autotest_2.rds?dl=0.

Could you please help me and explain why `autotest` increases its number of issues from 16 to 48?   Your help will be most appreciated.

*** MWE ***

```{r}
# install dev version
devtools::install_github("e-sensing/sits@dev")
# enable examples and tests
Sys.setenv("SITS_RUN_EXAMPLES" = "YES")
Sys.setenv("SITS_RUN_TESTS" = "YES")
# first run of autotest
autotest_1 <- autotest::autotest_package(package = "sits", test = TRUE)
# second run of autotest
autotest_2 <- autotest::autotest_package(package = "sits", test = TRUE)
```
Best regards
Gilberto
