+++ R/api_vector.R
@@ -20,70 +20,23 @@
     UseMethod(".vector_check_package", pkg_class)
 }
 
-.vector_open_vec <- function(file, layer = NULL) {
+.vector_open_vec <- function(file, ...) {
     # check package
     pkg_class <- .vector_check_package()
 
     UseMethod(".vector_open_vec", pkg_class)
 }
 
-#' @title Vector package internal get values function
-#' @name .vector_get_values
-#' @keywords internal
-#' @noRd
-#' @author Felipe Carvalho, \email{felipe.carvalho@@inpe.br}
-#'
-#' @param v_obj   vector package object
-#' @param ...     additional parameters to be passed to vector package
-#'
-#' @return Numeric matrix associated to vector object
-.vector_get_values <- function(v_obj, ...) {
+.vector_read_vec <- function(file, ...) {
     # check package
     pkg_class <- .vector_check_package()
 
-    # call function
-    UseMethod(".vector_get_values", pkg_class)
+    UseMethod(".vector_read_vec", pkg_class)
 }
 
-#' @title vector package internal set values function
-#' @name .vector_set_values
-#' @keywords internal
-#' @noRd
-#' @author Felipe Carvalho, \email{felipe.carvalho@@inpe.br}
-#'
-#' @param v_obj   vector package object
-#' @param values  Numeric matrix to copy to vector object
-#' @param ...     additional parameters to be passed to vector package
-#'
-#' @return        vector object
-.vector_set_values <- function(v_obj, values, ...) {
+.vector_write_vec <- function(v_obj, file, ...) {
     # check package
     pkg_class <- .vector_check_package()
-
-    UseMethod(".vector_set_values", pkg_class)
-}
-
-#' @title Raster package internal write raster file function
-#' @name .raster_write_rast
-#' @keywords internal
-#' @noRd
-#' @author Rolf Simoes, \email{rolf.simoes@@inpe.br}
-#'
-#' @param r_obj         raster package object to be written
-#' @param file          file path to save raster file
-#' @param format        GDAL file format string (e.g. SHP)
-#' @param layer         ...
-#' @param overwrite     logical indicating if file can be overwritten
-#' @param ...           additional parameters to be passed to raster package
-#'
-#' @return              No value, called for side effects.
-.vector_write_vec <- function(r_obj,
-                              file,
-                              format,
-                              layer,
-                              overwrite, ...) {
-    # check package
-    pkg_class <- .raster_check_package()
 
     UseMethod(".raster_write_rast", pkg_class)
 }
@@ -100,498 +53,37 @@
 #' @return               A vector object.
 .vector_new_vec <- function( ...) {
     # check package
-    pkg_class <- .raster_check_package()
+    pkg_class <- .vector_check_package()
 
     UseMethod(".raster_new_rast", pkg_class)
 }
 
-#' @title Raster package internal object properties
-#' @name .vector_nrows
+#' @name .vector_crs
 #' @keywords internal
 #' @noRd
-#' @author Rolf Simoes, \email{rolf.simoes@@inpe.br}
-#'
-#' @param r_obj    raster package object
-#' @param ...      additional parameters to be passed to raster package
-#'
-#' @return Raster object spatial properties
-.vector_nrows <- function(r_obj, ...) {
+.vector_crs <- function(v_obj, wkt = TRUE, ...) {
     # check package
-    pkg_class <- .raster_check_package()
+    pkg_class <- .vector_check_package()
 
-    UseMethod(".raster_nrows", pkg_class)
+    UseMethod(".vector_crs", pkg_class)
 }
 
-#' @name .raster_ncols
+#' @name .vector_bbox
 #' @keywords internal
 #' @noRd
-.raster_ncols <- function(r_obj, ...) {
[…]
