+++ R/api_vector_sf.R
@@ -12,3 +12,31 @@
 
     return(invisible(pkg_name))
 }
+
+.vector_open_vec.sf <- function(file, ...) {
+    sf::st_read(dsn = file, ...)
+}
+
+.vector_read_vec.sf <- function(file, ...) {
+    sf::read_sf(dsn = file, ...)
+}
+
+.vector_write_vec <- function(v_obj, file, ...) {
+    sf::st_write(obj = v_obj, dsn = file, ...)
+}
+
+.vector_bbox.sf <- function(v_obj, ...) {
+    sf::st_bbox(v_obj, ...)
+}
+
+.vector_crs.sf <- function(v_obj, wkt = TRUE, ...) {
+    crs <- sf::st_crs(v_obj, ...)
+    if (wkt) {
+        return(crs[["wkt"]])
+    }
+    crs
+}
+
+.vector_reproject.sf <- function(v_obj, crs, ...) {
+    sf::st_transform(x = v_obj, crs = crs, ...)
+}
