+++ R/api_raster_terra.R
@@ -37,6 +37,15 @@
 #' @keywords internal
 #' @noRd
 #' @export
+.raster_set_na.terra <- function(r_obj, na_value, ...) {
+    terra::NAflag(x = r_obj) <- na_value
+
+    return(r_obj)
+}
+
+#' @keywords internal
+#' @noRd
+#' @export
 .raster_extract.terra <- function(r_obj, xy, ...) {
     terra::extract(x = r_obj, y = xy, ...)
 }
@@ -376,10 +385,16 @@
     terra::colFromX(r_obj, x)
 }
 
-
 #' @keywords internal
 #' @noRd
 #' @export
 .raster_row.terra <- function(r_obj, y) {
     terra::rowFromY(r_obj, y)
+}
+
+#' @keywords internal
+#' @noRd
+#' @export
+.raster_polygonize.terra <- function(r_obj, dissolve = TRUE, ...) {
+    terra::as.polygons(r_obj, dissolve = TRUE, ...)
 }
