### 2023-07-17 8c69f60
add a check for vector type
