+++ NAMESPACE
@@ -63,9 +63,11 @@ S3method(.raster_new_rast,terra)
 S3method(.raster_nlayers,terra)
 S3method(.raster_nrows,terra)
 S3method(.raster_open_rast,terra)
+S3method(.raster_polygonize,terra)
 S3method(.raster_rast,terra)
 S3method(.raster_read_rast,terra)
 S3method(.raster_row,terra)
+S3method(.raster_set_na,terra)
 S3method(.raster_set_values,terra)
 S3method(.raster_write_rast,terra)
 S3method(.raster_xmax,terra)
