### 2023-07-17 f6cdcbc
update config internals
