### 2023-07-17 23d1c1c
update segment function
