+++ R/sits_segment2.R
@@ -92,7 +92,7 @@ sits_segment <- function(cube,
                 "change 'output_dir' or 'version' parameters)"
             )
         }
-        seg_tile <- .tile_derived_from_segment(
+        seg_tile <- .segments_derived_from_file(
             file = out_file,
             band = band,
             base_tile = tile,
@@ -141,8 +141,10 @@ sits_segment <- function(cube,
         input_pixels <- nrow(values)
         # Apply segment function
         values <- seg_fn(values, block, bbox)
+        # Check if the result values is a vector object
+        .check_vector(values)
         # Prepare and save results as vector
-        .segment_write_block(file = block_file, values = values)
+        .vector_write_vec(v_obj = values, file = block_file)
         # Free memory
         gc()
         # Returned block file
@@ -150,9 +152,12 @@ sits_segment <- function(cube,
     }, progress = progress)
     # Merge blocks into a new segs_cube tile
     seg_tile <- .tile_segment_merge_blocks(
-        block_files = block_files, base_tile = tile, band = "segment",
-        derived_class = "seg_cube",
-        out_file = out_file, update_bbox = update_bbox
+        block_files = block_files,
+        base_tile = tile,
+        band = "segments",
+        derived_class = "segs_cube",
+        out_file = out_file,
+        update_bbox = update_bbox
     )
     # Delete segments blocks
     unlink(block_files)
@@ -170,10 +175,10 @@ sits_segment <- function(cube,
     )
     # Set base tile
     base_file <- if (update_bbox) NULL else .tile_path(base_tile)
-    # Create a template raster based on the first image of the tile
-    vec_segments <- purrr::map_dfr(block_files, sf::st_read)
-    sf::st_write(obj = vec_segments, dsn = out_file)
-
+    # Read all blocks file
+    vec_segments <- purrr::map_dfr(block_files, .vector_read_vec.sf)
+    # Write all segments
+    .vector_write_vec(v_obj = vec_segments, file = out_file)
     # Create tile based on template
     tile <- .segments_derived_from_file(
         file = file,
@@ -253,82 +258,75 @@ sits_segment <- function(cube,
     ))
 }
 
-.segment_write_block <- function(file, values) {
-    sf::st_write(obj = values, dsn = file)
-    file
-}
-
 sits_supercells_temp <- function(data = NULL,
                                  step = 50,
                                  compactness = 1,
                                  dist_fun = "euclidean",
                                  avg_fun = "mean",
                                  iter = 10,
-                                 minarea = 30) {
+                                 minarea = 30,
+                                 verbose = FALSE) {
     function(data, block, bbox) {
-        r_obj <- .raster_new_rast(
+        # Create a template rast
+        v_obj <- .raster_new_rast(
             nrows = block[["nrows"]], ncols = block[["ncols"]],
             xmin = bbox[["xmin"]], xmax = bbox[["xmax"]],
             ymin = bbox[["ymin"]], ymax = bbox[["ymax"]],
-            nlayers = ncol(data), crs = bbox[["crs"]]
+            nlayers = 1, crs = bbox[["crs"]]
         )
-        mat <- c(.raster_nrows(r_obj), .raster_ncols(r_obj))
-        mode(mat) <- "integer"
-        new_centers = matrix(c(0L, 0L), ncol = 2)
[…]
