### 2023-07-17 34a3034
return values for side effects functions
