+++ R/api_check.R
@@ -88,15 +88,14 @@
 #' }
 #' }
 #'
-#' @return \code{.check_set_caller()} returns \code{NULL}.
+#' @return Called for side effects
 .check_set_caller <- function(caller) {
     envir <- parent.frame()
     if (length(sys.frames()) > 1) {
         envir <- sys.frame(-1)
     }
     assign(".check_caller", caller, envir = envir)
-
-    return(invisible(NULL))
+    return(invisible(caller))
 }
 #' @rdname check_functions
 #' @name .check_identify_caller
@@ -110,13 +109,11 @@
             return(caller)
         }
     }
-
     # check parent frame
     if (exists(".check_caller", envir = parent.frame())) {
         caller <- get(".check_caller", envir = f)
         return(caller)
     }
-
     # no caller defined, get first function in calling stack
     caller <- sys.calls()[[1]]
     caller <- gsub(
@@ -187,8 +184,7 @@
 #' }
 #' @keywords internal
 #' @noRd
-#' @return The same input value if no error occurs
-#'
+#' @return  Called for side effects
 .check_that <- function(x, ...,
                         local_msg = NULL,
                         msg = NULL) {
@@ -213,7 +209,6 @@
         # process message
         stop(sprintf(msg, local_msg), call. = FALSE)
     }
-
     return(invisible(x))
 }
 #' @rdname check_functions
@@ -243,6 +238,16 @@
             msg = msg
         )
     }
+    return(invisible(x))
+}
+#' @rdname check_functions
+#' @keywords internal
+#' @noRd
+.check_valid <- function(x, ...) {
+    # check for NA
+    .check_na(x)
+    # check for NULL
+    .check_null(x)
     return(invisible(x))
 }
 #' @rdname check_functions
@@ -290,7 +295,6 @@
                           msg = NULL) {
     # make default message
     msg <- .check_default_message(x, msg)
-
     # pre-condition
     .check_num_type(
         len_min,
@@ -310,7 +314,6 @@
             msg = msg
         )
     }
-
     # these checks are separate because the messages are different
     .check_that(
         length(x) >= len_min,
@@ -323,7 +326,6 @@
         local_msg = paste0("length should be <= ", len_max),
         msg = msg
     )
-
     return(invisible(x))
 }
 #' @rdname check_functions
@@ -333,10 +335,8 @@
     if (!is.function(fn_check)) {
         stop(".check_apply: fn_check should be a function.", call. = TRUE)
     }
-
     # check all elements
     lapply(x, fn_check, ...)
-
     return(invisible(x))
 }
 #' @rdname check_functions
@@ -368,21 +368,18 @@
         local_msg = "value is not a number",
         msg = msg
     )
-
     # test integer
     if (is_integer) {
         # if length is zero there is nothing to check
         if (length(x) == 0) {
             return(invisible(x))
         }
-
         .check_that(
             is.numeric(x) && all(x == suppressWarnings(as.integer(x))),
             local_msg = "value is not integer",
             msg = msg
         )
     }
-
     return(invisible(x))
 }
 #' @rdname check_functions
@@ -478,7 +475,6 @@
     }
     # check names
     .check_names(x, is_named = is_named, msg = msg)
-
     return(invisible(x))
 }
 #' @rdname check_functions
@@ -546,7 +542,6 @@
     # remove NAs before check to test tolerance
     result <- x
     x <- x[!is.na(x)]
-
     # adjust min and max to tolerance
     if (!is.null(tolerance)) {
         min <- min - tolerance
@@ -554,7 +549,6 @@
         exclusive_min <- exclusive_min - tolerance
         exclusive_max <- exclusive_max + tolerance
     }
-
     # min and max checks
[…]
