### 2023-07-18 1350e20
update predictors API
