### 2023-07-18 2182a15
update tile API with segments functions
