+++ R/sits_segment2.R
@@ -13,6 +13,8 @@ sits_segment <- function(cube,
     .check_is_regular(cube)
     .check_memsize(memsize)
     .check_output_dir(output_dir)
+    .check_version(version)
+    .check_progress(progress)
 
     # Spatial filter
     if (.has(roi)) {
@@ -68,289 +70,4 @@ sits_segment <- function(cube,
         return(segs_tile)
     })
     return(segs_cube)
-}
-
-.segment_tile <- function(tile,
-                          seg_fn,
-                          band,
-                          block,
-                          roi,
-                          output_dir,
-                          version,
-                          progress) {
-    # Output file
-    out_file <- .file_derived_name(
-        tile = tile, band = band, version = version,
-        output_dir = output_dir, ext = "gpkg"
-    )
-    # Resume feature
-    if (file.exists(out_file)) {
-        if (.check_messages()) {
-            message("Recovery: tile '", tile[["tile"]], "' already exists.")
-            message(
-                "(If you want to produce a new segmentation, please ",
-                "change 'output_dir' or 'version' parameters)"
-            )
-        }
-        seg_tile <- .segments_derived_from_file(
-            file = out_file,
-            band = band,
-            base_tile = tile,
-            derived_class = "segs_cube",
-            update_bbox = TRUE
-        )
-        return(seg_tile)
-    }
-
-    # Create chunks as jobs
-    chunks <- .tile_chunks_create(tile = tile, overlap = 0, block = block)
-    # By default, update_bbox is FALSE
-    update_bbox <- FALSE
-    if (.has(roi)) {
-        # How many chunks there are in tile?
-        nchunks <- nrow(chunks)
-        # Intersecting chunks with ROI
-        chunks <- .chunks_filter_spatial(chunks, roi = roi)
-        # Should bbox of resulting tile be updated?
-        update_bbox <- nrow(chunks) != nchunks
-    }
-    # Process jobs in parallel
-    block_files <- .jobs_map_parallel_chr(chunks, function(chunk) {
-        # Job block
-        block <- .block(chunk)
-        bbox <- .bbox(chunk)
-        # Block file name
-        hash_bundle <- digest::digest(list(block, seg_fn), algo = "md5")
-        block_file <- .file_block_name(
-            pattern = paste0(hash_bundle, "_segment"),
-            block = block,
-            output_dir = output_dir,
-            ext = "gpkg"
-        )
-        # Resume processing in case of failure
-        if (.segment_is_valid(block_file)) {
-            return(block_file)
-        }
-        # Read and preprocess values
-        values <- .segment_data_read(tile = tile, block = block)
-        # Get mask of NA pixels
-        na_mask <- C_mask_na(values)
-        # Fill with zeros remaining NA pixels
-        values <- C_fill_na(values, 0)
-        # Used to check values (below)
-        input_pixels <- nrow(values)
-        # Apply segment function
-        values <- seg_fn(values, block, bbox)
-        # Check if the result values is a vector object
-        .check_vector(values)
-        # Prepare and save results as vector
-        .vector_write_vec(v_obj = values, file = block_file)
-        # Free memory
-        gc()
-        # Returned block file
-        block_file
-    }, progress = progress)
-    # Merge blocks into a new segs_cube tile
-    seg_tile <- .tile_segment_merge_blocks(
-        block_files = block_files,
-        base_tile = tile,
-        band = "segments",
-        derived_class = "segs_cube",
-        out_file = out_file,
[…]
