### 2023-07-18 af9eed4
add a segment API
