### 2023-07-18 #596 @gilbertocamara
Dear @maelle @mpadge Begging your indulgence for being insistent, I would like to ask if there is a detailed explanation of the types of diagnostics provided by `autotest`. Consider the following case. The `sits` packages deals with big data, processing time series of satellite images. All functions that produce new images need to specify a directory where the results are stored. This is achieved by a parameter called `output_dir`, which is used in 18 functions, with the same parameter name and the same use.

Out of these 18 instances, `autotest` produces a diagnostic in only two (2) cases. In both instances, it produces a `single_char_case` diagnostic. As I understand it, this diagnostic works on the premise that changing the case of a character parameter should yield the same result. Obviously, this expectation cannot be met by operating systems where directory names are case-dependent.

Since the condition to proceed with the revision for statistical packages submitted to ROpenSci is that `autotest` should not find any problems with the code (no diagnostics, no warnings, no errors), I am at a loss on how to proceed. Please advise on what can be done in this case.

Please also explain why `autotest` only flags this condition in 2 out of the 18 cases where the parameter `output_dir` is used.

Many thanks for your help,
Gilberto
