### 2023-07-18 #596 @maelle
@gilbertocamara regarding the output you mentioned in your comment https://github.com/ropensci/software-review/issues/596#issuecomment-1636781753, can you confirm it's gone? I don't see that exact error (I'm going through your comments chronologically).
