### 2023-07-18 #596 @gilbertocamara
Dear @maelle,  above you have shown the `autotest` output without running the actual tests. My comments above refer to the output with the parameter `test` set to TRUE. This is the result that counts.
