### 2023-07-18 #596 @maelle
> I would like to ask if there is a detailed explanation of the types of diagnostics provided by autotest

Good question. To me the best answer is currently https://docs.ropensci.org/autotest/reference/autotest_types.html, does it help? I opened an issue in autotest because I agree the documentation could be improved on this front https://github.com/ropensci-review-tools/autotest/issues/83
