### 2023-07-18 #596 @maelle
I updated the results I get. Are they the same as on your machine @gilbertocamara?
