### 2023-07-18 #596 @gilbertocamara
Unfortunately, no. Please wait a little bit. Yesterday, we made some changes to `sits` trying to match the expectations of `autotest`. We are currently running the latest test. Please give me until the end of the morning BRT to provide you with an update.
