### 2023-07-18 #596 @maelle
Ok. My answer might have a few days delay depending on my availability but I'll do my best. I'll re-read autotest docs. :grin:
