### 2023-07-18 dd9ca49
improve examples, documentation and check parameter assignment
