+++ tests/testthat/test-roi.R
@@ -172,6 +172,5 @@ test_that("Internal functions in ROI", {
     expect_equal(si[["nrows"]], 84)
 })
 test_that("CRS errors", {
-    expect_error(.as_crs(TRUE))
-    expect_equal(.as_crs(NA), NA_character_)
+    expect_error(.as_crs(NA))
 })
