### 2023-07-18 2ba8940
update segments function
