+++ DESCRIPTION
@@ -177,6 +177,8 @@ Collate:
     'api_uncertainty.R'
     'api_utils.R'
     'api_variance.R'
+    'api_vector.R'
+    'api_vector_sf.R'
     'api_view.R'
     'RcppExports.R'
     'data.R'
@@ -216,6 +218,7 @@ Collate:
     'sits_resnet.R'
     'sits_sample_functions.R'
     'sits_segmentation.R'
+    'sits_segment2.R'
     'sits_select.R'
     'sits_sf.R'
     'sits_smooth.R'
