### 2023-07-19 #596 @mpadge
@gilbertocamara Please accept my apologies as lead developer of autotest for issues here. I am currently away on holidays until start of August. Thank you for taking the autotest procedures and results so seriously, but please note that autotest is currently *recommended* and not *required* infrastructure. As such, its output should currently be considered (nothing more than)  a useful guide to increase general robustness and documentation of packages prior to submission. It is not necessary for packages to completely pass autotest in order for submissions to proceed.

In short: Please use autotest to help improve your package as much as possible. Once you are satisfied, feel free to ignore any remaining autotest issues and proceed on to documentation of statistical standards compliance.

That said, please also feel free to open any issues in the autotest repository, or to ask any further questions there. The package will undergo a major revision hopefully sometime later this year, which will include numerous improvements in functionality, documentation, and general useability. Again, thank you for engaging so sincerely with these results, and apologies for any confusion during the process.
