### 2023-07-19 864257f
fix and centroid points bug
