+++ R/api_tile.R
@@ -938,6 +938,8 @@ NULL
     base_tile <- .tile(base_tile)
     # Read all blocks file
     vec_segments <- purrr::map_dfr(block_files, .vector_read_vec.sf)
+    # Define an unique ID
+    vec_segments[["pol_id"]] <- seq_len(nrow(vec_segments))
     # Write all segments
     .vector_write_vec(v_obj = vec_segments, file = out_file)
     # Create tile based on template
