### 2023-07-20 b5b7beb
add quiet parameter in sf vector api
