### 2023-07-20 f6fea58
fix bug in sits_segments
