### 2023-07-23 0e3de43
update segment API
