+++ R/sits_segment2.R
@@ -1,5 +1,5 @@
 sits_segment <- function(cube,
-                         seg_fn = sits_supercells_temp(),
+                         seg_fn = sits_slic(),
                          roi = NULL,
                          start_date = NULL,
                          end_date = NULL,
@@ -60,7 +60,7 @@ sits_segment <- function(cube,
         segs_tile <- .segment_tile(
             tile = tile,
             seg_fn = seg_fn,
-            band = "segment",
+            band = "segments",
             block = block,
             roi = roi,
             output_dir = output_dir,
