### 2023-07-26 9ebafa3
fix bug in vector cube
