### 2023-07-26 6a4abec
add vi as vector info column in sits_cube
