+++ R/api_cluster.R
@@ -57,14 +57,19 @@
 #' @param samples         Time series data and metadata
 #'                        to be used to generate the dendrogram.
 #' @param bands           Vector of bands to be clustered.
-#' @param dist_method     One of the supported distance
-#'                        from proxy's dist, e.g. \code{TWDTW}.
-#' @param linkage         Agglomeration method to be used.
-#'                        Can be any `hclust` method (see `hclust`).
-#'                        Default is 'ward.D2'.
+#' @param dist_method     One of the supported distances (single char vector)
+#'                        "dtw": DTW with a Sakoe-Chiba constraint.
+#'                        "dtw2": DTW with L2 norm and Sakoe-Chiba constraint.
+#'                        "dtw_basic": A faster DTW with less functionality.
+#'                        "lbk": Keogh's lower bound for DTW.
+#'                        "lbi": Lemire's lower bound for DTW.
+#' @param linkage         Agglomeration method to be used (single char vector)
+#'                        One of "ward.D", "ward.D2", "single", "complete",
+#'                        "average", "mcquitty", "median" or "centroid".
 #' @param  ...            Any additional parameters to be passed
 #'                        to dtwclust::tsclust() function.
-#' @return                Full dendrogram tree for data analysis.
+#' @return                Full dendrogram tree for data analysis
+#'                        (class "dendrogram")
 #'
 .cluster_dendrogram <- function(samples,
                                 bands,
