+++ NAMESPACE
@@ -86,6 +86,8 @@ S3method(.cube_token_generator,default)
 S3method(.cube_token_generator,mpc_cube)
 S3method(.data_get_ts,class_cube)
 S3method(.data_get_ts,raster_cube)
+S3method(.gc_arrange_images,"mpc_cube_sentinel-1-grd")
+S3method(.gc_arrange_images,raster_cube)
 S3method(.mosaic_split_band_date,derived_cube)
 S3method(.mosaic_split_band_date,raster_cube)
 S3method(.raster_check_package,terra)
@@ -150,6 +152,7 @@ S3method(.source_items_tile,deafrica_cube)
 S3method(.source_items_tile,hls_cube)
 S3method(.source_items_tile,sdc_cube)
 S3method(.source_items_tile,usgs_cube)
+S3method(.source_tile_get_bbox,"mpc_cube_sentinel-1-grd")
 S3method(.source_tile_get_bbox,stac_cube)
 S3method(.tile,default)
 S3method(.tile,raster_cube)
