### 2023-08-02 9547e45
Fix s3class of cubes
