+++ NAMESPACE
@@ -323,6 +323,7 @@ S3method(sits_model_export,sits_model)
 S3method(sits_reclassify,class_cube)
 S3method(sits_reclassify,default)
 S3method(sits_reclassify,tbl_df)
+S3method(sits_regularize,"mpc_cube_sentinel-1-grd")
 S3method(sits_regularize,default)
 S3method(sits_regularize,derived_cube)
 S3method(sits_regularize,raster_cube)
