### 2023-08-14 3b4de4b
add clean options as a parameter in sits_label_classification
