+++ R/sits_label_classification.R
@@ -8,6 +8,11 @@
 #'
 #' @param  cube        Classified image data cube.
 #' @param  ...         Other parameters for specific functions.
+#' @param  clean       A logical value to apply a modal function to clean up
+#'                     possible noisy pixels keeping the most frequently
+#'                     values within the neighborhood. Default is TRUE.
+#' @param window_size  An odd integer representing the size of the
+#'                     sliding window of the modal function (min = 1, max = 15).
 #' @param  multicores  Number of workers to label the classification in
 #'                     parallel.
 #' @param  memsize     maximum overall memory (in GB) to label the
@@ -52,6 +57,8 @@
 #' }
 #' @export
 sits_label_classification <- function(cube, ...,
+                                      clean = TRUE,
+                                      window_size = 3L,
                                       memsize = 4L,
                                       multicores = 2L,
                                       output_dir,
@@ -64,6 +71,8 @@ sits_label_classification <- function(cube, ...,
 #' @rdname sits_label_classification
 #' @export
 sits_label_classification.probs_cube <- function(cube, ...,
+                                                 clean = TRUE,
+                                                 window_size = 3L,
                                                  memsize = 4L,
                                                  multicores = 2L,
                                                  output_dir,
@@ -71,6 +80,8 @@ sits_label_classification.probs_cube <- function(cube, ...,
                                                  progress = TRUE) {
     # Pre-conditions - Check parameters
     .check_cube_files(cube)
+    .check_lgl_parameter(clean)
+    .check_window_size(window_size = window_size, min = 3, max = 15)
     .check_memsize(memsize, min = 1, max = 16384)
     .check_multicores(multicores, min = 1, max = 2048)
     .check_output_dir(output_dir)
@@ -80,6 +91,10 @@ sits_label_classification.probs_cube <- function(cube, ...,
 
     # Get block size
     block <- .raster_file_blocksize(.raster_open_rast(.tile_path(cube)))
+    # Get image size
+    image_size <- .raster_size(.raster_open_rast(.tile_path(cube)))
+    # Overlapping pixels
+    overlap <- ceiling(window_size / 2) - 1
     # Check minimum memory needed to process one block
     job_memsize <- .jobs_memsize(
         job_size = .block_size(block = block, overlap = 0),
@@ -113,6 +128,21 @@ sits_label_classification.probs_cube <- function(cube, ...,
             version = version,
             progress = progress
         )
+        label_path <- .tile_path(class_tile)
+        if (clean) {
+            # Apply clean in data
+            class_tile <- .clean_tile(
+                tile = class_tile,
+                block = image_size,
+                band = "class",
+                window_size = window_size,
+                overlap = overlap,
+                output_dir = output_dir,
+                version = version
+            )
+            # Remove raster file without clean
+            unlink(label_path)
+        }
         return(class_tile)
     })
     return(class_cube)
@@ -145,4 +175,3 @@ sits_label_classification.tbl_df <- function(cube, ...){
 sits_label_classification.default <- function(cube, ...) {
     stop("Input should be a classified cube")
 }
-
