### 2023-08-14 a7fbee9
Refactoring sits_segment
