+++ R/sits_segmentation.R
@@ -6,10 +6,8 @@
 #' @author Felipe Carvalho, \email{felipe.carvalho@@inpe.br}
 #'
 #' @description
-#' Apply a spatial segmentation on a data cube based on a user defined
-#' segmentation function. The user defines the tiles of
-#' the cube to be segmented and informs the bands and the date to
-#' be used. The function applies the segmentation algorithm
+#' Apply a spatial-temporal segmentation on a data cube based on a user defined
+#' segmentation function. The function applies the segmentation algorithm
 #' "seg_fn" to each tile.
 #'
 #' Segmentation uses the following steps:
@@ -29,16 +27,27 @@
 #'        Use the functions available in \code{sf} for further analysis.}
 #'  }
 #'
-#' @param cube          Regular data cube
-#' @param tiles         Tiles to be segmented
-#' @param bands         Bands to include in the segmentation
-#' @param dates         Dates to consider to be segmented
-#' @param seg_fn        Function to apply the segmentation
-#' @param ...           Other params to be passed to segmentation function
+#' @param cube        Regular data cube
+#' @param seg_fn      Function to apply the segmentation
+#' @param roi         Region of interest (see below)
+#' @param  start_date Start date for the segmentation
+#' @param  end_date   End date for the segmentation.
+#' @param  memsize    Memory available for classification (in GB).
+#' @param  multicores Number of cores to be used for classification.
+#' @param  output_dir Directory for output file.
+#' @param  version    Version of the output (for multiple
+#' segmentations).
+#' @param  progress   Show progress bar?
 #'
-#' @return              A list of "sf" objects, indexed by tile.
-#'                      Each "sf" object contains the polygons that define
-#'                      the segments.
+#' @return          A tibble of class 'segs_cube' representing the
+#' segmentation.
+#'
+#' @note
+#'    The "roi" parameter defines a region of interest. It can be
+#'    an sf_object, a shapefile, or a bounding box vector with
+#'    named XY values ("xmin", "xmax", "ymin", "ymax") or
+#'    named lat/long values ("lon_min", "lat_min", "lon_max", "lat_max")
+#'
 #' @examples
 #' if (sits_run_examples()) {
 #'     data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
@@ -51,10 +60,7 @@
 #'     # segment the image
 #'     segments <- sits_segment(
 #'         cube = cube,
-#'         tile = "012010",
-#'         bands = "NDVI",
-#'         dates = sits_timeline(cube)[1],
-#'         seg_fn = sits_supercells(step = 20)
+#'         output_dir = tempdir()
 #'     )
 #'     # create a classification model
 #'     rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
@@ -76,51 +82,79 @@
 #' }
 #' @export
 sits_segment <- function(cube,
-                         tiles = NULL,
-                         bands = NULL,
-                         dates = NULL,
-                         seg_fn, ...) {
-    # is the cube regular?
+                         seg_fn = sits_slic(),
+                         roi = NULL,
+                         start_date = NULL,
+                         end_date = NULL,
+                         memsize = 8,
+                         multicores = 2,
+                         output_dir,
+                         version = "v1",
+                         progress = TRUE) {
+    # Preconditions
+    .check_is_raster_cube(cube)
     .check_is_regular(cube)
-    # does tile belong to the cube?
-    tiles <- .default(tiles, .cube_tiles(cube))
-    .check_chr_within(
-        x = tiles,
-        within = .cube_tiles(cube),
-        msg = "tiles not available in the cube"
+    .check_memsize(memsize)
+    .check_output_dir(output_dir)
[…]
