+++ R/sits_label_classification.R
@@ -51,6 +51,8 @@
 #' }
 #' @export
 sits_label_classification <- function(cube,
+                                      clean = TRUE,
+                                      window_size = 3L,
                                       memsize = 4,
                                       multicores = 2,
                                       output_dir,
