+++ NAMESPACE
@@ -189,7 +189,7 @@ S3method(sits_cube,stac_cube)
 S3method(sits_get_data,csv)
 S3method(sits_get_data,data.frame)
 S3method(sits_get_data,default)
-S3method(sits_get_data,segments)
+S3method(sits_get_data,segs_cube)
 S3method(sits_get_data,sf)
 S3method(sits_get_data,shp)
 S3method(sits_get_data,sits)
@@ -292,12 +292,12 @@ export(sits_segment)
 export(sits_select)
 export(sits_sgolay)
 export(sits_show_prediction)
+export(sits_slic)
 export(sits_smooth)
 export(sits_som_clean_samples)
 export(sits_som_evaluate_cluster)
 export(sits_som_map)
 export(sits_stats)
-export(sits_supercells)
 export(sits_svm)
 export(sits_tae)
 export(sits_tempcnn)
