### 2023-08-17 18ac0b2
implementing new get data
