+++ R/api_data.R
@@ -46,188 +46,30 @@
     } else {
         cld_band <- NULL
     }
-    # get cubes timeline
-    tl <- sits_timeline(cube)
-    # get tile-band combination
-    tiles_bands <- tidyr::expand_grid(
-        tile = .cube_tiles(cube),
-        band = bands
-    ) |>
-        purrr::pmap(function(tile, band) {
-            return(list(tile, band))
-        })
-    # set output_dir
-    output_dir <- tempdir()
-    if (Sys.getenv("SITS_SAMPLES_CACHE_DIR") != "") {
-        output_dir <- Sys.getenv("SITS_SAMPLES_CACHE_DIR")
+    # define parallelization strategy
+    # 1st case - ...
+    if (nrow(samples) > .conf("n_samples_to_split")) {
+        ts_tbl <- .get_data_by_chunks(
+            cube = cube,
+            samples = samples,
+            bands = bands,
+            multicores = multicores,
+            progress = progress
+        )
+        return(ts_tbl)
     }
-    # prepare parallelization
-    .parallel_start(workers = multicores)
-    on.exit(.parallel_stop(), add = TRUE)
-    # get the samples in parallel using tile-band combination
-    samples_tiles_bands <- .parallel_map(
-        tiles_bands,
-        function(tile_band) {
-            tile_id <- tile_band[[1]]
-            band <- tile_band[[2]]
-
-            tile <- sits_select(cube,
-                bands = c(band, cld_band),
-                tiles = tile_id
-            )
-            hash_bundle <- digest::digest(list(tile, samples), algo = "md5")
-            # create a file to store the samples
-            filename <- .file_path(
-                "samples", hash_bundle,
-                ext = ".rds",
-                output_dir = output_dir
-            )
-            # does the file exist?
-            if (file.exists(filename)) {
-                tryCatch(
-                    {
-                        # ensure that the file is not corrupted
-                        timeseries <- readRDS(filename)
-                        return(timeseries)
-                    },
-                    error = function(e) {
-                        unlink(filename)
-                        gc()
-                    }
-                )
-            }
-            # get XY
-            xy_tb <- .proj_from_latlong(
-                longitude = samples[["longitude"]],
-                latitude  = samples[["latitude"]],
-                crs       = .cube_crs(tile)
-            )
-            # join lat-long with XY values in a single tibble
-            samples <- dplyr::bind_cols(samples, xy_tb)
-            # filter the points inside the data cube space-time extent
-            samples <- dplyr::filter(
-                samples,
-                .data[["X"]] > tile$xmin & .data[["X"]] < tile$xmax &
-                    .data[["Y"]] > tile$ymin & .data[["Y"]] < tile$ymax &
-                    .data[["start_date"]] <= as.Date(tl[length(tl)]) &
-                    .data[["end_date"]] >= as.Date(tl[1])
-            )
-            # are there points to be retrieved from the cube?
-            if (nrow(samples) == 0) {
-                return(NULL)
-            }
-            # create a matrix to extract the values
-            xy <- matrix(
-                c(samples[["X"]], samples[["Y"]]),
-                nrow = nrow(samples),
-                ncol = 2
-            )
-            colnames(xy) <- c("X", "Y")
-            # build the sits tibble for the storing the points
-            samples_tbl <- slider::slide_dfr(samples, function(point) {
-                # get the valid timeline
-                dates <- .timeline_during(
-                    timeline   = tl,
[…]
