+++ man/sits_get_data.Rd
@@ -15,8 +15,8 @@ sits_get_data(
   cube,
   samples = NULL,
   ...,
-  start_date = as.Date(sits_timeline(cube)[1]),
-  end_date = as.Date(sits_timeline(cube)[length(sits_timeline(cube))]),
+  start_date = NULL,
+  end_date = NULL,
   label = "NoClass",
   bands = sits_bands(cube),
   crs = 4326L,
@@ -45,8 +45,8 @@ sits_get_data(
   samples,
   ...,
   label = "NoClass",
-  start_date = as.Date(sits_timeline(cube)[1]),
-  end_date = as.Date(sits_timeline(cube)[length(sits_timeline(cube))]),
+  start_date = NULL,
+  end_date = NULL,
   bands = sits_bands(cube),
   label_attr = NULL,
   n_sam_pol = 30,
