### 2023-08-21 d12c4c9
add combine strategy
