### 2023-08-21 43d91c7
add new strategy to extract time series
