### 2023-08-21 bb0e88e
update strategy chunk
