+++ DESCRIPTION
@@ -121,6 +121,7 @@ Collate:
     'api_csv.R'
     'api_cube.R'
     'api_data.R'
+    'api_data2.R'
     'api_debug.R'
     'api_download.R'
     'api_factory.R'
