### 2023-08-26 4bb4da3
configuration adjustments
