### 2023-09-05 cbc27ce
update segmentation code
