+++ R/sits_get_data.R
@@ -123,6 +123,7 @@ sits_get_data <- function(cube,
     .check_crs(crs)
     .check_multicores(multicores, min = 1, max = 2048)
     .check_progress(progress)
+    samples <- .default(samples, cube)
     if (is.character(samples)) {
         class(samples) <- c(.file_ext(samples), class(samples))
     }
@@ -143,8 +144,9 @@ sits_get_data.csv <- function(cube,
                               crs = 4326L,
                               multicores = 2,
                               progress = FALSE) {
-    # Get samples
+    # Extract a data frame from csv
     samples <- .csv_get_samples(samples)
+    # Extract time series from a cube given a data.frame
     data <- .data_get_ts(
         cube       = cube,
         samples    = samples,
@@ -206,10 +208,9 @@ sits_get_data.shp <- function(cube,
 sits_get_data.sf <- function(cube,
                              samples,
                              ...,
+                             start_date = NULL,
+                             end_date = NULL,
                              bands = sits_bands(cube),
-                             start_date = as.Date(sits_timeline(cube)[1]),
-                             end_date = as.Date(sits_timeline(cube)
-                                                [length(sits_timeline(cube))]),
                              label = "NoClass",
                              label_attr = NULL,
                              n_sam_pol = 30,
@@ -223,7 +224,10 @@ sits_get_data.sf <- function(cube,
         with the id for each polygon and include
         this column name in the 'pol_id' parameter."
     )
-    # extract a samples data.frame from sf object
+    # Get default start and end date
+    start_date <- .default(start_date, .cube_start_date(cube))
+    end_date <- .default(end_date, .cube_end_date(cube))
+    # Extract a samples data.frame from sf object
     samples <- .sf_get_samples(
         sf_object  = samples,
         label      = label,
@@ -233,7 +237,7 @@ sits_get_data.sf <- function(cube,
         n_sam_pol  = n_sam_pol,
         pol_id     = pol_id
     )
-    # extract time series from a cube given a data.frame
+    # Extract time series from a cube given a data.frame
     data <- .data_get_ts(
         cube       = cube,
         samples    = samples,
@@ -255,7 +259,7 @@ sits_get_data.sits <- function(cube,
                                bands = sits_bands(cube),
                                multicores = 2,
                                progress = FALSE) {
-    # extract time series from a cube given a data.frame
+    # Extract time series from a cube given a data.frame
     data <- .data_get_ts(
         cube       = cube,
         samples    = samples,
@@ -272,38 +276,36 @@ sits_get_data.sits <- function(cube,
 sits_get_data.data.frame <- function(cube,
                                      samples,
                                      ...,
-                                     start_date = as.Date(
-                                         sits_timeline(cube)[1]
-                                     ),
-                                     end_date = as.Date(
-                                         sits_timeline(cube)[
-                                             length(sits_timeline(cube))
-                                         ]
-                                     ),
-                                     label = "NoClass",
+                                     start_date = NULL,
+                                     end_date = NULL,
                                      bands = sits_bands(cube),
+                                     label = "NoClass",
                                      crs = 4326,
                                      multicores = 2,
                                      progress = FALSE) {
-    # check if samples contains all the required columns
+    # Check if samples contains all the required columns
     .check_chr_contains(
         x = colnames(samples),
         contains = c("latitude", "longitude"),
         discriminator = "all_of",
         msg = "missing lat/long information in data frame"
     )
-    # fill missing columns
[…]
