+++ man/sits_get_data.Rd
@@ -60,9 +60,9 @@ sits_get_data(
   cube,
   samples,
   ...,
+  start_date = NULL,
+  end_date = NULL,
   bands = sits_bands(cube),
-  start_date = as.Date(sits_timeline(cube)[1]),
-  end_date = as.Date(sits_timeline(cube)[length(sits_timeline(cube))]),
   label = "NoClass",
   label_attr = NULL,
   n_sam_pol = 30,
@@ -85,10 +85,10 @@ sits_get_data(
   cube,
   samples,
   ...,
-  start_date = as.Date(sits_timeline(cube)[1]),
-  end_date = as.Date(sits_timeline(cube)[length(sits_timeline(cube))]),
-  label = "NoClass",
+  start_date = NULL,
+  end_date = NULL,
   bands = sits_bands(cube),
+  label = "NoClass",
   crs = 4326,
   multicores = 2,
   progress = FALSE
