### 2023-09-06 6edfe1c
Add extra parameter in .gc_get_valid_timeline()
