### 2023-09-06 ff8571a
update time series extraction strategy for segments
