+++ inst/extdata/config_colors.yml
@@ -1,75 +1,96 @@
 # Default Palette includes:
 # Most classes of the "Brazilian Vegetation Manual" (IBGE,2001)
-# IPCC AFOLU, TerraClass project, IGBP Discover, Copernicus Global Land Cover
+# IPCC AFOLU, PRODES, TerraClass project, IGBP Discover, Copernicus Global Land Cover
 
 # Source https://htmlcolorcodes.com/color-chart/flat-design-color-chart/
+# Colours partially adapted for Africa Standardized Terrestrial Ecosystems
+# and 1970 U.S. Vegetation Map
 
 colors:
     Tropical_Forest:
-        # Based on "nephritis" palette of Flat Design Color Chart
-        Evergreen_Broadleaf_Forest  : &forest "#1E8449"
+        # Oak-hickory-pine forest in US Veg Map
+        Evergreen_Broadleaf_Forest  : &forest "#518940"
         # Forest
         Forest                      : *forest
         Closed_Forest               : *forest
+        # Mixed Forest
+        # Southern mixed forest in US Veg Map
+        Open_Forest                 : &openforest "#53A145"
+        Mixed_Forest                : *openforest
+        Sparse_Forest               : *openforest
+        # Southern floodplain forest in US Veg Map
+        Riparian_Forest             : &riparian_forest "#8FADB1"
+        Mountainside_Forest         : *forest
+        #
+        Seasonally_Flooded          : "#A3E4D7"
         # Dense Woodlands and similar
-        Woodland                    : &woodlands "#27AE60"
+        # Oak-hickory forest in US Veg Map
+        Woodland                    : &woodlands "#91AC41"
         Dense_Woodland              : *woodlands
         Woody_Savanna               : *woodlands
-        Open_Forest                 : *woodlands
         Cerradao                    : *woodlands
-        # Mixed Forest
-        Mixed_Forest                : *woodlands
-        Sparse_Forest               : *woodlands
+
+    Wetland:
+        # Based on "belize hole" palette from Flat Design Color Chart
+        Water                       : "#2980B9"
+        Wetland                     :  &wetlands "#96C8C8"
+        Permanent_Wetland           :  *wetlands
+        # Based on "peter river" palette from Flat Design Color Chart
+        Mangrove                    : "#D4E6F1"
 
     Savanna:
         # Savanna and Cerrado
-        # Based on "emerald" palette of Flat Design Color Chart
-        Savanna                     : &savanna "#58D68D"
+        # Bluestem praire in US Veg Map
+        Savanna                     : &savanna "#a2d43f"
         Wooded_Grassland            : *savanna
         Cerrado_Strictu_Sensu       : *savanna
         Cerrado                     : *savanna
-        Campo_Cerrado               : &campo_cerrado "#ABEBC6"
+        # Wheatgrass-bluestem-needlegrass
+        Campo_Cerrado               : &campo_cerrado "#c5ed42"
         Savanna_Parkland            : *campo_cerrado
         Open_Cerrado                : *campo_cerrado
 
     Deciduous_Tropical_Forests:
         # Caatinga and decidous tropical forest
         # Based on "green sea" palette of Flat Design Color Chart
-        Caatinga                    : &caatinga "#A2D9CE"
+        Caatinga                    : &caatinga "#ffc3ca"
         Deciduous_Forest            : *caatinga
 
     Planted_Forest:
         # Silviculture and Planted Forest
-        # Based on "green sea" palette of Flat Design Color Chart
-        Planted_Forest              : &planted_forest "#45B39D"
+        # Northern hardwood
+        Planted_Forest              : &planted_forest "#90cac8"
[…]
