### 2023-09-12 3d905bd
add dates parameter in sits_select function (Closes #1002)
