+++ NAMESPACE
@@ -35,6 +35,8 @@ S3method(.cube_end_date,default)
 S3method(.cube_end_date,raster_cube)
 S3method(.cube_filter_bands,default)
 S3method(.cube_filter_bands,raster_cube)
+S3method(.cube_filter_dates,default)
+S3method(.cube_filter_dates,raster_cube)
 S3method(.cube_filter_interval,default)
 S3method(.cube_filter_interval,raster_cube)
 S3method(.cube_filter_spatial,default)
