### 2023-09-12 504e629
update sits_segment function
