+++ NAMESPACE
@@ -274,6 +274,7 @@ S3method(sits_bbox,tbl_df)
 S3method(sits_classify,default)
 S3method(sits_classify,derived_cube)
 S3method(sits_classify,raster_cube)
+S3method(sits_classify,segs_cube)
 S3method(sits_classify,sits)
 S3method(sits_classify,tbl_df)
 S3method(sits_clean,class_cube)
