### 2023-09-13 51e02e2
include different legends in color tables
