+++ R/sits_colors.R
@@ -19,6 +19,7 @@ sits_colors <- function() {
 #' @name sits_colors_show
 #' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
 #' @description         Shows the default SITS colors
+#' @param legend  One of the accepted legend in sits
 #'
 #' @return  no return, called for side effects
 #'
@@ -29,9 +30,34 @@ sits_colors <- function() {
 #' }
 #' @export
 #'
-sits_colors_show <- function() {
-    g <- .colors_show(sits_colors())
-    return(g)
+sits_colors_show <- function(legend = NULL) {
+    # verifies if stringr package is installed
+    .check_require_packages("stringr")
+    legends <- .conf("legends")
+    if (purrr::is_null(legend)) {
+        print("Showing all available colors")
+        leg <- paste0(paste("Optional - select one of the legends: "),
+               paste(legends, collapse = ", "))
+        print(leg)
+        g <- .colors_show(sits_colors())
+        return(g)
+    } else {
+        if (legend %in% legends) {
+            colors <- .conf(legend)
+            color_table_legend <- sits_colors() |>
+                dplyr::filter(.data[["name"]] %in% colors)
+            color_table_legend <- color_table_legend[
+                match(colors, color_table_legend$name), ]
+            g <- .colors_show(color_table_legend)
+            return(g)
+        } else {
+            print("Selected map legend not available")
+            leg <- paste0(paste("Please select one of the legends: "),
+                          paste(legends, collapse = ", "))
+            print(leg)
+            return(NULL)
+        }
+    }
 }
 
 #' @title Function to set sits color table
@@ -84,138 +110,4 @@ sits_colors_set <- function(color_tb) {
 sits_colors_reset <- function() {
     .conf_load_color_table()
     return(invisible(NULL))
-}
-#' @title Get colors associated to the labels
-#' @name .colors_get
-#' @param  labels  labels associated to the training classes
-#' @param  palette palette from `grDevices::hcl.pals()`
-#'                  replaces default colors
-#'                  when labels are not included in the config palette
-#' @param  rev      revert the order of colors?
-#' @keywords internal
-#' @noRd
-#' @return colors required to display the labels
-.colors_get <- function(labels,
-                        legend,
-                        palette,
-                        rev) {
-    # Get the SITS Color table
-    color_tb <- .conf_colors()
-    # Try to find colors in the SITS color palette
-    names_tb <- dplyr::filter(color_tb, .data[["name"]] %in% labels)$name
-    # find the labels that exist in the color table
-    labels_exist <- labels[labels %in% names_tb]
-    # get the colors for the names that exist
-    colors <- purrr::map_chr(labels_exist, function(l) {
-        col <- color_tb |>
-            dplyr::filter(.data[["name"]] == l) |>
-            dplyr::pull(.data[["color"]])
-        return(col)
-    })
-    # get the names of the colors that exist in the SITS color table
-    names(colors) <- labels_exist
-
-    # if there is a legend?
-    if (!purrr::is_null(legend)) {
-        # what are the names in the legend that are in the labels?
-        labels_leg <- labels[labels %in% names(legend)]
-        # what are the color labels that are included in the legend?
-        colors_leg <- legend[labels_leg]
-        # join color names in the legend to those in default colors
-        colors <- c(
-            colors_leg,
-            colors[!names(colors) %in% names(colors_leg)]
-        )
-    }
-    # are there any colors missing?
-    if (!all(labels %in% names(colors))) {
-        missing <- labels[!labels %in% names(colors)]
[…]
