+++ inst/extdata/config_colors.yml
@@ -5,43 +5,262 @@
 # Source https://htmlcolorcodes.com/color-chart/flat-design-color-chart/
 # Colours partially adapted for Africa Standardized Terrestrial Ecosystems
 # and 1970 U.S. Vegetation Map
+#
+class_schemes:
+    legends: ["IGBP", "UMD", "ESA_CCI_LC", "WORLDCOVER", "PRODES"]
+
+    IGBP: ["Evergreen_Needleleaf_Forests", "Evergreen_Broadleaf_Forests",
+       "Deciduous_Needleleaf_Forests", "Deciduous_Broadleaf_Forests",
+       "Closed_Shrublands", "Open_Shrublands",
+       "Woody_Savannas", "Savannas", "Grasslands",
+       "Permanent_Wetlands", "Croplands", "Urban_Builtup_Lands",
+       "Cropland_Vegetation_Mosaics", "Permanent_Snow_Ice",
+       "Barren", "Water_Bodies"]
+
+# IBGP Classification
+# (1) Evergreen Needleleaf Forests: dominated by evergreen conifer trees (canopy >2m).
+#     Tree cover >60%.
+# (2) Evergreen Broadleaf Forests: dominated by evergreen broadleaf and palmate
+#     trees (canopy >2m). Tree cover >60%.
+# (3) Deciduous Needleleaf Forests: dominated by deciduous needleleaf (larch)
+#     trees (canopy >2m). Tree cover >60%.
+# (4) Deciduous Broadleaf Forests: dominated by deciduous
+#     broadleaf trees (canopy >2m). Tree cover >60%.
+# (5) Mixed Forests: dominated by neither deciduous nor evergreen
+#     (40-60% of each) tree type (canopy >2m). Tree cover >60%.
+# (6) Closed Shrublands: dominated by woody perennials (1-2m height) >60% cover.
+# (7) Open Shrublands: dominated by woody perennials (1-2m height) 10-60% cover.
+# (8) Woody Savannas: tree cover 30-60% (canopy >2m).
+# (9) Savannas: tree cover 10-30% (canopy >2m).
+# (10) Grasslands: dominated by herbaceous annuals (<2m).
+# (11) Permanent Wetlands: permanently inundated lands with 30-60% water cover and >10% vegetated cover.
+# (12) Croplands: at least 60% of area is cultivated cropland.
+# (13) Urban and Built-up Lands: at least 30% impervious surface area including building materials, asphalt and vehicles.
+# (14) Cropland/Natural Vegetation Mosaics: mosaics of small-scale cultivation 40-60% with natural tree, shrub, or herbaceous vegetation.
+# (15) Permanent Snow and Ice: at least 60% of area is covered by snow and ice for at least 10 months of the year.
+# (16) Barren: at least 60% of area is non-vegetated barren (sand, rock, soil) areas with less than 10% vegetation.
+# (17) Water Bodies: at least 60% of area is covered by permanent water bodies.
+#
+    UMD: [ "Water_Bodies", "Evergreen_Needleleaf_Forests",
+       "Evergreen_Broadleaf_Forests", "Deciduous_Needleleaf_Forests",
+       "Deciduous_Broadleaf_Forests", "Mixed_Forests",
+       "Closed_Shrublands", "Open_Shrublands",
+       "Woody_Savannas", "Savannas", "Grasslands",
+       "Permanent_Wetlands", "Croplands", "Urban_Builtup_Lands",
+       "Cropland_Vegetation_Mosaics", "Non_Vegetated_Lands"]
+
+# UMD Classification
+# (1) Water Bodies: at least 60% of area is covered by permanent water bodies.
[…]

[…]
