+++ inst/extdata/config_colors.yml
@@ -7,7 +7,8 @@
 # and 1970 U.S. Vegetation Map
 #
 class_schemes:
-    legends: ["IGBP", "UMD", "ESA_CCI_LC", "WORLDCOVER", "PRODES"]
+    legends: ["IGBP", "UMD", "ESA_CCI_LC", "WORLDCOVER", "PRODES",
+              "TERRA_CLASS"]
 
     IGBP: ["Evergreen_Needleleaf_Forests", "Evergreen_Broadleaf_Forests",
        "Deciduous_Needleleaf_Forests", "Deciduous_Broadleaf_Forests",
@@ -193,8 +194,13 @@ class_schemes:
              "Mining", "Water", "Wetland",
              "Urban_Area", "Cropland", "Pasture"]
 
+    TERRA_CLASS: ["Forest", "Secondary_Vegetation","Degradation",
+                  "Deforestation",
+                  "Cropland", "Pasture", "Cerradao", "Cerrado",
+                  "Campo_Cerrado", "Mining", "Water", "Wetland",
+                  "Urban_Area"]
+
 colors:
-    Tropical_Forest:
         # Oak-hickory-pine forest in US Veg Map
         Evergreen_Broadleaf_Forest   : &forest "#518940"
         Evergreen_Broadleaf_Forests  : *forest
@@ -227,7 +233,6 @@ colors:
         Woody_Savannas              : *woodlands
         Cerradao                    : *woodlands
 
-    Deciduous_Tropical_Forests:
         # Caatinga and decidous tropical forest
         Caatinga                    : &caatinga "#c7ebdb"
         Deciduous_Forest            : *caatinga
@@ -236,14 +241,13 @@ colors:
         Deciduous_Broadleaf_Forest   : *caatinga
         Deciduous_Broadleaf_Forests   : *caatinga
 
-    Planted_Forest:
         # Silviculture and Planted Forest
         # Northern hardwood
         Planted_Forest              : &planted_forest "#90cac8"
         Planted_Forests             : *planted_forest
         Silviculture                : *planted_forest
 
-    Temperate_Forest:
+        # Temperate Forests
         # Great lakes pine forest in US Veg Map
         Evergreen_Needleleaf_Forest  : &enf "#96b1b7"
         Evergreen_Needleleaf_Forests : *enf
@@ -256,7 +260,6 @@ colors:
         Mixed_Forests               : *mixedforest
         Tree_Cover_Mixed_Leaf       : *mixedforest
 
-    Savanna:
         # Savanna and Cerrado
         # Bluestem praire in US Veg Map
         Savanna                     : &savanna "#a2d43f"
@@ -269,7 +272,7 @@ colors:
         Savanna_Parkland            : *campo_cerrado
         Open_Cerrado                : *campo_cerrado
 
-    Grassland:
+        # Grasslands and similar
         # California steppe
         Grassland                   : &grassland "#f6cc41"
         Grasslands                  : *grassland
@@ -281,7 +284,7 @@ colors:
         Dense_Herbaceous            : *pasture
         Herbaceous_Vegetation       : *pasture
 
-    Shrubland:
+        # Shrublands
         # Based on "pumpkin" palette of Flat Design Color Chart
         Closed_Shrubland            : &closed_shrubland "#BA4A00"
         Closed_Shrublands           : *closed_shrubland
@@ -293,7 +296,7 @@ colors:
         Sparse_Shrubland            : *shrubland
         Rocky_Savanna               : *shrubland
 
-    Barren:
+        # Barren and Sparse Vegetation
         # Based on "orange" palette of Flat Design Color Chart
         Barren                      : &barren "#FDEBD0"
         Bare_Area                   : *barren
[…]
