### 2023-09-15 f3c5737
Update GDAL functions
