### 2023-09-15 f11d8cf
Fix GDAL functions
