+++ R/sits_colors.R
@@ -1,25 +1,45 @@
 #' @title Function to retrieve sits color table
 #' @name sits_colors
 #' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
+#' @param legend  One of the accepted legends in sits
 #' @description Returns a color table
 #' @return              A tibble with color names and values
 #'
 #'
 #' @examples
 #' if (sits_run_examples()) {
-#'     # show the names of the colors supported by SITS
+#'     # return the names of all colors supported by SITS
 #'     sits_colors()
 #' }
 #' @export
 #'
-sits_colors <- function() {
-    return(.conf_colors())
+sits_colors <- function(legend = NULL) {
+    legends <- .conf("legends")
+    if (purrr::is_null(legend)) {
+        print("Returning all available colors")
+        return(.conf_colors())
+    } else {
+        if (legend %in% legends) {
+            colors <- .conf(legend)
+            color_table_legend <- .conf_colors() |>
+                dplyr::filter(.data[["name"]] %in% colors)
+            color_table_legend <- color_table_legend[
+                match(colors, color_table_legend$name), ]
+            return(color_table_legend)
+        } else {
+            print("Selected map legend not available")
+            leg <- paste0(paste("Please select one of the legends: "),
+                          paste(legends, collapse = ", "))
+            print(leg)
+            return(NULL)
+        }
+    }
 }
 #' @title Function to show colors in SITS
 #' @name sits_colors_show
 #' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
 #' @description         Shows the default SITS colors
-#' @param legend  One of the accepted legend in sits
+#' @param legend  One of the accepted legends in sits
 #' @param font_family A font family loaded in SITS
 #'
 #' @return  no return, called for side effects
@@ -32,37 +52,34 @@ sits_colors <- function() {
 #' @export
 #'
 sits_colors_show <- function(legend = NULL,
-                             font_family = "opensans") {
+                             font_family = "roboto") {
     # verifies if sysfonts package is installed
     .check_require_packages("sysfonts")
     # checks if font family is available
     if (!font_family %in% sysfonts::font_families())
-        font_family <- "opensans"
+        font_family <- "roboto"
+    # legend must be valid
     legends <- .conf("legends")
-    if (purrr::is_null(legend)) {
-        print("Showing all available colors")
-        leg <- paste0(paste("Optional - select one of the legends: "),
-               paste(legends, collapse = ", "))
-        print(leg)
-        g <- .colors_show(sits_colors(), font_family)
-        return(g)
-    } else {
-        if (legend %in% legends) {
-            colors <- .conf(legend)
-            color_table_legend <- sits_colors() |>
-                dplyr::filter(.data[["name"]] %in% colors)
-            color_table_legend <- color_table_legend[
-                match(colors, color_table_legend$name), ]
-            g <- .colors_show(color_table_legend, font_family)
-            return(g)
-        } else {
-            print("Selected map legend not available")
-            leg <- paste0(paste("Please select one of the legends: "),
-                          paste(legends, collapse = ", "))
-            print(leg)
-            return(NULL)
-        }
+    if (purrr::is_null(legend))
+        legend <- "none"
+    if (!(legend %in% legends)) {
+        msg <- paste0(paste("Please select one of the legends: "),
+                      paste(legends, collapse = ", "))
+        print(msg)
+        return(invisible(NULL))
     }
+    # retrieve the color names associated to the legend
+    colors <- .conf(legend)
+    # retrive the HEX codes associated to each color
+    color_table_legend <- .conf_colors() |>
+        dplyr::filter(.data[["name"]] %in% colors)
[…]
