+++ R/api_colors.R
@@ -82,10 +82,21 @@
 #' @param color_tb A SITS color table
 #' @return a gglot2 object
 .colors_show <- function(color_tb, font_family) {
+    # strip "_" if they exist
     color_tb$name <- purrr::map_chr(color_tb$name, function(name)
         { paste(name = unlist(strsplit(name, split = "_")), collapse = " ")})
+    # find out how many lines to write per name
+    color_tb$lines <- purrr::map_int(color_tb$name, function(s)
+        { return(stringr::str_count(stringr::str_wrap(s, width = 12), "\n")
+                 + 1)
+        }
+    )
     n_colors <- nrow(color_tb)
-    n_rows_show <- n_colors %/% 4
+    if (n_colors <= 12)
+        n_rows_show <- 3
+    else
+        n_rows_show <- n_colors %/% 4
+    # add place locators to color table entries
     color_tb <- tibble::add_column(color_tb,
         y = seq(0, n_colors - 1) %% n_rows_show,
         x = seq(0, n_colors - 1) %/% n_rows_show
@@ -116,7 +127,7 @@
             data = color_tb,
             mapping = ggplot2::aes(
                 x = .data[["x"]] + 0.5,
-                y = .data[["y"]] + 0.8,
+                y = .data[["y"]] + 0.6 + 0.1 * (.data[["lines"]] - 1),
                 label = stringr::str_wrap(.data[["name"]], width = 12)
             ),
             family = font_family,
@@ -130,4 +141,57 @@
         panel.background = ggplot2::element_rect(fill = "#FFFFFF")
     )
     return(g)
+}
+#'
+#' @title Write a color table in QGIS Style format
+#' @name .colors_qml
+#' @keywords internal
+#' @noRd
+#' @param color_table color table to write to QGIS
+#' @param file the file name to be written to.
+#' @return Called for side effects.
+.colors_qml <- function(color_table, file) {
+    # check that color table has index, name and color
+    .check_chr_contains(colnames(color_table), c("index", "color", "name"))
+    # open the file connection
+    con <- file(file, "w")
+
+    # read the top part of QGIS style
+    top_qgis_style <- system.file("extdata/qgis/qgis_style_top.xml",
+                                  package = "sits")
+    top_lines <- readLines(top_qgis_style)
+    # write the top part of QGIS style in the output file
+    writeLines(top_lines, con = con)
+
+    # the palette entry goes after this part
+    # write start of color palette
+    writeLines("      <colorPalette>", con = con)
+    # write palette entries
+    index <- purrr::pmap_chr(list(color_table$index,
+                         color_table$color,
+                         color_table$name),
+                    function(ind, color, name){
+                        writeLines(paste0("         <paletteEntry ",
+                                          " value=", "\"", ind, "\"",
+                                          " color=", "\"", color, "\"",
+                                          " label=", "\"", name, "\"",
+                                          " alpha=", "\"255\"", "/>"),
+                                   con = con
+                        )
+                        return(ind)
+                    })
+    # write end of color palette
+    writeLines("     </colorPalette>", con = con)
+
+    # read the bottom part of QGIS style files
+    # this part goes after the palette entry
+    bottom_qgis_style <- system.file("extdata/qgis/qgis_style_bottom.xml",
+                                     package = "sits")
+    bottom_lines <- readLines(bottom_qgis_style)
+    # write the bottom part of QGIS style in the output file
+    writeLines(bottom_lines, con = con)
+
+    # close the file
+    close(con)
+    return(invisible(NULL))
 }
