+++ man/sits_colors_show.Rd
@@ -4,10 +4,10 @@
 \alias{sits_colors_show}
 \title{Function to show colors in SITS}
 \usage{
-sits_colors_show(legend = NULL, font_family = "opensans")
+sits_colors_show(legend = NULL, font_family = "roboto")
 }
 \arguments{
-\item{legend}{One of the accepted legend in sits}
+\item{legend}{One of the accepted legends in sits}
 
 \item{font_family}{A font family loaded in SITS}
 }
