+++ man/sits_colors_qgis.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sits_colors.R
+\name{sits_colors_qgis}
+\alias{sits_colors_qgis}
+\title{Function to save color table as QML style for data cube}
+\usage{
+sits_colors_qgis(cube, file)
+}
+\arguments{
+\item{cube}{a classified data cube}
+
+\item{file}{a QGIS style file to be written to}
+}
+\value{
+No return, called for side effects
+}
+\description{
+Saves a color table associated to a classified
+             data cube as a QGIS style file
+}
+\examples{
+if (sits_run_examples()) {
+    # reset the default colors supported by SITS
+    sits_colors_reset()
+}
+}
+\author{
+Gilberto Camara, \email{gilberto.camara@inpe.br}
+}
