+++ NAMESPACE
@@ -377,6 +377,7 @@ export(sits_cluster_clean)
 export(sits_cluster_dendro)
 export(sits_cluster_frequency)
 export(sits_colors)
+export(sits_colors_qgis)
 export(sits_colors_reset)
 export(sits_colors_set)
 export(sits_colors_show)

[…]
