### 2023-09-19 ab4f0f1
replace function expanse by freq in summary() (Close #1006)
