### 2023-09-19 52bc52d
improve plotting including IBM Plex Sans font
