+++ inst/extdata/config_internals.yml
@@ -1,37 +1,21 @@
 # Internal configuration parameters - relevant only to sits developers
+# default raster package
+raster_api_package      : "terra"
 
-# bands resulting from classification and post-classification
-sits_results_bands : ["probs", "bayes", "uncert", "entropy",
-                      "class", "least", "margin", "variance"]
+# sample size
+summary_sample_size : 10000
 
-# bands resulting from classification and smoothing
-sits_probs_bands : ["probs", "bayes"]
-# bands resulting from uncertainty
-sits_uncert_bands : ["uncert", "entropy", "least", "margin"]
+# estimated relative growth size of R memory relative to block size
+processing_bloat        : 5
+processing_bloat_smooth : 8
 
-# color palettes available in sits
-sits_color_palettes:    ["BrBG", "PiYG", "PRGn", "PuOr", "RdBu", "RdGy", "RdYlBu",
-                        "RdYlGn", "Spectral", "Accent", "Dark2", "Paired",
-                        "Pastel1", "Pastel2", "Set1", "Set2", "Set3",
-                        "BuGn", "BuPu", "GnBu", "PuBu", "PuBuGn", "PuRd", "RdPu",
-                        "YlGn", "YlGnBu", "YlOrRd", "YlOrBr", "OrRd",
-                        "Blues", "Greens", "Oranges", "Reds", "Greys", "Purples"]
+# number of items returned by stac service
+rstac_pagination_limit  : 100
 
-sits_color_table_cols: ["name", "color"]
-
-# bands resulting from classification and post-classification
-sits_results_s3_class:
-                    probs: "probs_cube"
-                    bayes: "probs_cube"
-                    uncert: "uncertainty_cube"
-                    entropy: "uncertainty_cube"
-                    variance: "variance_cube"
-                    class:  "class_cube"
-
-# parsing info contents for results cube
-results_parse_info_def : ["X1", "X2", "tile", "start_date", "end_date",
-"band", "version"]
-results_parse_info_col : ["tile", "start_date", "end_date", "band"]
+# configuration for generate tokens for mpc cubes
+cube_token_generator_n_tries: 3
+cube_token_generator_sleep_time: 10
+cube_token_generator_min_remaining_time: 5
 
 # mandatory columns in data cube description
 sits_cube_cols :    ["source", "collection", "satellite", "sensor",
@@ -53,9 +37,6 @@ sample_get_data_formats : ["data.frame", "tbl_df", "sf", "sits"]
 # mandatory columns in CSV input files
 df_sample_columns : ["longitude", "latitude", "start_date", "end_date", "label"]
 
-# possible som evaluations
-som_outcomes    : ["clean", "analyze", "remove"]
-
 # supported shp types
 sf_geom_types_supported : ["POINT", "POLYGON", "MULTIPOLYGON"]
 
@@ -65,6 +46,32 @@ local_file_extensions   : ["tiff", "tif", "jp2", "hdf", "vrt"]
 local_parse_info_col    : ["tile", "band", "date"]
 local_parse_info_def    : ["X1", "X2", "tile", "band", "date"]
 
+# parse info of generated regular cubes
+reg_file_parse_info: ["cube", "tile", "band", "date"]
+
+# parsing info contents for results cube
+results_parse_info_def : ["X1", "X2", "tile", "start_date", "end_date",
+"band", "version"]
+results_parse_info_col : ["tile", "start_date", "end_date", "band"]
+
+# bands resulting from classification and post-classification
+sits_results_bands : ["probs", "bayes", "uncert", "entropy",
[…]
