### 2023-09-20 bec6c93
Labels in sits_reclassify() should consider the rules values
