+++ NAMESPACE
@@ -219,6 +219,8 @@ S3method(.tile_start_date,default)
 S3method(.tile_start_date,raster_cube)
 S3method(.tile_timeline,default)
 S3method(.tile_timeline,raster_cube)
+S3method(.tile_update_label,class_cube)
+S3method(.tile_update_label,default)
 S3method(.tile_within,default)
 S3method(.tile_within,raster_cube)
 S3method(.tile_xres,default)
