### 2023-09-20 757b923
sits_accuracy() produces error when samples labels do not match cube labels (Closes #1004)
