+++ R/api_data.R
@@ -332,3 +332,359 @@
     })
     return(ts)
 }
+
+.get_data_by_tile <- function(cube,
+                              samples,
+                              bands,
+                              cld_band,
+                              multicores,
+                              progress) {
+    # Get cube timeline
+    tl <- sits_timeline(cube)
+    # Get tile-band combination
+    tiles_bands <- .cube_split_tiles_bands(cube = cube, bands = bands)
+    # Set output_dir
+    output_dir <- tempdir()
+    if (Sys.getenv("SITS_SAMPLES_CACHE_DIR") != "") {
+        output_dir <- Sys.getenv("SITS_SAMPLES_CACHE_DIR")
+    }
+    # To avoid open more process than tiles and bands combinations
+    if (multicores > length(tiles_bands)) {
+        multicores <- length(tiles_bands)
+    }
+    # Prepare parallelization
+    .parallel_start(workers = multicores)
+    on.exit(.parallel_stop(), add = TRUE)
+    # Get the samples in parallel using tile-band combination
+    samples_tiles_bands <- .parallel_map(tiles_bands, function(tile_band) {
+        tile_id <- tile_band[[1]]
+        band <- tile_band[[2]]
+
+        tile <- sits_select(
+            data = cube,
+            bands = c(band, cld_band),
+            tiles = tile_id
+        )
+        hash_bundle <- digest::digest(list(tile, samples), algo = "md5")
+        # Create a file to store the samples
+        filename <- .file_path(
+            "samples", hash_bundle,
+            ext = ".rds",
+            output_dir = output_dir
+        )
+        # Does the file exist?
+        if (.ts_is_valid(filename)) {
+            ts <- readRDS(filename)
+            return(ts)
+        }
+        # get XY
+        xy_tb <- .proj_from_latlong(
+            longitude = samples[["longitude"]],
+            latitude  = samples[["latitude"]],
+            crs       = .cube_crs(tile)
+        )
+        # join lat-long with XY values in a single tibble
+        samples <- dplyr::bind_cols(samples, xy_tb)
+        # filter the points inside the data cube space-time extent
+        samples <- dplyr::filter(
+            samples,
+            .data[["X"]] > tile$xmin & .data[["X"]] < tile$xmax &
+                .data[["Y"]] > tile$ymin & .data[["Y"]] < tile$ymax &
+                .data[["start_date"]] <= as.Date(tl[length(tl)]) &
+                .data[["end_date"]] >= as.Date(tl[1])
+        )
+        # are there points to be retrieved from the cube?
+        if (nrow(samples) == 0) {
+            return(NULL)
+        }
+        # create a matrix to extract the values
+        xy <- matrix(
+            c(samples[["X"]], samples[["Y"]]),
+            nrow = nrow(samples),
+            ncol = 2
+        )
+        colnames(xy) <- c("X", "Y")
+        # build the sits tibble for the storing the points
+        samples_tbl <- slider::slide_dfr(samples, function(point) {
+            # get the valid timeline
+            dates <- .timeline_during(
+                timeline   = tl,
+                start_date = as.Date(point[["start_date"]]),
+                end_date   = as.Date(point[["end_date"]])
+            )
+            sample <- tibble::tibble(
+                longitude  = point[["longitude"]],
+                latitude   = point[["latitude"]],
+                start_date = dates[[1]],
+                end_date   = dates[[length(dates)]],
+                label      = point[["label"]],
+                cube       = tile[["collection"]],
+                polygon_id = point[["polygon_id"]]
[…]
