### 2023-09-22 8d8060d
add cube vector and support to vector cube classification
