+++ R/sits_get_data.R
@@ -315,43 +315,3 @@ sits_get_data.data.frame <- function(cube,
     )
     return(data)
 }
-
-#' @rdname sits_get_data
-#' @export
-sits_get_data.segs_cube <- function(cube,
-                                    samples,
-                                    ...,
-                                    bands = sits_bands(cube),
-                                    aggreg_fn = "mean",
-                                    pol_id = "pol_id",
-                                    n_sam_pol = NULL,
-                                    multicores = 1,
-                                    progress = FALSE) {
-    .check_that(
-        x = !(purrr::is_null(aggreg_fn) && purrr::is_null(n_sam_pol)),
-        msg = "either aggreg_fun or n_sam_pol must not be both NULL"
-    )
-    # extract time series from a cube from a set of segments
-    # case 1 - one point per polygon with an aggregation function
-    if (is.null(n_sam_pol)) {
-        data <- .segments_get_summary(
-            cube = cube,
-            bands = bands,
-            aggreg_fn = aggreg_fn,
-            pol_id = pol_id,
-            multicores = multicores,
-            progress = progress
-        )
-    } else {
-        data <- .segments_get_data(
-            cube = cube,
-            bands = bands,
-            pol_id = pol_id,
-            n_sam_pol = n_sam_pol,
-            multicores = multicores,
-            progress = progress
-        )
-    }
-
-    return(data)
-}
