+++ inst/extdata/config_internals.yml
@@ -156,9 +156,19 @@ derived_cube         :
 # Vector cube definitions
 vector_cube          :
     segs_cube        :
-        s3_class     : [ "segs_cube", "vector_cube"]
+        s3_class     : ["segs_cube", "vector_cube"]
         bands        :
             segments :
+    probs_vector_cube:
+        s3_class     : ["probs_vector_cube", "derived_vector_cube", "segs_cube",
+                        "vector_cube"]
+        bands        :
+            probs    :
+    class_vector_cube:
+        s3_class     : ["class_vector_cube", "derived_vector_cube", "segs_cube",
+                        "vector_cube"]
+        bands        :
+            class    :
 
 # configuration for classified cubes
 class_cube_data_type    : "INT1U"
