+++ man/sits_segment.Rd
@@ -80,28 +80,23 @@ if (sits_run_examples()) {
         collection = "MOD13Q1-6",
         data_dir = data_dir
     )
-    # segment the image
+    # segment the vector cube
     segments <- sits_segment(
         cube = cube,
         output_dir = tempdir()
     )
     # create a classification model
     rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
-    # get the average value per segment
-    samples_seg <- sits_get_data(
-        cube = cube,
-        samples = segments,
-        n_sam_pol = 10
-    )
     # classify the segments
-    seg_class <- sits_classify(
-        data = samples_seg,
-        ml_model = rfor_model
+    seg_probs <- sits_classify(
+        data = segments,
+        ml_model = rfor_model,
+        output_dir = tempdir()
     )
-    # add a column to the segments by class
-    sf_seg <- sits_join_segments(
-        data = seg_class,
-        segments = segments
+    # label the probability segments
+    seg_label <- sits_label_classification(
+        cube = seg_probs,
+        output_dir = tempdir()
     )
 }
 }
