+++ NAMESPACE
@@ -303,10 +303,12 @@ S3method(sits_get_data,sits)
 S3method(sits_label_classification,default)
 S3method(sits_label_classification,derived_cube)
 S3method(sits_label_classification,probs_cube)
+S3method(sits_label_classification,probs_vector_cube)
 S3method(sits_label_classification,raster_cube)
 S3method(sits_label_classification,tbl_df)
 S3method(sits_labels,default)
 S3method(sits_labels,derived_cube)
+S3method(sits_labels,derived_vector_cube)
 S3method(sits_labels,patterns)
 S3method(sits_labels,raster_cube)
 S3method(sits_labels,sits)
@@ -401,7 +403,6 @@ export(sits_formula_linear)
 export(sits_formula_logref)
 export(sits_geo_dist)
 export(sits_get_data)
-export(sits_join_segments)
 export(sits_kfold_validate)
 export(sits_label_classification)
 export(sits_labels)
