### 2023-09-22 86f79dd
New API to regularize
