+++ R/api_regularize.R
@@ -1,0 +1,170 @@
+.reg_tile <- function(cube, res, roi, period, output_dir, progress) {
+    # Save input cube class
+    cube_class <- class(cube)
+
+    # Create assets as jobs
+    cube_assets <- .reg_cube_split_assets(cube = cube, period = period)
+
+    # Process each tile sequentially
+    cube_assets <- .jobs_map_parallel_dfr(cube_assets, function(asset) {
+        .reg_merge_asset(
+            asset = asset,
+            res = res,
+            roi = roi,
+            output_dir = output_dir
+        )
+    }, progress = progress)
+    # Check result
+    .check_empty_data_frame(cube_assets,
+        msg = "no intersection between roi and cube"
+    )
+    # Prepare cube output
+    cube <- .cube_merge_tiles(cube_assets)
+    .set_class(cube, cube_class)
+}
+
+#' @title create assets for a data cube by assigning a unique ID using a period
+#' @noRd
+#' @param  cube  data cube
+#' @param  period period
+#' @return a data cube with assets of the same period (file ID)
+.reg_cube_split_assets <- function(cube, period) {
+    # Get timeline for the
+    timeline <- .gc_get_valid_timeline(
+        cube = cube, period = period, extra_date_step = TRUE
+    )
+    # Create assets data cube
+    .cube_foreach_tile(cube, function(tile) {
+        fi <- .fi_filter_interval(
+            fi = .fi(tile),
+            start_date = timeline[[1]],
+            end_date = timeline[[length(timeline)]]
+        )
+        groups <- cut(
+            x = .fi_timeline(fi),
+            breaks = timeline,
+            labels = FALSE
+        )
+        fi_groups <- unname(tapply(fi, groups, list))
+        assets <- .common_size(
+            .discard(tile, "file_info"),
+            feature = timeline[unique(groups)],
+            file_info = fi_groups
+        )
+        assets <- assets[, c("tile", "feature", "file_info")]
+        assets <- tidyr::unnest(assets, "file_info")
+        assets[["asset"]] <- assets[["band"]]
+        assets <- tidyr::nest(
+            assets,
+            file_info = -c("tile", "feature", "asset")
+        )
+        .common_size(
+            .discard(tile, "file_info"),
+            .discard(assets, "tile")
+        )
+    })
+}
+#' @title merges assets of a asset data cube
+#' @noRd
+#' @param  asset  assets data cube
+#' @param  period period
+#' @return a data cube with assets of the same period (file ID)
+.reg_merge_asset <- function(asset, res, roi, output_dir) {
+    # Get band conf missing value
+    band_conf <- .conf_eo_band(
+        source = "MPC",
+        collection = "SENTINEL-1-GRD",
+        band = asset[["asset"]]
+    )
+    # Prepare output file name
+    out_file <- .file_eo_name(
+        tile = asset,
+        band = asset[["asset"]],
+        date = asset[["feature"]],
+        output_dir = output_dir
+    )
+    # Resume feature
+    if (file.exists(out_file)) {
+        .check_recovery(asset[["tile"]])
+        asset <- .tile_eo_from_files(
+            files = out_file,
+            fid = .file_base(out_file),
+            bands = asset[["asset"]],
+            date = asset[["feature"]],
+            base_tile = .discard(asset, cols = c("asset", "feature")),
+            update_bbox = TRUE
+        )
+        return(asset)
+    }
+
+    # Create template based on tile metadata
+    block <- list(ncols = floor((.xmax(asset) - .xmin(asset)) / res),
[…]
