### 2023-09-22 2ed4487
Remove redundant code
