### 2023-09-22 e8c690f
Implement sits_regularize() for sentinel1 cube
