### 2023-09-22 2674b93
Change function name .reg_tile() -> .reg_cube()
