### 2023-09-22 55c5523
Remove -crop_to_cutline parameter
