+++ NAMESPACE
@@ -296,7 +296,6 @@ S3method(sits_cube,stac_cube)
 S3method(sits_get_data,csv)
 S3method(sits_get_data,data.frame)
 S3method(sits_get_data,default)
-S3method(sits_get_data,segs_cube)
 S3method(sits_get_data,sf)
 S3method(sits_get_data,shp)
 S3method(sits_get_data,sits)
