+++ R/api_reclassify.R
@@ -25,8 +25,11 @@
             band = band,
             base_tile = tile,
             derived_class = "class_cube",
+            labels = labels,
             update_bbox = FALSE
         )
+        # Update tile labels
+        class_tile <- .tile_update_label(class_tile, labels)
         return(class_tile)
     }
     # Create chunks as jobs
