### 2023-09-25 4ca5343
add tests of issue for 1005
