### 2023-09-25 26dd004
add test for issue 1004
