### 2023-09-26 581b37e
add tests for segmentation
