+++ tests/testthat/test-segmentation.R
@@ -1,6 +1,6 @@
 test_that("Segmentation", {
     # Example of classification of a data cube
-    # create a data cube from local files
+    # Create a data cube from local files
     set.seed(29031956)
     data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
     cube <- sits_cube(
@@ -9,62 +9,73 @@ test_that("Segmentation", {
         data_dir = data_dir,
         progress = FALSE
     )
-    # test sits_segments
+    # Segment the cube
     segments <- sits_segment(
         cube = cube,
-        tile = "012010",
-        bands = "NDVI",
-        date = sits_timeline(cube)[1],
-        seg_fn = sits_supercells(step = 20)
+        output_dir = tempdir(),
+        multicores = 2,
+        memsize = 4
     )
-    hash_bundle <- digest::digest(list(cube, segments), algo = "md5")
-    filename <- .file_path(
-        "samples", hash_bundle,
-        ext = ".rds",
-        output_dir = tempdir()
+    expect_s3_class(object = segments, class = "vector_cube")
+    expect_true("vector_info" %in% colnames(segments))
+    # Read segments as sf object
+    vector_segs <- .segments_read_vec(segments)
+    expect_equal(
+        as.character(unique(sf::st_geometry_type(vector_segs))),
+        expected = "POLYGON"
     )
-    file.create(filename)
-
-    sf_seg <- segments[[1]]
-    bbox <- sits_bbox(cube)
-    expect_true(all(sf_seg$x > bbox[["xmin"]]))
-    expect_true(all(sf_seg$x < bbox[["xmax"]]))
-    expect_true(all(sf_seg$y > bbox[["ymin"]]))
-    expect_true(all(sf_seg$y < bbox[["ymax"]]))
-
-    samples <- sits_get_data(
-        cube = cube,
-        samples = segments,
-        multicores = 1
+    # Train a rf model
+    rf_model <- sits_train(samples_modis_ndvi, ml_method = sits_rfor)
+    # Create a probability vector cube
+    probs_segs <- sits_classify(
+        data = segments,
+        ml_model = rf_model,
+        output_dir = tempdir(),
+        multicores = 2,
+        memsize = 4
     )
-    expect_equal(nrow(samples$time_series[[1]]), 12)
-
-    rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
-    # get the average value per segment
-    # classify the segments
-    seg_class <- sits_classify(
-        data = samples,
-        ml_model = rfor_model,
-        progress = FALSE
+    expect_s3_class(object = probs_segs, class = "probs_vector_cube")
+    expect_true(
+        "vector_info" %in% colnames(probs_segs)
     )
-    # add a column to the segments by class
-    sf_seg <- sits_join_segments(
-        data = seg_class,
-        segments = segments
+    # Read segments of a probability cube
+    vector_probs <- .segments_read_vec(probs_segs)
+    expect_true(
+        all(sits_labels(probs_segs) %in% colnames(vector_probs))
     )
-    sf_obj <- sf_seg[[1]]
-    plot(sf_obj["class"])
-    sf_obj <- sf_obj |>
-        dplyr::group_by(class) |>
-        dplyr::summarise()
-    plot(sf_obj["class"])
-
-    names(segments) <- "02010"
-    expect_error(
-        samples <- sits_get_data(
-            cube = cube,
-            samples = segments,
-            multicores = 1
-        )
+    # Create a classified vector cube
+    class_segs <- sits_label_classification(
+        cube = probs_segs,
+        output_dir = tempdir(),
+        multicores = 2,
+        memsize = 4
+    )
[…]
