+++ R/api_source_stac.R
@@ -423,7 +423,7 @@
     cube <- cube |>
         dplyr::mutate(crs2 = .data[["crs"]]) |>
         tidyr::nest(file_info = -dplyr::matches(c("tile", "crs2"))) |>
-        dplyr::rename(crs = .data[["crs2"]]) |>
+        dplyr::rename(crs = "crs2") |>
         slider::slide_dfr(function(tile) {
             # get file_info
             file_info <- tile[["file_info"]][[1]]
