+++ tests/testthat/test-raster.R
@@ -641,65 +641,6 @@ test_that("Classification with post-processing", {
     expect_true(all(file.remove(unlist(sinop_uncert$file_info[[1]]$path))))
 })
 
-test_that("Classification with clean processing", {
-    rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
-
-    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
-    sinop <- sits_cube(
-        source = "BDC",
-        collection = "MOD13Q1-6",
-        data_dir = data_dir,
-        progress = FALSE
-    )
-    output_dir <- paste0(tempdir(), "/bayes")
-    if (!dir.exists(output_dir)) {
-        dir.create(output_dir)
-    }
-
-    sinop_probs <- sits_classify(
-        data = sinop,
-        ml_model = rfor_model,
-        output_dir = output_dir,
-        memsize = 4,
-        multicores = 1,
-        progress = FALSE
-    )
-    expect_true(all(file.exists(unlist(sinop_probs$file_info[[1]]$path))))
-    sinop_class <- sits_label_classification(
-        sinop_probs,
-        output_dir = output_dir,
-        progress = FALSE
-    )
-    expect_error(
-        sits_clean(
-            cube = sinop_probs,
-            window_size = 3,
-            output_dir = output_dir,
-            memsize = 4,
-            multicores = 1,
-            progress = FALSE
-        )
-    )
-
-    sinop_clean <- sits_clean(
-        cube = sinop_class,
-        window_size = 3,
-        output_dir = output_dir,
-        memsize = 4,
-        multicores = 1,
-        progress = FALSE
-    )
-    expect_true(all(file.exists(unlist(sinop_clean$file_info[[1]]$path))))
-    expect_true(length(sits_timeline(sinop_class)) ==
-                    length(sits_timeline(sinop_clean)))
-
-    r_obj <- .raster_open_rast(sinop_clean$file_info[[1]]$path[[1]])
-    max_lab <- max(.raster_get_values(r_obj))
-    min_lab <- min(.raster_get_values(r_obj))
-    expect_true(max_lab == 4)
-    expect_true(min_lab == 1)
-})
-
 test_that("Raster GDAL datatypes", {
     gdal_type <- .raster_gdal_datatype("INT2U")
     expect_equal(gdal_type, "UInt16")
