+++ DESCRIPTION
@@ -202,7 +202,6 @@ Collate:
     'sits_csv.R'
     'sits_cube.R'
     'sits_cube_copy.R'
-    'sits_clean.R'
     'sits_cluster.R'
     'sits_factory.R'
     'sits_filters.R'

[…]
