### 2023-09-27 7902239
fix cloud cover bug
