+++ man/sits_view.Rd
@@ -164,70 +164,6 @@ if (sits_run_examples()) {
     )
     # view the uncertainty cube
     sits_view(modis_uncert)
-
-    # segment the image
-    segments <- sits_segment(
-        cube = modis_cube,
-        tile = "012010",
-        bands = "NDVI",
-        date = sits_timeline(modis_cube)[1],
-        seg_fn = sits_supercells(step = 20)
-    )
-    # view image and segments
-    sits_view(
-        modis_cube,
-        band = "NDVI",
-        segments = segments
-    )
-    # view image, classified image and segments
-    sits_view(
-        modis_cube,
-        red = "NDVI",
-        green = "NDVI",
-        blue = "NDVI",
-        class_cube = modis_label,
-        segments = segments
-    )
-    # view B/W image, classified image and segments
-    sits_view(
-        modis_cube,
-        band = "NDVI",
-        class_cube = modis_label,
-        segments = segments
-    )
-    # get the average value per segment
-    samples_seg <- sits_get_data(
-        cube = modis_cube,
-        samples = segments
-    )
-    # train a model
-    rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
-    # classify the segments
-    seg_class <- sits_classify(
-        data = samples_seg,
-        ml_model = rfor_model
-    )
-
-    # add a column to the segments by class
-    segments <- sits_join_segments(
-        data = seg_class,
-        segments = segments
-    )
-    # view image and classified segments
-    sits_view(
-        modis_cube,
-        band = "NDVI",
-        segments = segments
-    )
-    # view image, classified image and segments
-    sits_view(
-        modis_cube,
-        red = "NDVI",
-        green = "NDVI",
-        blue = "NDVI",
-        class_cube = modis_label,
-        segments = segments
-    )
 }
 }
 \author{
