### 2023-10-02 8bffaf0
fix NA bug
