### 2023-10-05 0355b48
plot vector cubes
