+++ R/sits_plot.R
@@ -315,124 +315,6 @@ plot.predicted <- function(x, y, ...,
     )
     return(invisible(p))
 }
-#' @title  Plot Segments
-#' @name plot.segments
-#' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
-#'
-#' @description Plot RGB raster cube
-#'
-#' @param  x             Object of class "segments".
-#' @param  ...           Further specifications for \link{plot}.
-#' @param  tile           Tile to be plotted.
-#' @param  legend        Named vector that associates labels to colors.
-#' @param  palette       Alternative RColorBrewer palette
-#' @param  tmap_options  List with optional tmap parameters
-#'                       tmap_max_cells (default: 1e+06)
-#'                       tmap_graticules_labels_size (default: 0.7)
-#'                       tmap_legend_title_size (default: 1.5)
-#'                       tmap_legend_text_size (default: 1.2)
-#'                       tmap_legend_bg_color (default: "white")
-#'                       tmap_legend_bg_alpha (default: 0.5)
-#'
-#' @return               A plot object with an RGB image
-#'                       or a B/W image on a color
-#'                       scale using the pallete
-#'
-#' @note To see which color palettes are supported, please run
-#' @examples
-#' if (sits_run_examples()) {
-#'     data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
-#'     cube <- sits_cube(
-#'         source = "BDC",
-#'         collection = "MOD13Q1-6",
-#'         data_dir = data_dir
-#'     )
-#'     # segment the image
-#'     segments <- sits_segment(
-#'         cube = cube,
-#'         tile = "012010",
-#'         bands = "NDVI",
-#'         date = sits_timeline(cube)[1],
-#'         seg_fn = sits_supercells(step = 20)
-#'     )
-#'     # create a classification model
-#'     rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
-#'     # get the average value per segment
-#'     samples_seg <- sits_get_data(
-#'         cube = cube,
-#'         samples = segments
-#'     )
-#'     # classify the segments
-#'     seg_class <- sits_classify(
-#'         data = samples_seg,
-#'         ml_model = rfor_model
-#'     )
-#' }
-#' @export
-plot.segments <- function(x, ...,
-                          tile = NULL,
-                          legend = NULL,
-                          palette = "Spectral",
-                          tmap_options = NULL) {
-    # check for color_palette parameter (sits 1.4.1)
-    dots <- list(...)
-    if (missing(palette) && "color_palette" %in% names(dots)) {
-        warning("please use palette in place of color_palette")
-        palette <- dots[["color_palette"]]
-    }
-    if (purrr::is_null(tile)) {
-        x <- .tile(x)
-    }
-    .check_chr_parameter(tile)
-    # retrieve the segments for this tile
-    sf_seg <- .segments_read_vec(tile)
-    # check that segments have been classified
-    .check_that("class" %in% colnames(sf_seg),
-        msg = "segments have not been classified"
-    )
-    # get the labels
-    labels <- sf_seg |>
-        sf::st_drop_geometry() |>
-        dplyr::select("class") |>
-        dplyr::distinct() |>
-        dplyr::pull()
-    names(labels) <- seq_along(labels)
-    # obtain the colors
-    colors <- .colors_get(
-        labels = labels,
-        legend = legend,
-        palette = palette,
-        rev = TRUE
-    )
-    # set the tmap options
-    tmap_params <- .plot_tmap_params(tmap_options)
-    # name the colors to match the labels
-    names(colors) <- labels
-    # join sf geometries
-    sf_seg <- sf_seg |>
-        dplyr::group_by(.data[["class"]]) |>
-        dplyr::summarise()
-    # plot the data using tmap
[…]
