+++ man/sits_slic.Rd
@@ -1,17 +1,16 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/sits_segmentation.R
-\name{sits_supercells}
-\alias{sits_supercells}
+\name{sits_slic}
 \alias{sits_slic}
-\title{Segment an image using supercells}
+\title{Segment an image using SLIC}
 \usage{
 sits_slic(
   data = NULL,
   step = 50,
   compactness = 1,
   dist_fun = "euclidean",
-  avg_fun = "mean",
-  iter = 10,
+  avg_fun = "median",
+  iter = 20,
   minarea = 30,
   verbose = FALSE
 )

[…]
