### 2023-10-05 33bbbea
fix slic defaults
