### 2023-10-05 4e53b01
fix bug in merging multiple time series bands (Closes #1017)
