### 2023-10-05 045e002
add new checker function
