+++ R/api_data.R
@@ -518,8 +518,11 @@
                                 progress) {
     # Get cube timeline
     tl <- sits_timeline(cube)
+    # transform sits tibble to sf
+    samples_sf <- sits_as_sf(samples)
     # Get chunks samples
-    chunks_samples <- .cube_split_chunks_samples(cube = cube, samples = samples)
+    chunks_samples <- .cube_split_chunks_samples(cube = cube,
+                                                 samples = samples_sf)
     # Set output_dir
     output_dir <- tempdir()
     if (Sys.getenv("SITS_SAMPLES_CACHE_DIR") != "") {
