+++ R/sits_classify.R
@@ -42,10 +42,8 @@
 #'                           (integer, min = 1, max = 2048).
 #' @param  aggreg_fn         Function to compute a summary of each segment
 #'                           (object of class "function").
-#' @param n_sam_pol          Number of samples per polygon to be read
-#'                           for POLYGON or MULTIPOLYGON shapefiles or sf
-#'                           objects.
-#'                           (single integer).
+#' @param  n_sam_pol         Number of time series per segment to be classified
+#'                           (integer, min = 10, max = 50).
 #' @param  output_dir        Valid directory for output file.
 #'                           (character vector of length 1).
 #' @param  version           Version of the output
@@ -74,6 +72,15 @@
 #'    processing. The "memsize" parameter  controls the amount of memory
 #'    available for classification. We recommend using a 4:1 relation between
 #'    "memsize" and "multicores".
+#'
+#'    For classifying vector data cubes created by
+#'    \code{\link[sits]{sits_segmentation}}, two parameters can be used:
+#'    \code{n_sam_pol}, which is the number of time series to be classified
+#'    per segment, or \code{aggreg_fn}, which is a function to aggregate the
+#'    values of all pixels in the segment for each time step. The choice
+#'    of \code{n_sam_pol} prevails over the choice of \code{aggreg_fn}. Thus,
+#'    to use \code{aggreg_fn}, the parameter \code{n_sam_pol} should be set to
+#'    NULL.
 #'
 #' @note
 #' Please refer to the sits documentation available in
@@ -115,6 +122,38 @@
 #'     )
 #'     # plot the classified image
 #'     plot(label_cube)
+#'     # segmentation
+#'     # segment the image
+#'     segments <- sits_segment(
+#'         cube = cube,
+#'         seg_fn = sits_slic(step = 5,
+#'                        compactness = 1,
+#'                        dist_fun = "euclidean",
+#'                        avg_fun = "median",
+#'                        iter = 50,
+#'                        minarea = 10,
+#'                        verbose = FALSE
+#'                        ),
+#'         output_dir = tempdir()
+#'     )
+#'     # Create a classified vector cube
+#'     probs_segs <- sits_classify(
+#'         data = segments,
+#'         ml_model = rf_model,
+#'         output_dir = tempdir(),
+#'         aggreg_fn = NULL,
+#'         n_sam_pol = 20,
+#'         multicores = 4
+#'     )
+#'     # Create a labelled vector cube
+#'     class_segs <- sits_label_classification(
+#'         cube = probs_segs,
+#'         output_dir = tempdir(),
+#'         multicores = 2,
+#'         memsize = 4
+#'     )
+#'     # plot class_segs
+#'     plot(class_segs)
 #' }
 #'
 #' @export
@@ -307,6 +346,7 @@ sits_classify.segs_cube <- function(data,
     # preconditions
     .check_is_vector_cube(data)
     .check_is_sits_model(ml_model)
+    .check_int_parameter(n_sam_pol, min = 5, max = 50)
     .check_memsize(memsize, min = 1, max = 16384)
     .check_multicores(multicores, min = 1, max = 2048)
     .check_output_dir(output_dir)
