+++ R/sits_plot.R
@@ -508,7 +508,9 @@ plot.vector_cube <- function(x, ...,
         palette <- dots[["color_palette"]]
     }
     # BW or color?
-    if (!purrr::is_null(red) && !purrr::is_null(green) && !purrr::is_null(blue))
+    if (!purrr::is_null(red) &&
+        !purrr::is_null(green) &&
+        !purrr::is_null(blue))
         bw <-  FALSE
     else
         bw <-  TRUE
@@ -578,7 +580,7 @@ plot.vector_cube <- function(x, ...,
 #' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
 #' @description plots a probability cube using stars
 #'
-#' @param  x             Object of class "probs_image".
+#' @param  x             Object of class "probs_cube".
 #' @param  ...           Further specifications for \link{plot}.
 #' @param tile           Tile to be plotted.
 #' @param labels         Labels to plot (optional).
@@ -643,6 +645,97 @@ plot.probs_cube <- function(x, ...,
 
     # plot the probs cube
     p <- .plot_probs(tile, labels, palette, rev, tmap_options)
+
+    return(p)
+}
+#' @title  Plot probability vector cubes
+#' @name   plot.probs_vector_cube
+#' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
+#' @description plots a probability cube using stars
+#'
+#' @param  x             Object of class "probs_vector_cube".
+#' @param  ...           Further specifications for \link{plot}.
+#' @param tile           Tile to be plotted.
+#' @param labels         Labels to plot (optional).
+#' @param palette        RColorBrewer palette
+#' @param rev            Reverse order of colors in palette?
+#' @param tmap_options   List with optional tmap parameters
+#'                       tmap_max_cells (default: 1e+06)
+#'                       tmap_graticules_labels_size (default: 0.7)
+#'                       tmap_legend_title_size (default: 1.5)
+#'                       tmap_legend_text_size (default: 1.2)
+#'                       tmap_legend_bg_color (default: "white")
+#'                       tmap_legend_bg_alpha (default: 0.5)
+#' @return               A plot containing probabilities associated
+#'                       to each class for each pixel.
+#'
+#'
+#' @examples
+#' if (sits_run_examples()) {
+#'     # create a random forest model
+#'     rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
+#'     # create a data cube from local files
+#'     data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
+#'     cube <- sits_cube(
+#'         source = "BDC",
+#'         collection = "MOD13Q1-6",
+#'         data_dir = data_dir
+#'     )
+#'     # segment the image
+#'     segments <- sits_segment(
+#'         cube = cube,
+#'         seg_fn = sits_slic(step = 5,
+#'                            compactness = 1,
+#'                            dist_fun = "euclidean",
+#'                            avg_fun = "median",
+#'                            iter = 20,
+#'                            minarea = 10,
+#'                            verbose = FALSE),
+#'         output_dir = tempdir()
+#'     )
+#'     # classify a data cube
+#'     probs_vector_cube <- sits_classify(
+#'         data = segments,
+#'         ml_model = rfor_model,
+#'         output_dir = tempdir()
+#'     )
+#'     # plot the resulting probability cube
+#'     plot(probs_vector_cube)
+#' }
+#'
+#' @export
+#'
+plot.probs_vector_cube <- function(x, ...,
+                                   tile = x$tile[[1]],
+                                   labels = NULL,
+                                   palette = "YlGnBu",
+                                   rev = FALSE,
[…]
