+++ R/api_plot_vector.R
@@ -1,0 +1,155 @@
+#' @title  Plot a classified vector cube
+#' @name   .plot_class_vector
+#' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
+#' @description plots a classified vector cube
+#' @keywords internal
+#' @noRd
+#' @param  tile          Tile to be plotted.
+#' @param  legend        Legend for the classes
+#' @param  palette       A sequential RColorBrewer palette
+#' @param  tmap_options  List with optional tmap parameters
+#'                       tmap max_cells (default: 1e+06)
+#'                       tmap_graticules_labels_size (default: 0.7)
+#'                       tmap_legend_title_size (default: 1.5)
+#'                       tmap_legend_text_size (default: 1.2)
+#'                       tmap_legend_bg_color (default: "white")
+#'                       tmap_legend_bg_alpha (default: 0.5)
+#'
+#' @return               A plot object
+#'
+.plot_class_vector  <- function(tile,
+                                legend,
+                                palette,
+                                tmap_options) {
+    # retrieve the segments for this tile
+    sf_seg <- .segments_read_vec(tile)
+    # check that segments have been classified
+    .check_that("class" %in% colnames(sf_seg),
+                msg = "segments have not been classified"
+    )
+    # get the labels
+    labels <- sf_seg |>
+        sf::st_drop_geometry() |>
+        dplyr::select("class") |>
+        dplyr::distinct() |>
+        dplyr::pull()
+    names(labels) <- seq_along(labels)
+    # obtain the colors
+    colors <- .colors_get(
+        labels = labels,
+        legend = legend,
+        palette = palette,
+        rev = TRUE
+    )
+    # set the tmap options
+    tmap_params <- .plot_tmap_params(tmap_options)
+    # name the colors to match the labels
+    names(colors) <- labels
+    # join sf geometries
+    sf_seg <- sf_seg |>
+        dplyr::group_by(.data[["class"]]) |>
+        dplyr::summarise()
+    # plot the data using tmap
+    p <- tmap::tm_shape(sf_seg) +
+        tmap::tm_fill(
+            col = "class",
+            palette = colors
+        ) +
+        tmap::tm_graticules(
+            labels.size = tmap_params[["labels_size"]]
+        ) +
+        tmap::tm_compass() +
+        tmap::tm_layout(
+            legend.show = TRUE,
+            legend.outside = FALSE,
+            legend.bg.color = tmap_params[["bg_color"]],
+            legend.bg.alpha = tmap_params[["bg_alpha"]],
+            legend.title.size = tmap_params[["title_size"]],
+            legend.text.size = tmap_params[["text_size"]]
+        ) +
+        tmap::tm_borders(lwd = 0.2)
+    return(p)
+}
+#' @title  Plot a probs vector cube
+#' @name   .plot_probs_vector
+#' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
+#' @description plots a classified vector cube
+#' @keywords internal
+#' @noRd
+#' @param  tile          Tile to be plotted.
+#' @param  labels_plot   Labels to be plotted
+#' @param  palette       A sequential RColorBrewer palette
+#' @param  rev           Revert the color of the palette?
+#' @param  tmap_options  List with optional tmap parameters
+#'                       tmap max_cells (default: 1e+06)
+#'                       tmap_graticules_labels_size (default: 0.7)
+#'                       tmap_legend_title_size (default: 1.5)
+#'                       tmap_legend_text_size (default: 1.2)
+#'                       tmap_legend_bg_color (default: "white")
+#'                       tmap_legend_bg_alpha (default: 0.5)
+#'
[…]
