+++ NAMESPACE
@@ -246,6 +246,7 @@ S3method(plot,geo_distances)
 S3method(plot,patterns)
 S3method(plot,predicted)
 S3method(plot,probs_cube)
+S3method(plot,probs_vector_cube)
 S3method(plot,raster_cube)
 S3method(plot,rfor_model)
 S3method(plot,sits)

[…]
