### 2023-10-15 48c411b
Improve documentation of sits_cube
