### 2023-10-16 2cbcfcb
update .check_chr_parameter function
