### 2023-10-16 6779e4d
add cloud coverage information in raster cube summary (Close #1003)
