+++ R/sits_cube.R
@@ -235,7 +235,7 @@
 #'         end_date = "2019-07-23"
 #'     )
 #'
-#'     # -- Creating Sentinel cube from MPC"
+#'     # -- Creating Sentinel cube from MPC
 #'     s2_cube <- sits_cube(
 #'         source = "MPC",
 #'         collection = "SENTINEL-2-L2A",
