+++ R/sits_classify.R
@@ -212,7 +212,7 @@ sits_classify.raster_cube <- function(data,
     .check_memsize(memsize, min = 1, max = 16384)
     .check_multicores(multicores, min = 1, max = 2048)
     .check_output_dir(output_dir)
-    .check_version(version)
+    version <- .check_version(version)
     .check_progress(progress)
 
     # version is case-insensitive in sits
@@ -351,7 +351,7 @@ sits_classify.segs_cube <- function(data,
     .check_memsize(memsize, min = 1, max = 16384)
     .check_multicores(multicores, min = 1, max = 2048)
     .check_output_dir(output_dir)
-    .check_version(version)
+    version <- .check_version(version)
     .check_progress(progress)
 
     # version is case-insensitive in sits
