+++ R/sits_smooth.R
@@ -76,7 +76,7 @@ sits_smooth <- function(cube, window_size = 7L, neigh_fraction = 0.5,
     output_dir <- path.expand(output_dir)
     .check_output_dir(output_dir)
     # Check version
-    .check_version(version)
+    version <- .check_version(version)
     # get nlabels
     nlabels <- length(sits_labels(cube))
     # Check smoothness
