### 2023-10-20 71f8a61
update segments code
