### 2023-10-20 98fd3ad
support for local vector cubes
