+++ R/api_source_local.R
@@ -6,6 +6,8 @@
 #' @param collection   Image collection in data source (To find out
 #'  the supported collections, use \code{\link{sits_list_collections}()}).
 #' @param data_dir     Local directory where images are stored.
+#' @param vector_dir    Local director where vector files are stored
+#'                     (for local vector cubes - character vector of length 1)
 #' @param parse_info   Parsing information for local files.
 #' @param version      Version id for local files.
 #' @param delim        Delimiter for parsing local files.
@@ -13,6 +15,7 @@
 #'                     the cube (see details below).
 #' @param bands        Spectral bands and indices to be included
 #'                     in the cube (optional).
+#' @param vector_band  Band for vector data cube
 #' @param labels       Labels associated to the classes (only for result cubes)
 #' @param start_date,end_date Initial and final dates to include
 #'                     images from the collection in the cube (optional).
@@ -23,11 +26,13 @@
 .local_cube <- function(source,
                         collection,
                         data_dir,
+                        vector_dir,
                         parse_info,
                         version,
                         delim,
                         tiles,
                         bands,
+                        vector_band,
                         labels,
                         start_date,
                         end_date,
@@ -35,6 +40,8 @@
                         progress, ...) {
     # set caller to show in errors
     .check_set_caller(".local_cube")
+    # initialize vector items
+    vector_items <- NULL
     # is this a cube with results?
     results_cube <- .check_cube_is_results_cube(bands, labels)
 
@@ -45,7 +52,7 @@
     bands <- .band_set_case(bands)
 
     # make query and retrieve items
-    items <- .local_cube_items_new(
+    raster_items <- .local_cube_items_raster_new(
         data_dir = data_dir,
         parse_info = parse_info,
         version = version,
@@ -54,52 +61,71 @@
         end_date = end_date,
         bands = bands
     )
+    if (!purrr::is_null(vector_dir)) {
+        # set the correct parse_info
+        parse_info <- .conf("results_parse_info_def")
+
+        vector_items <- .local_cube_items_vector_new(
+            vector_dir = vector_dir,
+            parse_info = parse_info,
+            version = version,
+            delim = delim,
+            start_date = start_date,
+            end_date = end_date,
+            vector_band = vector_band
+        )
+    }
 
     # filter bands in items (only for raw image cube)
     if (!results_cube) {
-        items <- .local_cube_items_bands_select(
+        raster_items <- .local_cube_items_bands_select(
             source = source,
             collection = collection,
             bands = bands,
-            items = items
+            items = raster_items
         )
     }
     # filter tiles
     if (!purrr::is_null(tiles)) {
-        items <- .local_cube_items_tiles_select(
+        raster_items <- .local_cube_items_tiles_select(
             tiles = tiles,
-            items = items
+            items = raster_items
         )
+        if (!purrr::is_null(vector_items)) {
+            vector_items <- .local_cube_items_tiles_select(
+                tiles = tiles,
+                items = vector_items
+            )
+        }
     }
-
     # build file_info for the items
     if (results_cube) {
-        items <- .local_results_cube_file_info(
-            items = items,
+        raster_items <- .local_results_cube_file_info(
+            items = raster_items,
             multicores = multicores,
             progress = progress
         )
     } else {
-        items <- .local_cube_file_info(
-            items = items,
+        raster_items <- .local_cube_file_info(
+            items = raster_items,
             multicores = multicores,
             progress = progress
         )
     }
 
     # get all tiles
[…]
