+++ R/api_check.R
@@ -1132,7 +1132,8 @@
 #' @return Called for side effects.
 #' @keywords internal
 #' @noRd
-.check_chr_parameter <- function(param, len_min = 1, len_max = 1) {
+.check_chr_parameter <- function(param,
+                                 len_min = 1, len_max = 1, msg = NULL) {
     .check_chr(
         param,
         len_min = len_min,
@@ -1140,6 +1141,7 @@
         allow_na = FALSE,
         allow_empty = FALSE,
         allow_null = FALSE,
+        msg = msg
     )
     return(invisible(param))
 }
@@ -1457,7 +1459,7 @@
         )
 
         # is label parameter was provided in labelled cubes?
-        if (bands %in% c("probs", "bayes", "probs-vector")) {
+        if (bands %in% c("probs", "bayes")) {
             .check_chr(
                 labels,
                 len_min = 1,
@@ -1467,7 +1469,7 @@
             )
         }
         # labels should be named in class cubes?
-        if (bands %in% c("class", "class-vector")) {
+        if (bands %in% c("class")) {
             .check_length(
                 labels,
                 len_min = 2,
