### 2023-10-21 a4632d3
new functions for sampling points
