### 2023-10-22 ae228ad
Revert "update segments code"
This reverts commit 71f8a6113dbd8a57ee9563aee5ccf8f009b4fb61.
