+++ R/api_source_local.R
@@ -76,7 +76,6 @@
     if (results_cube) {
         items <- .local_results_cube_file_info(
             items = items,
-            bands = bands,
             multicores = multicores,
             progress = progress
         )
@@ -169,8 +168,8 @@
     )
     # post-condition
     .check_chr(img_files,
-               allow_empty = FALSE, len_min = 1,
-               msg = "no file found in provided directory"
+        allow_empty = FALSE, len_min = 1,
+        msg = "no file found in provided directory"
     )
 
     # remove the extension
@@ -204,7 +203,7 @@
     # joint the list into a tibble and convert bands name to upper case
     items <- suppressMessages(
         tibble::as_tibble(img_files_mx,
-                          .name_repair = "universal"
+            .name_repair = "universal"
         )
     )
     # read the image files into a tibble with added parse info
@@ -212,7 +211,7 @@
     # joint the list into a tibble and convert bands name to upper case
     items <- suppressMessages(
         tibble::as_tibble(img_files_mx,
-                          .name_repair = "universal"
+            .name_repair = "universal"
         )
     )
     if (!purrr::is_null(bands)) {
@@ -308,7 +307,7 @@
     }
     # post-condition
     .check_that(nrow(items) > 0,
-                msg = "no files found in the interval"
+        msg = "no files found in the interval"
     )
     return(items)
 }
@@ -344,8 +343,8 @@
     if (!purrr::is_null(bands)) {
         # verify that the requested bands exist
         .check_chr_within(bands,
-                          within = unique(items[["band"]]),
-                          msg = "invalid 'bands' value"
+            within = unique(items[["band"]]),
+            msg = "invalid 'bands' value"
         )
         # select the requested bands
         items <- dplyr::filter(items, .data[["band"]] %in% !!bands)
@@ -366,8 +365,8 @@
     # filter tiles
     # verify that the requested tiles exist
     .check_chr_within(tiles,
-                      within = unique(items[["tile"]]),
-                      msg = "invalid 'tiles' value"
+        within = unique(items[["tile"]]),
+        msg = "invalid 'tiles' value"
     )
     # select the requested tiles
     items <- dplyr::filter(items, .data[["tile"]] %in% !!tiles)
@@ -388,7 +387,7 @@
     .check_set_caller(".local_cube_file_info")
     # post-condition
     .check_that(nrow(items) > 0,
-                msg = "invalid 'items' parameter"
+        msg = "invalid 'items' parameter"
     )
     # add feature id (fid)
     items <- dplyr::group_by(items, .data[["tile"]], .data[["date"]]) |>
@@ -437,8 +436,8 @@
     errors <- unlist(purrr::map(results_lst, `[[`, "error"))
     if (length(errors) > 0) {
         warning("cannot open file(s): ",
-                paste0("'", errors, "'", collapse = ", "),
-                call. = FALSE, immediate. = TRUE
+            paste0("'", errors, "'", collapse = ", "),
+            call. = FALSE, immediate. = TRUE
         )
     }
     items <- dplyr::bind_rows(items) |>
@@ -450,17 +449,16 @@
 #' @keywords internal
 #' @noRd
 #' @param tiles        Tiles in data cube.
-#' @param bands        Cube bands
 #' @param multicores   Number of workers for parallel processing
 #' @param progress     Show a progress bar?
 #' @return  Items with file info information
-.local_results_cube_file_info <- function(items, bands, multicores, progress) {
+.local_results_cube_file_info <- function(items, multicores, progress) {
     # set caller to show in errors
     .check_set_caller(".local_results_cube_file_info")
 
     # post-condition
     .check_that(nrow(items) > 0,
-                msg = "invalid 'items' parameter"
[…]
