+++ R/sits_segmentation.R
@@ -57,7 +57,7 @@
 #'     # segment the vector cube
 #'     segments <- sits_segment(
 #'         cube = cube,
-#'         output_dir = "/home/sits/test_seg/"
+#'         output_dir = tempdir()
 #'     )
 #'     # create a classification model
 #'     rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
@@ -65,12 +65,12 @@
 #'     seg_probs <- sits_classify(
 #'         data = segments,
 #'         ml_model = rfor_model,
-#'         output_dir = "/home/sits/test_seg/"
+#'         output_dir = tempdir()
 #'     )
 #'     # label the probability segments
 #'     seg_label <- sits_label_classification(
 #'         cube = seg_probs,
-#'         output_dir = = "/home/sits/test_seg/"
+#'         output_dir = tempdir()
 #'     )
 #' }
 #' @export
