### 2023-10-22 a615bf8
new sample method using inclusion
