+++ src/RcppExports.cpp
@@ -249,15 +249,39 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// sample_points
-NumericMatrix sample_points(const NumericMatrix& polymatrix, const int n_sam_pol);
-RcppExport SEXP _sits_sample_points(SEXP polymatrixSEXP, SEXP n_sam_polSEXP) {
+// sample_points_inclusion
+NumericMatrix sample_points_inclusion(const NumericMatrix& polymatrix, const int n_sam_pol);
+RcppExport SEXP _sits_sample_points_inclusion(SEXP polymatrixSEXP, SEXP n_sam_polSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const NumericMatrix& >::type polymatrix(polymatrixSEXP);
     Rcpp::traits::input_parameter< const int >::type n_sam_pol(n_sam_polSEXP);
-    rcpp_result_gen = Rcpp::wrap(sample_points(polymatrix, n_sam_pol));
+    rcpp_result_gen = Rcpp::wrap(sample_points_inclusion(polymatrix, n_sam_pol));
+    return rcpp_result_gen;
+END_RCPP
+}
+// sample_points_crossings
+NumericMatrix sample_points_crossings(const NumericMatrix& polymatrix, const int n_sam_pol);
+RcppExport SEXP _sits_sample_points_crossings(SEXP polymatrixSEXP, SEXP n_sam_polSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const NumericMatrix& >::type polymatrix(polymatrixSEXP);
+    Rcpp::traits::input_parameter< const int >::type n_sam_pol(n_sam_polSEXP);
+    rcpp_result_gen = Rcpp::wrap(sample_points_crossings(polymatrix, n_sam_pol));
+    return rcpp_result_gen;
+END_RCPP
+}
+// sample_points_grid
+NumericMatrix sample_points_grid(const NumericMatrix& polymatrix, const int n_sam_pol);
+RcppExport SEXP _sits_sample_points_grid(SEXP polymatrixSEXP, SEXP n_sam_polSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const NumericMatrix& >::type polymatrix(polymatrixSEXP);
+    Rcpp::traits::input_parameter< const int >::type n_sam_pol(n_sam_polSEXP);
+    rcpp_result_gen = Rcpp::wrap(sample_points_grid(polymatrix, n_sam_pol));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -429,7 +453,9 @@ static const R_CallMethodDef CallEntries[] = {
     {"_sits_C_nnls_solver_batch", (DL_FUNC) &_sits_C_nnls_solver_batch, 5},
     {"_sits_C_normalize_data", (DL_FUNC) &_sits_C_normalize_data, 3},
     {"_sits_C_normalize_data_0", (DL_FUNC) &_sits_C_normalize_data_0, 3},
-    {"_sits_sample_points", (DL_FUNC) &_sits_sample_points, 2},
+    {"_sits_sample_points_inclusion", (DL_FUNC) &_sits_sample_points_inclusion, 2},
[…]
