+++ src/sample_points.cpp
@@ -17,8 +17,20 @@ using namespace Rcpp;
 struct Point {
     double x, y;
 };
-
-bool is_inside(const std::vector<Point>& polygon, const Point& p) {
+bool is_inside_inclusion(const std::vector<Point>& polygon, const Point& p) {
+    int i, j, c = 0;
+    int nvert = polygon.size();
+    for (i = 0, j = nvert - 1; i < nvert; j = i++) {
+        if ( ((polygon[i].y > p.y) != (polygon[j].y > p.y)) &&
+             (p.x < (polygon[j].x - polygon[i].x) *
+             (p.y - polygon[i].y) / (polygon[j].y - polygon[i].y) +
+             polygon[i].x)
+        )
+            c = !c;
+    }
+    return c;
+}
+bool is_inside_crossings(const std::vector<Point>& polygon, const Point& p) {
     int crossings = 0;
     size_t num_vertices = polygon.size();
     for (size_t i = 0; i < num_vertices; ++i) {
@@ -32,8 +44,83 @@ bool is_inside(const std::vector<Point>& polygon, const Point& p) {
     }
     return (crossings % 2 > 0);
 }
+//****************************************************************************
+//****************************************************************************80
+//
+//  Purpose:
+//
+//    POLYGON_GRID_POINTS computes points on a polygonal grid.
+//
+//  Licensing:
+//    This code is distributed under the GNU LGPL license.
+//
+//  Modified:
+//    11 May 2015
+//  Author:
+//    John Burkardt
+//    (original code in C)
+//    Modified to C++ by Gilberto Camara
+//
+//  Parameters:
+//    int n_intervals,              number of subintervals.
+//    std::vector<Point>& polygon   input polygon
+//    int n_grid_points             number of points to be obtained
+//
+//    Output
+//    std::vector<Point> with the grid points.
+//
+std::vector<Point> polygon_grid_points (int n_intervals,
+                                        const std::vector<Point>& polygon,
+                                        int n_grid_points) {
+    int p = 0;
+    Point vc;
+    std::vector<Point> points(n_grid_points);  //vector of points
+    int n_vertices = polygon.size();
+    //
+    //  Determine the centroid.
+    //
+    vc.x = 0.0;
+    vc.y = 0.0;
+    for (int j = 0; j < n_vertices; j++)
+    {
+        vc.x = vc.x + polygon[j].x;
+        vc.y = vc.y + polygon[j].y;
+    }
+    vc.x = vc.x / n_vertices;
+    vc.y = vc.y / n_vertices;
+    std::cout << "vc.x " << vc.x;
+    std::cout << "vc.y " << vc.y;
+    //
+    //  The centroid is the first point.
+    //
+    points[0].x = vc.x;
+    points[0].y = vc.y;
+    p = p + 1;
+    //
+    //  Consider each triangle formed by two consecutive vertices and the centroid,
+    //  but skip the first line of points.
+    //
+    for (int l = 0; l < n_vertices; l++)
+    {
+        int lp1 = ((l + 1) % n_vertices);
+        for (int i = 1; i <= n_intervals; i++)
+        {
+            for (int j = 0; j <= n_intervals - i; j++)
+            {
+                int k = n_intervals - i - j;
+                points[p].x = (i * polygon[l].x + j * polygon[lp1].x + k * vc.x)
+                    / n_intervals;
+                points[p].y = (i * polygon[l].y + j * polygon[lp1].y + k * vc.y)
+                    / n_intervals;
+                p = p + 1;
+                if (p == n_grid_points) break;
+            }
+        }
+    }
[…]
