### 2023-10-22 f3ea0b0
raster classification with GPU
