+++ tests/testthat/test-summary.R
@@ -49,12 +49,7 @@ test_that("summary sits area accuracy", {
         progress = FALSE
     )
     sum_probs <- capture.output(suppressWarnings(summary(probs_cube)))
-    expect_true(grepl("class", sum_probs[1]))
-    expect_true(grepl("dimensions", sum_probs[2]))
-    expect_true(grepl("resolution", sum_probs[3]))
-    expect_true(grepl("extent", sum_probs[4]))
-    expect_true(grepl("coord", sum_probs[5]))
-    expect_true(grepl("Min", sum_probs[7]))
+    expect_true(any(grepl("Min", sum_probs)))
 
     # get the variance cube
     variance_cube <- sits_variance(
@@ -62,12 +57,7 @@ test_that("summary sits area accuracy", {
         output_dir = tempdir()
     )
     sum_var <- capture.output(suppressWarnings(summary(variance_cube)))
-    expect_true(grepl("class", sum_var[1]))
-    expect_true(grepl("dimensions", sum_var[2]))
-    expect_true(grepl("resolution", sum_var[3]))
-    expect_true(grepl("extent", sum_var[4]))
-    expect_true(grepl("coord", sum_var[5]))
-    expect_true(grepl("Min", sum_var[7]))
+    expect_true(any(grepl("Min", sum_var)))
 
     # label the probability cube
     label_cube <- sits_label_classification(
@@ -76,12 +66,7 @@ test_that("summary sits area accuracy", {
         progress = FALSE
     )
     sum_label <- capture.output(suppressWarnings(summary(label_cube)))
-    expect_true(grepl("class", sum_label[1]))
-    expect_true(grepl("dimensions", sum_label[2]))
-    expect_true(grepl("resolution", sum_label[3]))
-    expect_true(grepl("extent", sum_label[4]))
-    expect_true(grepl("coord", sum_label[5]))
-    expect_true(grepl("area_km2", sum_label[6]))
+    expect_true(any(grepl("area_km2", sum_label)))
 
     # obtain the ground truth for accuracy assessment
     ground_truth <- system.file("extdata/samples/samples_sinop_crop.csv",
