+++ R/sits_classify.R
@@ -249,8 +249,8 @@ sits_classify.raster_cube <- function(data,
         .check_int_parameter(gpu_memory, min = 1, max = 16384,
                              msg = "Using GPU: gpu_memory must be informed")
 
-        jobs_memsize = gpu_memory
-        multicores = 1
+        job_memsize <-  gpu_memory
+        multicores  <- 1
     } else {
         # Check minimum memory needed to process one block
         job_memsize <- .jobs_memsize(
