### 2023-10-23 a8deaf9
fix memsize for GPU
