### 2023-10-23 18f7b5a
fix raster processing in GPU
