+++ R/sits_smooth.R
@@ -110,7 +110,7 @@ sits_smooth.probs_cube <- function(cube, window_size = 7L, neigh_fraction = 0.5,
         job_size = .block_size(block = block, overlap = overlap),
         npaths = length(.tile_labels(cube)) * 2,
         nbytes = 8,
-        proc_bloat = .conf("processing_bloat")
+        proc_bloat = .conf("processing_bloat_cpu")
     )
     # Update multicores parameter
     multicores <- .jobs_max_multicores(
