+++ R/sits_mixture_model.R
@@ -188,7 +188,7 @@ sits_mixture_model.raster_cube <- function(data, endmembers, ...,
     job_memsize <- .jobs_memsize(
         job_size = .block_size(block = block, overlap = 0),
         npaths = length(bands) + length(out_fracs),
-        nbytes = 8, proc_bloat = .conf("processing_bloat")
+        nbytes = 8, proc_bloat = .conf("processing_bloat_cpu")
     )
     # Update multicores parameter
     multicores <- .jobs_max_multicores(
