+++ inst/extdata/config_internals.yml
@@ -13,8 +13,10 @@ n_samples_to_split: 1000
 
 # estimated relative growth size of R memory relative to block size
 processing_bloat        : 5
+processing_bloat_cpu    : 5
 processing_bloat_smooth : 8
 processing_bloat_seg    : 2
+processing_bloat_gpu    : 2
 
 # number of items returned by stac service
 rstac_pagination_limit  : 100
