+++ src/sample_points.cpp
@@ -1,6 +1,7 @@
 #include <Rcpp.h>
 #include <iostream>
 #include <vector>
+#include <list>
 #include <cstdlib>
 #include <ctime>
 using namespace Rcpp;
@@ -119,6 +120,83 @@ std::vector<Point> polygon_grid_points (int n_intervals,
     }
     return points;
 }
+// ======= Trapezoid (bin) algorithm =======================================
+
+// Split polygons along set of y bins and sorts the edge fragments.  Testing
+// is done against these fragments.
+
+struct Edge {
+    Point p1, p2;
+};
+
+class BinnedPolygon {
+public:
+    BinnedPolygon(const std::vector<Point>& vertices, int binSize)
+        : binSize(binSize) {
+        // Find the bounding box of the polygon
+        for (const auto& v : vertices) {
+            minX = std::min(minX, v.x);
+            minY = std::min(minY, v.y);
+            maxX = std::max(maxX, v.x);
+            maxY = std::max(maxY, v.y);
+        }
+
+        // Create the bins
+        int binsX = (maxX - minX) / binSize + 1;
+        int binsY = (maxY - minY) / binSize + 1;
+        bins.resize(binsX, std::vector<std::list<Edge>>(binsY));
+
+        // Insert the edges into the bins
+        for (size_t i = 0; i < vertices.size(); ++i) {
+            size_t j = (i + 1) % vertices.size();
+            insertEdge({vertices[i], vertices[j]});
+        }
+    }
+
+    bool isInside(const Point& point) const {
+        int binX = (point.x - minX) / binSize;
+        int binY = (point.y - minY) / binSize;
+
+        if (binX < 0 || binX >= bins.size() || binY < 0 || binY >= bins[0].size())
+            return false;
+
+        // Ray casting algorithm
+        int intersections = 0;
+        for (const auto& edge : bins[binX][binY]) {
+            if ((edge.p1.y > point.y) != (edge.p2.y > point.y) &&
+                point.x < (edge.p2.x - edge.p1.x) * (point.y - edge.p1.y) /
+                    (edge.p2.y - edge.p1.y) + edge.p1.x) {
+                ++intersections;
+            }
+        }
+        return (intersections % 2) == 1;
+    }
+
+private:
+    void insertEdge(const Edge& edge) {
+        // Find the bins that this edge intersects with
+        int minXBin = std::min(edge.p1.x, edge.p2.x);
+        int maxXBin = std::max(edge.p1.x, edge.p2.x);
+        int minYBin = std::min(edge.p1.y, edge.p2.y);
+        int maxYBin = std::max(edge.p1.y, edge.p2.y);
+
+        for (int x = minXBin; x <= maxXBin; x += binSize) {
+            for (int y = minYBin; y <= maxYBin; y += binSize) {
+                int binX = (x - minX) / binSize;
+                int binY = (y - minY) / binSize;
+                bins[binX][binY].push_back(edge);
+            }
+        }
+    }
+
+    double minX = std::numeric_limits<double>::infinity();
+    double minY = std::numeric_limits<double>::infinity();
+    double maxX = -std::numeric_limits<double>::infinity();
+    double maxY = -std::numeric_limits<double>::infinity();
+    int binSize;
+    std::vector<std::vector<std::list<Edge>>> bins;
+};
+
 // [[Rcpp::export]]
 NumericMatrix sample_points_inclusion(const NumericMatrix& polymatrix,
                             const int n_sam_pol){
@@ -220,4 +298,44 @@ NumericMatrix sample_points_grid(const NumericMatrix& polymatrix,
[…]
