### 2023-10-25 e4f8e5f
fix extraction of samples inside segments
