+++ R/api_segments.R
@@ -93,298 +93,6 @@
     seg_tile
 }
 
-#' @title Extract set of time series from supercells
-#'
-#' @name .segments_get_summary
-#' @keywords internal
-#' @noRd
-#' @description     Using the segments as polygons, get all time series
-#'
-#' @param cube       regular data cube
-#' @param bands      bands used in time series
-#' @param aggreg_fn  Function to compute a summary of each segment
-#' @param pol_id     ID attribute for polygons.
-#' @param multicores Number of cores to use for processing
-#' @param progress   Show progress bar?
-#'
-.segments_get_summary <- function(cube,
-                                  bands,
-                                  aggreg_fn,
-                                  pol_id,
-                                  multicores,
-                                  progress) {
-    # Verify if exactextractr is installed
-    .check_require_packages("exactextractr")
-    # Get start and end dates
-    start_date <- .cube_start_date(cube)
-    end_date <- .cube_end_date(cube)
-    # Get chunks samples
-    chunks_samples <- slider::slide(cube, function(tile) {
-        # Get segments from tile
-        segments_sf <- .get_segments_from_cube(tile)
-        .cube_split_chunks_samples(cube = tile, samples_sf = segments_sf)
-    })
-    chunks_samples <- unlist(chunks_samples, recursive = FALSE)
-
-    # Set output_dir
-    output_dir <- tempdir()
-    if (Sys.getenv("SITS_SAMPLES_CACHE_DIR") != "") {
-        output_dir <- Sys.getenv("SITS_SAMPLES_CACHE_DIR")
-    }
-    # To avoid open more process than chunks and samples combinations
-    if (multicores > length(chunks_samples)) {
-        multicores <- length(chunks_samples)
-    }
-    .parallel_start(workers = multicores)
-    on.exit(.parallel_stop(), add = TRUE)
-
-    samples_tiles_bands <- .parallel_map(chunks_samples, function(chunk) {
-        tile <- sits_select(
-            data = cube,
-            bands = bands,
-            tiles = chunk[["tile"]]
-        )
-        # Get chunk segments
-        segs_tile <- chunk[["samples"]][[1]]
-        # create hash for combination of tile and samples
-        hash_bundle <- digest::digest(list(tile, segs_tile), algo = "md5")
-        # create a file with a hash code
-        filename <- .file_path(
-            "samples", hash_bundle,
-            ext = ".rds",
-            output_dir = output_dir
-        )
-        # test if file exists
-        if (file.exists(filename)) {
-            tryCatch(
-                {
-                    # ensure that the file is not corrupted
-                    timeseries <- readRDS(filename)
-                    return(timeseries)
-                },
-                error = function(e) {
-                    unlink(filename)
-                    gc()
-                }
-            )
-        }
-        # build the sits tibble for the storing the points
-        samples_tbl <- purrr::pmap_dfr(
-            list(segs_tile$x, segs_tile$y, segs_tile[[pol_id]]),
-            function(x, y, pid) {
-                # convert XY to lat long
-                lat_long <- .proj_to_latlong(x, y, .crs(cube))
-
-                # create metadata for the polygons
-                sample <- tibble::tibble(
-                    longitude  = lat_long[1, "longitude"],
-                    latitude   = lat_long[1, "latitude"],
-                    start_date = start_date,
-                    end_date   = end_date,
-                    label      = "NoClass",
-                    cube       = tile[["collection"]],
-                    polygon_id = pid
-                )
-                # store them in the sample tibble
[…]
