+++ man/plot.raster_cube.Rd
@@ -7,7 +7,7 @@
 \method{plot}{raster_cube}(
   x,
   ...,
-  band = sits_bands(x)[1],
+  band = NULL,
   red = NULL,
   green = NULL,
   blue = NULL,

[…]
