### 2023-10-25 4b51716
add gpu_memory parameter in sits_classify.sits functionn
