### 2023-10-26 69c71a4
sample extraction from segments in parallel
