+++ DESCRIPTION
@@ -187,7 +187,6 @@ Collate:
     'api_variance.R'
     'api_vector.R'
     'api_vector_info.R'
-    'api_vector_sf.R'
     'api_view.R'
     'RcppExports.R'
     'data.R'
