+++ NAMESPACE
@@ -237,7 +237,6 @@ S3method(.tile_yres,raster_cube)
 S3method(.values_ts,bands_cases_dates)
 S3method(.values_ts,bands_dates_cases)
 S3method(.values_ts,cases_dates_bands)
-S3method(.vector_check_package,sf)
 S3method(.view_add_overlay_grps,derived_cube)
 S3method(.view_add_overlay_grps,raster_cube)
 S3method(.view_add_overlay_grps,vector_cube)
