### 2023-10-26 c00c1aa
Fix cloud_coverage for SAR cube
