+++ R/sits_regularize.R
@@ -85,9 +85,9 @@
 #'         cube = s1_cube_open,
 #'         period = "P12D",
 #'         res = 60,
-#'         roi = roi
+#'         roi = roi,
 #'         multicores = 2,
-#'         output_dir = tempdir(),
+#'         output_dir = tempdir()
 #'     )
 #' }
 #'
