+++ R/api_plot_raster.R
@@ -105,12 +105,14 @@
 #' @param  legend        Legend for the classes
 #' @param  palette       A sequential RColorBrewer palette
 #' @param  tmap_options  List with optional tmap parameters
-#'                       tmap max_cells (default: 1e+06)
-#'                       tmap_graticules_labels_size (default: 0.7)
-#'                       tmap_legend_title_size (default: 1.5)
-#'                       tmap_legend_text_size (default: 1.2)
-#'                       tmap_legend_bg_color (default: "white")
-#'                       tmap_legend_bg_alpha (default: 0.5)
+#'                       max_cells (default: 1e+06)
+#'                       scale (default: 0.8)
+#'                       font_family (default: "plex_sans")
+#'                       graticules_labels_size (default: 0.7)
+#'                       legend_title_size (default: 0.8)
+#'                       legend_text_size (default: 0.8)
+#'                       legend_bg_color (default: "white")
+#'                       legend_bg_alpha (default: 0.5)
 #'
 #' @return               A plot object
 #'
@@ -162,23 +164,26 @@
                 labels = labels
             ) +
             tmap::tm_graticules(
-                labels.size = tmap_params[["labels_size"]]
+                labels.size = tmap_params[["graticules_labels_size"]]
             ) +
             tmap::tm_compass() +
             tmap::tm_layout(
-                legend.show = TRUE,
-                legend.outside = FALSE,
-                scale = tmap_params[["scale"]],
-                fontfamily = tmap_params[["font_family"]],
-                legend.bg.color = tmap_params[["bg_color"]],
-                legend.bg.alpha = tmap_params[["bg_alpha"]],
-                legend.title.size = tmap_params[["title_size"]],
-                legend.text.size = tmap_params[["text_size"]],
-                legend.width = tmap_params[["legend_width"]]
+                scale           = tmap_params[["scale"]],
+                fontfamily      = tmap_params[["font_family"]],
+                legend.show     = TRUE,
+                legend.outside  = tmap_params[["legend_outside"]],
+                legend.bg.color = tmap_params[["legend_bg_color"]],
+                legend.bg.alpha = tmap_params[["legend_bg_alpha"]],
+                legend.title.size = tmap_params[["legend_title_size"]],
+                legend.text.size = tmap_params[["legend_text_size"]],
+                legend.width     = tmap_params[["legend_width"]],
+                legend.height    = tmap_params[["legend_height"]],
+                legend.position  = tmap_params[["legend_position"]]
             )
     )
     return(p)
 }
+
 #' @title  Plot probs
 #' @name   .plot_probs
 #' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
@@ -449,10 +454,10 @@
 #'
 .plot_read_size <- function(tile, tmap_options) {
     # get the maximum number of bytes to be displayed
-    if (!purrr::is_null(tmap_options[["tmap_max_cells"]])) {
-        max_cells <- tmap_options[["tmap_max_cells"]]
+    if (!purrr::is_null(tmap_options[["max_cells"]])) {
+        max_cells <- tmap_options[["max_cells"]]
     } else {
-        max_cells <- as.numeric(.conf("tmap", "tmap_max_cells"))
+        max_cells <- as.numeric(.conf("tmap", "max_cells"))
     }
     max_raster <- c(plot = max_cells, view = max_cells)
     # set the options for tmap
@@ -483,72 +488,85 @@
 #' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
 #'
[…]
