### 2023-10-26 9ae364e
fix bugs in plot
