### 2023-10-26 06618b2
extract large segments by iterations
