+++ R/sits_classify.R
@@ -409,6 +409,7 @@ sits_classify.segs_cube <- function(data,
             filter_fn = filter_fn,
             n_sam_pol = n_sam_pol,
             multicores = multicores,
+            memsize = memsize,
             gpu_memory = gpu_memory,
             version = version,
             output_dir = output_dir,
