+++ man/plot.raster_cube.Rd
@@ -40,12 +40,13 @@
 \item{rev}{Reverse the color order in the palette?}
 
 \item{tmap_options}{List with optional tmap parameters
-tmap_max_cells (default: 1e+06)
-tmap_graticules_labels_size (default: 0.7)
-tmap_legend_title_size (default: 1.5)
-tmap_legend_text_size (default: 1.2)
-tmap_legend_bg_color (default: "white")
-tmap_legend_bg_alpha (default: 0.5)}
+max_cells (default: 1e+06)
+scale (default: 0.5)
+graticules_labels_size (default: 0.7)
+legend_title_size (default: 1.0)
+legend_text_size (default: 1.0)
+legend_bg_color (default: "white")
+legend_bg_alpha (default: 0.5)}
 }
 \value{
 A plot object with an RGB image
@@ -56,7 +57,9 @@ A plot object with an RGB image
 Plot RGB raster cube
 }
 \note{
-To see which color palettes are supported, please run
+To see which colors are supported, please run \code{sits_colors()}
+      Use \code{scale} parameter for general output control.
+      If required, then set the other params individually
 }
 \examples{
 if (sits_run_examples()) {
