### 2023-10-27 fecb1f5
fix bug in chunks for reading data
