### 2023-10-27 5c5c28d
add S1 test - part 1
