+++ tests/testthat/test-regularize.R
@@ -274,3 +274,26 @@ test_that("Regularizing local cubes without CLOUD BAND", {
     )
     expect_equal(lubridate::time_length(int, "month"), 2)
 })
+test_that("Regularizing SENTINEL-1 data",{
+        ## Sentinel-1 SAR
+        roi <- c("lon_min" = -50.410, "lon_max" = -50.379,
+                 "lat_min" = -10.1910, "lat_max" = -10.1573)
+        s1_cube_open <- sits_cube(
+            source = "MPC",
+            collection = "SENTINEL-1-GRD",
+            bands = c("VV", "VH"),
+            roi = roi,
+            start_date = "2020-06-01",
+            end_date = "2020-09-28"
+        )
+        # regularize the cube
+        rg_cube <- sits_regularize(
+            cube = s1_cube_open,
+            period = "P12D",
+            res = 60,
+            roi = roi_sar,
+            multicores = 2,
+            output_dir = tempdir()
+        )
+
+})
