+++ man/sits_view.Rd
@@ -62,6 +62,7 @@ sits_view(x, ...)
   class_cube = NULL,
   legend = NULL,
   palette = "Blues",
+  opacity = 0.7,
   view_max_mb = NULL
 )
 
