+++ R/sits_combine_predictions.R
@@ -29,10 +29,6 @@
 #' The supported types of ensemble predictors are 'average' and
 #' 'uncertainty'.
 #'
-#' @note
-#' Please refer to the sits documentation available in
-#' <https://e-sensing.github.io/sitsbook/> for detailed examples.
-#'
 #' @examples
 #' if (sits_run_examples()) {
 #'     # create a data cube from local files
