+++ R/sits_cube.R
@@ -4,20 +4,10 @@
 #' @description Creates a data cube based on spatial and temporal restrictions
 #' in collections available in cloud services or local repositories.
 #' The following cloud providers are supported, based on the STAC protocol:
-#' \itemize{
-#'   \item{\code{"AWS"}: }{Amazon Web Services (AWS),
-#'   see https://registry.opendata.aws/ }
-#'   \item{\code{"BDC"}: }{Brazil Data Cube (BDC),
-#'   see http://brazildatacube.org/}
-#'   \item{\code{"DEAFRICA"}: }{Digital Earth Africa,
-#'   see https://www.digitalearthafrica.org/}
-#'   \item{\code{"MPC"}: }{Microsoft Planetary Computer,
-#'   see https://planetarycomputer.microsoft.com/}
-#'   \item{\code{"USGS"}:}{USGS LANDSAT collection,
-#'   see https://registry.opendata.aws/usgs-landsat/}
-#'  }
-#'
-#' Data cubes can also be created using local files (see details).
+#' Amazon Web Services (AWS), Brazil Data Cube (BDC),
+#' Digital Earth Africa (DEAFRICA), Microsoft Planetary Computer (MPC),
+#' Nasa Harmonized Landsat/Sentinel (HLS), USGS Landsat (USGS), and
+#' Swiss Data Cube (SDC). Data cubes can also be created using local files.
 #'
 #'
 #' @param source       Data source (one of \code{"AWS"}, \code{"BDC"},
@@ -65,40 +55,41 @@
 #' @param progress     Logical: show a progress bar?
 #' @return A \code{tibble} describing the contents of a data cube.
 #'
-#' @details
+#' @note{
 #' To create cubes from cloud providers, users need to inform:
-#' \itemize{
-#' \item{\code{source}: }{One of \code{"AWS"}, \code{"BDC"}, \code{"DEAFRICA"},
-#' \code{"MPC"}, \code{"USGS"}, \code{"SDC"} and \code{"HLS"}}.
-#' \item{\code{collection}: }{Use \code{sits_list_collections()} to see which
-#'   collections are supported.}
-#' \item{\code{tiles}: }{A set of tiles defined according to the collection
-#'   tiling grid.}
-#' \item{\code{roi}: }{Region of interest in WGS84 coordinates.}
+#' \enumerate{
+#'  \item \code{source}: One of "AWS", "BDC", "DEAFRICA", "HLS", "MPC",
+#' "SDC" or "USGS";
+#'  \item \code{collection}: Collection available in the cloud provider.
+#'         Use \code{sits_list_collections()} to see which
+#'         collections are supported;
+#'  \item \code{tiles}: A set of tiles defined according to the collection
+#'         tiling grid;
+#'  \item \code{roi}: Region of interest. Either
+#'        a named \code{vector} (\code{"lon_min"}, \code{"lat_min"},
+#'        \code{"lon_max"}, \code{"lat_max"}) in WGS84, a \code{sfc}
+#'        or \code{sf} object from sf package in WGS84 projection.
 #' }
 #' Either \code{tiles} or  \code{roi} must be informed.
 #' The parameters \code{bands}, \code{start_date}, and
 #' \code{end_date} are optional for cubes created from cloud providers.
-#' #' The \code{roi} parameter allows a selection of an area of interest,
-#' either using a named \code{vector} (\code{"lon_min"}, \code{"lat_min"},
-#' \code{"lon_max"}, \code{"lat_max"}) in WGS84, a \code{sfc} or \code{sf}
-#' object from sf package in WGS84 projection.
+#'
 #' GeoJSON geometries (RFC 7946) and shapefiles should be converted to
 #' \code{sf} objects before being used to define a region of interest.
 #' This parameter does not crop a region; it only selects images that
[…]
