+++ R/sits_predictors.R
@@ -4,7 +4,7 @@
 #' @description Predictors are X-Y values required for machine learning
 #' algorithms, organized as a data table where each row corresponds
 #' to a training sample. The first two columns of the predictors table
-#' are categorical ("label_id" and "label"). The other columns are
+#' are categorical (\code{label_id} and \code{label}). The other columns are
 #' the values of each band and time, organized first by band and then by time.
 #'
 #' @param  samples     Time series in sits format (tibble of class "sits")
