+++ man/sits_model_export.Rd
@@ -19,7 +19,7 @@ machine learning or deep learning package.
 \description{
 Given a trained machine learning or deep learning model,
 exports the model as an object for further exploration outside the
-"sits" package
+\code{sits} package.
 }
 \examples{
 if (sits_run_examples()) {

[…]
