+++ README.Rmd
@@ -94,12 +94,16 @@ devtools::install_github("e-sensing/sits", dependencies = TRUE)
 library(sits)
 ```
 
+### Support for GPU 
+
+Classification using torch-based deep learning models in `sits` uses CUDA compatible NVIDIA GPUs if available, which provides up 10-fold speed-up compared to using CPUs only. Please see the [installation instructions](https://torch.mlverse.org/docs/articles/installation) for more information on how to install the required drivers. 
+
 
 ## Building Earth Observation Data Cubes
 
 ### Image Collections Accessible by `sits`
 
-The `sits` package allows users to created data cubes from analysis-ready data (ARD) image collections available in cloud services. The collections accessible in `sits` `r packageVersion("sits")` are:
+Users create data cubes from analysis-ready data (ARD) image collections available in cloud services. The collections accessible in `sits` `r packageVersion("sits")` are:
 
 1. Brazil Data Cube ([BDC](http://brazildatacube.org/en/home-page-2/#dataproducts)): Open data collections of Sentinel-2, Landsat-8 and CBERS-4 images.
 2. Microsoft Planetary Computer ([MPC](https://planetarycomputer.microsoft.com/catalog)): Open data collection of Sentinel-2/2A and Landsat-8
@@ -273,7 +277,9 @@ plot(label_cube,
 
 ## Additional information
 
-For more information, please see the on-line book ["SITS: Data analysis and machine learning for data cubes using satellite image time series"](https://e-sensing.github.io/sitsbook/).
+Since version 1.4.2, `sits` support OBIA analysis of image time series, using an extension of R package `supercells`.  
+
+The package is described in detail in on-line book ["SITS: Data analysis and machine learning for data cubes using satellite image time series"](https://e-sensing.github.io/sitsbook/).
 
 
 ### References
@@ -314,11 +320,15 @@ We thank the authors of these papers for making their code available to be used 
 
 - [12] Maja Schneider, Marco Körner, "[Re] Satellite Image Time Series Classification with Pixel-Set Encoders and Temporal Self-Attention." ReScience C 7 (2), 2021. <doi:10.5281/zenodo.4835356>.
 
-#### R packages used in sits
+- [13] Jakub Nowosad, Tomasz Stepinski, "Extended SLIC superpixels algorithm for applications to non-imagery geospatial rasters". International Journal of Applied Earth Observation and Geoinformation, 112, 102935, 2022.
 
-The authors are thankful for the contributions of Marius Appel, Tim Appelhans, Henrik Bengtsson, Robert Hijmans, Edzer Pebesma, and Ron Wehrens, respectively chief developers of the packages `gdalcubes`, `leafem`, `data.table`, `terra/raster`, `sf`/`stars`, and `kohonen`. The `sits` package  is also much indebted to the work of the RStudio team, including the `tidyverse`. We are indepted to Daniel Falbel for his and the `torch` packages. We thank Charlotte Pelletier and Hassan Fawaz for sharing the python code that has been reused for the TempCNN and ResNet machine learning models. We would like to thank Maja Schneider for sharing the python code that helped the implementation of the `sits_lighttae()` and `sits_tae()` model. We recognise the importance of the work by Chris Holmes and Mattias Mohr on the STAC specification and API. 
+- [14] Martin Tennekes, “tmap: Thematic Maps in R.” Journal of Statistical Software, 84(6), 1–39, 2018.
 
-## Acknowledgements for Financial and Material Support
+### Acknowledgements for community support
+
[…]
