### 2023-11-09 6b055c1
re-introduces sits_clean function to sits API
