+++ man/sits_label_classification.Rd
@@ -12,8 +12,6 @@
 \usage{
 sits_label_classification(
   cube,
-  clean = TRUE,
-  window_size = 3L,
   memsize = 4,
   multicores = 2,
   output_dir,
@@ -24,8 +22,6 @@ sits_label_classification(
 \method{sits_label_classification}{probs_cube}(
   cube,
   ...,
-  clean = TRUE,
-  window_size = 3L,
   memsize = 4L,
   multicores = 2L,
   output_dir,
@@ -51,13 +47,6 @@ sits_label_classification(
 }
 \arguments{
 \item{cube}{Classified image data cube.}
-
-\item{clean}{A logical value to apply a modal function to clean up
-possible noisy pixels keeping the most frequently
-values within the neighborhood. Default is TRUE.}
-
-\item{window_size}{An odd integer representing the size of the
-sliding window of the modal function (min = 1, max = 15).}
 
 \item{memsize}{maximum overall memory (in GB) to label the
 classification.}
