+++ NAMESPACE
@@ -290,6 +290,11 @@ S3method(sits_classify,raster_cube)
 S3method(sits_classify,segs_cube)
 S3method(sits_classify,sits)
 S3method(sits_classify,tbl_df)
+S3method(sits_clean,class_cube)
+S3method(sits_clean,default)
+S3method(sits_clean,derived_cube)
+S3method(sits_clean,raster_cube)
+S3method(sits_clean,tbl_df)
 S3method(sits_cluster_dendro,default)
 S3method(sits_cluster_dendro,sits)
 S3method(sits_cluster_dendro,tbl_df)
@@ -389,6 +394,7 @@ export(sits_as_sf)
 export(sits_bands)
 export(sits_bbox)
 export(sits_classify)
+export(sits_clean)
 export(sits_cluster_clean)
 export(sits_cluster_dendro)
 export(sits_cluster_frequency)
