+++ DESCRIPTION
@@ -203,6 +203,7 @@ Collate:
     'sits_csv.R'
     'sits_cube.R'
     'sits_cube_copy.R'
+    'sits_clean.R'
     'sits_cluster.R'
     'sits_factory.R'
     'sits_filters.R'
