### 2023-11-10 4dd8273
fix bug in samples_validation
