+++ R/sits_lighttae.R
@@ -184,7 +184,7 @@ sits_lighttae <- function(samples = NULL,
                 timeline = timeline, bands = bands
             )
             # Test samples are extracted from validation data
-            test_samples <- .predictors(samples)
+            test_samples <- .predictors(samples_validation)
             test_samples <- .pred_normalize(
                 pred = test_samples, stats = ml_stats
             )
