+++ README.md
@@ -77,14 +77,13 @@ to the package. We recommend running them in the following order:
     SITS](https://www.kaggle.com/esensing/working-with-time-series-in-sits)
 3.  [Creating data cubes in
     SITS](https://www.kaggle.com/esensing/creating-data-cubes-in-sits)
-4.  [Machine learning for data
+4.  [Improving quality of training samples](https://www.kaggle.com/code/esensing/improving-quality-of-training-samples)
+5.  [Machine learning for data
     cubes](https://www.kaggle.com/esensing/machine-learning-for-data-cubes)
-5.  [Raster classification in
-    SITS](https://www.kaggle.com/esensing/raster-classification-in-sits)
-6.  [Object-based time series classification using
-    GPU](https://www.kaggle.com/esensing/object-based-time-series-classification-using-gpu)
-7.  [Using SOM for sample quality control in
-    SITS](https://www.kaggle.com/esensing/using-som-for-sample-quality-control-in-sits)
+6.  [Classification of raster data cubes](https://www.kaggle.com/code/esensing/classification-of-raster-data-cubes)
+7.  [Bayesian smooting for post-processing](https://www.kaggle.com/code/esensing/bayesian-smoothing-for-post-processing)
+8.  [Uncertainty and active learning](https://e-sensing.github.io/sitsbook/uncertainty-and-active-learning.html)
+9.  [Object-based time series classification](https://www.kaggle.com/esensing/object-based-image-time-series-classification)
 
 ## Installation
 
