+++ R/sits_machine_learning.R
@@ -284,6 +284,7 @@ sits_xgboost <- function(samples = NULL, learning_rate = 0.15,
                          min_split_loss = 1, max_depth = 5,
                          min_child_weight = 1, max_delta_step = 1,
                          subsample = 0.8, nfold = 5, nrounds = 100,
+                         nthread = 6,
                          early_stopping_rounds = 20, verbose = FALSE) {
     # Function that trains a xgb model
     train_fun <- function(samples) {
@@ -307,13 +308,28 @@ sits_xgboost <- function(samples = NULL, learning_rate = 0.15,
             eval_metric = "mlogloss", eta = learning_rate,
             gamma = min_split_loss, max_depth = max_depth,
             min_child_weight = min_child_weight,
-            max_delta_step = max_delta_step, subsample = subsample
+            max_delta_step = max_delta_step, subsample = subsample,
+            nthread = nthread
         )
         # Train a xgboost model
-        model <- xgboost::xgboost(
+        # model <- xgboost::xgboost(
+        #     data = as.matrix(.pred_features(train_samples)),
+        #     label = references, num_class = length(labels), params = params,
+        #     nrounds = nrounds, verbose = FALSE
+        # )
+        #
+        if (verbose)
+            verbose = 1
+        else
+            verbose = 0
+        # transform predictors in a xgb.DMatrix
+        xgb_matrix <- xgboost::xgb.DMatrix(
             data = as.matrix(.pred_features(train_samples)),
-            label = references, num_class = length(labels), params = params,
-            nrounds = nrounds, verbose = FALSE
+            label = references)
+        # train the model
+        model <- xgboost::xgb.train(xgb_matrix,
+            num_class = length(labels), params = params,
+            nrounds = nrounds, verbose = verbose
         )
         # Get best ntreelimit
         ntreelimit <- model$best_ntreelimit
