+++ README.Rmd
@@ -61,13 +61,16 @@ Detailed documentation on how to use `sits` is available in the e-book ["Satelli
 
 Those that want to evaluate the `sits` package before installing are invited to run the examples available on [Kaggle](https://www.kaggle.com/esensing/code). If you are new on kaggle, please follow the [instructions](https://gist.github.com/OldLipe/814089cc5792c9c0c989d870a22910f4) to set up your account. These examples provide a fast-track introduction to the package. We recommend running them in the following order: 
 
-1. [Introduction to SITS](https://www.kaggle.com/esensing/introduction-to-sits)
-2. [Working with time series in SITS](https://www.kaggle.com/esensing/working-with-time-series-in-sits)
-3. [Creating data cubes in SITS](https://www.kaggle.com/esensing/creating-data-cubes-in-sits)
-4. [Machine learning for data cubes](https://www.kaggle.com/esensing/machine-learning-for-data-cubes)
-5. [Raster classification in SITS](https://www.kaggle.com/esensing/raster-classification-in-sits)
-6. [Object-based time series classification using GPU](https://www.kaggle.com/esensing/object-based-time-series-classification-using-gpu)
-7. [Using SOM for sample quality control in SITS](https://www.kaggle.com/esensing/using-som-for-sample-quality-control-in-sits)
+1.  [Introduction to SITS](https://www.kaggle.com/esensing/introduction-to-sits)
+2.  [Working with time series in SITS](https://www.kaggle.com/esensing/working-with-time-series-in-sits)
+3.  [Creating data cubes in SITS](https://www.kaggle.com/esensing/creating-data-cubes-in-sits)
+4.  [Improving the quality of training samples](https://www.kaggle.com/code/esensing/improving-quality-of-training-samples)
+5.  [Machine learning for data cubes](https://www.kaggle.com/esensing/machine-learning-for-data-cubes)
+6.  [Classification of raster data cubes](https://www.kaggle.com/code/esensing/classification-of-raster-data-cubes)
+7.  [Bayesian smoothing for post-processing](https://www.kaggle.com/code/esensing/bayesian-smoothing-for-post-processing)
+8.  [Uncertainty and active learning](https://www.kaggle.com/code/esensing/uncertainty-and-active-learning)
+9.  [Object-based time series classification](https://www.kaggle.com/esensing/object-based-image-time-series-classification)
+
 
 ## Installation
 
