+++ NAMESPACE
@@ -356,6 +356,8 @@ S3method(sits_timeline,tbl_df)
 S3method(sits_to_csv,default)
 S3method(sits_to_csv,sits)
 S3method(sits_to_csv,tbl_df)
+S3method(sits_to_xlsx,list)
+S3method(sits_to_xlsx,sits_accuracy)
 S3method(sits_uncertainty,default)
 S3method(sits_uncertainty,entropy)
 S3method(sits_uncertainty,least)
@@ -434,6 +436,7 @@ export(sits_rfor)
 export(sits_run_examples)
 export(sits_run_tests)
 export(sits_sample)
+export(sits_sampling_design)
 export(sits_segment)
 export(sits_select)
 export(sits_sgolay)
