### 2023-11-28 cfe8fca
update hotfix version metadata
