### 2023-11-28 2ad3d7b
fix bug in view classified cube
