### 2023-11-30 8c7a7f0
reduce segmentation memory size
