### 2023-12-05 4be8ab6
add uncertainty for vector cubes; improve tmap-based plots
