+++ R/sits_plot.R
@@ -322,9 +322,9 @@ plot.predicted <- function(x, y, ...,
 #' @param  date          Date to be plotted.
 #' @param  palette       An RColorBrewer palette
 #' @param  rev           Reverse the color order in the palette?
-#' @param  tmap_options  List with optional tmap parameters
+#' @param  tmap_options  Named list with optional tmap parameters
 #'                       max_cells (default: 1e+06)
-#'                       scale (default: 0.5)
+#'                       scale (default: 1.0)
 #'                       graticules_labels_size (default: 0.7)
 #'                       legend_title_size (default: 1.0)
 #'                       legend_text_size (default: 1.0)
@@ -452,13 +452,14 @@ plot.raster_cube <- function(x, ...,
 #' @param  line_width    Line width to plot the segments boundary (in pixels)
 #' @param  palette       An RColorBrewer palette
 #' @param  rev           Reverse the color order in the palette?
-#' @param  tmap_options  List with optional tmap parameters
-#'                       tmap_max_cells (default: 1e+06)
-#'                       tmap_graticules_labels_size (default: 0.7)
-#'                       tmap_legend_title_size (default: 1.5)
-#'                       tmap_legend_text_size (default: 1.2)
-#'                       tmap_legend_bg_color (default: "white")
-#'                       tmap_legend_bg_alpha (default: 0.5)
+#' @param  tmap_options  Named list with optional tmap parameters
+#'                       max_cells (default: 1e+06)
+#'                       scale (default: 1.0)
+#'                       graticules_labels_size (default: 0.7)
+#'                       legend_title_size (default: 1.0)
+#'                       legend_text_size (default: 1.0)
+#'                       legend_bg_color (default: "white")
+#'                       legend_bg_alpha (default: 0.5)
 #'
 #' @return               A plot object with an RGB image
 #'                       or a B/W image on a color
@@ -582,13 +583,14 @@ plot.vector_cube <- function(x, ...,
 #' @param labels         Labels to plot (optional).
 #' @param palette        RColorBrewer palette
 #' @param rev            Reverse order of colors in palette?
-#' @param tmap_options   List with optional tmap parameters
-#'                       tmap_max_cells (default: 1e+06)
-#'                       tmap_graticules_labels_size (default: 0.7)
-#'                       tmap_legend_title_size (default: 1.5)
-#'                       tmap_legend_text_size (default: 1.2)
-#'                       tmap_legend_bg_color (default: "white")
-#'                       tmap_legend_bg_alpha (default: 0.5)
+#' @param  tmap_options  Named list with optional tmap parameters
+#'                       max_cells (default: 1e+06)
+#'                       scale (default: 1.0)
+#'                       graticules_labels_size (default: 0.7)
+#'                       legend_title_size (default: 1.0)
+#'                       legend_text_size (default: 1.0)
+#'                       legend_bg_color (default: "white")
+#'                       legend_bg_alpha (default: 0.5)
 #' @return               A plot containing probabilities associated
 #'                       to each class for each pixel.
 #'
@@ -655,13 +657,14 @@ plot.probs_cube <- function(x, ...,
 #' @param labels         Labels to plot (optional).
 #' @param palette        RColorBrewer palette
 #' @param rev            Reverse order of colors in palette?
-#' @param tmap_options   List with optional tmap parameters
-#'                       tmap_max_cells (default: 1e+06)
-#'                       tmap_graticules_labels_size (default: 0.7)
-#'                       tmap_legend_title_size (default: 1.5)
-#'                       tmap_legend_text_size (default: 1.2)
-#'                       tmap_legend_bg_color (default: "white")
-#'                       tmap_legend_bg_alpha (default: 0.5)
+#' @param  tmap_options  Named list with optional tmap parameters
[…]
