+++ R/api_plot_raster.R
@@ -12,13 +12,14 @@
 #' @param  line_width    Line width to plot the segments boundary
 #' @param  palette       A sequential RColorBrewer palette
 #' @param  rev           Reverse the color palette?
-#' @param  tmap_options  List with optional tmap parameters
-#'                       tmap max_cells (default: 1e+06)
-#'                       tmap_graticules_labels_size (default: 0.7)
-#'                       tmap_legend_title_size (default: 1.5)
-#'                       tmap_legend_text_size (default: 1.2)
-#'                       tmap_legend_bg_color (default: "white")
-#'                       tmap_legend_bg_alpha (default: 0.5)
+#' @param  tmap_options  Named vector with optional tmap parameters:
+#'                       scale (default = 1.0)
+#'                       max_cells (default: 1e+06)
+#'                       graticules_labels_size (default: 0.7)
+#'                       legend_title_size (default: 1.5)
+#'                       legend_text_size (default: 1.2)
+#'                       legend_bg_color (default: "white")
+#'                       legend_bg_alpha (default: 0.5)
 #'
 #' @return               A plot object
 #'
@@ -104,7 +105,7 @@
 #' @param  tile          Tile to be plotted.
 #' @param  legend        Legend for the classes
 #' @param  palette       A sequential RColorBrewer palette
-#' @param  tmap_options  List with optional tmap parameters
+#' @param  tmap_options  Named vector with optional tmap parameters
 #'                       max_cells (default: 1e+06)
 #'                       scale (default: 0.8)
 #'                       font_family (default: "plex_sans")
@@ -193,14 +194,14 @@
 #' @param  labels_plot   Labels to be plotted
 #' @param  palette       A sequential RColorBrewer palette
 #' @param  rev           Reverse the color palette?
-#' @param  tmap_options  List with optional tmap parameters
-#'                       tmap max_cells (default: 1e+06)
-#'                       tmap_graticules_labels_size (default: 0.7)
-#'                       tmap_legend_title_size (default: 1.5)
-#'                       tmap_legend_text_size (default: 1.2)
-#'                       tmap_legend_bg_color (default: "white")
-#'                       tmap_legend_bg_alpha (default: 0.5)
-#'
+#' @param  tmap_options  Named vector with optional tmap parameters
+#'                       max_cells (default: 1e+06)
+#'                       graticules_labels_size (default: 0.7)
+#'                       legend_title_size (default: 1.5)
+#'                       legend_text_size (default: 1.2)
+#'                       legend_bg_color (default: "white")
+#'                       legend_bg_alpha (default: 0.5)
+#'                       scale (default: 1.0)
 #' @return               A plot object
 #'
 .plot_probs <- function(tile,
@@ -260,7 +261,7 @@
     bds <- as.numeric(names(labels[labels %in% labels_plot]))
 
     # set the tmap options
-    tmap_params <- .plot_tmap_params(tmap_options)
+    tmap_options <- .plot_tmap_params(tmap_options)
 
     p <- tmap::tm_shape(probs_st[, , , bds]) +
         tmap::tm_raster(
@@ -274,10 +275,10 @@
         tmap::tm_layout(
             legend.show = TRUE,
             legend.outside = FALSE,
-            legend.bg.color = tmap_params[["legend_bg_color"]],
-            legend.bg.alpha = tmap_params[["legend_bg_alpha"]],
-            legend.title.size = tmap_params[["legend_title_size"]],
-            legend.text.size = tmap_params[["legend_text_size"]],
+            legend.bg.color = tmap_options[["legend_bg_color"]],
[…]
