+++ man/sits_bands.Rd
@@ -4,6 +4,7 @@
 \alias{sits_bands}
 \alias{sits_bands.sits}
 \alias{sits_bands.raster_cube}
+\alias{sits_bands.vector_cube}
 \alias{sits_bands.patterns}
 \alias{sits_bands.sits_model}
 \alias{sits_bands.tbl_df}
@@ -19,6 +20,8 @@ sits_bands(x)
 \method{sits_bands}{sits}(x)
 
 \method{sits_bands}{raster_cube}(x)
+
+\method{sits_bands}{vector_cube}(x)
 
 \method{sits_bands}{patterns}(x)
 
