+++ NAMESPACE
@@ -257,6 +257,7 @@ S3method(plot,som_evaluate_cluster)
 S3method(plot,som_map)
 S3method(plot,torch_model)
 S3method(plot,uncertainty_cube)
+S3method(plot,uncertainty_vector_cube)
 S3method(plot,variance_cube)
 S3method(plot,vector_cube)
 S3method(plot,xgb_model)
@@ -281,6 +282,7 @@ S3method(sits_bands,raster_cube)
 S3method(sits_bands,sits)
 S3method(sits_bands,sits_model)
 S3method(sits_bands,tbl_df)
+S3method(sits_bands,vector_cube)
 S3method(sits_bbox,default)
 S3method(sits_bbox,raster_cube)
 S3method(sits_bbox,sits)
