+++ inst/extdata/config_internals.yml
@@ -250,7 +250,7 @@ tmap:
         legend_width: 1.0
         legend_position: ["left", "bottom"]
         legend_height: 1.0
-        scale: 1.5
+        scale: 1.0
         font_family: "plex_sans"
 
 # maxbytes for leaflet (in MB)

[…]
