### 2023-12-08 64dac1d
fix segments_probs plotting
