### 2023-12-19 ddaac56
remove loading of google fonts
