### 2023-12-19 7edf7c6
working in new paralelism strategy
