### 2023-12-19 af06f3e
update NEWS
