+++ NEWS.md
@@ -2,6 +2,9 @@
 
 # What's new in SITS version 1.4
 
+### Hotfix version 1.4.2-3
+* Fix font download in package initialization
+
 ### Hotfix version 1.4.2-2
 * Fix integer overflow bug in `sits_classify()` segments
 
