+++ tests/testthat/test-plot.R
@@ -81,7 +81,7 @@ test_that("Plot Time Series and Images", {
     expect_equal(p_probs$tm_layout$legend.bg.color, "white")
 
     p_probs_f <- plot(sinop_probs, labels = "Forest")
-    expect_equal(p_probs_f$tm_raster$palette, "YlGnBu")
+    expect_equal(p_probs_f$tm_raster$palette, "YlGn")
     expect_equal(length(p_probs_f$tm_raster$title), 1)
     expect_equal(p_probs_f$tm_layout$legend.bg.color, "white")
 
