### 2023-12-26 6c08c1a
avoid warnings in sits_apply when doing documentation
