+++ R/sits_apply.R
@@ -135,10 +135,13 @@ sits_apply.raster_cube <- function(data, ...,
     out_band <- names(expr)
     # Check if band already exists in cube
     if (out_band %in% .cube_bands(data)) {
-        warning(
-            paste0("The provided band '", out_band, "' already exists in cube."),
-            call. = FALSE
-        )
+        if (.check_messages()){
+            warning(
+                paste0("The provided band '", out_band,
+                       "' already exists in cube."),
+                call. = FALSE
+            )
+        }
         return(data)
     }
     # Get all input bands in cube data
