### 2024-01-10 d37ac73
Introduces sits_reduce functionm
