+++ DESCRIPTION
@@ -150,6 +150,7 @@ Collate:
     'api_raster_sub_image.R'
     'api_raster_terra.R'
     'api_reclassify.R'
+    'api_reduce.R'
     'api_regularize.R'
     'api_roi.R'
     'api_s2tile.R'
@@ -223,6 +224,7 @@ Collate:
     'sits_plot.R'
     'sits_predictors.R'
     'sits_reclassify.R'
+    'sits_reduce.R'
     'sits_regularize.R'
     'sits_resnet.R'
     'sits_sample_functions.R'
