### 2024-01-11 7653daa
Update sits_reduce function
