+++ R/sits_reduce.R
@@ -13,10 +13,6 @@
 #' from the function.
 #'
 #' @param data          Valid sits tibble or cube
-#' @param window_size   An odd number representing the size of the
-#'                      sliding window of sits kernel functions
-#'                      used in expressions (for a list of supported
-#'                      kernel functions, please see details).
 #' @param memsize       Memory available for classification (in GB).
 #' @param multicores    Number of cores to be used for classification.
 #' @param output_dir    Directory where files will be saved.
@@ -27,41 +23,46 @@
 #' \code{sits_reduce()} allow any valid R expression to compute new bands.
 #' Use R syntax to pass an expression to this function.
 #' Besides arithmetic operators, you can use virtually any R function
-#' that can be applied to elements of a matrix (functions that are
-#' unaware of matrix sizes, e.g. \code{sqrt()}, \code{sin()},
-#' \code{log()}).
+#' that can be applied to elements of a matrix.
+#' The provided functions must operate at line level in order to perform
+#' temporal reduction on a pixel.
 #'
-#' Also, \code{sits_apply()} accepts a predefined set of kernel functions
-#' (see below) that can be applied to pixels considering its
-#' neighborhood. \code{sits_apply()} considers a neighborhood of a
-#' pixel as a set of pixels equidistant to it (including itself)
-#' according the Chebyshev distance. This neighborhood form a
-#' square window (also known as kernel) around the central pixel
-#' (Moore neighborhood). Users can set the \code{window_size}
-#' parameter to adjust the size of the kernel window.
-#' The image is conceptually mirrored at the edges so that neighborhood
-#' including a pixel outside the image is equivalent to take the
-#' 'mirrored' pixel inside the edge.
+#' \code{sits_reduce()} Applies a function to each row of a matrix.
+#' In this matrix, each row represents a pixel and each column
+#' represents a single date. We provide some operations already
+#' implemented in the package to perform the reduce operation.
+#' See the list of available functions below:
 #'
-#' \code{sits_apply()} applies a function to the kernel and its result
-#' is assigned to a corresponding central pixel on a new matrix.
-#' The kernel slides throughout the input image and this process
-#' generates an entire new matrix, which is returned as a new band
-#' to the cube. The kernel functions ignores any \code{NA} values
-#' inside the kernel window. Central pixel is \code{NA} just only
-#' all pixels in the window are \code{NA}.
-#'
-#' @section Summarizing kernel functions:
+#' @section Summarizing temporal functions:
 #' \itemize{
-#' \item{\code{w_median()}: returns the median of the neighborhood's values.}
-#' \item{\code{w_sum()}: returns the sum of the neighborhood's values.}
-#' \item{\code{w_mean()}: returns the mean of the neighborhood's values.}
-#' \item{\code{w_sd()}: returns the standard deviation of the neighborhood's
-#'   values.}
-#' \item{\code{w_min()}: returns the minimum of the neighborhood's values.}
-#' \item{\code{w_max()}: returns the maximum of the neighborhood's values.}
-#' \item{\code{w_var()}: returns the variance of the neighborhood's values.}
-#' \item{\code{w_modal()}: returns the modal of the neighborhood's values.}
+#'  \item{\code{t_max()}: Returns the maximum value of the series.}
+#'  \item{\code{t_min()}: Returns the minimum value of the series}
+#'  \item{\code{t_mean()}: Returns the mean of the series.}
+#'  \item{\code{t_median()}: Returns the median of the series.}
+#'  \item{\code{t_sum()}: Returns the sum of all the points in the series.}
[…]
