### 2024-01-11 eb15c6f
Update sits_reduce docs
