### 2024-01-11 5c58cff
add temporal operations fns
