+++ src/RcppExports.cpp
@@ -249,6 +249,193 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// C_temp_max
+arma::vec C_temp_max(const arma::mat& mtx);
+RcppExport SEXP _sits_C_temp_max(SEXP mtxSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type mtx(mtxSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_temp_max(mtx));
+    return rcpp_result_gen;
+END_RCPP
+}
+// C_temp_min
+arma::vec C_temp_min(const arma::mat& mtx);
+RcppExport SEXP _sits_C_temp_min(SEXP mtxSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type mtx(mtxSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_temp_min(mtx));
+    return rcpp_result_gen;
+END_RCPP
+}
+// C_temp_mean
+arma::vec C_temp_mean(const arma::mat& mtx);
+RcppExport SEXP _sits_C_temp_mean(SEXP mtxSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type mtx(mtxSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_temp_mean(mtx));
+    return rcpp_result_gen;
+END_RCPP
+}
+// C_temp_median
+arma::vec C_temp_median(const arma::mat& mtx);
+RcppExport SEXP _sits_C_temp_median(SEXP mtxSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type mtx(mtxSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_temp_median(mtx));
+    return rcpp_result_gen;
+END_RCPP
+}
+// C_temp_sum
+arma::vec C_temp_sum(const arma::mat& mtx);
+RcppExport SEXP _sits_C_temp_sum(SEXP mtxSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type mtx(mtxSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_temp_sum(mtx));
+    return rcpp_result_gen;
+END_RCPP
+}
+// C_temp_std
+arma::vec C_temp_std(const arma::mat& mtx);
+RcppExport SEXP _sits_C_temp_std(SEXP mtxSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type mtx(mtxSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_temp_std(mtx));
+    return rcpp_result_gen;
+END_RCPP
+}
[…]
