### 2024-01-12 4d96795
update internal reduce api
