### 2024-01-17 eedb21c
add bands parameter in .apply_input_bands function
