### 2024-01-17 f1a2435
add .samples_timeline function in internal API
