### 2024-01-17 7239d8c
add support to samples reduce
