### 2024-01-17 a59bf71
update usage of .apply_input_bands function
