### 2024-01-17 90ff772
add tests for sits_reduce function
