### 2024-01-17 f466d94
Fix sits_accuracy bug (Related to #1042)
