### 2024-01-22 16c5f4b
Fix default crs to roi of sits_regularize() function
