### 2024-01-27 0a2c0cb
introduces roi parameter in sits_classify segments
