+++ R/api_classify.R
@@ -207,6 +207,7 @@
 #' @param  tile       Single tile of a data cube.
 #' @param  ml_model   Model trained by \code{\link[sits]{sits_train}}.
 #' @param  block      Optimized block to be read into memory.
+#' @param  roi        Region of interest.
 #' @param  filter_fn  Smoothing filter function to be applied to the data.
 #' @param  n_sam_pol  Number of samples per polygon to be read
 #'                    for POLYGON or MULTIPOLYGON shapefiles or sf objects.
@@ -219,6 +220,7 @@
 .classify_vector_tile <- function(tile,
                                   ml_model,
                                   block,
+                                  roi,
                                   filter_fn,
                                   n_sam_pol,
                                   multicores,
@@ -256,12 +258,7 @@
         return(probs_tile)
     }
     # Create chunks as jobs
-    chunks <- .tile_
-    #
-    # Create chunks as jobs
-    chunks_samples <- .cube_split_chunks_samples(
-        cube = cube, samples_sf = samples
-    )
+    chunks <- .tile_chunks_create(tile = tile, overlap = 0, block = block)
     # By default, update_bbox is FALSE
     update_bbox <- FALSE
     if (.has(roi)) {
@@ -272,37 +269,37 @@
         # Should bbox of resulting tile be updated?
         update_bbox <- nrow(chunks) != nchunks
     }
-    #
-    # # Create chunks as jobs
-    # chunks <- .tile_chunks_create(tile = tile, overlap = 0, block = block)
+    chunks <- .chunks_filter_segments(chunks, tile)
 
-    # Get tile bands
-    bands <- .tile_bands(tile, add_cloud = FALSE)
-    segments_ts <- .segments_extract_multicores(
-        tile = tile,
-        bands = bands,
-        pol_id = "pol_id",
-        n_sam_pol = n_sam_pol,
-        multicores = multicores,
-        memsize = memsize,
-        output_dir = output_dir,
-        progress = FALSE
-    )
-    # Classify segments
-    segments_ts <- .classify_ts(
-        samples = segments_ts,
-        ml_model = ml_model,
-        filter_fn = filter_fn,
-        multicores = multicores,
-        gpu_memory = gpu_memory,
-        progress = progress
-    )
-    # Join probability values with segments
-    segments_ts <- .segments_join_probs(
-        data = segments_ts,
-        segments = .segments_read_vec(tile)
-    )
-
+    # Process jobs in parallel
+    segments_ts <- .jobs_map_parallel(chunks, function(chunk) {
+        # Job block
+        block <- .block(chunk)
+        # Block file name
+        block_file <- .file_block_name(
+            pattern = .file_pattern(out_file),
+            block = block,
+            output_dir = output_dir
+        )
+        # Extract segments time series
+        segments_ts <- .segments_poly_read(tile, chunk, n_sam_pol)
+        # Classify segments
+        segments_ts <- .classify_ts(
+            samples = segments_ts,
+            ml_model = ml_model,
+            filter_fn = filter_fn,
+            multicores = multicores,
+            gpu_memory = gpu_memory,
+            progress = progress
+        )
+        # Join probability values with segments
+        segments_ts <- .segments_join_probs(
+            data = segments_ts,
+            segments = .segments_read_vec(tile)
+        )
+        return(segments_ts)
+    }, progress = progress)
+    segments_ts <- dplyr::bind_rows(segments_ts)
     # Write all segments
     .vector_write_vec(v_obj = segments_ts, file_path = out_file)
     # Create tile based on template
