### 2024-01-27 de2ef5b
introduces roi parameter in sits_classify segments
