### 2024-01-27 4e7cdb7
refactory function to extract time series from segments
