+++ R/api_segments.R
@@ -135,6 +135,7 @@
     }
     return(TRUE)
 }
+
 #' @name .segments_data_read
 #' @keywords internal
 #' @noRd
@@ -299,69 +300,66 @@
     dplyr::left_join(segments, data_id, by = c("pol_id" = "polygon_id"))
 }
 #'
-#' @name .segments_extract_data
+#' @name .segments_data_read
 #' @keywords internal
 #' @noRd
 #' @description     Using the segments as polygons, get all time series
 #'
-#' @param seg_tile_bands tibble to be processed in parallel
 #' @param tile       tile of regular data cube
-#' @param bands      bands used in time series
-#' @param pol_id     ID attribute for polygons.
+#' @param chunk      A chunk to be read.
 #' @param n_sam_pol  Number of samples per polygon to be read.
-#' @param multicores Number of cores to use for processing
-#' @param progress   Show progress bar?
 #'
 #' @return  samples associated to segments
-#'
-.segments_extract_data <- function(seg_tile_bands,
-                                   tile,
-                                   bands,
-                                   pol_id,
-                                   n_sam_pol,
-                                   multicores,
-                                   progress) {
-
-    # multicores is less or equal to number of bands
-    multicores <- min(multicores, length(bands))
-    # prepare parallelization
-    .parallel_start(workers = multicores)
-    on.exit(.parallel_stop(), add = TRUE)
-
-    # Extract band values
-    ts_bands <- .jobs_map_parallel(seg_tile_bands, function(seg_tile_band) {
-        # get the scale factors, max, min and missing values
-        band_params <- seg_tile_band[["params"]][[1]]
-        missing_value <- .miss_value(band_params)
-        minimum_value <- .min_value(band_params)
-        maximum_value <- .max_value(band_params)
-        scale_factor <- .scale(band_params)
-        offset_value <- .offset(band_params)
-        # extract the values
-        values <- .tile_extract_segments(seg_tile_band)
+.segments_poly_read <- function(tile, chunk, n_sam_pol) {
+    # For cubes that have a time limit to expire (MPC cubes only)
+    tile <- .cube_token_generator(tile)
+    # Read and preprocess values of cloud
+    # Get tile bands
+    tile_bands <- .tile_bands(tile, add_cloud = FALSE)
+    # Read and preprocess values of each band
+    ts_bands <- purrr::map(tile_bands, function(band) {
+        # extract band values
+        values <- .tile_extract_segments(tile, band, chunk)
         pol_id <- values[,"pol_id"]
         values <- values[, -1:0]
-        # adjust maximum and minimum values
-        values[values == missing_value] <- NA
-        values[values < minimum_value] <- NA
-        values[values > maximum_value] <- NA
+        # Correct missing, minimum, and maximum values and
+        # apply scale and offset.
+        band_conf <- .tile_band_conf(tile = tile, band = band)
+        miss_value <- .miss_value(band_conf)
+        if (.has(miss_value)) {
+            values[values == miss_value] <- NA
+        }
+        min_value <- .min_value(band_conf)
+        if (.has(min_value)) {
+            values[values < min_value] <- NA
+        }
+        max_value <- .max_value(band_conf)
+        if (.has(max_value)) {
+            values[values > max_value] <- NA
+        }
+        scale <- .scale(band_conf)
+        if (.has(scale) && scale != 1) {
+            values <- values * scale
+        }
+        offset <- .offset(band_conf)
+        if (.has(offset) && offset != 0) {
+            values <- values + offset
+        }
         # use linear imputation
         impute_fn <- .impute_linear()
         # are there NA values? interpolate them
         if (any(is.na(values))) {
             values <- impute_fn(values)
         }
-        # correct the values using the scale factor
-        values <- values * scale_factor + offset_value
         # Returning extracted time series
[…]
