### 2024-02-05 4823f44
test to deal with non normalized cubes
