### 2024-02-05 6a0f8cc
Fix eProblems with spherical geometry when extracting MODIS time series using sf #1065
