### 2024-02-05 04c2090
adjust data type and properties to new computed indexes
