+++ inst/extdata/config_internals.yml
@@ -101,44 +101,24 @@ probs_cube_class        : ["probs_cube", "raster_cube", "sits_cube", "tbl_df",
                            "tbl", "data.frame"]
 
 # Default values for non-registered bands
-default_values       :
-    INT2S            : &conf_default_int2s
-        data_type    : "INT2S"
-        missing_value: -32768
-        minimum_value: -10000
-        maximum_value: 10000
-        offset_value : 0
-        scale_factor : 0.0001
-
-    INT4S            : &conf_default_int4s
-        data_type    : "INT4S"
-        missing_value: -2147483648
-        minimum_value: -2147483647
-        maximum_value:  2147483647
-        offset_value : 0
-        scale_factor : 0.0001
-
-    FLT4S            :
-        data_type    : "INT4S"
-        missing_value: -Inf
-        minimum_value: -Inf
-        maximum_value:  Inf
-        offset_value : 0
-        scale_factor : 1
-
-    NDVI             :
-                   <<: *conf_default_int2s
-
-    EVI              :
-                   <<: *conf_default_int2s
-
-    eo_cube          :
-        data_type    : "INT2S"
-        missing_value: -32768
-        minimum_value: -10000
-        maximum_value: 10000
-        offset_value : 0
-        scale_factor : 0.0001
+default_values           :
+    eo_cube              :
+        NDVI             : &ndvi_int2s_eo
+            data_type    : "INT2S"
+            missing_value: -32767
+            minimum_value: -10000
+            maximum_value: 10000
+            offset_value : 0
+            scale_factor : 0.0001
+        EVI              :
+            <<: *ndvi_int2s_eo
+        default          :
+            data_type    : "FLT4S"
+            missing_value: -3.402823466385288e+38
+            minimum_value: -3.402823466385288e+38
+            maximum_value: 1.7014118346015974e+38
+            offset_value : 0
+            scale_factor : 1
 
 # Derived cube definitions
 derived_cube         :
