### 2024-02-05 ca94dba
add more indices in config_internals
