### 2024-02-08 5a96838
Improve S1 cube
