+++ R/api_source_mpc.R
@@ -97,7 +97,9 @@
 #' @param source     Name of the STAC provider.
 #' @param collection Collection to be searched in the data source.
 #' @param stac_query Query that follows the STAC protocol
+#' @param bands      Names of the bands to filter
 #' @param ...        Other parameters to be passed for specific types.
+#' @param orbit      Name of the orbit (e.g. "Ascending" or "Descending")
 #' @param tiles      Selected tiles (optional)
 #' @param platform   Satellite platform (optional).
 #' @return An object referring the images of a sits cube.
@@ -106,33 +108,39 @@
         source,
         collection,
         bands, ...,
+        orbit = "Descending",
         start_date = NULL,
         end_date = NULL,
         dry_run = TRUE) {
 
     # require package
     .check_require_packages("rstac")
+    orbits <- .conf("sources", source, "collections", collection, "orbits")
+    .check_chr_within(
+        x = orbit,
+        within = orbits,
+        msg = "Invalid `orbit` parameter"
+    )
 
     stac_query <- .stac_create_items_query(
         source = source,
         collection = collection,
         roi = list(
-            "xmin" = -50.379,
-            "ymin" = -10.1573,
+            "xmin" = -50.479,
+            "ymin" = -10.1973,
             "xmax" = -50.410,
-            "ymax" = -10.1910,
+            "ymax" = -10.1510,
             "crs"  = "EPSG:4386"
         ),
         start_date = start_date,
         end_date = end_date,
         limit = 1
     )
-
     stac_query <- rstac::ext_filter(
         stac_query,
         `sar:frequency_band` == "C" &&
             `sar:instrument_mode` == "IW" &&
-            `sat:orbit_state` == "descending"
+            `sat:orbit_state` == {{orbit}}
     )
 
     # assert that service is online
@@ -235,15 +243,23 @@
 `.source_items_new.mpc_cube_sentinel-1-grd` <- function(source,
                                                         collection,
                                                         stac_query, ...,
-                                                        tiles = NULL) {
+                                                        tiles = NULL,
+                                                        orbit = "ascending") {
 
     # set caller to show in errors
     .check_set_caller(".source_items_new.mpc_cube_sentinel-1-grd")
+    orbits <- .conf("sources", source, "collections", collection, "orbits")
+    .check_chr_within(
+        x = orbit,
+        within = orbits,
+        msg = "Invalid `orbit` parameter"
+    )
 
     stac_query <- rstac::ext_filter(
         stac_query,
         `sar:frequency_band` == "C" &&
-        `sar:instrument_mode` == "IW"
+            `sar:instrument_mode` == "IW" &&
+            `sat:orbit_state` == {{orbit}}
     )
 
     # mpc does not support %in% operator, so we have to
