+++ R/api_source_mpc.R
@@ -99,7 +99,7 @@
 #' @param stac_query Query that follows the STAC protocol
 #' @param bands      Names of the bands to filter
 #' @param ...        Other parameters to be passed for specific types.
-#' @param orbit      Name of the orbit (e.g. "Ascending" or "Descending")
+#' @param orbit      Name of the orbit (e.g. "ascending" or "descending")
 #' @param tiles      Selected tiles (optional)
 #' @param platform   Satellite platform (optional).
 #' @return An object referring the images of a sits cube.
@@ -108,7 +108,7 @@
         source,
         collection,
         bands, ...,
-        orbit = "Descending",
+        orbit = "descending",
         start_date = NULL,
         end_date = NULL,
         dry_run = TRUE) {
@@ -201,6 +201,26 @@
     }
     return(invisible(NULL))
 }
+
+`.source_collection_access_test.mpc_cube_sentinel-1-rtc` <- function(
+        source,
+        collection,
+        bands, ...,
+        orbit = "descending",
+        start_date = NULL,
+        end_date = NULL,
+        dry_run = TRUE) {
+
+    `.source_collection_access_test.mpc_cube_sentinel-1-grd`(
+        source = source,
+        collection = collection,
+        bands = bands, ...,
+        orbit = orbit,
+        start_date = start_date,
+        end_date = end_date,
+        dry_run = dry_run
+    )
+}
 #' @title Get bbox from file info
 #' @keywords internal
 #' @noRd
@@ -236,6 +256,15 @@
     # create a bbox
     bbox <- c(xmin = xmin, ymin = ymin, xmax = xmax, ymax = ymax)
     return(bbox)
+}
+`.source_tile_get_bbox.mpc_cube_sentinel-1-rtc` <- function(source,
+                                                            file_info, ...,
+                                                            collection = NULL) {
+    `.source_tile_get_bbox.mpc_cube_sentinel-1-grd`(
+        source = source,
+        file_info = file_info, ...,
+        collection = collection
+    )
 }
 #' @keywords internal
 #' @noRd
@@ -304,7 +333,19 @@
     )
     return(items_info)
 }
-
+`.source_items_new.mpc_cube_sentinel-1-rtc` <- function(source,
+                                                        collection,
+                                                        stac_query, ...,
+                                                        tiles = NULL,
+                                                        orbit = "ascending") {
+    `.source_items_new.mpc_cube_sentinel-1-grd`(
+        source = source,
+        collection = collection,
+        stac_query = stac_query, ...,
+        tiles = tiles,
+        orbit = orbit
+    )
+}
 #' @keywords internal
 #' @noRd
 #' @export
@@ -312,6 +353,15 @@
                                                          items, ...,
                                                          collection = NULL) {
     rep("NoTilingSystem", rstac::items_length(items))
+}
+`.source_items_tile.mpc_cube_sentinel-1-rtc` <- function(source,
+                                                         items, ...,
+                                                         collection = NULL) {
+    `.source_items_tile.mpc_cube_sentinel-1-grd`(
+        source = source,
+        items = items, ...,
+        collection = collection
+    )
 }
 
 #' @keywords internal
