+++ NAMESPACE
@@ -304,6 +304,7 @@ S3method(sits_combine_predictions,average)
 S3method(sits_combine_predictions,default)
 S3method(sits_combine_predictions,uncertainty)
 S3method(sits_cube,"mpc_cube_sentinel-1-grd")
+S3method(sits_cube,"mpc_cube_sentinel-1-rtc")
 S3method(sits_cube,default)
 S3method(sits_cube,local_cube)
 S3method(sits_cube,stac_cube)
