### 2024-02-11 9e6f9fe
Fix scale_factor value
