### 2024-02-12 14796e5
Fix scale_factor
