+++ inst/extdata/config.yml
@@ -1240,7 +1240,7 @@ sources:
                         missing_value : 0
                         minimum_value : 1
                         maximum_value : 65534
-                        scale_factor  : 1
+                        scale_factor  : 0.001
                         offset_value  : 0
                         resolution    : 10
                         band_name     : "vv"
