+++ NAMESPACE
@@ -16,6 +16,8 @@ S3method(.band_rename,sits)
 S3method(.check_samples,default)
 S3method(.check_samples,sits)
 S3method(.check_samples,tbl_df)
+S3method(.cube_adjust_crs,"mpc_cube_sentinel-1-grd")
+S3method(.cube_adjust_crs,default)
 S3method(.cube_as_sf,default)
 S3method(.cube_as_sf,raster_cube)
 S3method(.cube_bands,default)
