### 2024-02-14 538a1b7
Revert "Adjust data type and properties to new computed indexes in `sits_apply()`"
