### 2024-02-14 ec47e06
Revert "Revert "Adjust data type and properties to new computed indexes in `sits_apply()`""
