+++ R/api_conf.R
@@ -428,7 +428,8 @@
 #' @keywords internal
 #' @noRd
 #' @return   names associated to the chosen access key
-.conf_names <- function(key) {
+.conf_names <- function(...) {
+    key <- c(...)
     res <- tryCatch(
         {
             names(sits_env$config[[key]])
@@ -928,41 +929,23 @@ NULL
 #' @param source  Data source.
 #' @param collection  Collection in the data source.
 #' @param band  Band name
-#' @param tile  Tile
 #' @details
 #' If the band is not found, a default value will be returned from config.
 #' If neither source nor collection entries are found in configuration file,
 #' an error is thrown.
 #' @returns  A value in config.
-.conf_eo_band <- function(source, collection, band, tile = NULL) {
+.conf_eo_band <- function(source, collection, band) {
     # Format band name
     band <- .band_eo(band)
     # Return a default value if band does not exists in config
     if (!.conf_eo_band_exists(source, collection, band)) {
-        data_type <- "INT2S"
-        # does the file exist?
-        if (!purrr::is_null(tile)) {
-            band_path <- .tile_path(tile, band)
-            if (!purrr::is_null(band_path)) {
-                rast <- terra::rast(band_path)
-                data_type <- terra::datatype(rast)
-                conf_band <- .conf("default_vales", data_type)
-            } else {
-            # file does not exist
-            if (!.has(.conf("default_vales", band)))
-                conf_band <- .conf("default_vales", band)
-            else
-                conf_band <- .conf("default_vales", "INT4S")
-            }
+        if (band %in% .conf_names("default_values", "eo_cube")) {
+            return(.conf("default_values", "eo_cube", band))
         }
+        return(.conf("default_values", "eo_cube", "default"))
     }
-    else {
-        # Get band config value and return it
-        conf_band <- .conf("sources", source,
-                           "collections", collection,
-                           "bands", band)
-    }
-    return(conf_band)
+    # Get band config value and return it
+    .conf("sources", source, "collections", collection, "bands", band)
 }
 #' @title Config functions for derived_cube
 #' @noRd
