+++ R/sits_apply.R
@@ -12,16 +12,19 @@
 #' cubes, it materializes a new band in \code{output_dir} using
 #' \code{gdalcubes}.
 #'
-#' @param data          Valid sits tibble or cube
-#' @param window_size   An odd number representing the size of the
-#'                      sliding window of sits kernel functions
-#'                      used in expressions (for a list of supported
-#'                      kernel functions, please see details).
-#' @param memsize       Memory available for classification (in GB).
-#' @param multicores    Number of cores to be used for classification.
-#' @param output_dir    Directory where files will be saved.
-#' @param progress      Show progress bar?
-#' @param ...           Named expressions to be evaluated (see details).
+#' @param data              Valid sits tibble or cube
+#' @param window_size       An odd number representing the size of the
+#'                          sliding window of sits kernel functions
+#'                          used in expressions (for a list of supported
+#'                          kernel functions, please see details).
+#' @param normalized        An logical value representing the
+#'                          output generate index. Default is TRUE.
+#'                          (see details)
+#' @param memsize           Memory available for classification (in GB).
+#' @param multicores        Number of cores to be used for classification.
+#' @param output_dir        Directory where files will be saved.
+#' @param progress          Show progress bar?
+#' @param ...               Named expressions to be evaluated (see details).
 #'
 #' @details
 #' \code{sits_apply()} allow any valid R expression to compute new bands.
@@ -50,6 +53,12 @@
 #' to the cube. The kernel functions ignores any \code{NA} values
 #' inside the kernel window. Central pixel is \code{NA} just only
 #' all pixels in the window are \code{NA}.
+#'
+#' By default, the indexes generated by the \code{sits_apply()} function are
+#' normalized between -1 and 1, scaled by a factor of 0.0001.
+#' Normalized indexes are saved as INT2S (Integer with sign).
+#' If the \code{normalized} parameter is FALSE, no scaling factor will be
+#' applied and the index will be saved as FLT4S (Float with sign).
 #'
 #' @section Summarizing kernel functions:
 #' \itemize{
