+++ man/sits_apply.Rd
@@ -17,6 +17,7 @@ sits_apply(data, ...)
   data,
   ...,
   window_size = 3L,
+  normalized = TRUE,
   memsize = 4L,
   multicores = 2L,
   output_dir,
@@ -38,6 +39,10 @@ sits_apply(data, ...)
 sliding window of sits kernel functions
 used in expressions (for a list of supported
 kernel functions, please see details).}
+
+\item{normalized}{An logical value representing the
+output generate index. Default is TRUE.
+(see details)}
 
 \item{memsize}{Memory available for classification (in GB).}
 
@@ -84,6 +89,12 @@ generates an entire new matrix, which is returned as a new band
 to the cube. The kernel functions ignores any \code{NA} values
 inside the kernel window. Central pixel is \code{NA} just only
 all pixels in the window are \code{NA}.
+
+By default, the indexes generated by the \code{sits_apply()} function are
+normalized between -1 and 1, scaled by a factor of 0.0001.
+Normalized indexes are saved as INT2S (Integer with sign).
+If the \code{normalized} parameter is FALSE, no scaling factor will be
+applied and the index will be saved as FLT4S (Float with sign).
 }
 \section{Summarizing kernel functions}{
 
