### 2024-02-15 a169734
update .tile_band_conf function to accept sits_apply() changes
