### 2024-02-15 184dad8
update .tile_band_conf call in sits_mixture function
