### 2024-02-15 08d65f7
add .raster_datatype function
