### 2024-02-15 9f3f339
add tests for the  parameter in sits_apply
