+++ inst/extdata/config_internals.yml
@@ -126,20 +126,6 @@ default_values       :
         offset_value : 0
         scale_factor : 1
 
-    NDVI             :
-                   <<: *conf_default_int2s
-
-    EVI              :
-                   <<: *conf_default_int2s
-
-    eo_cube          :
-        data_type    : "INT2S"
-        missing_value: -32768
-        minimum_value: -10000
-        maximum_value: 10000
-        offset_value : 0
-        scale_factor : 0.0001
-
 # Derived cube definitions
 derived_cube         :
     probs_cube       :
