### 2024-02-15 aa8357d
fix legend error in sits_reclassify
