+++ tests/testthat/test-reclassify.R
@@ -11,17 +11,9 @@ test_that("One-year, multicores processing reclassify", {
         ),
         bands = "class",
         version = "v20220606",
-        labels = c("1" = "Forest", "2" = "Water", "3" = "NonForest",
-                   "4" = "NonForest2", "6" = "d2007", "7" = "d2008",
-                   "8" = "d2009", "9" = "d2010", "10" = "d2011",
-                   "11" = "d2012", "12" = "d2013", "13" = "d2014",
-                   "14" = "d2015", "15" = "d2016", "16" = "d2017",
-                   "17" = "d2018", "18" = "r2010", "19" = "r2011",
-                   "20" = "r2012", "21" = "r2013", "22" = "r2014",
-                   "23" = "r2015", "24" = "r2016", "25" = "r2017",
-                   "26" = "r2018", "27" = "d2019", "28" = "r2019",
-                   "29" = "d2020", "31" = "r2020", "32" = "Clouds2021",
-                   "33" = "d2021", "34" = "r2021"),
+        labels = c("1" = "Forest", "11" = "d2012", "16" = "d2017",
+                   "17" = "d2018", "27" = "d2019", "29" = "d2020",
+                   "32" = "Clouds2021", "33" = "d2021"),
         progress = FALSE
     )
     # Open classification map
@@ -47,18 +39,9 @@ test_that("One-year, multicores processing reclassify", {
         mask = prodes2021,
         rules = list(
             "Old_Deforestation" = mask %in% c(
-                "d2007", "d2008", "d2009",
-                "d2010", "d2011", "d2012",
-                "d2013", "d2014", "d2015",
-                "d2016", "d2017", "d2018",
-                "r2010", "r2011", "r2012",
-                "r2013", "r2014", "r2015",
-                "r2016", "r2017", "r2018",
-                "d2019", "r2019", "d2020",
-                "r2020", "r2021"
-            ),
-            "Water_Mask" = mask == "Water",
-            "NonForest_Mask" = mask %in% c("NonForest", "NonForest2")
+                "d2012", "d2017", "d2018",
+                "d2019", "d2020", "d2021"
+            )
         ),
         memsize = 4,
         multicores = 2,
@@ -71,7 +54,7 @@ test_that("One-year, multicores processing reclassify", {
         c(
             "1" = "ClearCut_Fire", "2" =  "ClearCut_Soil",
             "3" =  "ClearCut_Veg", "4" = "Forest",
-            "5" = "Old_Deforestation", "7" = "NonForest_Mask"
+            "5" = "Old_Deforestation"
         )
     )
     ro_class_obj <- .raster_open_rast(.tile_path(ro_class))
@@ -87,7 +70,7 @@ test_that("One-year, multicores processing reclassify", {
     # prodes2021 is ""d2018"
     expect_equal(vls_prodes2021[2000], 1)
     # ro_class is "Old_Deforestation"
-    expect_equal(vls_ro_mask[2000], 5)
+    expect_equal(vls_ro_mask[2000], 3)
 
     out <- capture_messages({
         expect_message(
@@ -97,19 +80,9 @@ test_that("One-year, multicores processing reclassify", {
                     mask = prodes2021,
                     rules = list(
                         "Old_Deforestation" = mask %in% c(
-                            "d2007", "d2008", "d2009",
-                            "d2010", "d2011", "d2012",
-                            "d2013", "d2014", "d2015",
-                            "d2016", "d2017", "d2018",
-                            "r2010", "r2011", "r2012",
-                            "r2013", "r2014", "r2015",
-                            "r2016", "r2017", "r2018",
-                            "d2019", "r2019", "d2020",
-                            "r2020", "r2021"
-                        ),
-                        "Water_Mask" = mask == "Water",
[…]
