### 2024-02-19 c404b54
add extraction of segments time series using chunks
