+++ man/sits_regularize.Rd
@@ -148,6 +148,7 @@ if (sits_run_examples()) {
         source = "MPC",
         collection = "SENTINEL-1-GRD",
         bands = c("VV", "VH"),
+        orbit = "descending",
         roi = roi,
         start_date = "2020-06-01",
         end_date = "2020-09-28"
