### 2024-02-27 db23810
add dtw distance in sits_som_map
