+++ src/RcppExports.cpp
@@ -1,6 +1,7 @@
 // Generated by using Rcpp::compileAttributes() -> do not edit by hand
 // Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+#include "sits_types.h"
 #include <RcppArmadillo.h>
 #include <Rcpp.h>
 
@@ -137,6 +138,16 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type band(bandSEXP);
     Rcpp::traits::input_parameter< int >::type window_size(window_sizeSEXP);
     rcpp_result_gen = Rcpp::wrap(C_kernel_modal(x, ncols, nrows, band, window_size));
+    return rcpp_result_gen;
+END_RCPP
+}
+// dtw
+Rcpp::XPtr<DistanceFunctionPtr> dtw();
+RcppExport SEXP _sits_dtw() {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    rcpp_result_gen = Rcpp::wrap(dtw());
     return rcpp_result_gen;
 END_RCPP
 }
@@ -444,6 +455,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_sits_C_kernel_max", (DL_FUNC) &_sits_C_kernel_max, 5},
     {"_sits_C_kernel_var", (DL_FUNC) &_sits_C_kernel_var, 5},
     {"_sits_C_kernel_modal", (DL_FUNC) &_sits_C_kernel_modal, 5},
+    {"_sits_dtw", (DL_FUNC) &_sits_dtw, 0},
     {"_sits_C_label_max_prob", (DL_FUNC) &_sits_C_label_max_prob, 1},
     {"_sits_linear_interp", (DL_FUNC) &_sits_linear_interp, 1},
     {"_sits_linear_interp_vec", (DL_FUNC) &_sits_linear_interp_vec, 1},
