### 2024-03-04 79a3931
improvement in plotting to include different styles and to have scale as a parameter
