+++ R/sits_plot.R
@@ -321,15 +321,12 @@ plot.predicted <- function(x, y, ...,
 #' @param  tile          Tile to be plotted.
 #' @param  date          Date to be plotted.
 #' @param  palette       An RColorBrewer palette
+#' @param  style         Method to process the color scale
+#'                       ("cont", "order", "quantile", "fisher",
+#'                        "jenks", "log10")
+#' @param  n_colors      Number of colors to be shown
 #' @param  rev           Reverse the color order in the palette?
-#' @param  tmap_options  Named list with optional tmap parameters
-#'                       max_cells (default: 1e+06)
-#'                       scale (default: 1.0)
-#'                       graticules_labels_size (default: 0.7)
-#'                       legend_title_size (default: 1.0)
-#'                       legend_text_size (default: 1.0)
-#'                       legend_bg_color (default: "white")
-#'                       legend_bg_alpha (default: 0.5)
+#' @param  scale         Scale to plot map (0.4 to 1.0)
 #'
 #' @return               A plot object with an RGB image
 #'                       or a B/W image on a color
@@ -359,8 +356,10 @@ plot.raster_cube <- function(x, ...,
                              tile = x$tile[[1]],
                              date = NULL,
                              palette = "RdYlGn",
+                             style = "cont",
+                             n_colors = 10,
                              rev = FALSE,
-                             tmap_options = NULL) {
+                             scale = 0.8) {
     # check for color_palette parameter (sits 1.4.1)
     dots <- list(...)
     if (missing(palette) && "color_palette" %in% names(dots)) {
@@ -414,8 +413,10 @@ plot.raster_cube <- function(x, ...,
             sf_seg    = NULL,
             seg_color = NULL,
             palette = palette,
+            style = style,
+            n_colors = n_colors,
             rev = rev,
-            tmap_options = tmap_options
+            scale = scale
         )
     } else {
         # plot RGB image
@@ -428,8 +429,7 @@ plot.raster_cube <- function(x, ...,
             blue = blue,
             date = date,
             sf_seg    = NULL,
-            seg_color = NULL,
-            tmap_options = tmap_options
+            seg_color = NULL
         )
     }
     return(p)
@@ -451,15 +451,12 @@ plot.raster_cube <- function(x, ...,
 #' @param  seg_color     Color to show the segment boundaries
 #' @param  line_width    Line width to plot the segments boundary (in pixels)
 #' @param  palette       An RColorBrewer palette
+#' @param  style         Method to process the color scale
+#'                       ("cont", "order", "quantile", "fisher",
+#'                        "jenks", "log10")
+#' @param  n_colors      Number of colors to be shown
 #' @param  rev           Reverse the color order in the palette?
-#' @param  tmap_options  Named list with optional tmap parameters
-#'                       max_cells (default: 1e+06)
-#'                       scale (default: 1.0)
-#'                       graticules_labels_size (default: 0.7)
-#'                       legend_title_size (default: 1.0)
-#'                       legend_text_size (default: 1.0)
-#'                       legend_bg_color (default: "white")
-#'                       legend_bg_alpha (default: 0.5)
+#' @param  scale         Scale to plot map (0.4 to 1.0)
 #'
 #' @return               A plot object with an RGB image
 #'                       or a B/W image on a color
@@ -496,8 +493,10 @@ plot.vector_cube <- function(x, ...,
                              seg_color = "black",
                              line_width = 1,
                              palette = "RdYlGn",
+                             style = "cont",
+                             n_colors = 10,
                              rev = FALSE,
-                             tmap_options = NULL) {
+                             scale = 0.8) {
     # check for color_palette parameter (sits 1.4.1)
     dots <- list(...)
[…]
