+++ tests/testthat/test-color.R
@@ -37,6 +37,17 @@ test_that("sits colors", {
     sits_colors_reset()
 })
 
+test_that("color errors", {
+    colors <- sits_colors(legend = "IGBP")
+    expect_equal(nrow(colors), 16)
+    expect_equal(colors[16,1]$name, "Water_Bodies")
+    out <- capture.output(sits_colors("New"))
+    expect_true(grepl("Please", out[2]))
+    out <- capture.output(sits_colors_show())
+    expect_true(grepl("Please", out[1]))
+
+})
+
 test_that("plot colors", {
     data_dir <- system.file("extdata/raster/classif", package = "sits")
     ro_class <- sits_cube(
