### 2024-03-04 14a48ec
fix documentation error is plot.vector_cube
