+++ tests/testthat/test-space-time-operations.R
@@ -15,7 +15,7 @@ test_that("Time Series Dates", {
 
     cerrado_tb <- cerrado_2classes
     class(cerrado_tb) <- "tbl_df"
-    times2 <- sits_timeline(cerrado_2classes)
+    times2 <- sits_timeline(cerrado_tb)
     expect_true(length(times2) == 23)
 })
 test_that("Timeline format", {
