### 2024-03-07 e2390e9
add new function in torch api
