### 2024-03-07 b4402f6
update classify to accelerate cuda classificatins
