### 2024-03-07 c2e7ff7
update tempcnn to support inference with batch
