+++ tests/testthat/test-summary.R
@@ -18,6 +18,28 @@ test_that("summary cube",{
     sum <- capture.output(summary(cube))
     expect_true(grepl("MODIS", sum[1]))
     expect_true(grepl("Median", sum[4]))
+
+    tiles <- c("007004", "007005")
+    start_date <- "2022-05-01"
+    end_date <- "2022-08-29"
+    bands <- c("NDVI", "EVI", "B13", "B14", "B15", "B16", "CLOUD")
+    # create a raster cube file from BDC
+    cbers_cube_8d <- .try(
+        {
+            sits_cube(
+                source = "BDC",
+                collection = "CBERS-WFI-8D",
+                tiles = tiles,
+                start_date = start_date,
+                end_date = end_date,
+                progress = FALSE
+            )
+        },
+        .default = NULL
+    )
+    sum2 <- capture.output(summary(cbers_cube_8d, tile = "007004"))
+    expect_true(grepl("007004", sum2[4]))
+    expect_true(grepl("007004", sum2[48]))
 })
 
 test_that("summary sits accuracy", {
