+++ inst/extdata/cran/sits_codecov.R
@@ -1,1 +1,2 @@
 withr::with_envvar(new = c("SITS_RUN_TESTS" = "YES"), covr::codecov(token = "53fbd031-6dab-40a4-98a5-84885b45531e"))
+
