+++ R/api_clean.R
@@ -85,8 +85,6 @@
         # Returned block files for each fraction
         block_files
     })
-    # Remove raster without clean
-    unlink(.tile_path(tile))
     # Merge blocks into a new class_cube tile
     band_tile <- .tile_derived_merge_blocks(
         file = out_file,
