+++ R/sits_label_classification.R
@@ -162,7 +162,7 @@ sits_label_classification.derived_cube <- function(cube, ...) {
 
 #' @rdname sits_label_classification
 #' @export
-sits_label_classification.tbl_df <- function(cube, ...){
+sits_label_classification.default <- function(cube, ...){
     cube <- tibble::as_tibble(cube)
     if (all(.conf("sits_cube_cols") %in% colnames(cube))) {
         cube <- .cube_find_class(cube)
@@ -170,10 +170,4 @@ sits_label_classification.tbl_df <- function(cube, ...){
         stop("Input should be a classified cube")
     class_cube <- sits_label_classification(cube, ...)
     return(class_cube)
-}
-
-#' @rdname sits_label_classification
-#' @export
-sits_label_classification.default <- function(cube, ...) {
-    stop("Input should be a classified cube")
 }

[…]
