+++ src/RcppExports.cpp
@@ -456,42 +456,6 @@ BEGIN_RCPP
     return R_NilValue;
 END_RCPP
 }
-// sample_points_inclusion
-NumericMatrix sample_points_inclusion(const NumericMatrix& polymatrix, const int n_sam_pol);
-RcppExport SEXP _sits_sample_points_inclusion(SEXP polymatrixSEXP, SEXP n_sam_polSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const NumericMatrix& >::type polymatrix(polymatrixSEXP);
-    Rcpp::traits::input_parameter< const int >::type n_sam_pol(n_sam_polSEXP);
-    rcpp_result_gen = Rcpp::wrap(sample_points_inclusion(polymatrix, n_sam_pol));
-    return rcpp_result_gen;
-END_RCPP
-}
-// sample_points_crossings
-NumericMatrix sample_points_crossings(const NumericMatrix& polymatrix, const int n_sam_pol);
-RcppExport SEXP _sits_sample_points_crossings(SEXP polymatrixSEXP, SEXP n_sam_polSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const NumericMatrix& >::type polymatrix(polymatrixSEXP);
-    Rcpp::traits::input_parameter< const int >::type n_sam_pol(n_sam_polSEXP);
-    rcpp_result_gen = Rcpp::wrap(sample_points_crossings(polymatrix, n_sam_pol));
-    return rcpp_result_gen;
-END_RCPP
-}
-// sample_points_bin
-NumericMatrix sample_points_bin(const NumericMatrix& polymatrix, const int n_sam_pol);
-RcppExport SEXP _sits_sample_points_bin(SEXP polymatrixSEXP, SEXP n_sam_polSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const NumericMatrix& >::type polymatrix(polymatrixSEXP);
-    Rcpp::traits::input_parameter< const int >::type n_sam_pol(n_sam_polSEXP);
-    rcpp_result_gen = Rcpp::wrap(sample_points_bin(polymatrix, n_sam_pol));
-    return rcpp_result_gen;
-END_RCPP
-}
 // C_max_sampling
 DataFrame C_max_sampling(const NumericVector& x, int nrows, int ncols, int window_size);
 RcppExport SEXP _sits_C_max_sampling(SEXP xSEXP, SEXP nrowsSEXP, SEXP ncolsSEXP, SEXP window_sizeSEXP) {
@@ -503,22 +467,6 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type ncols(ncolsSEXP);
     Rcpp::traits::input_parameter< int >::type window_size(window_sizeSEXP);
     rcpp_result_gen = Rcpp::wrap(C_max_sampling(x, nrows, ncols, window_size));
-    return rcpp_result_gen;
-END_RCPP
-}
-// bayes_smoother
-arma::mat bayes_smoother(const arma::mat& m, const arma::uword m_nrow, const arma::uword m_ncol, const arma::mat& w, const arma::mat& sigma, const double neigh_fraction);
[…]
