+++ NAMESPACE
@@ -334,7 +334,6 @@ S3method(sits_mixture_model,tbl_df)
 S3method(sits_model_export,sits_model)
 S3method(sits_reclassify,class_cube)
 S3method(sits_reclassify,default)
-S3method(sits_reclassify,tbl_df)
 S3method(sits_reduce,raster_cube)
 S3method(sits_reduce,sits)
 S3method(sits_regularize,"mpc_cube_sentinel-1-grd")
