### 2024-03-13 78415ad
fix additional problems in testing segmentation
