### 2024-03-13 d68d1e9
Implement function to filter non-empty cube/tile
