### 2024-03-14 b2a4cbe
new function sits_stratified_sampling
