+++ R/sits_mixture_model.R
@@ -249,5 +249,7 @@ sits_mixture_model.tbl_df <- function(data, endmembers, ...) {
 #' @rdname sits_mixture_model
 #' @export
 sits_mixture_model.default <- function(data, endmembers, ...){
-    stop("Input should be an object of class cube or class sits")
+    data <- tibble::as_tibble(data)
+    data <- sits_mixture_model(data, endmembers, ...)
+    return(data)
 }
