+++ man/sits_classify.Rd
@@ -5,7 +5,6 @@
 \alias{sits_classify.sits}
 \alias{sits_classify.raster_cube}
 \alias{sits_classify.segs_cube}
-\alias{sits_classify.derived_cube}
 \alias{sits_classify.default}
 \title{Classify time series or data cubes}
 \usage{
@@ -61,8 +60,6 @@ sits_classify(
   verbose = FALSE,
   progress = TRUE
 )
-
-\method{sits_classify}{derived_cube}(data, ml_model, ...)
 
 \method{sits_classify}{default}(data, ml_model, ...)
 }
