### 2024-03-14 7b14a0f
examples stratified sampling
