### 2024-03-17 376181a
update sits_reduce docs
