+++ man/sits_reduce.Rd
@@ -56,6 +56,12 @@ represents a single date. We provide some operations already
 implemented in the package to perform the reduce operation.
 See the list of available functions below:
 }
+\note{
+The \code{t_sum()}, \code{t_std()}, \code{t_skewness()},
+ \code{t_kurtosis}, \code{t_mse} indexes generate values greater than
+ the limit of a two-byte integer. Therefore, we save the images
+ generated by these as Float-32 with no scale.
+}
 \section{Summarizing temporal functions}{
 
 \itemize{
@@ -77,8 +83,6 @@ See the list of available functions below:
  The energy of the time series is distributed by frequency.}
  \item{\code{t_fqr()}: Returns the value of the first quartile of the
  series (0.25).}
- \item{\code{t_sqr()}: Returns the value of the second quartile of the
- series (0.50).}
  \item{\code{t_tqr()}: Returns the value of the third quartile of the
  series (0.75).}
  \item{\code{t_iqr()}: Returns the interquartile range
