### 2024-03-17 00a11dc
update sits_reduces tests
