### 2024-03-17 9b309bd
add batch classification feature in torch models
