### 2024-03-18 e01fc23
remove unused chunks in sits_classify segments
