### 2024-03-18 78df500
add support to mosaic derived cubes in sits_mosaic
