### 2024-03-18 b8e770e
remove cropped images from .sits directory in sits_mosaic
