### 2024-03-18 1c71866
update docs on GPU usage in sits_classify
