### 2024-03-18 290f4cb
update GPU support to segments classification
