### 2024-03-21 d34a67c
support for tiles in DEAfrica and HLS
