+++ tests/testthat/test-cube.R
@@ -426,22 +426,30 @@ test_that("Creating cubes from DEA", {
     expect_equal(dea_cube$xmax[[1]], .raster_xmax(r), tolerance = 1)
     expect_equal(dea_cube$xmin[[1]], .raster_xmin(r), tolerance = 1)
 })
-test_that("Creating cubes from DEA - error using tiles", {
-    expect_error(
-        object = {
-            dea_cube <-
-                sits_cube(
+test_that("Creating cubes from DEA - using tiles", {
+    dea_cube <- .try(
+        {
+            sits_cube(
                     source = "DEAFRICA",
                     collection = "s2_l2a",
-                    bands = c("B01", "B04", "B05"),
-                    tiles = "37MEP",
+                    bands = c("B02", "B8A", "B11"),
+                    tiles = c("37MDT","37MET"),
                     start_date = "2019-01-01",
                     end_date = "2019-10-28",
                     progress = FALSE
                 )
         },
-        regexp = "DEAFRICA cubes do not support searching for tiles"
+        .default = NULL
     )
+    testthat::skip_if(purrr::is_null(dea_cube),
+                      message = "DEAFRICA is not accessible"
+    )
+    expect_true(all(sits_bands(dea_cube) %in% c("B02", "B8A", "B11")))
+    r <- .raster_open_rast(.tile_path(dea_cube))
+    expect_equal(dea_cube$xmax[[1]], .raster_xmax(r), tolerance = 1)
+    expect_equal(dea_cube$xmin[[1]], .raster_xmin(r), tolerance = 1)
+    expect_true(all(dea_cube$tile %in% c("37MDT","37MET")))
+
 })
 test_that("Creating Sentinel cubes from MPC", {
     mpc_token <- Sys.getenv("MPC_TOKEN")
@@ -522,20 +530,69 @@ test_that("Creating Sentinel cubes from MPC with ROI", {
     cube_nrows <- .tile_nrows(s2_cube_mpc)
     expect_true(.raster_nrows(r_obj) == cube_nrows)
 })
-test_that("Creating Sentinel-1 cubes from MPC", {
-    cube_s1 <-  sits_cube(
+test_that("Creating Sentinel-1 RTC cubes from MPC", {
+    roi <- sits_mgrs_to_roi("20LKP")
+    cube_s1_rtc <-  sits_cube(
         source = "MPC",
         collection = "SENTINEL-1-RTC",
         bands = c("VV", "VH"),
         orbit = "descending",
-        roi = c(lat_max = 3.619058, lon_max = -74.01153, lat_min = 2.625301, lon_min = -75.00018),
+        roi = roi,
         start_date = "2023-03-01",
         end_date = "2023-09-30"
     )
-    bbox <- sits_bbox(cube_s1)
+    bbox <- sits_bbox(cube_s1_rtc)
     expect_true(grepl("18N", bbox[["crs"]]))
     expect_equal(381340, bbox[["xmin"]])
     expect_equal(701860, bbox[["xmax"]])
+
+    # output_dir <- paste0(tempdir(), "/s1rtc2")
+    # if (!dir.exists(output_dir)) {
+    #     dir.create(output_dir)
+    # }
+    #
+    # cube_s1_20LKP_rtc <- sits_regularize(
+    #     cube = cube_s1_rtc,
+    #     period = "P12D",
+    #     res = 120,
+    #     roi = roi,
+    #     multicores = 4,
+    #     output_dir = output_dir,
+    #     progress = TRUE
+    # )
+})
[…]

[…]
