### 2024-03-21 d8f423b
Fix sits_regularize() for mpc sentinel1 rtc
