+++ R/sits_regularize.R
@@ -213,6 +213,50 @@ sits_regularize.raster_cube <- function(cube,
 }
 #' @rdname sits_regularize
 #' @export
+`sits_regularize.mpc_cube_sentinel-1-rtc` <- function(cube,
+                                                      period,
+                                                      res,
+                                                      output_dir,
+                                                      roi = NULL,
+                                                      multicores = 2L,
+                                                      progress = TRUE) {
+    # Preconditions
+    .check_raster_cube_files(cube)
+    .period_check(period)
+    .check_num_parameter(res, exclusive_min = 0)
+    output_dir <- .file_normalize(output_dir)
+    .check_output_dir(output_dir)
+    .check_multicores(multicores, min = 1, max = 2048)
+    .check_progress(progress)
+    .check_null(roi, msg = "invalid roi parameter")
+    roi <- .roi_as_sf(roi)
+    # Display warning message in case STAC cube
+    if (!.cube_is_local(cube)) {
+        if (.check_warnings()) {
+            warning("Regularization works better when data store locally. ",
+                    "Please, use 'sits_cube_copy()' to copy data locally ",
+                    "before regularization",
+                    call. = FALSE, immediate. = TRUE
+            )
+        }
+    }
+    # Prepare parallel processing
+    .parallel_start(workers = multicores)
+    on.exit(.parallel_stop(), add = TRUE)
+    # Convert input sentinel1 cube to sentinel2 grid
+    cube <- .reg_s2tile_convert(cube = cube, roi = roi)
+    # Call regularize in parallel
+    .reg_cube(
+        cube = cube,
+        res = res,
+        roi = roi,
+        period = period,
+        output_dir = output_dir,
+        progress = progress
+    )
+}
+#' @rdname sits_regularize
+#' @export
 sits_regularize.derived_cube <- function(cube,
                                          period,
                                          res,
