### 2024-03-22 b2b3ca8
improve test cube
