### 2024-03-22 4985069
filter cubes in S1 regularize
