### 2024-03-23 c6c7423
improve test cube
