### 2024-03-23 0350bfa
document sits_bbox
