### 2024-03-25 dc1d539
Fix bug in sits_regularize() for mpc_sentinel-1-rtc cube
