### 2024-03-25 c5dfcf7
improve test cube
