### 2024-03-25 56c304a
fix error when creating local cubes with NA
