+++ R/api_cube.R
@@ -330,13 +330,24 @@ NULL
 }
 #' @export
 .cube_s3class.raster_cube <- function(cube) {
-    s3_class <- .source_s3class(source = .cube_source(cube = cube))
+
+    source <-  .cube_source(cube = cube)
+    collection <- .tile_collection(cube)
+    s3_class <- .source_s3class(source = source)
     col_class <- paste(
         s3_class[[1]],
-        tolower(.tile_collection(cube)),
+        tolower(collection),
         sep = "_"
     )
-    unique(c(col_class, s3_class, class(cube)))
+    sar_cube <- .try({
+        .conf("sources", source, "collections", collection, "sar_cube")
+    },
+    .default = FALSE
+    )
+    if (sar_cube)
+        unique(c(col_class, "sar_cube", s3_class, class(cube)))
+    else
+        unique(c(col_class, s3_class, class(cube)))
 }
 #' @export
 .cube_s3class.default <- function(cube) {
