+++ R/sits_regularize.R
@@ -171,14 +171,14 @@ sits_regularize.raster_cube <- function(cube, ...,
 }
 #' @rdname sits_regularize
 #' @export
-`sits_regularize.mpc_cube_sentinel-1-grd` <- function(cube, ...,
-                                                      period,
-                                                      res,
-                                                      output_dir,
-                                                      roi = NULL,
-                                                      tiles = NULL,
-                                                      multicores = 2L,
-                                                      progress = TRUE) {
+sits_regularize.sar_cube <- function(cube, ...,
+                                     period,
+                                     res,
+                                     output_dir,
+                                     roi = NULL,
+                                     tiles = NULL,
+                                     multicores = 2L,
+                                     progress = TRUE) {
     # Preconditions
     .check_raster_cube_files(cube)
     .period_check(period)
@@ -189,10 +189,6 @@ sits_regularize.raster_cube <- function(cube, ...,
     .check_progress(progress)
     # Check for ROI and tiles
     .check_roi_tiles(roi, tiles)
-    if (is.character(tiles)) {
-        roi <- .s2_mgrs_to_roi(tiles)
-    }
-    roi <- .roi_as_sf(roi)
     # Display warning message in case STAC cube
     if (!.cube_is_local(cube)) {
         if (.check_warnings()) {
@@ -207,7 +203,7 @@ sits_regularize.raster_cube <- function(cube, ...,
     .parallel_start(workers = multicores)
     on.exit(.parallel_stop(), add = TRUE)
     # Convert input sentinel1 cube to sentinel2 grid
-    cube <- .reg_s2tile_convert(cube = cube, roi = roi)
+    cube <- .reg_s2tile_convert(cube = cube, roi = roi, tiles = tiles)
     .check_that(
         nrow(cube) > 0,
         msg = "Spatial region does not intersect cube"
@@ -226,27 +222,6 @@ sits_regularize.raster_cube <- function(cube, ...,
         progress = progress
     )
     return(cube)
-}
-#' @rdname sits_regularize
-#' @export
-`sits_regularize.mpc_cube_sentinel-1-rtc` <- function(cube, ...,
-                                                      period,
-                                                      res,
-                                                      output_dir,
-                                                      roi = NULL,
-                                                      tiles = NULL,
-                                                      multicores = 2L,
-                                                      progress = TRUE) {
-    `sits_regularize.mpc_cube_sentinel-1-grd`(
-        cube = cube,
-        period = period,
-        res = res,
-        output_dir = output_dir,
-        roi = roi,
-        tiles  = tiles,
-        multicores = multicores,
-        progress = progress, ...
-    )
 }
 #' @rdname sits_regularize
 #' @export
