+++ man/sits_cube.Rd
@@ -2,29 +2,14 @@
 % Please edit documentation in R/sits_cube.R
 \name{sits_cube}
 \alias{sits_cube}
-\alias{sits_cube.mpc_cube_sentinel-1-grd}
-\alias{sits_cube.mpc_cube_sentinel-1-rtc}
+\alias{sits_cube.sar_cube}
 \alias{sits_cube.stac_cube}
 \alias{sits_cube.local_cube}
 \title{Create data cubes from image collections}
 \usage{
 sits_cube(source, collection, ...)
 
-\method{sits_cube}{`mpc_cube_sentinel-1-grd`}(
-  source,
-  collection,
-  ...,
-  orbit = "ascending",
-  bands = NULL,
-  tiles = NULL,
-  roi = NULL,
-  start_date = NULL,
-  end_date = NULL,
-  platform = NULL,
-  progress = TRUE
-)
-
-\method{sits_cube}{`mpc_cube_sentinel-1-rtc`}(
+\method{sits_cube}{sar_cube}(
   source,
   collection,
   ...,
