+++ inst/extdata/config.yml
@@ -1255,6 +1255,7 @@ sources:
                     SENTINEL-1A: "Sentinel-1A"
                     SENTINEL-1B: "Sentinel-1B"
                 collection_name: "sentinel-1-grd"
+                sar_cube: true
                 open_data: true
                 open_data_token: false
                 metadata_search: "feature"
@@ -1281,6 +1282,7 @@ sources:
                     SENTINEL-1A: "Sentinel-1A"
                     SENTINEL-1B: "Sentinel-1B"
                 collection_name: "sentinel-1-rtc"
+                sar_cube: true
                 open_data: false
                 open_data_token: false
                 metadata_search: "feature"

[…]
