### 2024-03-26 61a1b15
Improve SAR regularization functions
