### 2024-03-27 541e48a
Improve SAR regularization functions
