### 2024-03-27 51186b1
use .block_regulate_size in user input function to control block size
