+++ NAMESPACE
@@ -16,6 +16,7 @@ S3method(.check_samples,default)
 S3method(.check_samples,sits)
 S3method(.check_samples,tbl_df)
 S3method(.cube_adjust_crs,"mpc_cube_sentinel-1-grd")
+S3method(.cube_adjust_crs,"mpc_cube_sentinel-1-rtc")
 S3method(.cube_adjust_crs,default)
 S3method(.cube_as_sf,default)
 S3method(.cube_as_sf,raster_cube)
@@ -291,6 +292,7 @@ S3method(sits_bands,raster_cube)
 S3method(sits_bands,sits)
 S3method(sits_bands,sits_model)
 S3method(sits_bbox,"mpc_cube_sentinel-1-grd")
+S3method(sits_bbox,"mpc_cube_sentinel-1-rtc")
 S3method(sits_bbox,default)
 S3method(sits_bbox,raster_cube)
 S3method(sits_bbox,sits)
