### 2024-03-27 c75d411
remove unused parameter from .cube_split_chunks_samples
