### 2024-03-28 2aaaa17
remove bbox for GRD cube
