+++ man/sits_bbox.Rd
@@ -4,7 +4,6 @@
 \alias{sits_bbox}
 \alias{sits_bbox.sits}
 \alias{sits_bbox.raster_cube}
-\alias{sits_bbox.mpc_cube_sentinel-1-grd}
 \alias{sits_bbox.tbl_df}
 \alias{sits_bbox.default}
 \title{Get the bounding box of the data}
@@ -14,8 +13,6 @@ sits_bbox(data, crs = "EPSG:4326", as_crs = NULL)
 \method{sits_bbox}{sits}(data, crs = "EPSG:4326", as_crs = NULL)
 
 \method{sits_bbox}{raster_cube}(data, crs = "EPSG:4326", as_crs = NULL)
-
-\method{sits_bbox}{`mpc_cube_sentinel-1-grd`}(data, crs = "EPSG:4326", as_crs = NULL)
 
 \method{sits_bbox}{tbl_df}(data, crs = "EPSG:4326", as_crs = NULL)
 
