### 2024-04-02 914935d
remove multithreads in gdal_warp
