+++ R/sits_classify.R
@@ -27,6 +27,7 @@
 #'                           ("lon_min", "lat_min", "lon_max", "lat_max").
 #' @param  filter_fn         Smoothing filter to be applied - optional
 #'                           (clousure containing object of class "function").
+#' @param  impute_fn         Imputation function to remove NA.
 #' @param  start_date        Start date for the classification
 #'                           (Date in YYYY-MM-DD format).
 #' @param  end_date          End date for the classification
@@ -167,6 +168,7 @@ sits_classify.sits <- function(data,
                                ml_model,
                                ...,
                                filter_fn = NULL,
+                               impute_fn = impute_linear(),
                                multicores = 2L,
                                gpu_memory = 16,
                                progress = TRUE) {
@@ -184,6 +186,7 @@ sits_classify.sits <- function(data,
         samples = data,
         ml_model = ml_model,
         filter_fn = filter_fn,
+        impute_fn = impute_fn,
         multicores = multicores,
         gpu_memory = gpu_memory,
         progress = progress
@@ -197,6 +200,7 @@ sits_classify.raster_cube <- function(data,
                                       ml_model, ...,
                                       roi = NULL,
                                       filter_fn = NULL,
+                                      impute_fn = impute_linear(),
                                       start_date = NULL,
                                       end_date = NULL,
                                       memsize = 8L,
@@ -315,6 +319,7 @@ sits_classify.raster_cube <- function(data,
             block = block,
             roi = roi,
             filter_fn = filter_fn,
+            impute_fn = impute_fn,
             output_dir = output_dir,
             version = version,
             verbose = verbose,
@@ -361,6 +366,7 @@ sits_classify.segs_cube <- function(data,
                                     ml_model, ...,
                                     roi = NULL,
                                     filter_fn = NULL,
+                                    impute_fn = impute_linear(),
                                     start_date = NULL,
                                     end_date = NULL,
                                     memsize = 8L,
@@ -465,6 +471,7 @@ sits_classify.segs_cube <- function(data,
             block = block,
             roi = roi,
             filter_fn = filter_fn,
+            impute_fn = impute_fn,
             n_sam_pol = n_sam_pol,
             multicores = multicores,
             memsize = memsize,
