+++ R/api_tibble.R
@@ -271,7 +271,7 @@
     .check_set_caller(".tibble_bands_check")
     # check the bands are available
     sp_bands <- sits_bands(samples)
-    if (purrr::is_null(bands)) {
+    if (.has_not(bands)) {
         bands <- toupper(sp_bands)
     } else {
         bands <- toupper(bands)
