+++ R/api_radd.R
@@ -1,0 +1,7 @@
+.pdf_fun <- function(dist_name) {
+    switch(
+        dist_name,
+        "gaussian" = dnorm,
+        "weibull" = dweibull
+    )
+}
