+++ tests/testthat/test-internals.R
@@ -1,5 +1,5 @@
 test_that("Period functions check", {
-    expect_error(.period_check("P16"))
+    expect_error(.check_period("P16"))
     expect_equal(object = .period_val("P16D"), expected = 16)
     expect_equal(object = .period_val("P2M"), expected = 2)
     expect_equal(object = .period_val("P1Y"), expected = 1)
