### 2024-04-22 3b52d8f
fix empty blocks bug in .raster_merge_blocks
