### 2024-04-22 c7d5db7
add support to new deafrica products
