+++ R/api_source_cdse.R
@@ -1,0 +1,634 @@
+
+# ---- general utilities ----
+#' @title Extract item type of a given `collection`.
+#' @keywords internal
+#' @noRd
+#'
+#' @description
+#'  This function reads the item type of a `collection` and prepares it for use
+#'  in the CDSE Source API's internal methods. This is required as for some
+#'  `collection`, CDSE returns multiple kinds of items.
+#' @param source     Data source.
+#' @param collection Image collection.
+#' @return           item type of a given `collection`.
+.cdse_item_type <- function(source, collection) {
+    item_type <- .conf(
+        "sources",
+        source,
+        "collections",
+        collection,
+        "item_type"
+    )
+    class(item_type) <- c("character", item_type)
+    item_type
+}
+
+# ---- stac utilities ----
+#' @title Extract datetime from a STAC Query.
+#' @keywords internal
+#' @noRd
+#'
+#' @param stac_query Query that follows the STAC protocol.
+#' @return           List with `start_date` and `end_date` properties.
+.stac_cdse_datetime_as_dates <- function(stac_query) {
+    query_datetime <- stringr::str_split(
+        stac_query[["params"]][["datetime"]], "/"
+    )
+    list(
+        start_date = query_datetime[[1]][1],
+        end_date = query_datetime[[1]][2]
+    )
+}
+
+#' @title Extract bounding box from a STAC Query.
+#' @keywords internal
+#' @noRd
+#'
+#' @param stac_query Query that follows the STAC protocol.
+#' @return           List with `bbox` property.
+.stac_cdse_intersects_as_bbox <- function(stac_query) {
+    result <- list(bbox = NULL)
+    # Extract spatial reference from STAC object
+    intersects <- stac_query[["params"]][["intersects"]]
+    coordinates <- intersects[["coordinates"]]
+    # Check if query is valid
+    if (is.null(coordinates)) {
+        return(result)
+    }
+    # Extract x-coordinates and y-coordinates
+    coordinates_x <- coordinates[,,1]
+    coordinates_y <- coordinates[,,2]
+    # Calculate bounding box
+    min_x <- min(coordinates_x)
+    max_x <- max(coordinates_x)
+    min_y <- min(coordinates_y)
+    max_y <- max(coordinates_y)
+    # Create bbox object
+    result[["bbox"]] <- c(min_x, min_y, max_x, max_y)
+    result
+}
+
+#' @title Fix STAC Items from CDSE with assets metadata.
+#' @keywords internal
+#' @noRd
+#'
+#' @description
+#'  This auxiliary function creates `STAC Assets` for each of the bands selected
+#'  by user. This is necessary since CDSE STAC does not return band files as
+#'  `STAC Assets` The information is extracted from files available in the
+#'  CDSE S3 API.
+#' @param source     Data source.
+#' @param items      STAC items.
+#' @param bands      Band names.
+#' @param collection Image collection.
+#' @return           STAC Items updated with `assets` property.
+.stac_cdse_fix_items <- function(source, items, bands, collection, multicores) {
+    # Start parallel workers
+    .parallel_start(workers = multicores)
+    on.exit(.parallel_stop(), add = TRUE)
+    # Define path used to extract the product prefix in CDSE S3
+    s3_path <- c("properties", "productIdentifier")
+    # Define name of the CDSE products bucket
+    s3_bucket <- "/eodata/"
+    # Define protocol to access S3 data
+    s3_protocol <- "/vsis3"
+    # CDSE does not provide files directly in the `assets` property`. As a
+    # workaround, we build the `assets` property using files in the CDSE S3 API
+    features_fixed <- .parallel_map(items[["features"]], function(feature) {
[…]
