+++ NAMESPACE
@@ -126,6 +126,7 @@ S3method(.reg_s2tile_convert,grd_cube)
 S3method(.reg_s2tile_convert,rtc_cube)
 S3method(.slice_dfr,numeric)
 S3method(.source_collection_access_test,"mpc_cube_sentinel-1-grd")
+S3method(.source_collection_access_test,cdse_cube)
 S3method(.source_collection_access_test,mpc_cube)
 S3method(.source_collection_access_test,stac_cube)
 S3method(.source_collection_access_test,usgs_cube)
@@ -133,13 +134,17 @@ S3method(.source_cube,stac_cube)
 S3method(.source_filter_tiles,"mpc_cube_sentinel-1-grd")
 S3method(.source_filter_tiles,stac_cube)
 S3method(.source_item_get_bands,stac_cube)
+S3method(.source_item_get_cloud_cover,cdse_cube)
 S3method(.source_item_get_cloud_cover,sdc_cube)
 S3method(.source_item_get_cloud_cover,stac_cube)
+S3method(.source_item_get_date,cdse_cube)
+S3method(.source_item_get_date,deafrica_cube)
 S3method(.source_item_get_date,stac_cube)
 S3method(.source_item_get_hrefs,bdc_cube)
 S3method(.source_item_get_hrefs,sdc_cube)
 S3method(.source_item_get_hrefs,stac_cube)
 S3method(.source_item_get_hrefs,usgs_cube)
+S3method(.source_items_bands_select,cdse_cube)
 S3method(.source_items_bands_select,stac_cube)
 S3method(.source_items_cube,stac_cube)
 S3method(.source_items_fid,stac_cube)
@@ -150,25 +155,31 @@ S3method(.source_items_new,"mpc_cube_sentinel-1-rtc")
 S3method(.source_items_new,"mpc_cube_sentinel-2-l2a")
 S3method(.source_items_new,aws_cube)
 S3method(.source_items_new,bdc_cube)
+S3method(.source_items_new,cdse_cube)
 S3method(.source_items_new,deafrica_cube)
 S3method(.source_items_new,deafrica_cube_s2_l2a)
 S3method(.source_items_new,hls_cube)
 S3method(.source_items_new,sdc_cube)
 S3method(.source_items_new,usgs_cube)
 S3method(.source_items_tile,"aws_cube_landsat-c2-l2")
+S3method(.source_items_tile,"cdse_cube_sentinel-1-rtc")
 S3method(.source_items_tile,"mpc_cube_landsat-c2-l2")
 S3method(.source_items_tile,"mpc_cube_sentinel-1-grd")
 S3method(.source_items_tile,"mpc_cube_sentinel-1-rtc")
 S3method(.source_items_tile,"mpc_cube_sentinel-2-l2a")
 S3method(.source_items_tile,aws_cube)
 S3method(.source_items_tile,bdc_cube)
+S3method(.source_items_tile,cdse_cube)
 S3method(.source_items_tile,deafrica_cube)
+S3method(.source_items_tile,deafrica_cube_rainfall_chirps_daily)
+S3method(.source_items_tile,deafrica_cube_rainfall_chirps_monthly)
 S3method(.source_items_tile,hls_cube)
 S3method(.source_items_tile,sdc_cube)
 S3method(.source_items_tile,usgs_cube)
 S3method(.source_roi_tiles,"mpc_cube_landsat-c2-l2")
 S3method(.source_roi_tiles,sdc_cube)
 S3method(.source_roi_tiles,stac_cube)
+S3method(.source_tile_get_bbox,"cdse_cube_sentinel-1-rtc")
[…]
