+++ DESCRIPTION
@@ -64,6 +64,7 @@ Imports:
     torch (>= 0.11.0),
     utils
 Suggests:
+    aws.s3,
     caret,
     cli,
     covr,
@@ -166,6 +167,7 @@ Collate:
     'api_source.R'
     'api_source_aws.R'
     'api_source_bdc.R'
+    'api_source_cdse.R'
     'api_source_deafrica.R'
     'api_source_hls.R'
     'api_source_local.R'
