### 2024-04-23 fc3de2b
add environment api to enable cdse tests
