+++ DESCRIPTION
@@ -131,6 +131,7 @@ Collate:
     'api_data.R'
     'api_debug.R'
     'api_download.R'
+    'api_environment.R'
     'api_factory.R'
     'api_file_info.R'
     'api_file.R'
