+++ NAMESPACE
@@ -131,6 +131,7 @@ S3method(.source_collection_access_test,mpc_cube)
 S3method(.source_collection_access_test,stac_cube)
 S3method(.source_collection_access_test,usgs_cube)
 S3method(.source_cube,stac_cube)
+S3method(.source_filter_tiles,"cdse_cube_sentinel-1-rtc")
 S3method(.source_filter_tiles,"mpc_cube_sentinel-1-grd")
 S3method(.source_filter_tiles,stac_cube)
 S3method(.source_item_get_bands,stac_cube)
