### 2024-04-24 3a0f8e9
add cdse examples
