+++ R/api_source_cdse.R
@@ -560,7 +560,7 @@
 #' @export
 .source_items_bands_select.cdse_cube <- function(source, ...,
                                                  items, bands,
-                                                 collection, multicores) {
+                                                 collection, multicores = 2L) {
     # CDSE does not provide files in the `assets` property. So, it is
     # required to fix this using content from CDSE S3 API.
     items <- .stac_cdse_fix_items(source, items, bands, collection, multicores)
