### 2024-04-24 f20f76f
update cdse examples
