+++ inst/extdata/sources/config_source_deafrica.yml
@@ -254,7 +254,7 @@ sources:
                     AWS_NO_SIGN_REQUEST: true
                 open_data: true
                 open_data_token: false
-                metadata_search : "feature"
+                metadata_search : "tile"
                 ext_tolerance: 0
                 grid_system  : "WRS-2"
             LS9_SR          :
@@ -407,18 +407,18 @@ sources:
                         maximum_value : 10000
                         scale_factor  : 0.0001
                         offset_value  : 0
-                        resolution   :  60
+                        resolution    :  60
                         band_name     : "B01"
-                        data_type    : "INT2S"
+                        data_type     : "INT2S"
                     B02     : &deafrica_msi_10m
                         missing_value : -9999
                         minimum_value : 0
                         maximum_value : 10000
                         scale_factor  : 0.0001
                         offset_value  : 0
-                        resolution   :  10
+                        resolution    :  10
                         band_name     : "B02"
-                        data_type    : "INT2S"
+                        data_type     : "INT2S"
                     B03     :
                         <<: *deafrica_msi_10m
                         band_name     : "B03"
@@ -431,9 +431,9 @@ sources:
                         maximum_value : 10000
                         scale_factor  : 0.0001
                         offset_value  : 0
-                        resolution   :  20
+                        resolution    :  20
                         band_name     : "B05"
-                        data_type    : "INT2S"
+                        data_type     : "INT2S"
                     B06     :
                         <<: *deafrica_msi_20m
                         band_name     : "B06"
@@ -485,4 +485,32 @@ sources:
                 open_data_token: false
                 metadata_search : "tile"
                 ext_tolerance: 0
-                grid_system  : "MGRS"
+                grid_system : "MGRS"
+            SENTINEL-1-RTC  :
+                bands       :
+                    VV      : &deafrica_rtc_10m
+                        missing_value : -32768.0
+                        minimum_value : 0
+                        maximum_value : 65534
+                        scale_factor  : 1
+                        offset_value  : 0
+                        resolution    : 10
+                        band_name     : "VV"
+                        data_type     : "FLT4S"
+                    VH      :
+                        <<: *deafrica_rtc_10m
+                        band_name     : "VH"
+                satellite   : "SENTINEL-1"
+                sensor      : "RTC"
+                orbits      : ["ascending", "descending"]
+                access_vars :
+                    AWS_DEFAULT_REGION  : "af-south-1"
+                    AWS_S3_ENDPOINT     : "s3.af-south-1.amazonaws.com"
+                    AWS_NO_SIGN_REQUEST : true
+                collection_name : "s1_rtc"
+                sar_cube        : true
+                open_data       : true
+                open_data_token : false
+                metadata_search : "tile"
+                ext_tolerance   : 0
+                grid_system     : "MGRS"
