### 2024-04-24 75b0a45
update rtc resolution
