### 2024-04-25 7ce79ad
new error messages and lintr corrections
