+++ R/api_check.R
@@ -198,10 +198,10 @@
         if (is.null(msg))
             msg <- .conf("messages", caller)
         # include local message if available
-        if (!(is.null(local_msg)))
-            msg <- paste0(caller,": ", local_msg, " - ", msg)
+        if (is.null(local_msg))
+            msg <- paste0(caller, ": ", msg)
         else
-            msg <- paste0(caller,": ", msg)
+            msg <- paste0(caller, ": ", local_msg, " - ", msg)
         # process message
         stop(msg, call. = FALSE)
     }
@@ -226,7 +226,7 @@
 .check_na <- function(x, ..., allow_na = FALSE, local_msg = NULL, msg = NULL) {
     if (!allow_na) {
         .check_that(
-            !any(is.na(x)),
+            !anyNA(x),
             local_msg = local_msg,
             msg = msg
         )
@@ -248,22 +248,22 @@
     }
     if (is_named) {
         .check_that(
-            !(is.null(names(x))) && !any(is.na(names(x))),
+            .has(names(x)) && !anyNA(names(x)),
             local_msg = local_msg,
-            msg = .conf("messages", ".check_names_is_named"),
+            msg = .conf("messages", ".check_names_is_named")
         )
         if (is_unique) {
             .check_that(
                 length(names(x)) == length(unique(names(x))),
                 local_msg = local_msg,
-                msg = .conf("messages", ".check_names_is_unnamed")
+                msg = .conf("messages", ".check_names_unique" )
             )
         }
     } else {
         .check_that(
             is.null(names(x)),
             local_msg = local_msg,
-            msg = .conf("messages", ".check_names_unique")
+            msg = .conf("messages", ".check_names_is_unnamed")
         )
     }
     return(invisible(x))
@@ -313,12 +313,12 @@
 #' (if \code{is_integer=TRUE}).
 #' }
 #' \item{
-#' \code{.check_chr_type()} checks for \code{character} type and empty strings (if
+#' \code{.check_chr_type()} checks for \code{character} and empty strings (if
 #' \code{allow_empty=FALSE}). It also checks strings through regular
 #' expression (if \code{regex} parameter is defined).
 #' }
 #' \item{
-#' \code{.check_lst_type()} checks for \code{list} type. By default, it checks if
+#' \code{.check_lst_type()} checks for \code{list} type. By default, checks if
 #' the list is named. Additionally, a function can be passed to
 #' \code{fn_check} parameter to check its elements. This enables to pass
 #' other checking functions like \code{.check_num()} to verify the type of
@@ -587,7 +587,6 @@
     .check_null(x, local_msg = local_msg, msg = msg)
     # check type
     .check_chr_type(x, local_msg = local_msg, msg = msg)
-    browser
     # check length
     .check_length(x, len_min = len_min, len_max = len_max,
                   local_msg = local_msg, msg = msg)
@@ -606,7 +605,7 @@
     # check duplicate
     if (!allow_duplicate) {
         .check_that(
-            all(!duplicated(x)),
+            anyDuplicated(x) == 0,
             local_msg = local_msg,
             msg = msg
         )
@@ -979,7 +978,7 @@
             expr
         },
         error = function(e) {
-            warning(e$message, call. = FALSE)
+            warning(e[["message"]], call. = FALSE)
         }
     )
     return(invisible(result))
@@ -1003,7 +1002,7 @@
             expr
         },
         error = function(e) {
-            .check_that(FALSE, local_msg = e$message, msg = msg)
+            .check_that(FALSE, local_msg = e[["message"]], msg = msg)
         }
     )
[…]
