+++ demo/evaluate_samples_Kohonen.R
@@ -4,12 +4,9 @@ library(sits)
 
 # load the sitsdata library
 if (!requireNamespace("sitsdata", quietly = TRUE)) {
-    stop(
-        paste0(
-            "Please install package sitsdata\n",
-            "Please call devtools::install_github('e-sensing/sitsdata')"
-        ),
-        call. = FALSE
+    stop("Please install package sitsdata\n",
+         "Please call devtools::install_github('e-sensing/sitsdata')",
+         call. = FALSE
     )
 }
 library(sitsdata)

[…]
