### 2024-04-25 a38e8e7
add cpp fns to calc probability of non forest
