+++ src/RcppExports.cpp
@@ -260,6 +260,32 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// C_dnorm
+arma::mat C_dnorm(const arma::mat& mtx, const double mean, const double std);
+RcppExport SEXP _sits_C_dnorm(SEXP mtxSEXP, SEXP meanSEXP, SEXP stdSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type mtx(mtxSEXP);
+    Rcpp::traits::input_parameter< const double >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< const double >::type std(stdSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_dnorm(mtx, mean, std));
+    return rcpp_result_gen;
+END_RCPP
+}
+// C_radd_calc_nf
+arma::vec C_radd_calc_nf(const arma::mat& ts, const arma::mat& mean, const arma::mat& std);
+RcppExport SEXP _sits_C_radd_calc_nf(SEXP tsSEXP, SEXP meanSEXP, SEXP stdSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type ts(tsSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type std(stdSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_radd_calc_nf(ts, mean, std));
+    return rcpp_result_gen;
+END_RCPP
+}
 // C_temp_max
 arma::vec C_temp_max(const arma::mat& mtx);
 RcppExport SEXP _sits_C_temp_max(SEXP mtxSEXP) {
@@ -609,6 +635,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_sits_C_nnls_solver_batch", (DL_FUNC) &_sits_C_nnls_solver_batch, 5},
     {"_sits_C_normalize_data", (DL_FUNC) &_sits_C_normalize_data, 3},
     {"_sits_C_normalize_data_0", (DL_FUNC) &_sits_C_normalize_data_0, 3},
+    {"_sits_C_dnorm", (DL_FUNC) &_sits_C_dnorm, 3},
+    {"_sits_C_radd_calc_nf", (DL_FUNC) &_sits_C_radd_calc_nf, 3},
     {"_sits_C_temp_max", (DL_FUNC) &_sits_C_temp_max, 1},
     {"_sits_C_temp_min", (DL_FUNC) &_sits_C_temp_min, 1},
     {"_sits_C_temp_mean", (DL_FUNC) &_sits_C_temp_mean, 1},
