### 2024-04-25 b8a2a5f
hls configuration
