### 2024-04-25 9efac95
add resume operation in sits_uncertainty for polygons
