### 2024-04-26 f9464f6
add n_tries parameter in sits_cube_copy
