+++ man/sits_cube_copy.Rd
@@ -8,9 +8,11 @@ sits_cube_copy(
   cube,
   roi = NULL,
   res = NULL,
+  n_tries = 3,
   multicores = 2L,
   output_dir,
-  progress = TRUE
+  progress = TRUE,
+  ...
 )
 }
 \arguments{
@@ -26,6 +28,8 @@ named lat/long values
 \item{res}{An integer value corresponds to the output
 spatial resolution of the images. Default is NULL.}
 
+\item{n_tries}{Number of tries to download the same image. Default is 3.}
+
 \item{multicores}{Number of cores for parallel downloading
 (integer, min = 1, max = 2048).}
 
@@ -33,6 +37,8 @@ spatial resolution of the images. Default is NULL.}
 (character vector of length 1).}
 
 \item{progress}{Logical: show progress bar?}
+
+\item{...}{Additional parameters to httr package.}
 }
 \value{
 Copy of input data cube (class "raster cube").
