### 2024-04-26 #596 @gilbertocamara
Dear @mpadge Apologies for the long delay in responding. First of all, kudos to ROpenSci for your work! In the last two months, we have been working on release 1.5.0 of the `sits` package which is due April 30th. In this release, we have tried to incorporate in `sits` the guidelines for Statistical Software proposed by ROpenSci in connection with the `srr` package. In particular, we took the guidelines associated with machine learning, spatial and time series packages.

A key point here is that arguably `sits` is currently unique in the R package landscape. It provides an end-to-end environment for ML/DL analysis of big Earth observation. We are not aware of any similar package. Thus, many recommendations that would apply for packages that implement improved versions of existing ML algorithms do not apply. A second point is that there is a full book on `sits` available in https://e-sensing.github.io/sitsbook/, which allows users to perform extended tests and experiments with medium-sized datasets that cannot be loaded in CRAN. We also provide large data sets in github to serve as basis for user experiments.

The `sits` package is supported by a set of packages available in the github repository (https://github.com/e-sensing) which include `sitsdata` (medium-sized data sets used in the book) and `rondonia20LMR` a data set of 28 GB to test ML methods for image time series in a big data context.

Overall, we found the guidelines to be very useful. To avoid lengthy issues, I will post my thoughts on the SRR guidelines in a set of comments below
