### 2024-04-26 #596 @gilbertocamara
Dear @mpadge Some comments on the SRR Generic Guidelines.

Generic guidelines are quite good points, especially regarding documentation and error messages associated with variable checking. They encouraged us to:

(a) improve documentation of internal functions (G1.4a);
(b) provide a CONTRIBUTING.md statement (G1.2);
(c) include assertions on all input parameters (G2.0);
(d) handle missing data and provide explicit imputation function (G2.14);
(e) include specific messages for each different error, including indicating parameter names (G5.2);
(f) include tests that support a 94% code coverage (G5.4);
(g) include edge-condition tests and associated messages (G5.8);

We could not fully understand the scope of G5.6 (parameter recovery) and G5.9 (noise susceptibility tests) so we consider that they do not apply to `sits.

We missed explicit support in the SRR Guidelines regarding the `tidyverse`. While we understand that there is resistance to `tidyverse` in certain quarters, in software engennering terms the `tidyverse` is much better than tradional R *apply methods for data handling. We could not have developed a reliable and efficient package without the `tidyverse`.
