### 2024-04-26 5eb655f
rename deafrica products
