### 2024-04-26 25202ee
add deafrica product tests
