+++ tests/testthat/test-cube.R
@@ -96,6 +96,7 @@ test_that("Reading a raster cube", {
     expect_true(params_2$ncols == 255)
     expect_true(params_2$xres >= 231.5)
 })
+
 test_that("Creating cubes from BDC - CBERS-WFI-16D", {
     tiles <- c("007004", "007005")
     start_date <- "2021-09-01"
@@ -442,13 +443,114 @@ test_that("Creating cubes from BDC - SENTINEL-2 - tile", {
     cube_nrows <- .tile_nrows(bdc_s2_cube_t)
     expect_true(.raster_nrows(r_obj) == cube_nrows)
 })
+
+test_that("Creating LS5-SR cubes from DEA", {
+    landsat_cube <-  sits_cube(
+        source = "DEAFRICA",
+        collection = "LS5-SR",
+        bands = c("B05", "CLOUD"),
+        roi   = c(
+            lon_min = 33.546,
+            lon_max = 34.999,
+            lat_min = 1.427,
+            lat_max = 3.726
+        ),
+        start_date = "2009-01-01",
+        end_date = "2010-01-01",
+        progress = FALSE
+    )
+    expect_true(all(sits_bands(landsat_cube) %in% c("B05", "CLOUD")))
+    expect_equal(nrow(landsat_cube), 6)
+    bbox_cube <- sits_bbox(landsat_cube, as_crs = "EPSG:4326")
+    bbox_cube_1 <- sits_bbox(.tile(landsat_cube), as_crs = "EPSG:4326")
+    expect_true(bbox_cube["xmax"] >= bbox_cube_1["xmax"])
+    expect_true(bbox_cube["ymax"] >= bbox_cube_1["ymax"])
+    r_obj <- .raster_open_rast(landsat_cube$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(landsat_cube)
+    expect_true(.raster_nrows(r_obj) == cube_nrows)
+})
+test_that("Creating LS7-SR cubes from DEA", {
+    landsat_cube <-  sits_cube(
+        source = "DEAFRICA",
+        collection = "LS7-SR",
+        bands = c("B05", "CLOUD"),
+        roi   = c(
+            lon_min = 33.546,
+            lon_max = 34.999,
+            lat_min = 1.427,
+            lat_max = 3.726
+        ),
+        start_date = "2009-01-01",
+        end_date = "2010-01-01",
+        progress = FALSE
+    )
+    expect_true(all(sits_bands(landsat_cube) %in% c("B05", "CLOUD")))
+    expect_equal(nrow(landsat_cube), 6)
+    bbox_cube <- sits_bbox(landsat_cube, as_crs = "EPSG:4326")
+    bbox_cube_1 <- sits_bbox(.tile(landsat_cube), as_crs = "EPSG:4326")
+    expect_true(bbox_cube["xmax"] >= bbox_cube_1["xmax"])
+    expect_true(bbox_cube["ymax"] >= bbox_cube_1["ymax"])
+    r_obj <- .raster_open_rast(landsat_cube$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(landsat_cube)
+    expect_true(.raster_nrows(r_obj) == cube_nrows)
+})
+test_that("Creating LS8-SR cubes from DEA", {
+    landsat_cube <-  sits_cube(
+        source = "DEAFRICA",
+        collection = "LS8-SR",
+        bands = c("B05", "CLOUD"),
+        roi   = c(
+            lon_min = 33.546,
+            lon_max = 34.999,
+            lat_min = 1.427,
+            lat_max = 3.726
+        ),
[…]
