### 2024-04-26 #596 @gilbertocamara
Dear @mpadge comments on SRR Guidelines on Machine Learning

From the perspective of the `sits` package and Earth observation in general, the first part of the ML guidelines (ML1.0 to ML1.5) seems to have as an excessive focus on the differentiation between training and test data. In the case of EO data, R packages have to deal with CRAN limitations on example data sets. We tried to overcome this limitation by providing a specific chapter in the on-line book (https://e-sensing.github.io/sitsbook/validation-and-accuracy-measurements.html) and with additional packages with are available in github, as explained above.

Guidelines ML1.6 to ML1.8 deal with missing values. They were useful for us as reminders, taking in account that missing values in EO data arise in a different context than tabular data.

We consider guideline ML2.0 quite important. In `sits`, we actually developed a single interface which encapsulates different models using closures. In fact, the guideline ML2.0 is better elaborated in ML4.0 and later in ML5.0. Could one consider merging them?

We also agree with ML2.2, although we have not yet implemented it. That said, we did not understand the difference between guidelines ML2.2, ML2.3, and ML2.4, ML2.5, and ML2.6. Perhaps they could be grouped together for brevity's sake. We did not understand the context of ML3.0 and its subpoints. We fail to see in which context such separation between specification and training might be useful. This is probably due to the specific nature of EO data analysis.

As for item ML3.4 and subitems, we consider that requiring developers to provide functions for tuning hyperparameters might be a better approach, especially with deep learning.

Although we understand the rationale for ML6.0, we advise against using training and test data for model assessment in the case of EO data. The community has developed a specific set of best practices of quality assessment. See Olofsson et al., (2014) <doi:10.1016/j.rse.2014.02.015>.

In short, the specific case of applying ML/DL for EO data has issues which are impossible to cover in the scope of generic guidelines as provided in SSR for ML.
