### 2024-04-27 #596 @gilbertocamara
Dear @mpadge Comments on SRR Guidelines on Spatial and Time Series.

The SSR Spatial Guidelines are very good and generally applicable. The emphasis on the `sf` package (SP2.1) is welcome. However, we missed guidelines on handling raster data.  In our work, we found that `terra` to be better and easier to use than `stars`. In any case, SRR should discourage the use of `raster` as these packages have superseded it. We also missed guidelines regarding visualisation of vector and raster data. For your reference, we found that `tmap`, `leaflet` and `leafem` to be excellent packages. Note that both `tmap` and `leafem` require raster data to be handled by `stars`. In `sits`, we use `stars` for plotting and visualisation, and `terra` for access to raster values.

We suggest the inclusion of a guideline regarding the installation of GDAL and PROJ, following the instructions associated with the `sf` package. See more at https://r-spatial.github.io/sf/#installing. We also suggest that you consider mentioning the desirability of combining `sf` with the `tidyverse`. As acknowledged by Edzer Pebesma, the design of `sf`has been influenced by the `tidyverse` to the extent that some functions for `tidyverse` can be applied to the output of `sf` ones. Thus, arguably `sf` users will find it easier to combine it with the `tidyverse`.

As for time series, we fully agree with guidelines TS1.0 to T2.1c and have implemented them in `sits`. As for guidelines TS2.2 to TS2.4b, we considered they do not apply, since in general time series derived from satellite data are not stationary. Also, since satellite image series analysis is about classification and prediction rather than forecasting, we considered that guidelines TS3.0 to TS4.7c do not apply for `sits`.
