### 2024-04-27 #596 @gilbertocamara
Dear @mpadge Final considerations and a plea for support.

Overall, the SRR Guidelines deserve high praise. The ROpenSci team has provided an excellent service to the community by working hard to develop them. While the guidelines are aimed at small, focused R packages, they are also relevant to larger packages such as `sits`.
