+++ inst/extdata/srr/srr-stats-standards.R
@@ -1,0 +1,280 @@
+#' srr_stats
+#'
+#' All of the following standards initially have `@srrstatsTODO` tags.
+#' These may be moved at any time to any other locations in your code.
+#' Once addressed, please modify the tag from `@srrstatsTODO` to `@srrstats`,
+#' or `@srrstatsNA`, ensuring that references to every one of the following
+#' standards remain somewhere within your code.
+#' (These comments may be deleted at any time.)
+#'
+#' @srrstatsVerbose TRUE
+#'
+#' @srrstats {G1.0} *Described in Simoes et al (2021) <doi:10.3390/rs13132428>.*
+#' @srrstats {G1.1} *Full description of package available in the [sits book](https://e-sensing.github.io/sitsbook/)*
+#' @srrstats {G1.2} *Life Cycle Statement available in CONTRIBUTING.md file*.
+#' @srrstats {G1.3} *See [sits book](https://e-sensing.github.io/sitsbook/)*
+#' @srrstats {G1.4} *All functions are documented in [`roxygen2`](https://roxygen2.r-lib.org/)*
+#' @srrstats {G1.4a} *All relevant internal (non-exported) functions are also be documented in standard [`roxygen2`](https://roxygen2.r-lib.org/) format*
+#' @srrstats {G1.5} *Software includes all code necessary to reproduce results which form the basis of performance claims made in associated publications.*
+#' @srrstatsNA {G1.6} *Software should include code necessary to compare performance claims with alternative implementations in other R packages.*
+#' @srrstats {G2.0} *The package implements assertions on all input parameters.*
+#' @srrstats {G2.0a} *Provide explicit secondary documentation of any expectations on lengths of inputs*
+#' @srrstats {G2.1} *Implement assertions on types of inputs (see the initial point on nomenclature above)*
+#' @srrstats {G2.1a} *Provide explicit secondary documentation of expectations on data types of all vector inputs.*
+#' @srrstats {G2.2} *Appropriately prohibit or restrict submission of multivariate input to parameters expected to be univariate.*
+#' @srrstats {G2.3} *For univariate character input:*
+#' @srrstats {G2.3a} *Use `match.arg()` or equivalent where applicable to only permit expected values.*
+#' @srrstats {G2.3b} *Either: use `tolower()` or equivalent to ensure input of character parameters is not case dependent; or explicitly document that parameters are strictly case-sensitive.*
+#' @srrstats {G2.4} *Provide appropriate mechanisms to convert between different data types, potentially including:*
+#' @srrstats {G2.4a} *explicit conversion to `integer` via `as.integer()`*
+#' @srrstats {G2.4b} *explicit conversion to continuous via `as.numeric()`*
+#' @srrstats {G2.4c} *explicit conversion to character via `as.character()` (and not `paste` or `paste0`)*
+#' @srrstatsNA {G2.4d} *explicit conversion to factor via `as.factor()`*
+#' @srrstatsNA {G2.4e} *explicit conversion from factor via `as...()` functions*
+#' @srrstatsNA {G2.5} *Where inputs are expected to be of `factor` type, secondary documentation should explicitly state whether these should be `ordered` or not, and those inputs should provide appropriate error or other routines to ensure inputs follow these expectations.*
+#' @srrstatsNA {G2.6} *Software which accepts one-dimensional input should ensure values are appropriately pre-processed regardless of class structures.*
+#' @srrstats {G2.7} *Software should accept as input as many of the above standard tabular forms as possible, including extension to domain-specific forms.*
+#' @srrstats {G2.8} *Software should provide appropriate conversion or dispatch routines as part of initial pre-processing to ensure that all other sub-functions of a package receive inputs of a single defined class or type.*
[…]
