### 2024-04-27 a4a6576
improved documentation of internal functions
