+++ R/api_data.R
@@ -57,7 +57,7 @@
     if (.raster_nrows(rast) == block[["nrows"]] &&
             .raster_ncols(rast) == block[["ncols"]]) {
         # split samples by bands and tile
-        ts_tbl <- .get_data_by_tile(
+        ts_tbl <- .data_by_tile(
             cube = cube,
             samples = samples,
             bands = bands,
@@ -68,7 +68,7 @@
         )
     } else {
         # get data by chunks
-        ts_tbl <- .get_data_by_chunks(
+        ts_tbl <- .data_by_chunks(
             cube = cube,
             samples = samples,
             bands = bands,
@@ -330,28 +330,26 @@
     return(data_avg)
 }
 
-.ts_is_valid <- function(filename) {
-    if (!file.exists(filename)) {
-        return(FALSE)
-    }
-    ts <- tryCatch({
-        ts <- readRDS(filename)
-        return(TRUE)
-    }, error = function(e) {
-        unlink(filename)
-        gc()
-        return(FALSE)
-    })
-    return(ts)
-}
-
-.get_data_by_tile <- function(cube,
-                              samples,
-                              bands,
-                              impute_fn,
-                              cld_band,
-                              multicores,
-                              progress) {
+#' @title get time series from data cubes on tile by tile bassis
+#' @name .data_by_tile
+#' @keywords internal
+#' @noRd
+#' @param cube            Data cube from where data is to be retrieved.
+#' @param samples         Samples to be retrieved.
+#' @param bands           Bands to be retrieved (optional).
+#' @param impute_fn       Imputation function to remove NA.
+#' @param cld_band        Cloud band
+#' @param multicores      Number of threads to process the time series.
+#' @param progress        A logical value indicating if a progress bar
+#'                        should be shown.
+.data_by_tile <- function(cube,
+                          samples,
+                          bands,
+                          impute_fn,
+                          cld_band,
+                          multicores,
+                          progress) {
+    .check_set_caller(".data_by_tile")
     # Get cube timeline
     tl <- sits_timeline(cube)
     # Get tile-band combination
@@ -379,16 +377,24 @@
             tiles = tile_id
         )
         hash_bundle <- digest::digest(list(tile, samples), algo = "md5")
-        # Create a file to store the samples
+        # File to store the samples
         filename <- .file_path(
             "samples", hash_bundle,
             ext = ".rds",
             output_dir = output_dir
         )
         # Does the file exist?
-        if (.ts_is_valid(filename)) {
-            ts <- readRDS(filename)
-            return(ts)
+        if (file.exists(filename)) {
+            tryCatch(
+                { # ensure that the file is not corrupted
+                    timeseries <- readRDS(filename)
+                    return(timeseries)
+                },
+                error = function(e) {
+                    unlink(filename)
+                    gc()
+                }
+            )
         }
         # get XY
         xy_tb <- .proj_from_latlong(
@@ -461,7 +467,7 @@
     # bind rows to get a melted tibble of samples
     ts_tbl <- dplyr::bind_rows(samples_tiles_bands)
     if (!.has_ts(ts_tbl)) {
-        warning(.conf("messages", ".get_data_by_tile"),
+        warning(.conf("messages", ".data_by_tile"),
                 immediate. = TRUE, call. = FALSE
         )
         return(.tibble())
@@ -524,14 +530,25 @@
     }
     return(ts_tbl)
 }
-
-.get_data_by_chunks <- function(cube,
-                                samples,
-                                bands,
-                                impute_fn,
-                                cld_band,
-                                multicores,
-                                progress) {
+#' @title get time series from data cubes using chunks
+#' @name .data_by_tile
+#' @keywords internal
+#' @noRd
+#' @param cube            Data cube from where data is to be retrieved.
+#' @param samples         Samples to be retrieved.
[…]

[…]
