### 2024-04-28 526fb87
include platform search for CDSE products
