### 2024-04-28 b332f22
include platform search for DEAfrica Sentinel 1 and 2 products
