+++ tests/testthat/test-cube.R
@@ -459,6 +459,11 @@ test_that("Creating LS5-SR cubes from DEA", {
         end_date = "2010-01-01",
         progress = FALSE
     )
+
+    testthat::skip_if(purrr::is_null(landsat_cube),
+                      message = "DEAFRICA is not accessible"
+    )
+
     expect_true(all(sits_bands(landsat_cube) %in% c("B05", "CLOUD")))
     expect_equal(nrow(landsat_cube), 6)
     bbox_cube <- sits_bbox(landsat_cube, as_crs = "EPSG:4326")
@@ -484,6 +489,11 @@ test_that("Creating LS7-SR cubes from DEA", {
         end_date = "2010-01-01",
         progress = FALSE
     )
+
+    testthat::skip_if(purrr::is_null(landsat_cube),
+                      message = "DEAFRICA is not accessible"
+    )
+
     expect_true(all(sits_bands(landsat_cube) %in% c("B05", "CLOUD")))
     expect_equal(nrow(landsat_cube), 6)
     bbox_cube <- sits_bbox(landsat_cube, as_crs = "EPSG:4326")
@@ -509,6 +519,11 @@ test_that("Creating LS8-SR cubes from DEA", {
         end_date = "2021-04-29",
         progress = FALSE
     )
+
+    testthat::skip_if(purrr::is_null(landsat_cube),
+                      message = "DEAFRICA is not accessible"
+    )
+
     expect_true(all(sits_bands(landsat_cube) %in% c("B05", "CLOUD")))
     expect_equal(nrow(landsat_cube), 6)
     bbox_cube <- sits_bbox(landsat_cube, as_crs = "EPSG:4326")
@@ -534,6 +549,11 @@ test_that("Creating LS9-SR cubes from DEA", {
         end_date = "2022-04-01",
         progress = FALSE
     )
+
+    testthat::skip_if(purrr::is_null(landsat_cube),
+                      message = "DEAFRICA is not accessible"
+    )
+
     expect_true(all(sits_bands(landsat_cube) %in% c("B05", "CLOUD")))
     expect_equal(nrow(landsat_cube), 6)
     bbox_cube <- sits_bbox(landsat_cube, as_crs = "EPSG:4326")
@@ -565,10 +585,13 @@ test_that("Creating S2 cubes from DEA using ROI", {
         },
         .default = NULL
     )
+
     testthat::skip_if(purrr::is_null(dea_cube),
         message = "DEAFRICA is not accessible"
     )
+
     expect_true(all(sits_bands(dea_cube) %in% c("B01", "B04", "B05")))
+    expect_equal(nrow(dea_cube), 1)
     r <- .raster_open_rast(.tile_path(dea_cube))
     expect_equal(dea_cube$xmax[[1]], .raster_xmax(r), tolerance = 1)
     expect_equal(dea_cube$xmin[[1]], .raster_xmin(r), tolerance = 1)
@@ -582,21 +605,23 @@ test_that("Creating S2 cubes from DEA using tiles", {
                 bands = c("B02", "B8A", "B11"),
                 tiles = c("37MDT","37MET"),
                 start_date = "2019-01-01",
-                end_date = "2019-10-28",
+                end_date = "2019-08-28",
                 progress = FALSE
             )
         },
         .default = NULL
     )
+
     testthat::skip_if(purrr::is_null(dea_cube),
                       message = "DEAFRICA is not accessible"
     )
+
     expect_true(all(sits_bands(dea_cube) %in% c("B02", "B8A", "B11")))
[…]
