+++ tests/testthat/test-cube-cdse.R
@@ -23,7 +23,13 @@ test_that("Creating S2 cubes from CDSE with ROI", {
         },
         .default = NULL
     )
-    testthat::skip_if(purrr::is_null(s2_cube_cdse), "CDSE is not accessible")
+
+    if (purrr::is_null(s2_cube_cdse)) {
+        .environment_rollback(cdse_env_config)
+
+        testthat::skip("CDSE is not accessible")
+    }
+
     expect_true(all(sits_bands(s2_cube_cdse) %in% c("B05", "CLOUD")))
     expect_equal(nrow(s2_cube_cdse), 3)
     bbox_cube <- sits_bbox(s2_cube_cdse, as_crs = "EPSG:4326")
@@ -57,6 +63,13 @@ test_that("Creating S2 cubes from CDSE with tiles", {
         },
         .default = NULL
     )
+
+    if (purrr::is_null(s2_cube)) {
+        .environment_rollback(cdse_env_config)
+
+        testthat::skip("CDSE is not accessible")
+    }
+
     expect_true(all(sits_bands(s2_cube) %in% c("B05", "CLOUD")))
     r <- .raster_open_rast(.tile_path(s2_cube))
     expect_equal(s2_cube$xmax[[1]], .raster_xmax(r), tolerance = 1)
@@ -64,6 +77,7 @@ test_that("Creating S2 cubes from CDSE with tiles", {
     r_obj <- .raster_open_rast(s2_cube$file_info[[1]]$path[1])
     cube_nrows <- .tile_nrows(s2_cube)
     expect_true(.raster_nrows(r_obj) == cube_nrows)
+    expect_true(s2_cube$tile == "20LKP")
     # Rollback environment changes
     .environment_rollback(cdse_env_config)
 })
@@ -74,16 +88,29 @@ test_that("Creating Sentinel-1 RTC cubes from CDSE", {
     # Patch environment variables
     .environment_patch(cdse_env_config)
     # Test
-    cube_s1_rtc <-  sits_cube(
-        source = "CDSE",
-        collection = "SENTINEL-1-RTC",
-        bands = c("VV"),
-        orbit = "descending",
-        tiles = c("36NWH"),
-        start_date = "2021-07-01",
-        end_date = "2021-09-30",
-        multicores = 1L
+    cube_s1_rtc <- .try(
+        {
+            sits_cube(
+                source = "CDSE",
+                collection = "SENTINEL-1-RTC",
+                bands = c("VV"),
+                orbit = "descending",
+                tiles = c("36NWH"),
+                start_date = "2021-07-01",
+                end_date = "2021-09-30",
+                progress = FALSE,
+                multicores = 1L
+            )
+        },
+        .default = NULL
     )
+
+    if (purrr::is_null(cube_s1_rtc)) {
+        .environment_rollback(cdse_env_config)
+
+        testthat::skip("CDSE is not accessible")
+    }
+
     bbox <- sits_bbox(cube_s1_rtc[1,])
     expect_true(grepl("4326", bbox[["crs"]]))
     expect_equal(32, bbox[["xmin"]])
