### 2024-04-28 f417234
define alos-palsar data as sar
