### 2024-04-28 c763bf7
replace warning texts by those in config_messages
