+++ R/api_samples.R
@@ -71,10 +71,9 @@
 #' @param samples Data.frame with samples
 #' @return Time series for the first sample
 .samples_ts <- function(samples) {
+    .check_set_caller(".samples_ts")
     # Check time_series column
-    if (!.has_ts(samples)) {
-        stop("time_series column not found")
-    }
+    .check_that(.has_ts(samples))
     # Return time series of the first sample
     samples[["time_series"]][[1]]
 }
@@ -155,16 +154,15 @@
 #' @param samples Data.frame with samples
 #' @return Samples with the same number of temporal intervals as the first
 .samples_prune <- function(samples) {
+    .check_set_caller(".samples_prune")
     # Get the time series length for the first sample
     ntimes <- .samples_ntimes(samples)
     # Prune time series according to the first time series length and return
-    .samples_foreach_ts(samples, function(ts) {
-        if (nrow(ts) >= ntimes) {
-            ts[seq_len(ntimes), ]
-        } else {
-            stop("time series length is smaller than the first sample")
-        }
+    new_samples <- .samples_foreach_ts(samples, function(ts) {
+        .check_that(nrow(ts) >= ntimes)
+        ts[seq_len(ntimes), ]
     })
+    return(new_samples)
 }
 #' @title Get sample statistics
 #' @noRd
