+++ R/api_source_cdse.R
@@ -23,50 +23,9 @@
     item_type
 }
 
-# ---- stac utilities ----
-#' @title Extract datetime from a STAC Query.
-#' @keywords internal
-#' @noRd
-#'
-#' @param stac_query Query that follows the STAC protocol.
-#' @return           List with `start_date` and `end_date` properties.
-.stac_cdse_datetime_as_dates <- function(stac_query) {
-    query_datetime <- stringr::str_split(
-        stac_query[["params"]][["datetime"]], "/"
-    )
-    list(
-        start_date = query_datetime[[1]][1],
-        end_date = query_datetime[[1]][2]
-    )
-}
 
-#' @title Extract bounding box from a STAC Query.
-#' @keywords internal
-#' @noRd
-#'
-#' @param stac_query Query that follows the STAC protocol.
-#' @return           List with `bbox` property.
-.stac_cdse_intersects_as_bbox <- function(stac_query) {
-    result <- list(bbox = NULL)
-    # Extract spatial reference from STAC object
-    intersects <- stac_query[["params"]][["intersects"]]
-    coordinates <- intersects[["coordinates"]]
-    # Check if query is valid
-    if (is.null(coordinates)) {
-        return(result)
-    }
-    # Extract x-coordinates and y-coordinates
-    coordinates_x <- coordinates[,,1]
-    coordinates_y <- coordinates[,,2]
-    # Calculate bounding box
-    min_x <- min(coordinates_x)
-    max_x <- max(coordinates_x)
-    min_y <- min(coordinates_y)
-    max_y <- max(coordinates_y)
-    # Create bbox object
-    result[["bbox"]] <- c(min_x, min_y, max_x, max_y)
-    result
-}
+
+
 
 #' @title Fix STAC Items from CDSE with assets metadata.
 #' @keywords internal
@@ -82,7 +41,8 @@
 #' @param bands      Band names.
 #' @param collection Image collection.
 #' @return           STAC Items updated with `assets` property.
-.stac_cdse_fix_items <- function(source, items, bands, collection, multicores) {
+.cdse_stac_fix_items <- function(source, items, bands, collection, multicores) {
+    .check_set_caller(".cdse_stac_fix_items")
     # Start parallel workers
     .parallel_start(workers = multicores)
     on.exit(.parallel_stop(), add = TRUE)
@@ -116,8 +76,7 @@
                 dplyr::filter(item_s3_content,
                               stringr::str_detect(.data[["Key"]], band_pattern))
             # Check if the correct file was selected.
-            .check_that(nrow(band_item) == 1,
-                        msg = paste0("Invalid band: ", band))
+            .check_that(nrow(band_item) == 1)
             # Prepare the file address
             band_path_s3 <- paste0(s3_protocol, s3_bucket, band_item[["Key"]])
             # Prepare result and return it
@@ -135,267 +94,6 @@
     # Update items object
     items[["features"]] <- features_fixed
     items
-}
-
-# ---- open search utilities ----
-#' @title Prepare Open Search feature as STAC Item (Compatible with `rstac`).
-#' @keywords internal
-#' @noRd
-#'
-#' @description
-#'  This function prepares an Open Search feature to act as an STAC Item,
-#'  compatible with the `rstac` package. This is required to avoid changes in
-#'  other parts of the `sits` validation structure and to use features
-#'  from `rstac`.
-#' @param features     List of features.
-#' @param product_type Type of product associated with the features.
-#' @return             List of features compatible with
-#'                     `rstac` (`rstac::doc_item`).
-.opensearch_as_stac_item <- function(features, product_type) {
-    purrr::map(features, function(feature) {
-        # Use `rstac` class, including `product type`
[…]
