### 2024-04-29 f9aa0ca
Working in RADD implementatiom
