+++ R/api_radd.R
@@ -1,3 +1,225 @@
+.radd_detect_events <- function(data,
+                                threshold = 0.5,
+                                start_date = NULL,
+                                end_date = NULL) {
+    data <- .radd_filter_changes(
+        data = data, threshold = threshold, start_date = start_date,
+        end_date = end_date
+    )
+    data <- .radd_add_dummy(data)
+
+    data <- .radd_start_monitoring(data, threshold)
+}
+
+.radd_start_monitoring <- function(data, threshold, chi = 0.9) {
+    prob_nf <- tidyr::unnest(data, "prob_nf")
+    prob_nf <- dplyr::select(
+        prob_nf, dplyr::all_of(c("sample_id", "NF", "Index", "Flag", "PChange"))
+    )
+    prob_nf <- dplyr::group_by(prob_nf, .data[["sample_id"]])
+    prob_nf[prob_nf$NF < threshold, "Flag"] <- "0"
+    prob_nf <- dplyr::group_modify(prob_nf, ~ {
+        # Filter observations to monitoring and remove first dummy data
+        valid_idxs <- which(.x$NF >= threshold)[-1] - 1
+        for (r in seq_len(length(valid_idxs))) {
+            for (t in seq(valid_idxs[r], nrow(.x))) {
+                # step 2.1: Update Flag and PChange for current time step (i)
+                # (case 1) No confirmed or flagged change:
+                if (nrow(.x[t - 1, "Flag"]) > 0 && !is.na(.x[t - 1, "Flag"])[[1]]) {
+                    if (.x[t - 1, "Flag"] == "0" || .x[t - 1, "Flag"] == "oldFlag") {
+                        i <- 0
+                        prior <- .x[t - 1, "NF"]
+                        likelihood <- .x[t, "NF"]
+                        posterior <- .radd_calc_post(prior, likelihood)
+                        .x[t, "Flag"] <- "Flag"
+                        .x[t, "PChange"] <- posterior
+                    }
+                    # (case 2) Flagged change at previous time step: update PChange
+                    if (.x[t - 1, "Flag"] == "Flag") {
+                        prior <- .x[t - 1, "PChange"]
+                        likelihood <- .x[t, "NF"]
+                        posterior <- .radd_calc_post(prior, likelihood)
+                        .x[t, "Flag"] <- "Flag"
+                        .x[t, "PChange"] <- posterior
+                        i <- i + 1
+                    }
+                }
+                # step 2.2: Confirm and reject flagged changes
+                if (nrow(.x[t - 1, "Flag"]) > 0 && !is.na(.x[t, "Flag"]) && .x[t, "Flag"] == "Flag") {
+                    if ((i > 0)) {
+                        if (.x[t, "PChange"] < 0.5) {
+                            .x[(t - i):t, "Flag"] <- "0"
+                            .x[(t - i), "Flag"] <- "oldFlag"
+                            break
+                        }
+                    }
+                }
+                # confirm change in case PChange >= chi
+                if (nrow(.x[t - 1, "Flag"]) > 0 && !is.na(.x[t, "PChange"]) && .x[t, "PChange"] >= chi) {
+                    if (.x[t, "NF"] >= threshold) {
+                        min_idx <- min(which(.x$Flag == "Flag"))
+                        .x[min_idx:t, "Flag"] <- "Change"
[…]
