+++ R/sits_plot.R
@@ -1722,3 +1722,98 @@ plot.sits_cluster <- function(x, ...,
     )
     return(invisible(dend))
 }
+
+#' @title  ...
+#' @name   plot.radd_model
+#' @description   ...
+#'
+#'
+#' @param  x             Object of class "patterns".
+#' @param  y             Ignored.
+#' @return               A plot object produced by ggplot2
+#'                       with one average pattern per label.
+#'
+#' @note
+#' ....
+#' @examples
+#' if (sits_run_examples()) {
+#'     # plot patterns
+#'
+#' }
+#' @export
+plot.radd_model <- function(x, y) {
+    stopifnot(missing(y))
+    locs <- dplyr::distinct(x, .data[["longitude"]], .data[["latitude"]])
+
+    plots <- purrr::pmap(
+        list(locs$longitude, locs$latitude),
+        function(long, lat) {
+            dplyr::filter(
+                x,
+                .data[["longitude"]] == long,
+                .data[["latitude"]] == lat
+            ) |>
+                .plot_ggplot_radd() |>
+                graphics::plot()
+        }
+    )
+    return(invisible(plots[[1]]))
+}
+
+.plot_ggplot_radd <- function(row) {
+    # create the plot title
+    plot_title <- .plot_title(row$latitude, row$longitude, row$label)
+    colors <- grDevices::hcl.colors(
+        n = 20,
+        palette = "Harmonic",
+        alpha = 1,
+        rev = TRUE
+    )
+    # extract the time series
+    data_ts <- dplyr::bind_rows(row$time_series)
+    # extract the time series
+    data_prob <- dplyr::bind_rows(row$prob_nf)
+    # melt the data into long format
+    melted_ts <- data_ts |>
+        tidyr::pivot_longer(cols = -"Index", names_to = "variable") |>
+        as.data.frame()
+    data_prob <- data_prob[data_prob$Flag == "Change",]
+    change_occur <- NA
+    if (nrow(data_prob) > 0) {
+        melted_prob_pts <- dplyr::filter(
+            melted_ts, .data[["Index"]] %in% data_prob[["Index"]]
+        )
+        change_occur <- max(data_prob$Index, na.rm = TRUE)
+    }
+    # plot the data with ggplot
+    g <- ggplot2::ggplot(melted_ts, ggplot2::aes(
+        x = .data[["Index"]],
+        y = .data[["value"]],
+        group = .data[["variable"]]
+    )) +
+        ggplot2::geom_line(ggplot2::aes(color = .data[["variable"]])) +
+        ggplot2::labs(title = plot_title) +
+        ggplot2::scale_fill_manual(palette = colors)
+
+    if (!is.na(change_occur)) {
+        g <- g +
+            ggplot2::geom_point(
+                data = melted_prob_pts,
+                ggplot2::aes(
+                    x = .data[["Index"]],
+                    y = .data[["value"]],
+                    group = .data[["variable"]]
+                ),
+                colour = "#EBA423"
+            ) +
+            ggplot2::geom_vline(
+                ggplot2::aes(xintercept = change_occur,
+                             linetype = "dotdash"),
+                size = 1,
+                colour = "#EBB023"
+
+            ) +
+            ggplot2::scale_linetype_manual(values = "dotdash", name = "break")
+    }
+    return(g)
+}
