+++ NAMESPACE
@@ -260,6 +260,7 @@ S3method(plot,patterns)
 S3method(plot,predicted)
 S3method(plot,probs_cube)
 S3method(plot,probs_vector_cube)
+S3method(plot,radd_model)
 S3method(plot,raster_cube)
 S3method(plot,rfor_model)
 S3method(plot,sits)
