### 2024-04-29 666f9f4
improve cdse search filters
