### 2024-04-29 0253588
Introduces tolerance parameter in sits_merge function
