+++ man/sits_merge.Rd
@@ -2,16 +2,13 @@
 % Please edit documentation in R/sits_merge.R
 \name{sits_merge}
 \alias{sits_merge}
-\alias{sits_merge.sits}
 \alias{sits_merge.raster_cube}
 \alias{sits_merge.default}
 \title{Merge two data sets (time series or cubes)}
 \usage{
 sits_merge(data1, data2, ...)
 
-\method{sits_merge}{sits}(data1, data2, ..., suffix = c(".1", ".2"))
-
-\method{sits_merge}{raster_cube}(data1, data2, ...)
+\method{sits_merge}{raster_cube}(data1, data2, ..., tolerance = NULL, output_dir = NULL)
 
 \method{sits_merge}{default}(data1, data2, ...)
 }
