+++ NAMESPACE
@@ -353,6 +353,7 @@ S3method(sits_labels,sits_model)
 S3method(sits_labels_summary,sits)
 S3method(sits_merge,default)
 S3method(sits_merge,raster_cube)
+S3method(sits_merge,sits)
 S3method(sits_mixture_model,default)
 S3method(sits_mixture_model,derived_cube)
 S3method(sits_mixture_model,raster_cube)
