+++ tests/testthat/test-cube.R
@@ -9,6 +9,7 @@ test_that("List collections", {
     expect_true(grepl("CBERS-WFI-16D", col_bdc))
     expect_true(grepl("CBERS-WFI-8D", col_bdc))
 })
+
 test_that("api_source", {
     res_s2_b2 <- .source_bands_resolution(
         source = "CDSE",
@@ -63,6 +64,7 @@ test_that("api_source", {
     )
     expect_true(token_bdc)
 })
+
 test_that("Reading a raster cube", {
     data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
     raster_cube <- sits_cube(
@@ -97,1326 +99,6 @@ test_that("Reading a raster cube", {
     expect_true(params_2$xres >= 231.5)
 })
 
-test_that("Creating cubes from BDC - CBERS-WFI-16D", {
-    tiles <- c("007004", "007005")
-    start_date <- "2021-09-01"
-    end_date <- "2021-09-30"
-    bands <- c("NDVI", "EVI", "B13", "B14", "B15", "B16", "CLOUD")
-    # create a raster cube
-    cbers_cube_16d <- .try(
-        {
-            sits_cube(
-                source = "BDC",
-                collection = "CBERS-WFI-16D",
-                tiles = tiles,
-                start_date = start_date,
-                end_date = end_date,
-                progress = FALSE
-            )
-        },
-        .default = NULL
-    )
-    testthat::skip_if(purrr::is_null(cbers_cube_16d),
-                      message = "BDC is not accessible"
-    )
-    # test bands and bbox
-    expect_true(all(sits_bands(cbers_cube_16d) %in% bands))
-    bbox <- sits_bbox(cbers_cube_16d)
-    int_bbox <- .bbox_intersection(bbox, .tile_bbox(cbers_cube_16d))
-    expect_true(all(int_bbox == sits_bbox(.tile(cbers_cube_16d))))
-    # test timeline
-    timeline <- sits_timeline(cbers_cube_16d)
-    expect_true(timeline[1] <= as.Date(start_date))
-    expect_true(timeline[length(timeline)] <= as.Date(end_date))
-    # test raster obj
-    r_obj <- .raster_open_rast(cbers_cube_16d$file_info[[1]]$path[1])
-    cube_nrows <- .tile_nrows(cbers_cube_16d)
-    expect_true(.raster_nrows(r_obj) == cube_nrows)
-})
-test_that("Creating cubes from BDC - CBERS-WFI-8D", {
-    tiles <- c("007004", "007005")
-    start_date <- "2022-05-01"
-    end_date <- "2022-08-29"
-    bands <- c("NDVI", "EVI", "B13", "B14", "B15", "B16", "CLOUD")
-    # create a raster cube file from BDC
-    cbers_cube_8d <- .try(
-        {
-            sits_cube(
-                source = "BDC",
-                collection = "CBERS-WFI-8D",
-                tiles = tiles,
-                start_date = start_date,
-                end_date = end_date,
-                progress = FALSE
-            )
-        },
-        .default = NULL
-    )
-    testthat::skip_if(purrr::is_null(cbers_cube_8d),
-        message = "BDC is not accessible"
-    )
-    expect_true(all(sits_bands(cbers_cube_8d) %in% bands))
-    bbox <- sits_bbox(cbers_cube_8d)
-    int_bbox <- .bbox_intersection(bbox, .tile_bbox(cbers_cube_8d))
-    expect_true(all(int_bbox == sits_bbox(.tile(cbers_cube_8d))))
-
-    timeline <- sits_timeline(cbers_cube_8d)
-    expect_true(timeline[1] <= as.Date(start_date))
[…]
