### 2024-04-29 e515b04
fix bug in cube copy for SAR images
