+++ NEWS.md
@@ -2,11 +2,13 @@
 
 # What's new in SITS version 1.5
 ### New features in SITS version 1.5.0
-* Improved support for GPU-based classification of deep learning models
+* Support for SENTINEL-1-RTC and SENTINEL-2-L2A in CDSE
+* Include support for DEA products SENTINEL-1-RTC, LS5-SR, LS7-SR, LS9-SR, ALOS-PALSAR-MOSAIC, NDVI ANOMALY, DAILY CHIRPS, MONTHLY CHIRPS and DEM-30
 * Support for Sentinel-1 GRD and RTC collections from Planetary Computer
 * Include parameter tile to select data from Sentinel-1 (MPC) 
-* Include parameter tile to select data from Sentinel-2 (DEAFRICA)
+* Include parameter tile to select data from Sentinel-1 and Sentinel-2 (DEAFRICA)
 * Include parameter tile to select data from HLS collections
+* Improved support for GPU-based classification of deep learning models
 * Support for non-normalized derived indexes
 * Support for shapefiles as ROI in `sits_cube()`
 * Fix inconsistencies in HLS scale factors
