### 2024-04-30 f6f769a
fix documentation errors in sits_merge
