### 2024-04-30 b590003
add docs in sits_merge
