### 2024-05-01 d12a306
add new checks on sits_merge
