### 2024-05-01 b994793
update error message in sits_merge
