+++ inst/extdata/config_messages.yml
@@ -373,7 +373,7 @@ sits_labels_summary: "this function is deprecated; please use summary()"
 sits_lighttae: "wrong input parameters - see example in documentation"
 sits_list_collections: "invalid source parameter as data provider"
 sits_merge_default: "input should be objects of class sits or class raster_cube"
-sits_merge_raster_cubes: "merge cubes requires same tiles in both cubes"
+sits_merge_raster_cubes: "merge cubes requires same tiles and timelines length in both cubes"
 sits_merge_sits: "input data is NULL or has different number of rows"
 sits_merge_sits_bands: "duplicated band names - merge only works if bands in inputs are different"
 sits_mixture_model: "wrong input parameters - see example in documentation"
