### 2024-05-02 #596 @mpadge
@gilbertocamara Thank you so much for your considered and very deep engagement with our statistical standards. The first thing I would like to ask would be for you to copy the above comments into separate issues within the repository for our [*Statistical Standards Book*](https://github.com/ropensci/statistical-software-review-book/issues?q=is%3Aissue+is%3Aopen+sort%3Aupdated-desc) - one for the *Machine Learning* and one for *Spatial* standards. We'll then incorporate your excellent feedback there via updates to our standards.

We are definitely keen to progress with peer-review here. The `sits` package is definitely within scope. I imagine the only problems `sits` may pose will be extra burden on reviewers of such a very large and comprehensive package. But we are definitely excited to learn from guiding this kind of package through our system, and from hopefully mutually beneficial feedback from both sides throughout the process.

Our current Editor-in-Chief @jooolia will take it from here. Thank you for all of your work!
