### 2024-05-02 #596 @gilbertocamara
Dear @mpadge Many thanks for your response. I will include the relevant part of my comments as issues in the github repository for the book. I shall be waiting for instructions from @jooolia on how to proceed.
