### 2024-05-02 bdf0080
Change in the requirement for the 'output_dir' parameter in sits_merge
