### 2024-05-05 ce1d3f7
improve plot and view using COG overviews
