+++ tests/testthat/test-vignette.R
@@ -1,5 +1,5 @@
 test_that("Test Data Handling", {
-    expect_true(all(samples_modis_ndvi[1:3, ][["label"]] %in% "Pasture"))
+    expect_true(all(samples_modis_ndvi[1:3, ][["label"]] == "Pasture"))
     expect_identical(sum(summary(samples_modis_ndvi)[["prop"]]), 1)
-    expect_true(all(sits_bands(samples_modis_ndvi) %in% "NDVI"))
+    expect_true(all(sits_bands(samples_modis_ndvi) == "NDVI"))
 })
