### 2024-05-05 #596 @gilbertocamara
Dear @jooolia

Many thanks for starting the review process of R package `sits`.

Regarding the items mentioned:

✖️ Package name is not available (on CRAN):
version 1.5.0 will be submitted to CRAN tomorrow. I will inform you when it is accepted.

✖️ Package has no HTML vignettes:
The package has a full on-line book (see https://e-sensing.github.io/sitsbook/) and so there is no need for HTML vignettes.

✖️ These functions do not have examples: [sits_run_examples, sits_run_tests].
These functions are auxiliary functions, to avoid CRAN checks.

✖️ Package coverage is 0.1% (should be at least 75%).
In fact, package coverage is 94% (see https://app.codecov.io/gh/e-sensing/sits). SITS includes functions that access cloud services and functions which take a long time to run. We access seven cloud services and some of them may be off-line. For this reason, tests and examples are run off-line. To run tests and examples, please include the following environmental variables:
```{r}
Sys.setenv(SITS_RUN_TESTS = "YES")
Sys.setenv(SITS_RUN_EXAMPLES = "YES")
```
 The issues raised by `lintr` and `cyclocomp` have been addressed in version 1.5.0
