### 2024-05-06 227a294
merge cubes - part 2
