### 2024-05-06 1c71d92
update sits_merge function
