+++ R/sits_merge.R
@@ -34,13 +34,6 @@
 #' @param suffix     If there are duplicate bands in data1 and data2
 #'                   these suffixes will be added
 #'                   (character vector).
-#' @param tolerance  A period tolerance to merge both cubes.
-#'                   ISO8601-compliant time period for regular data cubes,
-#'                   with number and unit, where "D", "M" and "Y" stand
-#'                   for days, month and year; e.g., "P16D" for 16 days.
-#'                   The temporal tolerance parameter should be less than
-#'                   the time interval between two images of both cubes.
-#' @param output_dir Valid directory for storing merged images.
 #'
 #' @return merged data sets (tibble of class "sits" or
 #'         tibble of class "raster_cube")
@@ -109,96 +102,98 @@ sits_merge.sits <- function(data1, data2, ..., suffix = c(".1", ".2")) {
 
 #' @rdname sits_merge
 #' @export
-#'
-sits_merge.raster_cube <- function(data1, data2, ...,
-                                   tolerance = NULL,
-                                   output_dir = NULL) {
+sits_merge.sar_cube <- function(data1, data2) {
+    .check_set_caller("sits_merge_sar_cube")
+    # pre-condition - check cube type
+    .check_is_raster_cube(data1)
+    .check_is_raster_cube(data2)
+    # pre-condition for merge is having the same tiles
+    common_tiles <- intersect(data1[["tile"]], data2[["tile"]])
+    .check_that(length(common_tiles) > 0)
+    # filter cubes by common tiles and arrange them
+    data1 <- dplyr::arrange(
+        dplyr::filter(data1, .data[["tile"]] %in% common_tiles),
+        .data[["tile"]]
+    )
+    data2 <- dplyr::arrange(
+        dplyr::filter(data2, .data[["tile"]] %in% common_tiles),
+        .data[["tile"]]
+    )
+    if (inherits(data2, "sar_cube")) {
+        return(.merge_equal_cube(data1, data2))
+    } else {
+        return(.merge_distinct_cube(data1, data2))
+    }
+}
+
+#' @rdname sits_merge
+#' @export
+sits_merge.raster_cube <- function(data1, data2) {
     .check_set_caller("sits_merge_raster_cube")
     # pre-condition - check cube type
     .check_is_raster_cube(data1)
     .check_is_raster_cube(data2)
-    # precondition for merge is having the same tiles
-    # join cube tiles
+    # pre-condition for merge is having the same tiles
     common_tiles <- intersect(data1[["tile"]], data2[["tile"]])
     .check_that(length(common_tiles) > 0)
-    # filter cubes by common tiles
-    data1 <- dplyr::filter(data1, .data[["tile"]] %in% common_tiles)
-    data2 <- dplyr::filter(data2, .data[["tile"]] %in% common_tiles)
-    # Get cubes timeline
-    d1_tl <- unique(as.Date(unlist(.cube_timeline(data1))))
-    d2_tl <- unique(as.Date(unlist(.cube_timeline(data2))))
-    # get minimum interval
-    min_interval_data1 <- min(
-        lubridate::as.period(
-            lubridate::int_diff(d1_tl)
-        )
+    # filter cubes by common tiles and arrange them
+    data1 <- dplyr::arrange(
+        dplyr::filter(data1, .data[["tile"]] %in% common_tiles),
+        .data[["tile"]]
     )
-    min_interval_data2 <- min(
-        lubridate::as.period(
-            lubridate::int_diff(d2_tl)
-        )
+    data2 <- dplyr::arrange(
+        dplyr::filter(data2, .data[["tile"]] %in% common_tiles),
[…]
