+++ R/api_merge.R
@@ -2,7 +2,7 @@
     abs(as.Date(t1) - as.Date(t2))
 }
 
-.merge_fi <- function(data1, data2) {
+.cube_merge <- function(data1, data2) {
     data1 <- slider::slide2_dfr(data1, data2, function(x, y) {
         .fi(x) <- dplyr::arrange(
             dplyr::bind_rows(.fi(x), .fi(y)),
