+++ NAMESPACE
@@ -93,6 +93,10 @@ S3method(.data_get_ts,raster_cube)
 S3method(.gc_arrange_images,raster_cube)
 S3method(.mosaic_split_band_date,derived_cube)
 S3method(.mosaic_split_band_date,raster_cube)
+S3method(.opensearch_cdse_extract_tile,RTC)
+S3method(.opensearch_cdse_extract_tile,S2MSI2A)
+S3method(.opensearch_cdse_search,RTC)
+S3method(.opensearch_cdse_search,S2MSI2A)
 S3method(.raster_check_package,terra)
 S3method(.raster_col,terra)
 S3method(.raster_crop,terra)
@@ -414,7 +418,8 @@ S3method(summary,sits_accuracy)
 S3method(summary,sits_area_accuracy)
 export("sits_bands<-")
 export("sits_labels<-")
-export(.mpc_clean_token_cache)
+export(.opensearch_cdse_extract_tile)
+export(.opensearch_cdse_search)
 export(impute_linear)
 export(sits_accuracy)
 export(sits_accuracy_summary)
