### 2024-05-08 4bf743d
fix opensearch s3 exports
