+++ R/RcppExports.R
@@ -37,8 +37,32 @@ C_kernel_modal <- function(x, ncols, nrows, band, window_size) {
     .Call(`_sits_C_kernel_modal`, x, ncols, nrows, band, window_size)
 }
 
-dtw <- function() {
-    .Call(`_sits_dtw`)
+kohonen_dtw <- function() {
+    .Call(`_sits_kohonen_dtw`)
+}
+
+kohonen_euclidean <- function() {
+    .Call(`_sits_kohonen_euclidean`)
+}
+
+kohonen_object_distances <- function(data, numVars, numNAs, distanceFunction, weights) {
+    .Call(`_sits_kohonen_object_distances`, data, numVars, numNAs, distanceFunction, weights)
+}
+
+RcppMap <- function(data, numVars, numNAs, codes, weights, distanceFunction) {
+    .Call(`_sits_RcppMap`, data, numVars, numNAs, codes, weights, distanceFunction)
+}
+
+RcppSupersom <- function(data, codes, numVars, weights, distanceFunction, numNAs, neighbourhoodDistances, alphas, radii, numEpochs) {
+    .Call(`_sits_RcppSupersom`, data, codes, numVars, weights, distanceFunction, numNAs, neighbourhoodDistances, alphas, radii, numEpochs)
+}
+
+RcppBatchSupersom <- function(data, codes, numVars, weights, distanceFunction, numNAs, neighbourhoodDistances, radii, numEpochs) {
+    .Call(`_sits_RcppBatchSupersom`, data, codes, numVars, weights, distanceFunction, numNAs, neighbourhoodDistances, radii, numEpochs)
+}
+
+RcppParallelBatchSupersom <- function(data, codes, numVars, weights, distanceFunction, numNAs, neighbourhoodDistances, radii, numEpochs, numCores) {
+    .Call(`_sits_RcppParallelBatchSupersom`, data, codes, numVars, weights, distanceFunction, numNAs, neighbourhoodDistances, radii, numEpochs, numCores)
 }
 
 C_label_max_prob <- function(x) {
@@ -139,14 +163,6 @@ C_temp_tqr <- function(mtx) {
 
 C_temp_iqr <- function(mtx) {
     .Call(`_sits_C_temp_iqr`, mtx)
-}
-
-rlang_env_unlock <- function(env) {
-    invisible(.Call(`_sits_rlang_env_unlock`, env))
-}
-
-rlang_env_lock <- function(env) {
-    invisible(.Call(`_sits_rlang_env_lock`, env))
 }
 
 C_max_sampling <- function(x, nrows, ncols, window_size) {
