+++ src/kohonen_som.cpp
@@ -1,0 +1,835 @@
+
+
+#include <Rcpp.h>
+#include <Rmath.h>
+
+#ifdef _OPENMP
+#include <omp.h>
+#endif
+
+#include "./sits_types.h"
+
+#define EPS 1e-8 /* relative test of equality of distances */
+
+#define RANDIN  GetRNGstate()
+#define RANDOUT PutRNGstate()
+#define UNIF unif_rand()
+
+using namespace Rcpp;
+
+/**
+ * Gaussian neighborhood function.
+ *
+ * @description
+ * This function computes the gaussian neighborhood function.
+ *
+ * @param distance          A `double` with the distance between two objects.
+ * @param radius            A `double` with the radius of the neighbourhood.
+ *
+ * @note
+ * The implementation of this function was adapted from the `kohonen` R Package.
+ * The code is open-source, under the GPL license, and is available on
+ * GitHub (https://github.com/rwehrens/kohonen)
+ *
+ * @return The Gaussian neighborhood value.
+ */
+double Gaussian(double distance, double radius) {
+    return exp(-(distance * distance) / (2 * radius * radius));
+}
+
+/**
+ * Find best match unit (BMU).
+ *
+ * @description
+ * Finds the best matching codebook unit for the given data object and stores
+ * its index and distance in the specified nearest unit index and nearest unit
+ * distance references.
+ *
+ * @param object            A `double*` to the data object.
+ * @param codes             A `double*` to the codes object.
+ * @param offsets           A `int*` to the offset object.
+ * @param numNAs            A `int*` to the number of NA object.
+ * @param numCodes          A `int` with the number of codes.
+ * @param numLayers         A `int` with the number of layers.
+ * @param numVars           A `int*` to the number of variables to be used.
+ * @param totalVars         A `int` with the total of variables in the data.
+ * @param distanceFunction  A `DistanceFunctionPtr&` with the reference to the
+ *                          distance function.
+ * @param weights           A `double*` to the weights object.
+ * @param index             A `int&` referencing the nearest BMU.
+ * @param distance          A `double&` referencing the distance between the
+ *                          current object and the nearest BMU.
+ * @note
+ * The implementation of this function was adapted from the `kohonen` R Package.
+ * The code is open-source, under the GPL license, and is available on
+ * GitHub (https://github.com/rwehrens/kohonen)
+ */
+void FindBestMatchingUnit(
+    double *object,
+    double *codes,
+    int *offsets,
+    int *numNAs,
+    int numCodes,
+    int numLayers,
+    int *numVars,
+    int totalVars,
+    const DistanceFunctionPtr &distanceFunction,
+    double *weights,
+    int &index,
+    double &distance)
+{
+    int nind = 1;
+    double dist;
+
+    index = NA_INTEGER;
+    distance = DBL_MAX;
+    for (int cd = 0; cd < numCodes; ++cd) {
+
+        /* Calculate current unit distance */
+        dist = 0.0;
+        for (int l = 0; l < numLayers; ++l) {
+            dist += weights[l] * (*distanceFunction)(
+                &object[offsets[l]],
+                &codes[cd * totalVars + offsets[l]],
+                numVars[l],
+                numNAs[l]
+            );
+        }
+
+        /* Update best matching unit */
+        if (dist <= distance * (1 + EPS)) {
+            if (dist < distance * (1 - EPS)) {
+                nind = 1;
+                index = cd;
+            } else {
+                if (++nind * UNIF < 1.0) {
+                    index = cd;
+                }
[…]
