### 2024-05-09 5e226ed
fix batch supersom call
