+++ src/kohonen_som.cpp
@@ -828,8 +828,7 @@ List RcppParallelBatchSupersom(
 
   #else
     ::Rf_warning("OpenMP not available: cannot run in parallel mode");
-    return RcppBatchSupersom(data, codes, numVars, weights, distanceFunctions,
-			     numNAs, neighbourhoodDistances, neighbourhoodFct,
-			     radii, numEpochs);
+    return RcppBatchSupersom(data, codes, numVars, weights, distanceFunction,
+			     numNAs, neighbourhoodDistances, radii, numEpochs);
   #endif
 }
