### 2024-05-09 #596 @gilbertocamara
Dear @jooola, version 1.5.0 of  `sits` is now on CRAN. Whenever you feel appropriate, you can start the software review of the package. Please note that all issues raised by the automated bot have been responded above.
