### 2024-05-12 682fcd1
fix sits_reduce bug
