+++ src/probability_fns.cpp
@@ -11,42 +11,161 @@ arma::mat C_dnorm(const arma::mat& mtx,
 }
 
 arma::vec C_radd_calc_pcond(const arma::vec& p1,
-                           const arma::vec& p2) {
+                            const arma::vec& p2) {
     return p1 / (p1 + p2);
 }
 
 arma::vec C_radd_calc_pbayes(const arma::vec& prior,
                              const arma::vec& post) {
+    return (prior % post) / ((prior % post) + ((1 - prior) % (1 - post)));
+}
+
+double C_radd_calc_pbayes(const double& prior,
+                          const double& post) {
     return (prior * post) / ((prior * post) + ((1 - prior) * (1 - post)));
 }
 
 // [[Rcpp::export]]
-arma::vec C_radd_calc_nf(const arma::mat& ts,
+arma::mat C_radd_calc_nf(const arma::mat& ts,
                          const arma::mat& mean,
-                         const arma::mat& std) {
+                         const arma::mat& sd,
+                         const arma::uword& n_times,
+                         const double& threshold = 0.5) {
 
-    arma::vec pnfor(ts.n_rows, arma::fill::zeros);
-    arma::vec pfor(ts.n_rows, arma::fill::zeros);
-    arma::vec pres(ts.n_rows, arma::fill::zeros);
-    bool update_res = false;
-    for (int i = 0; i < ts.n_cols; i++) {
-        pfor = C_dnorm(ts.col(i), mean(0, i), std(0, i));
-        pnfor = C_dnorm(ts.col(i), mean(1, i), std(1, i));
+    arma::mat p_res(ts.n_rows, n_times, arma::fill::zeros);
+    arma::mat p_flag(ts.n_rows, n_times, arma::fill::value(arma::datum::nan));
+    arma::mat p_change(ts.n_rows, n_times, arma::fill::value(arma::datum::nan));
 
-        pnfor.elem(arma::find(pnfor < 0.00001)).zeros();
+    // for each pixel
+    for (int i = 0; i < ts.n_rows; i++) {
+        arma::colvec p_for(n_times, arma::fill::zeros);
+        arma::colvec p_nfor(n_times, arma::fill::zeros);
+        arma::colvec p_nfor_past(n_times, arma::fill::zeros);
 
-        pnfor.elem(arma::find(pnfor > 0)) = C_radd_calc_pcond(
-            pnfor.elem(arma::find(pnfor > 0)),
-            pfor.elem(arma::find(pnfor > 0))
-        );
+        arma::uword col_idx = 0;
 
-        pnfor.elem(arma::find(pnfor < 0)).zeros();
-        pnfor.elem(arma::find(pnfor > 1)).ones();
+        bool update_res = false;
+        // for each band
+        for (int c = 0; c < ts.n_cols; c = c + n_times) {
+            p_for = C_dnorm(
+                ts.submat(i, c, i, c + n_times - 1).t(),
+                mean(0, col_idx),
+                sd(0, col_idx)
+            );
+            p_nfor = C_dnorm(
+                ts.submat(i, c, i, c + n_times - 1).t(),
+                mean(1, col_idx),
+                sd(1, col_idx)
+            );
+            p_nfor.elem(arma::find(p_nfor < 0.00001)).zeros();
 
-        if (update_res) {
-            pnfor = C_radd_calc_pbayes(pnfor, pres);
+            p_nfor.elem(arma::find(p_nfor > 0)) = C_radd_calc_pcond(
+                p_nfor.elem(arma::find(p_nfor > 0)),
+                p_for.elem(arma::find(p_nfor > 0))
+            );
+
+            p_nfor.elem(arma::find(p_nfor < 0)).zeros();
[…]

[…]
