+++ man/plot.raster_cube.Rd
@@ -12,12 +12,10 @@
   green = NULL,
   blue = NULL,
   tile = x[["tile"]][[1]],
-  date = NULL,
+  dates = NULL,
   palette = "RdYlGn",
-  style = "cont",
-  n_colors = 10,
   rev = FALSE,
-  scale = 1
+  scale = 1.2
 )
 }
 \arguments{
@@ -35,15 +33,9 @@
 
 \item{tile}{Tile to be plotted.}
 
-\item{date}{Date to be plotted.}
+\item{dates}{Dates to be plotted.}
 
 \item{palette}{An RColorBrewer palette}
-
-\item{style}{Method to process the color scale
-("cont", "order", "quantile", "fisher",
- "jenks", "log10")}
-
-\item{n_colors}{Number of colors to be shown}
 
 \item{rev}{Reverse the color order in the palette?}
 
@@ -51,16 +43,18 @@
 }
 \value{
 A plot object with an RGB image
-                      or a B/W image on a color
-                      scale using the pallete
+                      or a B/W image on a color scale
 }
 \description{
 Plot RGB raster cube
 }
 \note{
-To see which colors are supported, please run \code{sits_colors()}
-      Use \code{scale} parameter for general output control.
-      If required, then set the other params individually
+Use \code{scale} parameter for general output control.
+      The \code{dates} parameter indicates the date allows plotting of different dates when
+      a single band and three dates are provided, `sits` will plot a
+      multi-temporal RGB image for a single band (useful in the case of
+      SAR data). For RGB bands with multi-dates, multiple plots will be
+      produced.
 }
 \examples{
 if (sits_run_examples()) {
@@ -72,7 +66,8 @@ if (sits_run_examples()) {
         data_dir = data_dir
     )
     # plot NDVI band of the second date date of the data cube
-    plot(cube, band = "NDVI", date = sits_timeline(cube)[1])
+    plot(cube, band = "NDVI", dates = sits_timeline(cube)[1])
+    # plot NDVI band as an RGB composite for the three bands
 }
 }
 \author{
