### 2024-05-13 7b2211a
Fix gdal_presets for sits_regularize
