### 2024-05-14 ce544aa
introduce detect changes api
