+++ man/sits_detect_change_method.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sits_detect_change_method.R
+\name{sits_detect_change_method}
+\alias{sits_detect_change_method}
+\title{Create detect change method.}
+\usage{
+sits_detect_change_method(samples, cd_method = sits_dtw())
+}
+\arguments{
+\item{samples}{Time series with the training samples.}
+
+\item{cd_method}{Change detection method.}
+}
+\value{
+Change detection method prepared
+                         to be passed to
+                         \code{\link[sits]{sits_detect_change}}
+}
+\description{
+Prepare detection change method. Currently, sits supports the
+following methods: 'dtw' (see \code{\link[sits]{sits_dtw}})
+}
+\author{
+Gilberto Camara, \email{gilberto.camara@inpe.br}
+
+Felipe Carlos, \email{efelipecarlos@gmail.com}
+}
