### 2024-05-15 71fae2d
fix errors when plots and graphics are required
