+++ man/plot.raster_cube.Rd
@@ -15,7 +15,8 @@
   dates = NULL,
   palette = "RdYlGn",
   rev = FALSE,
-  scale = 1.2
+  scale = 1,
+  style = "order"
 )
 }
 \arguments{
@@ -40,6 +41,8 @@
 \item{rev}{Reverse the color order in the palette?}
 
 \item{scale}{Scale to plot map (0.4 to 1.0)}
+
+\item{style}{Style for plotting continuous objects}
 }
 \value{
 A plot object with an RGB image
