+++ src/probability_fns.cpp
@@ -10,18 +10,15 @@ arma::mat C_dnorm(const arma::mat& mtx,
     return arma::normpdf(mtx, mean, std);
 }
 
-arma::vec C_radd_calc_pcond(const arma::vec& p1,
-                            const arma::vec& p2) {
+arma::vec C_radd_calc_pcond(const arma::vec& p1, const arma::vec& p2) {
     return p1 / (p1 + p2);
 }
 
-arma::vec C_radd_calc_pbayes(const arma::vec& prior,
-                             const arma::vec& post) {
+arma::vec C_radd_calc_pbayes(const arma::vec& prior, const arma::vec& post) {
     return (prior % post) / ((prior % post) + ((1 - prior) % (1 - post)));
 }
 
-double C_radd_calc_pbayes(const double& prior,
-                          const double& post) {
+double C_radd_calc_pbayes(const double& prior, const double& post) {
     return (prior * post) / ((prior * post) + ((1 - prior) * (1 - post)));
 }
 
@@ -29,64 +26,70 @@ double C_radd_calc_pbayes(const double& prior,
 arma::mat C_radd_calc_nf(const arma::mat& ts,
                          const arma::mat& mean,
                          const arma::mat& sd,
-                         const arma::uword& n_times,
-                         const double& threshold = 0.5) {
+                         const arma::uword& n_times) {
 
-    arma::mat p_res(ts.n_rows, n_times, arma::fill::zeros);
-    arma::mat p_flag(ts.n_rows, n_times, arma::fill::value(arma::datum::nan));
-    arma::mat p_change(ts.n_rows, n_times, arma::fill::value(arma::datum::nan));
-
-    // for each pixel
-    for (int i = 0; i < ts.n_rows; i++) {
+    // Using the first element as dummy value
+    arma::mat p_res(ts.n_rows, n_times + 1, arma::fill::value(0.5));
+    // For each pixel
+    for (arma::uword i = 0; i < ts.n_rows; i++) {
+        // Probability to be a Forest
         arma::colvec p_for(n_times, arma::fill::zeros);
+        // Probability to be a Non-Forest
         arma::colvec p_nfor(n_times, arma::fill::zeros);
+        // Probability to be a Non-Forest in the past
         arma::colvec p_nfor_past(n_times, arma::fill::zeros);
 
+        // Aux variables
         arma::uword col_idx = 0;
-
         bool update_res = false;
-        // for each band
-        for (int c = 0; c < ts.n_cols; c = c + n_times) {
+
+        // For each band
+        for (arma::uword c = 0; c < ts.n_cols; c = c + n_times) {
+            // Estimate a normal distribution based on Forest stats
             p_for = C_dnorm(
                 ts.submat(i, c, i, c + n_times - 1).t(),
                 mean(0, col_idx),
                 sd(0, col_idx)
             );
+            // Estimate a normal distribution based on Non-Forest stats
             p_nfor = C_dnorm(
                 ts.submat(i, c, i, c + n_times - 1).t(),
                 mean(1, col_idx),
                 sd(1, col_idx)
             );
+            // Clean values lower than 0.00001
             p_nfor.elem(arma::find(p_nfor < 0.00001)).zeros();
-
+            // Estimate a conditional prob for each positive distribution value
             p_nfor.elem(arma::find(p_nfor > 0)) = C_radd_calc_pcond(
                 p_nfor.elem(arma::find(p_nfor > 0)),
                 p_for.elem(arma::find(p_nfor > 0))
             );
-
+            // Fix the range of prob values between 0 and 1
             p_nfor.elem(arma::find(p_nfor < 0)).zeros();
             p_nfor.elem(arma::find(p_nfor > 1)).ones();
 
[…]
