+++ src/RcppExports.cpp
@@ -274,8 +274,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // C_radd_calc_nf
-arma::mat C_radd_calc_nf(const arma::mat& ts, const arma::mat& mean, const arma::mat& sd, const arma::uword& n_times, const double& threshold);
-RcppExport SEXP _sits_C_radd_calc_nf(SEXP tsSEXP, SEXP meanSEXP, SEXP sdSEXP, SEXP n_timesSEXP, SEXP thresholdSEXP) {
+arma::mat C_radd_calc_nf(const arma::mat& ts, const arma::mat& mean, const arma::mat& sd, const arma::uword& n_times);
+RcppExport SEXP _sits_C_radd_calc_nf(SEXP tsSEXP, SEXP meanSEXP, SEXP sdSEXP, SEXP n_timesSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -283,33 +283,21 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const arma::mat& >::type mean(meanSEXP);
     Rcpp::traits::input_parameter< const arma::mat& >::type sd(sdSEXP);
     Rcpp::traits::input_parameter< const arma::uword& >::type n_times(n_timesSEXP);
-    Rcpp::traits::input_parameter< const double& >::type threshold(thresholdSEXP);
-    rcpp_result_gen = Rcpp::wrap(C_radd_calc_nf(ts, mean, sd, n_times, threshold));
+    rcpp_result_gen = Rcpp::wrap(C_radd_calc_nf(ts, mean, sd, n_times));
     return rcpp_result_gen;
 END_RCPP
 }
-// seq_int
-arma::vec seq_int(const arma::uword& from, const arma::uword& to, const arma::uword& n);
-RcppExport SEXP _sits_seq_int(SEXP fromSEXP, SEXP toSEXP, SEXP nSEXP) {
+// C_radd_detect_changes
+arma::mat C_radd_detect_changes(const arma::mat& p_res, const double& threshold, const double& chi);
+RcppExport SEXP _sits_C_radd_detect_changes(SEXP p_resSEXP, SEXP thresholdSEXP, SEXP chiSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const arma::uword& >::type from(fromSEXP);
-    Rcpp::traits::input_parameter< const arma::uword& >::type to(toSEXP);
-    Rcpp::traits::input_parameter< const arma::uword& >::type n(nSEXP);
-    rcpp_result_gen = Rcpp::wrap(seq_int(from, to, n));
-    return rcpp_result_gen;
-END_RCPP
-}
-// C_radd_start_monitoring
-void C_radd_start_monitoring(const arma::mat& p_res, const double& threshold);
-RcppExport SEXP _sits_C_radd_start_monitoring(SEXP p_resSEXP, SEXP thresholdSEXP) {
-BEGIN_RCPP
-    Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const arma::mat& >::type p_res(p_resSEXP);
     Rcpp::traits::input_parameter< const double& >::type threshold(thresholdSEXP);
-    C_radd_start_monitoring(p_res, threshold);
-    return R_NilValue;
[…]

[…]
