### 2024-05-16 a4d7211
add multicores processing support for active learning sampling methods
