+++ R/sits_active_learning.R
@@ -31,6 +31,10 @@
 #' @param min_uncert      Minimum uncertainty value to select a sample.
 #' @param sampling_window Window size for collecting points (in pixels).
 #'                        The minimum window size is 10.
+#' @param multicores      Number of workers for parallel processing
+#'                        (integer, min = 1, max = 2048).
+#' @param memsize         Maximum overall memory (in GB) to run the
+#'                        function.
 #'
 #' @return
 #' A tibble with longitude and latitude in WGS84 with locations
@@ -75,23 +79,64 @@
 sits_uncertainty_sampling <- function(uncert_cube,
                                       n = 100L,
                                       min_uncert = 0.4,
-                                      sampling_window = 10L) {
+                                      sampling_window = 10L,
+                                      multicores = 1L,
+                                      memsize = 1L) {
     .check_set_caller("sits_uncertainty_sampling")
-
     # Pre-conditions
     .check_is_uncert_cube(uncert_cube)
     .check_int_parameter(n, min = 1, max = 10000)
     .check_num_parameter(min_uncert, min = 0.2, max = 1.0)
     .check_int_parameter(sampling_window, min = 10L)
-
+    .check_int_parameter(multicores, min = 1, max = 2048)
+    .check_int_parameter(memsize, min = 1, max = 16384)
+    # Get block size
+    block <- .raster_file_blocksize(.raster_open_rast(.tile_path(uncert_cube)))
+    # Overlapping pixels
+    overlap <- ceiling(sampling_window / 2) - 1
+    # Check minimum memory needed to process one block
+    job_memsize <- .jobs_memsize(
+        job_size = .block_size(block = block, overlap = overlap),
+        npaths = sampling_window,
+        nbytes = 8,
+        proc_bloat = .conf("processing_bloat_cpu")
+    )
+    # Update multicores parameter
+    multicores <- .jobs_max_multicores(
+        job_memsize = job_memsize,
+        memsize = memsize,
+        multicores = multicores
+    )
+    # Update block parameter
+    block <- .jobs_optimal_block(
+        job_memsize = job_memsize,
+        block = block,
+        image_size = .tile_size(.tile(uncert_cube)),
+        memsize = memsize,
+        multicores = multicores
+    )
+    # Prepare parallel processing
+    .parallel_start(workers = multicores)
+    on.exit(.parallel_stop(), add = TRUE)
     # Slide on cube tiles
     samples_tb <- slider::slide_dfr(uncert_cube, function(tile) {
-        path <- .tile_path(tile)
+        # Create chunks as jobs
+        chunks <- .tile_chunks_create(
+            tile = tile,
+            overlap = overlap,
+            block = block
+        )
+        # Tile path
+        tile_path <- .tile_path(tile)
         # Get a list of values of high uncertainty
-        top_values <- .raster_open_rast(path) |>
+        # Process jobs in parallel
+        top_values <- .jobs_map_parallel_dfr(chunks, function(chunk) {
+            # Read and preprocess values
+            .raster_open_rast(tile_path) |>
             .raster_get_top_values(
-                band = 1,
-                n = n,
+                block = .block(chunk),
+                band  = 1,
+                n     = n,
                 sampling_window = sampling_window
             ) |>
             dplyr::mutate(
@@ -105,6 +150,7 @@ sits_uncertainty_sampling <- function(uncert_cube,
                 c("longitude", "latitude", "value")
             )) |>
             tibble::as_tibble()
+        })
         # All the cube's uncertainty images have the same start & end dates.
         top_values[["start_date"]] <- .tile_start_date(tile)
         top_values[["end_date"]] <- .tile_end_date(tile)
@@ -174,6 +220,10 @@ sits_uncertainty_sampling <- function(uncert_cube,
 #' @param min_margin      Minimum margin of confidence to select a sample
[…]
