### 2024-05-18 ee6ecb4
include raster support in the detections api
