+++ NAMESPACE
@@ -339,6 +339,7 @@ S3method(sits_cube,local_cube)
 S3method(sits_cube,sar_cube)
 S3method(sits_cube,stac_cube)
 S3method(sits_detect_change,default)
+S3method(sits_detect_change,raster_cube)
 S3method(sits_detect_change,sits)
 S3method(sits_get_data,csv)
 S3method(sits_get_data,data.frame)
