+++ R/api_view.R
@@ -32,11 +32,6 @@
     cube <- .view_filter_tiles(cube, tiles)
     # get the dates
     dates <- .view_set_dates(cube, dates)
-    # find out if resampling is required (for big images)
-    output_size <- .view_resample_size(
-        cube = cube,
-        ndates = max(length(dates), 1)
-    )
     # create a leaflet and add providers
     leaf_map <- .view_add_basic_maps()
     # get names of basic maps
@@ -49,7 +44,6 @@
                 cube = cube,
                 band = band,
                 dates = dates,
-                output_size = output_size,
                 palette = palette
             )
     } else {
@@ -61,8 +55,7 @@
                 red = red,
                 green = green,
                 blue = blue,
-                dates = dates,
-                output_size = output_size
+                dates = dates
             )
     }
     # include class cube if available
@@ -72,8 +65,7 @@
             tiles = tiles,
             legend = legend,
             palette = palette,
-            opacity = opacity,
-            output_size = output_size
+            opacity = opacity
         )
     # get overlay groups
     overlay_groups <- .view_add_overlay_grps(
@@ -133,11 +125,6 @@
                                line_width) {
     # filter the tiles to be processed
     cube <- .view_filter_tiles(cube, tiles)
-    # find out if resampling is required (for big images)
-    output_size <- .view_resample_size(
-        cube = cube,
-        ndates = max(length(dates), 1)
-    )
     # create a leaflet and add providers
     leaf_map <- .view_add_basic_maps()
     # get names of basic maps
@@ -152,7 +139,6 @@
                 cube = cube,
                 band = band,
                 dates = dates,
-                output_size = output_size,
                 palette = palette
             )
     }
@@ -168,8 +154,7 @@
                 red = red,
                 green = green,
                 blue = blue,
-                dates = dates,
-                output_size = output_size
+                dates = dates
             )
     }
     # include segments (and class cube if available)
@@ -471,18 +456,14 @@
 #' @param  band          Band to be shown
 #' @param  dates         Dates to be plotted
 #' @param  palette       Palette to show false colors
-#' @param  output_size   Controls size of leaflet to be visualized
 #' @return               A leaflet object
 #
 .view_bw_band <- function(leaf_map,
                           cube,
                           band,
                           dates,
-                          palette,
-                          output_size) {
+                          palette) {
 
-    # adjust palette for SAR images
-    palette <- .view_adjust_palette(cube, palette)
     # obtain the raster objects for the dates chosen
     for (i in seq_along(dates)) {
         date <- as.Date(dates[[i]])
@@ -504,8 +485,7 @@
                 tile = tile,
                 band = band,
                 date = date,
-                palette = palette,
-                output_size = output_size
+                palette = palette
             )
         }
     }
@@ -523,7 +503,6 @@
 #' @param  green         Band to be shown in green color
 #' @param  blue          Band to be shown in blue color
 #' @param  dates         Dates to be plotted
-#' @param  output_size   Controls size of leaflet to be visualized
 #' @return               A leaflet object
 #
 .view_rgb_bands <- function(leaf_map,
@@ -531,8 +510,11 @@
                             red,
                             green,
                             blue,
-                            dates,
-                            output_size) {
+                            dates) {
[…]
