+++ NAMESPACE
@@ -274,8 +274,6 @@ S3method(.view_add_overlay_grps,class_cube)
 S3method(.view_add_overlay_grps,derived_cube)
 S3method(.view_add_overlay_grps,raster_cube)
 S3method(.view_add_overlay_grps,vector_cube)
-S3method(.view_adjust_palette,default)
-S3method(.view_adjust_palette,sar_cube)
 S3method(plot,class_cube)
 S3method(plot,class_vector_cube)
 S3method(plot,geo_distances)
