+++ DESCRIPTION
@@ -28,10 +28,12 @@ Description: An end-to-end toolkit for land use and land cover classification
     temporal convolutional neural networks proposed by Pelletier et al (2019) <doi:10.3390/rs11050523>, 
     residual networks by Fawaz et al (2019) <doi:10.1007/s10618-019-00619-1>, and temporal attention encoders
     by Garnot and Landrieu (2020) <doi:10.48550/arXiv.2007.00586>.
+    Supports GPU processing of deep learning models using torch <https://torch.mlverse.org/>.
     Performs efficient classification of big Earth observation data cubes and includes 
     functions for post-classification smoothing based on Bayesian inference, and 
-    methods for uncertainty assessment. Enables best
-    practices for estimating area and assessing accuracy of land change as 
+    methods for active learning and uncertainty assessment. Supports object-based 
+    time series analysis using package supercells <https://jakubnowosad.com/supercells/>.
+    Enables best practices for estimating area and assessing accuracy of land change as 
     recommended by Olofsson et al (2014) <doi:10.1016/j.rse.2014.02.015>.
     Minimum recommended requirements: 16 GB RAM and 4 CPU dual-core.
 Encoding: UTF-8
