### 2024-05-20 445a358
Introduce a new version of RADD algorithm
