+++ R/sits_radd.R
@@ -3,7 +3,6 @@ sits_radd <- function(data, pdf, ...,
                       chi = 0.9,
                       start_date = NULL,
                       end_date = NULL) {
-
     UseMethod("sits_radd", data)
 }
 
@@ -97,71 +96,55 @@ sits_radd.raster_cube <- function(data,
         if (!.has(stats_layer)) {
             stats_layer <- .radd_create_stats(data)
         }
-
-        # Check memory and multicores
-        # Get block size
-        block <- .raster_file_blocksize(.raster_open_rast(.tile_path(data)))
-        # Check minimum memory needed to process one block
-        # TODO: verify npaths param
-        job_memsize <- .jobs_memsize(
-            job_size = .block_size(block = block, overlap = 0),
-            npaths = length(.tile_paths(data)),
-            nbytes = 8,
-            proc_bloat = proc_bloat
-        )
-        # Update multicores parameter
-        multicores <- .jobs_max_multicores(
-            job_memsize = job_memsize,
-            memsize = memsize,
-            multicores = multicores
-        )
-        # Update block parameter
-        block <- .jobs_optimal_block(
-            job_memsize = job_memsize,
-            block = block,
-            image_size = .tile_size(.tile(data)),
-            memsize = memsize,
-            multicores = multicores
-        )
-        # Terra requires at least two pixels to recognize an extent as valid
-        # polygon and not a line or point
-        block <- .block_regulate_size(block)
-        # Prepare parallel processing
-        .parallel_start(workers = multicores)
-        on.exit(.parallel_stop(), add = TRUE)
-
-        # Calculate the probability of Non-Forest
-        # Process each tile sequentially
-        probs_cube <- .cube_foreach_tile(data, function(tile) {
-            # Classify the data
-            probs_tile <- .radd_calc_tile(
-                tile = tile,
-                band = "probs",
-                pdf_fn = pdf_fn,
-                stats_layer = stats_layer,
+        predict_fun <- function() {
+            # Check memory and multicores
+            # Get block size
+            block <- .raster_file_blocksize(.raster_open_rast(.tile_path(data)))
+            # Check minimum memory needed to process one block
+            job_memsize <- .jobs_memsize(
+                job_size = .block_size(block = block, overlap = 0),
+                npaths = length(.tile_paths(data)),
+                nbytes = 8,
+                proc_bloat = proc_bloat
+            )
+            # Update multicores parameter
+            multicores <- .jobs_max_multicores(
+                job_memsize = job_memsize,
+                memsize = memsize,
+                multicores = multicores
+            )
+            # Update block parameter
+            block <- .jobs_optimal_block(
+                job_memsize = job_memsize,
                 block = block,
-                impute_fn = impute_fn,
-                output_dir = output_dir,
-                version = version,
-                progress = TRUE
+                image_size = .tile_size(.tile(data)),
+                memsize = memsize,
+                multicores = multicores
             )
-            return(probs_tile)
-        })
+            # Terra requires at least two pixels to recognize an extent as valid
+            # polygon and not a line or point
+            block <- .block_regulate_size(block)
+            # Prepare parallel processing
+            .parallel_start(workers = multicores)
+            on.exit(.parallel_stop(), add = TRUE)
 
-        # Calculate probability for NF
-        data <- .radd_calc_pnf(
-            data = data,
-            pdf_fn = pdf_fn,
-            stats_layer = stats_layer
-        )
-        predict_fun <- function(start_date = NULL, end_date = NULL) {
-            # Now we need to detected the changes
-            data <- .radd_detect_events(
-                data = data,
-                threshold = 0.5,
-                start_date = start_date,
-                end_date = end_date
-            )
[…]
