### 2024-05-20 c67fadb
fix input_pixels argument for compatibility with old models
