+++ R/sits_mlp.R
@@ -277,7 +277,7 @@ sits_mlp <- function(samples = NULL,
             # Unserialize model
             torch_model[["model"]] <- .torch_unserialize_model(serialized_model)
             # Used to check values (below)
-            n_input_pixels <- nrow(values)
+            input_pixels <- nrow(values)
             # Performs data normalization
             values <- .pred_normalize(pred = values, stats = ml_stats)
             # Transform input into matrix
@@ -305,7 +305,7 @@ sits_mlp <- function(samples = NULL,
             )
             # Are the results consistent with the data input?
             .check_processed_values(
-                values = values, n_input_pixels = n_input_pixels
+                values = values, input_pixels = input_pixels
             )
             # Update the columns names to labels
             colnames(values) <- labels
