### 2024-05-22 0719c70
plot and view upgrades
