+++ DESCRIPTION
@@ -60,7 +60,7 @@ Imports:
     sysfonts,
     slider (>= 0.2.0),
     stats,
-    terra (>= 1.7-71),
+    terra (>= 1.7-65),
     tibble (>= 3.1),
     tidyr (>= 1.2.0),
     torch (>= 0.11.0),
