### 2024-05-22 abb39b5
fix M3 chip torch bug
