+++ R/sits_tae.R
@@ -240,6 +240,11 @@ sits_tae <- function(samples = NULL,
                 return(x)
             }
         )
+        # torch 12.0 not working with Apple MPS
+        if (torch::backends_mps_is_available())
+            cpu_train <-  TRUE
+        else
+            cpu_train <-  FALSE
         # train the model using luz
         torch_model <-
             luz::setup(
@@ -273,6 +278,7 @@ sits_tae <- function(samples = NULL,
                         gamma = lr_decay_rate
                     )
                 ),
+                accelerator = luz::accelerator(cpu = cpu_train),
                 dataloader_options = list(batch_size = batch_size),
                 verbose = verbose
             )
