+++ R/sits_dtw.R
@@ -19,8 +19,6 @@
 #'                       used in operations with data cubes.
 #' @return               Change detection method prepared to be passed to
 #'                       \code{\link[sits]{sits_detect_change_method}}
-#' @export
-#'
 sits_dtw <-
     function(samples    = NULL,
              ...,
