+++ NAMESPACE
@@ -337,9 +337,6 @@ S3method(sits_cube,default)
 S3method(sits_cube,local_cube)
 S3method(sits_cube,sar_cube)
 S3method(sits_cube,stac_cube)
-S3method(sits_detect_change,default)
-S3method(sits_detect_change,raster_cube)
-S3method(sits_detect_change,sits)
 S3method(sits_get_data,csv)
 S3method(sits_get_data,data.frame)
 S3method(sits_get_data,default)
@@ -443,9 +440,6 @@ export(sits_config)
 export(sits_config_show)
 export(sits_cube)
 export(sits_cube_copy)
-export(sits_detect_change)
-export(sits_detect_change_method)
-export(sits_dtw)
 export(sits_factory_function)
 export(sits_filter)
 export(sits_formula_linear)
