+++ tests/testthat/test-view.R
@@ -125,12 +125,6 @@ test_that("View", {
     expect_equal(v7$x$calls[[1]]$args[[1]], "GeoportailFrance.orthos")
     expect_equal(v7$x$calls[[5]]$method, "addRasterImage")
 
-    v8 <- sits_view(segments, band = "NDVI")
-    expect_true(grepl("EPSG3857", v8$x$options$crs$crsClass))
-    expect_identical(v8$x$calls[[1]]$method, "addProviderTiles")
-    expect_identical(v8$x$calls[[1]]$args[[1]], "GeoportailFrance.orthos")
-    expect_identical(v8$x$calls[[5]]$method, "addRasterImage")
-    expect_identical(v8$x$calls[[6]]$method, "addPolygons")
 
     probs_segs <- sits_classify(
         data = segments,
@@ -186,7 +180,10 @@ test_that("View", {
                     green = "B16",
                     blue = "B13",
                     dates = "2018-08-29")
-
+    v_cb2 <- sits_view(cbers_cube,
+                      tiles = c("007004", "007005"),
+                      band = "B16",
+                      dates = "2018-08-29")
     expect_identical(v_cb$x$options$crs$crsClass, "L.CRS.EPSG3857")
     expect_identical(v_cb$x$calls[[1]]$args[[1]], "GeoportailFrance.orthos")
     expect_identical(v_cb$x$calls[[5]]$method, "addRasterImage")
