### 2024-05-22 69e50bf
Update NEWS.md
