+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: sits
 Type: Package
-Version: 1.5.0-1
+Version: 1.5.1
 Title: Satellite Image Time Series Analysis for Earth Observation Data Cubes
 Authors@R: c(person('Rolf', 'Simoes', role = c('aut'), email = 'rolf.simoes@inpe.br'),
              person('Gilberto', 'Camara', role = c('aut', 'cre'), email = 'gilberto.camara.inpe@gmail.com'),
