### 2024-05-23 #596 @gilbertocamara
Thanks! Included rOpenSci badge in sits github site.
