### 2024-05-24 4cfb651
improve sits_dtw operations in data cubes
