+++ R/api_view.R
@@ -1,203 +1,4 @@
-#' @title  Include leaflet to view images (BW or RGB)
-#' @name .view_image_raster
-#' @keywords internal
-#' @noRd
-#' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
-#'
-#' @param  cube          Cube to be plotted (raster cube)
-#' @param  class_cube    Classified cube to be overlayed on top on image
-#' @param  tiles         Tiles to be plotted (in case of a multi-tile cube).
-#' @param  dates         Dates to be plotted.
-#' @param  band          For plotting grey images.
-#' @param  red           Band for red color.
-#' @param  green         Band for green color.
-#' @param  blue          Band for blue color.
-#' @param  legend        Named vector that associates labels to colors.
-#' @param  palette       Palette provided in the configuration file.
-#'
-#' @return               A leaflet object.
-#'
-.view_image_raster <- function(cube,
-                        class_cube,
-                        tiles,
-                        dates,
-                        band,
-                        red,
-                        green,
-                        blue,
-                        legend,
-                        palette,
-                        opacity) {
-    # filter the tiles to be processed
-    cube <- .view_filter_tiles(cube, tiles)
-    # get the dates
-    dates <- .view_set_dates(cube, dates)
-    # create a leaflet and add providers
-    leaf_map <- .view_add_base_maps()
-    # add B/W band if required
-    # create a leaflet for B/W bands
-    if (.has(band)) {
-        leaf_map <- leaf_map |>
-            .view_bw_band(
-                cube = cube,
-                band = band,
-                dates = dates,
-                palette = palette
-            )
-    } else {
-        # add RGB bands if required
-        # create a leaflet for RGB bands
-        leaf_map <- leaf_map |>
-            .view_rgb_bands(
-                cube = cube,
-                red = red,
-                green = green,
-                blue = blue,
-                dates = dates
-            )
-    }
-    # include class cube if available
-    leaf_map <- leaf_map |>
-        .view_class_cube(
-            class_cube = class_cube,
-            tiles = tiles,
-            legend = legend,
-            palette = palette,
-            opacity = opacity
-        )
-    # get overlay groups
-    overlay_groups <- .view_add_overlay_grps(
-        cube = cube,
-        dates = dates,
-        class_cube = class_cube
-    )
-    # add layers control to leafmap
-    leaf_map <- leaf_map |>
-        leaflet::addLayersControl(
-            baseGroups = c("ESRI", "GeoPortalFrance",
-                           "Sentinel-2-2020", "OSM"),
-            overlayGroups = overlay_groups,
-            options = leaflet::layersControlOptions(
-                collapsed = FALSE,
-                autoZIndex = TRUE)
-        ) |>
-        # add legend to leaf_map
-        .view_add_legend(
-            cube = class_cube,
-            legend = legend,
-            palette = palette
-        )
-    return(leaf_map)
-}
-#' @title  Include leaflet to view vector images
-#' @name .view_image_vector
-#' @keywords internal
-#' @noRd
-#' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
-#'
-#' @param  cube          Cube to be plotted (vector or raster cube)
-#' @param  tiles         Tiles to be plotted (in case of a multi-tile cube).
-#' @param  dates         Dates to be plotted.
-#' @param  band          For plotting grey images.
-#' @param  red           Band for red color.
-#' @param  green         Band for green color.
-#' @param  blue          Band for blue color.
-#' @param  class_cube    Classified cube to be overlayed on top on image
-#' @param  legend        Named vector that associates labels to colors.
-#' @param  opacity       Opacity of segment fill
-#' @param  palette       Palette provided in the configuration file.
[…]
