+++ NAMESPACE
@@ -271,10 +271,9 @@ S3method(.tile_yres,raster_cube)
 S3method(.values_ts,bands_cases_dates)
 S3method(.values_ts,bands_dates_cases)
 S3method(.values_ts,cases_dates_bands)
-S3method(.view_add_overlay_grps,class_cube)
-S3method(.view_add_overlay_grps,derived_cube)
-S3method(.view_add_overlay_grps,raster_cube)
-S3method(.view_add_overlay_grps,vector_cube)
+S3method(.view_add_overlay_group,derived_cube)
+S3method(.view_add_overlay_group,raster_cube)
+S3method(.view_add_overlay_group,vector_cube)
 S3method(plot,class_cube)
 S3method(plot,class_vector_cube)
 S3method(plot,geo_distances)
@@ -410,8 +409,6 @@ S3method(sits_view,probs_cube)
 S3method(sits_view,raster_cube)
 S3method(sits_view,sits)
 S3method(sits_view,som_map)
-S3method(sits_view,uncertainty_cube)
-S3method(sits_view,vector_cube)
 S3method(summary,class_cube)
 S3method(summary,derived_cube)
 S3method(summary,raster_cube)
