+++ R/sits_config.R
@@ -40,8 +40,18 @@ sits_config <- function(config_user_file = NULL) {
         input = config_internals_file,
         merge.precedence = "override"
     )
-    # set options defined in sits config
+    # set options defined in config_internals
     do.call(.conf_set_options, args = config_internals)
+
+    # load the user-relevant configuration parameters
+    config_file <- .config_file()
+    config_user <- yaml::yaml.load_file(
+        input = config_file,
+        merge.precedence = "override"
+    )
+    # set options defined in config_internals
+    do.call(.conf_set_options, args = config_user)
+
     # load sources configuration
     .conf_load_sources()
     # set the default color table
@@ -65,80 +75,27 @@ sits_config <- function(config_user_file = NULL) {
 #'
 #' @return No return value, called for side effects.
 #' @examples
-#' sits_config_show(source = "BDC")
-#' sits_config_show(source = "BDC", collection = "CBERS-WFI-16D")
+#' sits_config_show()
 #' @export
-sits_config_show <- function(source = NULL,
-                             collection = NULL) {
+sits_config_show <- function() {
     config <- sits_env[["config"]]
 
-    if (!is.null(source)) {
-        # check source value
-        .check_chr(source,
-            allow_empty = FALSE,
-            len_min = 1,
-            len_max = 1
-        )
-        # check source is available
-        source <- toupper(source)
-        .check_chr_within(source,
-            within = .sources(),
-            discriminator = "one_of"
-        )
-        # get the configuration values associated to the source
-        config <- config[[c("sources", source)]]
-        # check collection value
-        if (!is.null(collection)) {
-            .check_chr(collection,
-                allow_empty = FALSE,
-                len_min = 1,
-                len_max = 1
-            )
-            # check collection is available
-            collection <- toupper(collection)
-            .check_chr_within(collection,
-                within = .source_collections(source = source),
-                discriminator = "one_of"
-            )
-            config <- config[[c("collections", collection)]]
-        } else {
-            config <- lapply(config, function(x) {
-                if (is.atomic(x)) {
-                    return(x)
-                }
-                list(names(x))
-            })
-        }
-    } else {
-        config <- lapply(config, function(x) {
-            if (is.atomic(x)) {
-                return(x)
-            }
-            list(names(x))
-        })
-    }
-    config_txt <- yaml::as.yaml(config,
-        indent = 4,
-        handlers = list(
-            character = function(x) {
-                res <- toString(x)
-                class(res) <- "verbatim"
-                res
-            },
-            integer = function(x) {
-                res <- toString(x)
-                class(res) <- "verbatim"
-                res
-            },
-            numeric = function(x) {
-                res <- toString(x)
-                class(res) <- "verbatim"
-                res
-            }
-        )
-    )
-    cat(config_txt, sep = "\n")
-    return(invisible(config))
+    cat("Data sources and user configurable parameters in sits\n\n")
+    cat("Data sources available in sits\n")
+    cat(toString(.sources()))
+    cat("\n\n")
+    cat("Use sits_list_collections(<source>) to get details for each source\n\n")
+
+    cat("User configurable parameters for plotting\n")
+    config_plot <- sits_env[["config"]][["plot"]]
+    .conf_list_params(config_plot)
+
+    cat("User configurable parameters for visualisation\n")
+    config_view <- sits_env[["config"]][["view"]]
+    .conf_list_params(config_view)
+
+    cat("User sits_config_user_file() to create a user configuration file")
+    return(invisible(NULL))
 }
 
 #' @title List the cloud collections supported by sits
@@ -173,35 +130,61 @@ sits_list_collections <- function(source = NULL) {
[…]
