+++ NAMESPACE
@@ -438,6 +438,7 @@ export(sits_combine_predictions)
 export(sits_confidence_sampling)
 export(sits_config)
 export(sits_config_show)
+export(sits_config_user_file)
 export(sits_cube)
 export(sits_cube_copy)
 export(sits_detect_change)

[…]
