+++ R/sits_merge.R
@@ -204,17 +204,21 @@ sits_merge.raster_cube <- function(data1, data2, ...) {
 }
 
 .merge_single_timeline <- function(data1, data2){
-    # Get data1 timeline
+    # Get data1 timeline.
     d1_tl <- unique(as.Date(.cube_timeline(data1)[[1]]))
-    fi_new <- purrr::map(1:nrow(data2), function(idx) {
-        d <- data2[idx,]
-        fi <- .fi(data2[idx,])
-        fi[["date"]] <- as.Date(d1_tl[1:nrow(d)])
+    # Create new `file_info` using dates from `data1` timeline.
+    fi_new <- purrr::map(seq_len(nrow(data2)), function(row) {
+        data_row <- data2[row,]
+
+        fi <- .fi(data_row)
+        fi[["date"]] <- as.Date(d1_tl[1:nrow(data_row)])
+
         return(fi)
     })
+    # Assign the new `file_into` into `data2`
     data2[["file_info"]] <- fi_new
-    data1 <- .cube_merge(data1, data2)
-    return(data1)
+    # Merge cubes and return
+    .cube_merge(data1, data2)
 }
 
 #' @rdname sits_merge
