### 2024-05-28 16eacce
implement interp version of radd
