+++ src/RcppExports.cpp
@@ -273,30 +273,45 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// C_radd_calc_nf
-arma::mat C_radd_calc_nf(const arma::mat& ts, const arma::mat& mean, const arma::mat& sd, const arma::uword& n_times);
-RcppExport SEXP _sits_C_radd_calc_nf(SEXP tsSEXP, SEXP meanSEXP, SEXP sdSEXP, SEXP n_timesSEXP) {
+// C_radd_calc_sub
+arma::rowvec C_radd_calc_sub(const arma::mat& x, const arma::mat& y);
+RcppExport SEXP _sits_C_radd_calc_sub(SEXP xSEXP, SEXP ySEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const arma::mat& >::type ts(tsSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type x(xSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type y(ySEXP);
+    rcpp_result_gen = Rcpp::wrap(C_radd_calc_sub(x, y));
+    return rcpp_result_gen;
+END_RCPP
+}
+// C_radd_calc_nf
+arma::mat C_radd_calc_nf(arma::mat& ts, const arma::mat& mean, const arma::mat& sd, const arma::uword& n_times, const arma::mat& deseasonlize_values);
+RcppExport SEXP _sits_C_radd_calc_nf(SEXP tsSEXP, SEXP meanSEXP, SEXP sdSEXP, SEXP n_timesSEXP, SEXP deseasonlize_valuesSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat& >::type ts(tsSEXP);
     Rcpp::traits::input_parameter< const arma::mat& >::type mean(meanSEXP);
     Rcpp::traits::input_parameter< const arma::mat& >::type sd(sdSEXP);
     Rcpp::traits::input_parameter< const arma::uword& >::type n_times(n_timesSEXP);
-    rcpp_result_gen = Rcpp::wrap(C_radd_calc_nf(ts, mean, sd, n_times));
+    Rcpp::traits::input_parameter< const arma::mat& >::type deseasonlize_values(deseasonlize_valuesSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_radd_calc_nf(ts, mean, sd, n_times, deseasonlize_values));
     return rcpp_result_gen;
 END_RCPP
 }
 // C_radd_detect_changes
-arma::mat C_radd_detect_changes(const arma::mat& p_res, const double& threshold, const double& chi);
-RcppExport SEXP _sits_C_radd_detect_changes(SEXP p_resSEXP, SEXP thresholdSEXP, SEXP chiSEXP) {
+arma::mat C_radd_detect_changes(const arma::mat& p_res, const arma::uword& start, const arma::uword& end, const double& threshold, const double& chi);
+RcppExport SEXP _sits_C_radd_detect_changes(SEXP p_resSEXP, SEXP startSEXP, SEXP endSEXP, SEXP thresholdSEXP, SEXP chiSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
[…]
