### 2024-05-28 0710675
update raster API and add RADD metadata in config
