+++ inst/extdata/config_internals.yml
@@ -170,6 +170,16 @@ derived_cube         :
                 maximum_value: 254
                 offset_value : 0
                 scale_factor : 1
+    radd_cube        :
+        s3_class     : [ "radd_cube", "derived_cube", "raster_cube" ]
+        bands        :
+            radd     :
+                data_type    : "INT2S"
+                missing_value: 0
+                minimum_value: 1
+                maximum_value: 32768
+                offset_value : 0
+                scale_factor : 1
 
 # Vector cube definitions
 vector_cube          :
