+++ NAMESPACE
@@ -108,6 +108,7 @@ S3method(.raster_nlayers,terra)
 S3method(.raster_nrows,terra)
 S3method(.raster_open_rast,terra)
 S3method(.raster_polygonize,terra)
+S3method(.raster_quantile,terra)
 S3method(.raster_rast,terra)
 S3method(.raster_read_rast,terra)
 S3method(.raster_row,terra)
