### 2024-05-28 552c550
include dem product from mpc
