+++ NAMESPACE
@@ -137,6 +137,7 @@ S3method(.source_collection_access_test,usgs_cube)
 S3method(.source_cube,stac_cube)
 S3method(.source_filter_tiles,"cdse_cube_sentinel-1-rtc")
 S3method(.source_filter_tiles,"deafrica_cube_sentinel-1-rtc")
+S3method(.source_filter_tiles,"mpc_cube_cop-dem-glo-30")
 S3method(.source_filter_tiles,"mpc_cube_sentinel-1-grd")
 S3method(.source_filter_tiles,stac_cube)
 S3method(.source_item_get_bands,stac_cube)
@@ -157,6 +158,7 @@ S3method(.source_items_fid,stac_cube)
 S3method(.source_items_new,"aws_cube_landsat-c2-l2")
 S3method(.source_items_new,"deafrica_cube_sentinel-1-rtc")
 S3method(.source_items_new,"deafrica_cube_sentinel-2-l2a")
+S3method(.source_items_new,"mpc_cube_cop-dem-glo-30")
 S3method(.source_items_new,"mpc_cube_landsat-c2-l2")
 S3method(.source_items_new,"mpc_cube_sentinel-1-grd")
 S3method(.source_items_new,"mpc_cube_sentinel-1-rtc")
@@ -172,6 +174,7 @@ S3method(.source_items_tile,"aws_cube_landsat-c2-l2")
 S3method(.source_items_tile,"cdse_cube_sentinel-1-rtc")
 S3method(.source_items_tile,"deafrica_cube_rainfall-chirps-daily")
 S3method(.source_items_tile,"deafrica_cube_rainfall-chirps-monthly")
+S3method(.source_items_tile,"mpc_cube_cop-dem-glo-30")
 S3method(.source_items_tile,"mpc_cube_landsat-c2-l2")
 S3method(.source_items_tile,"mpc_cube_sentinel-1-grd")
 S3method(.source_items_tile,"mpc_cube_sentinel-1-rtc")
@@ -187,6 +190,7 @@ S3method(.source_roi_tiles,"mpc_cube_landsat-c2-l2")
 S3method(.source_roi_tiles,sdc_cube)
 S3method(.source_roi_tiles,stac_cube)
 S3method(.source_tile_get_bbox,"cdse_cube_sentinel-1-rtc")
+S3method(.source_tile_get_bbox,"mpc_cube_cop-dem-glo-30")
 S3method(.source_tile_get_bbox,"mpc_cube_sentinel-1-grd")
 S3method(.source_tile_get_bbox,"mpc_cube_sentinel-1-rtc")
 S3method(.source_tile_get_bbox,stac_cube)
