### 2024-05-29 fb85f44
add new function to remove specific samples of a class inside a neuron of other class
