+++ NAMESPACE
@@ -488,6 +488,7 @@ export(sits_smooth)
 export(sits_som_clean_samples)
 export(sits_som_evaluate_cluster)
 export(sits_som_map)
+export(sits_som_remove_samples)
 export(sits_stats)
 export(sits_stratified_sampling)
 export(sits_svm)
