### 2024-05-29 02d7160
add dem_cube regularization support
