+++ NAMESPACE
@@ -126,6 +126,7 @@ S3method(.raster_xres,terra)
 S3method(.raster_ymax,terra)
 S3method(.raster_ymin,terra)
 S3method(.raster_yres,terra)
+S3method(.reg_s2tile_convert,dem_cube)
 S3method(.reg_s2tile_convert,grd_cube)
 S3method(.reg_s2tile_convert,rtc_cube)
 S3method(.slice_dfr,numeric)
@@ -377,6 +378,7 @@ S3method(sits_reclassify,default)
 S3method(sits_reduce,raster_cube)
 S3method(sits_reduce,sits)
 S3method(sits_regularize,default)
+S3method(sits_regularize,dem_cube)
 S3method(sits_regularize,derived_cube)
 S3method(sits_regularize,raster_cube)
 S3method(sits_regularize,sar_cube)
