+++ inst/extdata/sources/config_source_chile.yml
@@ -1,0 +1,100 @@
+# These are configuration parameters that can be set by users
+# The parameters enable access to the cloud collections
+
+sources:
+    DCCHILE                 :
+        s3_class            : ["aws_cube", "stac_cube", "eo_cube",
+                               "raster_cube"]
+        service             : "STAC"
+        url                 : "https://earth-search.aws.element84.com/v1/"
+        collections         :
+            SENTINEL-2-L2A  :
+                bands       :
+                    B01     : &aws_msi_60m
+                        missing_value: -9999
+                        minimum_value: 0
+                        maximum_value: 10000
+                        scale_factor : 0.0001
+                        offset_value : 0
+                        resolution  :  60
+                        band_name    : "coastal"
+                        data_type     : "INT2S"
+                    B02     : &aws_msi_10m
+                        missing_value: -9999
+                        minimum_value: 0
+                        maximum_value: 10000
+                        scale_factor : 0.0001
+                        offset_value : 0
+                        resolution  :  10
+                        band_name    : "blue"
+                        data_type     : "INT2S"
+                    B03     :
+                        <<: *aws_msi_10m
+                        band_name    : "green"
+                    B04     :
+                        <<: *aws_msi_10m
+                        band_name    : "red"
+                    B05     : &aws_msi_20m
+                        missing_value: -9999
+                        minimum_value: 0
+                        maximum_value: 10000
+                        scale_factor : 0.0001
+                        offset_value : 0
+                        resolution  :  20
+                        band_name    : "rededge1"
+                        data_type     : "INT2S"
+                    B06     :
+                        <<: *aws_msi_20m
+                        band_name    : "rededge2"
+                    B07     :
+                        <<: *aws_msi_20m
+                        band_name    : "rededge3"
+                    B08     :
+                        <<: *aws_msi_10m
+                        band_name    : "nir"
+                    B8A     :
+                        <<: *aws_msi_20m
+                        band_name    : "nir08"
+                    B09     :
+                        <<: *aws_msi_60m
+                        band_name    : "nir09"
+                    B11     :
+                        <<: *aws_msi_20m
+                        band_name    : "swir16"
+                    B12     :
+                        <<: *aws_msi_20m
+                        band_name    : "swir22"
+                    CLOUD   :
+                        bit_mask     : false
+                        band_name    : "scl"
+                        values       :
+                            0        : "missing_data"
+                            1        : "defective pixel"
+                            2        : "shadows"
+                            3        : "cloud shadows"
+                            4        : "vegetation"
+                            5        : "non-vegetated"
+                            6        : "water"
+                            7        : "unclassified"
+                            8        : "cloud medium"
+                            9        : "cloud high"
+                            10       : "thin cirrus"
+                            11       : "snow or ice"
+                        interp_values: [0, 1, 2, 3, 8, 9, 10, 11]
+                        resolution  : 20
+                        data_type     : "INT1U"
+                satellite   : "SENTINEL-2"
+                sensor      : "MSI"
+                platforms   :
+                    SENTINEL-2A: "sentinel-2a"
+                    SENTINEL-2B: "sentinel-2b"
+                collection_name: "sentinel-2-l2a"
+                access_vars :
+                   AWS_DEFAULT_REGION   : "us-west-2"
+                   AWS_S3_ENDPOINT      : "s3.amazonaws.com"
+                   AWS_NO_SIGN_REQUEST  : true
+                open_data   : true
+                open_data_token: false
+                metadata_search : "tile"
+                ext_tolerance: 0
+                grid_system  : "MGRS"
