+++ DESCRIPTION
@@ -87,6 +87,7 @@ Suggests:
     leafem (>= 0.2.0),
     leaflet (>= 2.2.0),
     luz (>= 0.4.0),
+    magrittr, 
     methods,
     mgcv,
     nnet,
