### 2024-05-30 4bf6ba4
update RADD implementation code
