+++ src/radd_fns.cpp
@@ -32,7 +32,8 @@ arma::mat C_radd_calc_nf(arma::mat& ts,
                          const arma::mat& mean,
                          const arma::mat& sd,
                          const arma::uword& n_times,
-                         const arma::mat& deseasonlize_values) {
+                         const arma::mat& quantile_values,
+                         const arma::vec& bwf) {
 
     // Using the first element as dummy value
     arma::mat p_res(ts.n_rows, n_times + 1, arma::fill::value(0.5));
@@ -52,10 +53,10 @@ arma::mat C_radd_calc_nf(arma::mat& ts,
         // For each band
         for (arma::uword c = 0; c < ts.n_cols; c = c + n_times) {
             // Deseasonlize time series
-            if (deseasonlize_values.size() > 1) {
+            if (quantile_values.size() > 1) {
                 ts.submat(i, c, i, c + n_times - 1) = C_radd_calc_sub(
                     ts.submat(i, c, i, c + n_times - 1),
-                    deseasonlize_values.submat(0, c, 0, c + n_times - 1)
+                    quantile_values.submat(0, c, 0, c + n_times - 1)
                 );
             }
             // Estimate a normal distribution based on Forest stats
@@ -77,12 +78,11 @@ arma::mat C_radd_calc_nf(arma::mat& ts,
                 p_nfor.elem(arma::find(p_nfor > 0)),
                 p_for.elem(arma::find(p_nfor > 0))
             );
-            // Fix the range of prob values between 0 and 1
-            // TODO: use parameter bwf
-            p_nfor.elem(arma::find(p_nfor < 0.1)).fill(0.1);
-            p_nfor.elem(arma::find(p_nfor > 0.9)).fill(0.9);
+            // Fix the range of prob values
+            p_nfor.elem(arma::find(p_nfor < bwf(0))).fill(bwf(0));
+            p_nfor.elem(arma::find(p_nfor > bwf(1))).fill(bwf(1));
 
-            // Update NF prob with a Bayesian approach
+            // Update NF probabilities with a Bayesian approach
             if (update_res) {
                 p_nfor = C_radd_calc_pbayes(p_nfor, p_nfor_past);
             }
@@ -111,110 +111,6 @@ arma::vec seq_int(const arma::uword& from,
     return aux;
 }
 
-// [[Rcpp::export]]
-arma::mat C_radd_detect_changes(const arma::mat& p_res,
-                                const arma::uword& start,
-                                const arma::uword& end,
-                                const double& threshold = 0.5,
-                                const double& chi = 0.9) {
-
-    arma::mat res(
-            p_res.n_rows, 1, arma::fill::value(arma::datum::nan)
-    );
-    arma::mat p_flag(
-            p_res.n_rows, p_res.n_cols, arma::fill::value(arma::datum::nan)
-    );
-    arma::mat p_change(
-            p_res.n_rows, p_res.n_cols, arma::fill::value(arma::datum::nan)
-    );
-    arma::rowvec p_flag_aux(
-            p_res.n_cols, arma::fill::value(arma::datum::nan)
-    );
-    arma::uvec idx_value_res;
-    arma::uword v;
-    bool next_pixel;
-    for (arma::uword i = 0; i < p_res.n_rows; i++) {
-        // create an auxiliary matrix
-        p_flag_aux.fill(arma::datum::nan);
-        // set to zero in the past time
-        p_flag_aux.row(0).col(start - 1) = 0;
-        p_flag_aux.elem(
-            arma::find(p_res.submat(i, 0, i, p_res.n_cols - 1) < threshold)
-        ).zeros();
-        p_flag.row(i) = p_flag_aux;
-
[…]
