+++ R/RcppExports.R
@@ -85,24 +85,12 @@ C_radd_calc_sub <- function(x, y) {
     .Call(`_sits_C_radd_calc_sub`, x, y)
 }
 
-C_radd_calc_nf <- function(ts, mean, sd, n_times, deseasonlize_values) {
-    .Call(`_sits_C_radd_calc_nf`, ts, mean, sd, n_times, deseasonlize_values)
+C_radd_calc_nf <- function(ts, mean, sd, n_times, quantile_values, bwf) {
+    .Call(`_sits_C_radd_calc_nf`, ts, mean, sd, n_times, quantile_values, bwf)
 }
 
-C_radd_detect_changes <- function(p_res, start, end, threshold = 0.5, chi = 0.9) {
-    .Call(`_sits_C_radd_detect_changes`, p_res, start, end, threshold, chi)
-}
-
-C_select_cols <- function(m, row, idx) {
-    .Call(`_sits_C_select_cols`, m, row, idx)
-}
-
-C_vec_select_cols <- function(m, idx) {
-    .Call(`_sits_C_vec_select_cols`, m, idx)
-}
-
-C_radd_detect_changes_2 <- function(p_res, start, end, threshold = 0.5, chi = 0.9) {
-    .Call(`_sits_C_radd_detect_changes_2`, p_res, start, end, threshold, chi)
+C_radd_detect_changes <- function(p_res, start_detection, end_detection, threshold = 0.5, chi = 0.9) {
+    .Call(`_sits_C_radd_detect_changes`, p_res, start_detection, end_detection, threshold, chi)
 }
 
 C_temp_max <- function(mtx) {

[…]
