### 2024-05-31 504796b
update radd algorithm
