### 2024-05-31 f5c7d0e
add option to merge irregular cubes in sits_merge
