### 2024-05-31 92c706f
update config_internals
