+++ inst/extdata/config_internals.yml
@@ -144,10 +144,10 @@ derived_cube         :
         s3_class     : [ "radd_cube", "derived_cube", "raster_cube" ]
         bands        :
             radd     :
-                data_type    : "INT2S"
+                data_type    : "INT4U"
                 missing_value: 0
                 minimum_value: 1
-                maximum_value: 32768
+                maximum_value: 4999365
                 offset_value : 0
                 scale_factor : 1
 
