+++ src/RcppExports.cpp
@@ -413,8 +413,47 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// definitelyGreaterThan
+bool definitelyGreaterThan(float a, float b, float epsilon);
+RcppExport SEXP _sits_definitelyGreaterThan(SEXP aSEXP, SEXP bSEXP, SEXP epsilonSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< float >::type a(aSEXP);
+    Rcpp::traits::input_parameter< float >::type b(bSEXP);
+    Rcpp::traits::input_parameter< float >::type epsilon(epsilonSEXP);
+    rcpp_result_gen = Rcpp::wrap(definitelyGreaterThan(a, b, epsilon));
+    return rcpp_result_gen;
+END_RCPP
+}
+// approximatelyEqual
+bool approximatelyEqual(float a, float b, float epsilon);
+RcppExport SEXP _sits_approximatelyEqual(SEXP aSEXP, SEXP bSEXP, SEXP epsilonSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< float >::type a(aSEXP);
+    Rcpp::traits::input_parameter< float >::type b(bSEXP);
+    Rcpp::traits::input_parameter< float >::type epsilon(epsilonSEXP);
+    rcpp_result_gen = Rcpp::wrap(approximatelyEqual(a, b, epsilon));
+    return rcpp_result_gen;
+END_RCPP
+}
+// essentiallyEqual
+bool essentiallyEqual(float a, float b, float epsilon);
+RcppExport SEXP _sits_essentiallyEqual(SEXP aSEXP, SEXP bSEXP, SEXP epsilonSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< float >::type a(aSEXP);
+    Rcpp::traits::input_parameter< float >::type b(bSEXP);
+    Rcpp::traits::input_parameter< float >::type epsilon(epsilonSEXP);
+    rcpp_result_gen = Rcpp::wrap(essentiallyEqual(a, b, epsilon));
+    return rcpp_result_gen;
+END_RCPP
+}
 // C_radd_detect_changes
-arma::mat C_radd_detect_changes(const arma::mat& p_res, const arma::uword& start_detection, const arma::uword& end_detection, const double& threshold, const double& chi);
+arma::mat C_radd_detect_changes(const arma::mat& p_res, const arma::uword& start_detection, const arma::uword& end_detection, const double& threshold, float chi);
 RcppExport SEXP _sits_C_radd_detect_changes(SEXP p_resSEXP, SEXP start_detectionSEXP, SEXP end_detectionSEXP, SEXP thresholdSEXP, SEXP chiSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -423,7 +462,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const arma::uword& >::type start_detection(start_detectionSEXP);
     Rcpp::traits::input_parameter< const arma::uword& >::type end_detection(end_detectionSEXP);
[…]
