### 2024-06-03 11dfb09
new function sits_add_base_cube
