+++ DESCRIPTION
@@ -207,6 +207,7 @@ Collate:
     'RcppExports.R'
     'data.R'
     'sits-package.R'
+    'sits_add_base_cube.R'
     'sits_apply.R'
     'sits_accuracy.R'
     'sits_active_learning.R'
