+++ NAMESPACE
@@ -426,6 +426,7 @@ export("sits_labels<-")
 export(impute_linear)
 export(sits_accuracy)
 export(sits_accuracy_summary)
+export(sits_add_base_cube)
 export(sits_apply)
 export(sits_as_sf)
 export(sits_bands)
