+++ R/sits_radd.R
@@ -1,7 +1,80 @@
+#' @title Detection disturbance in combined time series or data cubes
+#' @name sits_radd
+#' @author Felipe Carvalho, \email{lipecaso@@gmail.com}
+#' @author Felipe Carlos, \email{efelipecarlos@@gmail.com}
+#' @author Gilberto Camara, \email{gilberto.camara@@inpe.br}
+#'
+#' @description
+#' This function implements the algorithm described by Johanes Reiche
+#' referenced below.
+#'
+#' @references Reiche J, De Bruin S, Hoekman D, Verbesselt J, Herold M,
+#' "A Bayesian approach to combine Landsat and ALOS PALSAR time
+#' series for near real-time deforestation detection.",
+#' Remote Sensing, 7, 2015 DOI: 10.3390/rs70504973.
+#'
+#'
+#' @param  data              Data cube (tibble of class "raster_cube")
+#' @param  mean_stats        A tibble with mean value of each band.
+#' @param sd_stats           A tibble with the standard deviation
+#'                           value of each band.
+#' @param  ...               Other parameters for specific functions.
+#' @param  impute_fn         Imputation function to remove NA.
+#' @param  roi               Region of interest (either an sf object, shapefile,
+#'                           or a numeric vector with named XY values
+#'                           ("xmin", "xmax", "ymin", "ymax") or
+#'                           named lat/long values
+#'                           ("lon_min", "lat_min", "lon_max", "lat_max").
+#' @param  start_date        Start date for the detection
+#'                           (Date in YYYY-MM-DD format).
+#' @param  end_date          End date for the dectection
+#'                           (Date im YYYY-MM-DD format).
+#' @param  memsize           Memory available for classification in GB
+#'                           (integer, min = 1, max = 16384).
+#' @param  multicores        Number of cores to be used for classification
+#'                           (integer, min = 1, max = 2048).
+#' @param  deseasonlize      A numeric value with the quantile percentage to
+#'                           deseasonlize time series using spatial
+#'                           normalization.
+#' @param threshold          A numeric value with threshold of the probability
+#'                           of Non-Forest above which the first observation
+#'                           is flagged. Default = 0.5.
+#' @param bwf                A numeric vector with the block weighting function
+#'                           to truncate the Non-Forest probability.
+#'                           Default = (0.1, 0.9).
+#' @param  chi               A numeric with threshold of the probability
+#'                           change at which the change is confirmed.
+#'                           Default = 0.5.
+#' @param  output_dir        Valid directory for output file.
+#'                           (character vector of length 1).
+#' @param  version           Version of the output
+#'                           (character vector of length 1).
+#' @param  verbose           Logical: print information about processing time?
+#' @param  progress          Logical: Show progress bar?
+#'
+#' @return                   Time series with detection dates for
+#'                           each point (tibble of class "sits")
+#'                           or a data cube with the detection day of the year
+#'                           for each pixel
+#'                           (tibble of class "radd_cube").
+#'
+#' @note
+#'    The \code{roi} parameter defines a region of interest. It can be
+#'    an sf_object, a shapefile, or a bounding box vector with
+#'    named XY values (\code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}) or
+#'    named lat/long values (\code{lon_min}, \code{lon_max},
+#'    \code{lat_min}, \code{lat_max})
+#'
+#'    Parameter \code{memsize} controls the amount of memory available
+#'    for classification, while \code{multicores}  defines the number of cores
+#'    used for processing. We recommend using as much memory as possible.
[…]
