+++ man/sits_radd.Rd
@@ -1,0 +1,130 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sits_radd.R
+\name{sits_radd}
+\alias{sits_radd}
+\alias{sits_radd.raster_cube}
+\title{Detection disturbance in combined time series or data cubes}
+\usage{
+sits_radd(
+  data,
+  mean_stats,
+  sd_stats,
+  ...,
+  chi = 0.9,
+  start_date = NULL,
+  end_date = NULL
+)
+
+\method{sits_radd}{raster_cube}(
+  data,
+  mean_stats,
+  sd_stats,
+  ...,
+  impute_fn = identity,
+  roi = NULL,
+  start_date = NULL,
+  end_date = NULL,
+  memsize = 8L,
+  multicores = 2L,
+  deseasonlize = 0.95,
+  threshold = 0.5,
+  bwf = c(0.1, 0.9),
+  chi = 0.9,
+  output_dir,
+  version = "v1",
+  progress = TRUE
+)
+}
+\arguments{
+\item{data}{Data cube (tibble of class "raster_cube")}
+
+\item{mean_stats}{A tibble with mean value of each band.}
+
+\item{sd_stats}{A tibble with the standard deviation
+value of each band.}
+
+\item{...}{Other parameters for specific functions.}
+
+\item{chi}{A numeric with threshold of the probability
+change at which the change is confirmed.
+Default = 0.5.}
+
+\item{start_date}{Start date for the detection
+(Date in YYYY-MM-DD format).}
+
+\item{end_date}{End date for the dectection
+(Date im YYYY-MM-DD format).}
+
+\item{impute_fn}{Imputation function to remove NA.}
+
+\item{roi}{Region of interest (either an sf object, shapefile,
+or a numeric vector with named XY values
+("xmin", "xmax", "ymin", "ymax") or
+named lat/long values
+("lon_min", "lat_min", "lon_max", "lat_max").}
+
+\item{memsize}{Memory available for classification in GB
+(integer, min = 1, max = 16384).}
+
+\item{multicores}{Number of cores to be used for classification
+(integer, min = 1, max = 2048).}
+
+\item{deseasonlize}{A numeric value with the quantile percentage to
+deseasonlize time series using spatial
+normalization.}
+
+\item{threshold}{A numeric value with threshold of the probability
+of Non-Forest above which the first observation
+is flagged. Default = 0.5.}
+
+\item{bwf}{A numeric vector with the block weighting function
+to truncate the Non-Forest probability.
+Default = (0.1, 0.9).}
+
+\item{output_dir}{Valid directory for output file.
+(character vector of length 1).}
+
+\item{version}{Version of the output
+(character vector of length 1).}
+
+\item{progress}{Logical: Show progress bar?}
+
+\item{verbose}{Logical: print information about processing time?}
+}
+\value{
+Time series with detection dates for
+                          each point (tibble of class "sits")
+                          or a data cube with the detection day of the year
+                          for each pixel
+                          (tibble of class "radd_cube").
+}
+\description{
+This function implements the algorithm described by Johanes Reiche
+referenced below.
+}
+\note{
+The \code{roi} parameter defines a region of interest. It can be
+   an sf_object, a shapefile, or a bounding box vector with
+   named XY values (\code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}) or
+   named lat/long values (\code{lon_min}, \code{lon_max},
+   \code{lat_min}, \code{lat_max})
+
+   Parameter \code{memsize} controls the amount of memory available
+   for classification, while \code{multicores}  defines the number of cores
[…]
