+++ NAMESPACE
@@ -375,6 +375,7 @@ S3method(sits_mixture_model,raster_cube)
 S3method(sits_mixture_model,sits)
 S3method(sits_mixture_model,tbl_df)
 S3method(sits_model_export,sits_model)
+S3method(sits_radd,raster_cube)
 S3method(sits_reclassify,class_cube)
 S3method(sits_reclassify,default)
 S3method(sits_reduce,raster_cube)
@@ -477,6 +478,7 @@ export(sits_pred_normalize)
 export(sits_pred_references)
 export(sits_pred_sample)
 export(sits_predictors)
+export(sits_radd)
 export(sits_reclassify)
 export(sits_reduce)
 export(sits_reduce_imbalance)
