### 2024-06-06 7498cb9
update RADD algorithm
