### 2024-06-06 e974f3e
fix bug in sits_radd pdf param
