### 2024-06-07 #596 @gilbertocamara
Dear @mpadge @paleolimbot @mikemahoney218
Many thanks to @mpadge for motivating two reviewers and to @paleolimbot and @mikemahoney for accepting to do it.

May I suggest to @paleolimbot and @mikemahoney  that they start the software review by browsing the online book (https://e-sensing.github.io/sitsbook/). The book is a self-learning document and takes a step-by-step approach with reproducible code.

Unlike most R packages, which are a collection of mostly independent functions, the `sits` API is an end-to-end solution for analysing big Earth observation data. Thus, the functions have to be executed in a well-defined order. Looking at the scripts in the online book will give one a better sense of how the package has been organised.

Again, many thanks to @paleolimbot and @mikemahoney for your generosity.
