### 2024-08-24 #596 @adamhsparks
Hi @paleolimbot and @mikemahoney218, just checking in to see how you're going with your reviews that were due at the beginning of this month as the current EIC. Please just let us know if you need some more time so we can update this issue.
