### 2024-09-11 #596 @gilbertocamara
Dear @mikemahoney218 many thanks for your time and effort! Please note that `sits` is now on version 1.5.2. on CRAN. Before looking at the code, may I suggest you browse the [online book](https://e-sensing.github.io/sitsbook/) that describes the package. Some additional tips that might help you:

(1) `sits` differs from most traditional R packages, which are collections of functions that can be used independently.
It is an end-to-end solution for Earth observation (EO) land use and land cover classification. As such, there is a certain preferred order of function execution. Please see the [Introduction](https://e-sensing.github.io/sitsbook/introduction.html) chapter of the online book for more details.

(2) `sits` is more than a collection of machine learning algorithms for EO analytics. Much effort went into providing access to EO cloud collection and organising analysis-ready data (ARD) images into data cubes. In `R`, there is nothing similar. Even compared with EO-ML packages in Python, `sits` is the only software that allows easy access to cloud collections.

(3) To access EO cloud collections, `sits` works together with [`rstac`](https://brazil-data-cube.github.io/rstac/), which is the main support in `R` for STAC collection access. `rstac` has also been developed by our team.

(4) A useful [comparison table](https://torchgeo.readthedocs.io/en/stable/user/alternatives.html) has been prepared by the Python TorchGEO team. It provides a fair appraisal of the ML capabilities of different EO-ML packages.

(5) As far as we know, the algorithms for [SOM-based training data quality control](https://e-sensing.github.io/sitsbook/improving-the-quality-of-training-samples.html) and [Bayesian inference for post-processing](https://e-sensing.github.io/sitsbook/bayesian-smoothing-for-post-processing.html) are only available on `sits`.

(6) Much thought and effort went into designing the `sits` API. Please see a [comparison of the `sits` API and that of the Google Earth Engine](https://e-sensing.github.io/sitsbook/technical-annex.html#sits-and-google-earth-engine-apis-a-side-by-side-exploration).
