### 2024-10-21 #596 @gilbertocamara
Dear @paleolimbot @mikemahoney218 any news on your review of the `sits` package? Many thanks in advance.
