### 2024-10-21 #596 @mikemahoney218
Hi @gilbertocamara , and deep apologies for the radio silence here. I'm about 8-10 hours into my review so far and am still making progress. That's not to say I'm finding a lot of concerns, but rather reviewing the book and the package has become quite a bit more of a task than anticipated.

Sorry to not have more for you this instant, other than "I've started and am making progress". I'll post comments on the book once they're finished and then will follow with package-specific comments after that.
