### 2024-10-30 #596 @mpadge
@paleolimbot Could you please respond to Gilberto's [question above](https://github.com/ropensci/software-review/issues/596#issuecomment-2426194034)? Thanks
