### 2025-01-30 #596 @paleolimbot
Apologies to all for not responding here for a very long time! As you may have suspected, my ability to put adequate time into a thorough review was really just not there, and I apologize for not responding earlier to at least say it 😬

I did take a severely time-boxed look through the book, package code, and issues this evening and I'm happy to share my thoughts (take them with a grain of salt given the amount of content that is here and the amount of time I spent reviewing it).

First, it's great! You have a well-organized repository with a number of contributors, a documented set of design principles, and files with names that make sense that contain reasonable amounts of code. Everything is documented (there's even a book, for crying out loud!), and the formatting of the code itself does not distract from its content. There's a clarity of vision here and that comes through in all aspects of the repo content.

One of the challenges with providing comments on a well-established package with several releases, users, and existing contributors, is that substantial comments about design are sort of unaddressable (because they would break compatibility for existing users). In addition to the fact that there is just a lot of content here to cover well, I have to admit that was also a factor in me taking so long to get here (i.e., do you need me!?).

The one challenge I will make here is one of scope: this is a package that does a *lot* of things. I don't think you should change that (for your existing users), but it may be worth thinking about whether there are some self-contained pieces of functionality that are useful outside of specifically satellite imagery or whose rate of change might differ from other parts of the package (e.g., collections of imagery are necessarily something that require updating as the APIs progress or new collections become available; a data frame-based raster time series model has much broader applicability).

Again with apologies for getting here in an inexcusably late (partial/cursory) review here 😬
