### 2025-01-30 #596 @gilbertocamara
Dear @paleolimbot, many thanks for your kind comments. Regarding the idea of breaking `sits` in smaller, self-contained packages, it is a question that the developers have discussed a lot. Ultimately, we decided to keep `sits` together for the following reasons:
- `sits` is an operational package for end-users, providing an end-to-end workflow. Many of these users have limited knowledge of `R`, so it is easier to follow examples such as those in the book. They don't have the mindset of `R` experts, which know to combine pieces from different packages.
- Considering the challenges of documentation, s/w maintenance, and dealing with the CRAN policy, it is easier to have a single package. Also, there is a single point of communication between users and developers, which is the `issues` interface in `github`.
- There is also the issue of branding. Many of perspective `sits` users are not part of the `R` community. In fact, most EO (Earth observation) package developers work in `Python`. To attract non-R experts to use `sits`, we must build a reputation in the global EO community. Arguably, having a single brand (`sits`) projects a stronger image to the global community than a number of smaller packages. Even Hadley Wickham acknowledged this problem when he created the `tidyverse` brand.
- Finally, there is the question of the book. As you have noticed, we put a lot of effort into producing a book which is not simply a description of the package's functions. It is a guide on how to use big EO data for land classification, where the actions are illustrated by examples using `sits`. The inspiration for the book was the work "R for Data Science" by Wickham and Golemund.

So, I hope to have managed to explain why `sits` is a single package. Many thanks again for your time and effort.
