### 2025-01-30 #596 @mpadge
Thanks @gilbertocamara. Please accept once again my apologies for this process taking so long. You are nevertheless very aware that this really is an enormous package to review. I have some ideas for how we can nevertheless proceed here, but will just wait for an update from @mikemahoney218.
