### 2025-02-02 #596 @maurolepore
Dear @gilbertocamara and @mpadge this is just to mark the start of my EiC rotation. As a note to myself, I see the communication is fluid and the last interaction was very recent. Things are taking long but it's expected given the size of the package. I'll step back and let things flow. Thanks!
