### 2025-02-02 #596 @gilbertocamara
Dear @maurolepore, many thanks for your contribution!! Since you are an expert on use of geospatial data for forest studies, it seems relevant to inform you that the `sits` package is being using operationally in Brazil to support the production of official maps of LUCC classification in the Amazon and Cerrado biomes by INPE (Brazil's National Institute for Space Research).
