### 2025-03-20 e-sensing/sits#1306 @mdsumner
quoting in 'sits_labels<-' replacement function
I found the quoting in sits_labels.R is possibly unnecessary:

```
#' @name sits_labels<-
```

it currently uses

```
#' @name `sits_labels<-`
```

I managed to build the reference docs with pkgdown by changing this and running (and Yes to the donwload torch and installs)

```R
pkgdown::build_reference()
```
