### 2025-03-20 e-sensing/sits#1307 @mdsumner
sits_regularize: "named" issue
I can't get this regularization process past the "named" check:

```R
library(sits)
s2_cube <- sits_cube(
    source = "DEAUSTRALIA",
    collection = "GA_S2BM_ARD_3",
    ## tiles = c("20LKP", "20LLP"),
    roi = c(lon_min = 147, lon_max = 147.3, lat_min = -42.9, lat_max = -42.5),
    bands = c("RED", "GREEN", "BLUE", "CLOUD"),
    start_date = as.Date("2023-07-01"),
    end_date = as.Date("2024-06-30"),
    progress = FALSE
)
gc_cube <- sits_regularize(
    cube          = s2_cube,
    output_dir    = tempdir(),
    period        = "P15D",
    res           = 560, multicores = 8)
```

I got it to work by putting in an avoid in `api_check.R'

```
return(invisible(unname(x)))
```

I don't know if it really matters if it's "named", but it otherwise worked so maybe can be relaxed here.
