### 2025-03-20 e-sensing/sits#1307 @gilbertocamara
Dear @mdsumner We tested your script in the latest development version of `sits` and it works well. Could you please download it using:
```{r}
devtools::install_github("e-sensing/sits@dev")
```
Please let us know if this version works for you.
