### 2025-03-20 e-sensing/sits#1308 @gilbertocamara
Improving detection of snow cover
**Describe the requested improvement**
In the current version of `sits`, @loreabad6 has indicated that the cloud mask is also indicating areas covered by snow. These areas should be left out of `sits` cloud mask.

**Associated sits API function**
`sits_regularize()`

**Additional context**
This improvement requires updating the configuration files for the image collections handled by `sits`.
