### 2025-03-20 e-sensing/sits#1308 @gilbertocamara
Dear @loreabad6, the issue has been addressed in the development version. Could you please download it using:
```{r}
devtools::install_github("e-sensing/sits@dev")
```
Please let us know if this version works for you.
