### 2025-03-21 e-sensing/sits#1309 @gilbertocamara
Improve function documentation
**Describe the requested improvement**
The ROpenSci review committee recommended a comprehensive review of function documentation. Particular attention should be devoted to functions that are part of the main `sits` workflow.

**Associated sits API function**
- `sits_cube()`
- `sits_cube_copy()`
- `sits_regularize()`
- `sits_merge()`
- `sits_mosaic()`
- `sits_get_data()`
- `sits_apply()`
-  `sits_mixture_model()`
-  `sits_reduce()`
-  `sits_texture()`
- `sits_som()`
- `sits_reduce_imbalance()`
- `sits_train()`
- `sits_tuning()`
- `sits_classify()`
- `sits_smooth()`
- `sits_label_classification()`
- `sits_uncertainty()`
-  `sits_uncertainty_sampling()`
- `sits_combine_predictions()`
- `sits_segment()`
- `sits_accuracy()`
- `sits_sampling_design()`
- `sits_stratified_sampling()`

**Additional context**
Progress on this work should be reported to @mdsumner @edzer @loreabad6 @nowosad @mpadge @nowosad @alexgleith
