### 2025-03-21 e-sensing/sits#1309 @gilbertocamara
Revised up to 21 March:
- `sits_classify()`
- `sits_cube()`
- `sits_get_data()`
- `sits_apply()`
