### 2025-03-21 e-sensing/sits#1310 @gilbertocamara
Improve and review `plot`
**Describe the requested improvement**
Reported by @nowosad:

The current plot() function on "raster_cube" returns an error:
`Error: .check_bw_rgb_bands: either 'band' parameter or 'red', 'green', and 'blue' parameters should be informed.`

Have you considered creating a new plotting mechanism that would work as a default and present information about the data cube (e.g., a time series of the available data or a simple map with the spatial extension of the data or a combination of both?)

I found plot.raster_cube confusing, until I looked at the documentation, it's only plotting the first available tile and seemingly can only do that. I wish it could add to an existing plot, and perhaps do that iteratively.

**Associated sits API function**
`sits_plot()`

**Additional context**
Progress on this work should be reported to @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith
