### 2025-03-21 e-sensing/sits#1311 @gilbertocamara
Improve creation of data cubes from PlanetScope data
**Describe the requested improvement**
As requested by @loreabad6:

Currently, for Planet images to be included in a `sits` data cube, file names have to be changed to contain information on tile, date and band. Doing so is cumbersome for many users who would like a simpler way to include these images.

**Associated sits API function**
`sits_cube()`

**Additional context**
Progress on this work should be reported to @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith
