### 2025-03-21 e-sensing/sits#1312 @gilbertocamara
Improve documentation of configuration files
**Describe the requested improvement**
Requested by @loreabad6 and @alexgleith

For each image collection available in a cloud provider (e.g., "SENTINEL-1-RTC" in "MPC") `sits` relies on configuration files which store information such as: minimum/maximum value, data quantization, scale and offset, cloud masks. There is one YAML configuration file for each provider (all located at `/inst/extdata/source`). For example, MPC collections are described in `config_source_mpc.yaml`. The technical appendix of the `sits` book should include detailed information on how to configure such files.

**Associated sits API function**
`sits_cube()`

**Additional context**
Progress on this work should be reported to @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith
