### 2025-03-21 e-sensing/sits#1313 @gilbertocamara
Provide conversion of `sits` objects to/from other `R` packages
**Describe the requested improvement**
Users would benefit from the capacity to convert of `sits` objects to/from other `R` spatial packages, such as `terra`, `stars` and `sf`. Converting from a stars object to a sits cube could facilitate user experience, particularly for local files.

**Associated sits API function**
`sits_cube()`
`sits_as_sf()`

**Additional context**
Progress on this work should be reported to @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith
