### 2025-03-21 e-sensing/sits#1314 @gilbertocamara
Improve the documentation by include an Overview chapter with concise, objective use cases
**Describe the requested improvement**
In their ROpenSci review of `sits`, @edzer, @nowosad, @mdsumner, @loreabad6, @alexgleith and @mpadge pointed out the need for for the documentation to provide a concise overview of the main workflows to be followed by `sits` users. The review committee express their concern that currently users need to read a lot of chapters to understand how to do an end-to-end procedure.  @edzer and @loreabad6 emphasized the importance of providing concise and objective use cases to increase adoption.

**Additional context**
Add any other context about the improvement here.
Progress on this work should be reported to @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith
