### 2025-03-22 e-sensing/sits#1310 @Nowosad
One possible idea is to have a more abstract representation of the data than a direct visualization of the rasters, e.g., it could be a set of bounding boxes (or one general bounding box) with some text information added (e.g., about the time frame, data source, spatial resolution, etc.).
