### 2025-03-22 e-sensing/sits#1310 @gilbertocamara
Thanks, @Nowosad! Have you tried to use `sits_view()`?  It would be valuable for us to get your comments on this function.
