### 2025-03-23 e-sensing/sits#1313 @gilbertocamara
For information to @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith:

`sits_as_stars()`  is already available in version 1.5.3-dev (development version)
