### 2025-03-24 #596 @alexgleith
Completed @ropensci-review-bot
