### 2025-03-24 e-sensing/sits#1309 @gilbertocamara
For information to @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith:

The `sits` function documentation has been revised and improved. Your comments will be most appreciated. To see the improvements, please download the "dev" version:

```{r}
devtools::install_github("e-sensing/sits@dev")
```
