### 2025-03-24 e-sensing/sits#1313 @gilbertocamara
For information to @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith:

`sits` now provides functions `sits_as_sf`, `sits_as_stars` and `sits_as_terra` to convert from `sits` objects to `stars`, `terra`, and `sf` one. As suggested by @edzer, we will also include a `st_as_stars.raster_cube` function so that `stars` users can directly import data cubes from `sits`

The new functions can be assessed and tested in the "dev" version:
```{r}
devtools::install_github("e-sensing/sits@dev")
```
