### 2025-03-24 e-sensing/sits#1307 @mdsumner
ok thanks!  sorry I thought I had used dev branch.

I'm getting other errors atm but I think it's a transient issue with the source, it's definitely not this
