### 2025-03-25 #596 @mdsumner
Here is a meeting summary from the committee review, after a few days grace to allow feedback.

### rOpenSci {sits} review 2025-03-19

Attended by Gilberto Camara, Mark Padgham, Lorena Abad, Jakub Nowosad, Alex Leith, Edzer Pebesma, Michael Sumner

## Meeting Summary and Package Discussion

In the meeting, Mark and Gilberto discussed the recording and summarization of their discussions. They agreed to have Gilberto present the most general comments and responses, with Mark controlling the screen.

## Package Splitting for Flexibility and Reusability

Jakub proposed splitting the current package into smaller ones, allowing for more flexibility and reusability of its components. He suggested creating satellite packages for specific functionalities, such as Bayesian smoothing, to be used in other processes. Gilberto raised the issue of branding and visibility, emphasizing the importance of maintaining a cohesive package for user convenience. Alex agreed with Jakub's proposal, highlighting the benefits of breaking out core functions for better maintainability and reusability. Gilberto also discussed the interaction between the package and other tools, particularly the issue of handling time series data in {sf}. Gilberto argued that major refactors would entail a massive duplication of effort,  Michael agreed with Gilberto's points and suggested that major refactoring is of interest for the long-term but out of current scope.

## Sits Cube Conversion Challenges Discussed

The team discussed the benefits and challenges of converting a sits cube to a terra object. Michael emphasized the need for a clear API to capture the raster image as a dataset, regardless of the format. Lorena suggested that converting from a stars object to a sits cube could facilitate user experience, particularly for local files. However, Gilberto pointed out the difficulty of specifying compression factors, bit depth, and masking values for each band, which is a requirement for many packages, including Google Earth Engine. The team agreed that this issue needs further exploration and resolution.

## Sentinel 2 Collection Mode Configuration

Gilberto discussed the collection mode and data storage process, highlighting the need for a configuration file for each provider. Lorena raised a concern about the sentinel 2 collection, specifically the issue of number 11 being used for both clouds and snow/ice (the Sentinel-2 SCL band, value 11 stands for “cloud” but has been erroneously included as part of the default cloud mask for this product in sits). Alex discussed that user-customized config file are a solution, but acknowledged that this could be complex and not user-friendly. Lorena proposed providing an example of the configuration file in the technical annex of the book as a simple solution.

## Addressing Planet Scope Data Challenges

The team discussed the challenges they face with the planet scope data, particularly the difficulty in finding the date of the image on the tiff file. They acknowledged the need for better documentation and an overview of the package, as well as improvements to the help functions. However, there was concern about the scope and timing of the project, with a suggestion to focus on capturing a review response collectively. The team agreed on the importance of addressing all comments and welcomed further input.

## Improving Package Alignment and Imagery

Alex suggested focusing on the issues that would prevent the review from proceeding, rather than discussing every comment. Michael agreed, expressing concern about the level of detail being discussed rather than the goal of planning the review process. Mark proposed a high-level discussion around the non-package authors, focusing on one or two key issues to improve the package. Michael expressed a personal interest in the details of how GDAL (warper API) is used internally. Edzer stated that the package uses {gdalcubes} for alignment and image regularization, Michael clarified that reference was the warp usage by the plot function in sits via sf's 'gdal_utils'.

## Suggestions of Key Package Improvements and Applications

 Edzer suggests creating simple, accessible use cases to increase adoption. Lorena recommends developing a technical annex for advanced users to customize default settings. Jakub highlights the need for better cloud coverage filtering and data cube visualization. Alex proposes simplifying the package by removing some features. The group also discusses the parallels with the {targets} package in R and the potential for generalizing workflows. Gilberto provides context on sits real-world applications in deforestation monitoring and mentions ongoing efforts to combine sits with OpenEO and develop a Python version of the package.

## Peer Review Process for Large Software

The meeting discusses a new approach to peer review for a large software package. Mark explains that due to the package's size, not all reviewer concerns can be addressed immediately. He suggests documenting unaddressable issues as GitHub issues for future consideration. Michael volunteers to summarize the meeting and post it in the review thread. Gilberto agrees to wait for the review before responding, but may address urgent user issues immediately. Mark emphasizes the importance of this committee review process as a potential new model for software peer review. He invites all participants to provide feedback on the process to help refine it for future use.

As  a part of this process the following issues have been opened:

https://github.com/e-sensing/sits/issues/1306

https://github.com/e-sensing/sits/issues/1307

https://github.com/e-sensing/sits/issues/1308

https://github.com/e-sensing/sits/issues/1309

https://github.com/e-sensing/sits/issues/1310

https://github.com/e-sensing/sits/issues/1311

https://github.com/e-sensing/sits/issues/1312

https://github.com/e-sensing/sits/issues/1313

https://github.com/e-sensing/sits/issues/1314
