### 2025-03-26 e-sensing/sits#1308 @loreabad6
Dear @gilbertocamara, thank you very much for the fast handling of this issue. I can confirm that the fix addresses the cloud masking as expected.
