### 2025-03-26 e-sensing/sits#1311 @mdsumner
@loreabad6 would you meet with me to show me what the process is like when you download these files?

I don't have planet scope access, but it's pretty likely someone I work with does and I will ask around the data centre next week
