### 2025-03-28 e-sensing/sits#1310 @Nowosad
I think that `sits_view()` has the same limitation as the `plot()` function -- it does not return anything if just a sits object is provide (without additional arguments).
