### 2025-03-28 #596 @mpadge
Thanks @mdsumner.

### Important next task

 @loreabad6, @Nowosad, @edzer, @alexgleith could each of you issue a statement here along the lines of:

> As part of the review committee, I also support the [above review comments](https://github.com/ropensci/software-review/issues/596#issuecomment-2752429783),

(Or whatever else you'd like to say).

This is a vital step to allow me as editor to log all of you as official reviewers of this package. We also log times spent reviewing, for which I'll put each of you at three hours (90 min review + time before and after going through package). If anybody devoted more time than that to going through the package, please note that in your comment how many hours should be logged. Thanks!

----

@gilbertocamara It seems that the above-linked issues cover most aspects discussed throughout the review process. Once all have been addressed, please post your response here, and ensure that you explicitly refer back to the review text posted above. Feel free to ask general questions of the reviewers here, or to discuss more specific issues in the repo itself.
