### 2025-04-02 #596 @loreabad6
As part of the review committee, I also support the https://github.com/ropensci/software-review/issues/596#issuecomment-2752429783.
