### 2025-04-03 #596 @Nowosad
As part of the review committee, I also support the [above review comments](https://github.com/ropensci/software-review/issues/596#issuecomment-2752429783). It took me about ~3,5 hour before the meeting to review the package and its book.
