### 2025-04-04 e-sensing/sits#1318 @mpadge
Update author list
Dear @gilbertocamara and package authors, could you please ensure that all email addresses in your DESCRIPTION file are valid, and that all authors are correctly listed at the top of [the rOpenSci software review issue](https://github.com/ropensci/software-review/issues/596#issue-1784982057)? All authors in DESCRIPTION have email addresses, yet not all of these match addresses currently affiliated with GitHub handles given in software review issue. Thanks! And ping @maelle
