### 2025-04-06 e-sensing/sits#1318 @gilbertocamara
Dear @mpadge @maelle I have updated the authors list and included them as requested. Note that the package has four main authors: @gilbertocamara, @rolfsimoes, @M3nin0, @lorenalves and @OldLipe. These are the main codes who are in charge of bug fixes and improving the API. The other people listed in the DESCRIPTION are contributors.
