### 2025-04-06 e-sensing/sits#1311 @gilbertocamara
Dear @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith, there has been a major update in the documentation of function `sits_cube` to clarify how to create different types of cubes from STAC services and local files. Please see the updates:
```{r}
devtools::install_github("e-sensing/sits@dev")
?sits_cube
```
