### 2025-04-06 e-sensing/sits#1310 @gilbertocamara
Dear @Nowosad, many thanks for your message. The new versions of `plot` and `sits_view` in version 1.5.3 address your concerns (we hope!), as follows.  Consider the following MWE
````
devtools::install_github("e-sensing/sits@dev")
library(sits)
s2_cube <- sits_cube(
     source = "MPC",
     collection = "SENTINEL-2-L2A",
     tiles = "20LKP",
     start_date = "2021-03-01",
     end_date = "2021-09-30"
)
plot(s2_cube)
sits_view(s2_cube)
```
There is no longer a need to specify  bands for plotting and visualisation. We provide a set of common color composites for all collections, and we choose whatever composite is available based on the requested bands. We have a priority list of composites. By default, we try to get one band from the visual spectrum, one from the NIR, and one from the SWIR.

I hope this solves your issue. For reference of our reviewers (@mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith) we consider to have solved this matter.
