### 2025-04-06 e-sensing/sits#1313 @gilbertocamara
For reference of @mdsumner @edzer @loreabad6 @Nowosad @mpadge @Nowosad @alexgleith

We have now functions that export `sits` objects to `terra` and `stars`. Next, we will try to figure out how to do the reverse conversion.
