### 2025-04-06 #596 @gilbertocamara
Dear @mpadge, @maelle, @Nowosad, @alexgleith, @mdsumner, @edzer, @loreabad6, this is an update on the improvements requested by the reviewers:

(a) [Improve function documentation](https://github.com/e-sensing/sits/issues/1309): major update of most sits functions, including `sits_cube`, `sits_get_data`, `sits_classify`, `sits_regularize`. Mention of the sits worflow has been included in the main sits functions, showing a step-by-step procedure.

(b) [Provide conversion of sits objects to/from other R packages](https://github.com/e-sensing/sits/issues/1313): conversion from `sits` to `stars` and to `terra` working.

(c) [Quoting in 'sits_labels<-' replacement function](https://github.com/e-sensing/sits/issues/1306): review done, we prefferred to keep the current version.

(d) [Improving detection of snow cover](https://github.com/e-sensing/sits/issues/1308): done and verified by @loreabad6.

(e)[Improve plot](https://github.com/e-sensing/sits/issues/1310): requested by @Nowosad, completed, waiting for verification by @Nowosad.

(f) [Improve creating local cubes](https://github.com/e-sensing/sits/issues/1311): major update on the documenation of `sits_cube`. Planet files remain a hard problem. As acknowledged by @loreabad6 and @mdsumner, some of these problems have to do with the data provider.

We consider that the requested changes in the code and in the function documentation have been mostly addressed. We are now going to focus on improving the online book, as requested.

Many thanks for your excellent contribution,
@gilbertocamara @rolfsimoes @m3nino @OldLipe
