+++ R/sits_cube_local.R
@@ -197,7 +197,7 @@ sits_cube.local_cube <- function(
 #'         parse_info = c("satellite", "sensor", "tile", "band", "date")
 #'     )
 #'     # segment the vector cube
-#'     segments <- sits_segment(
+#'     segs_cube <- sits_segment(
 #'         cube = modis_cube,
 #'         seg_fn = sits_slic(
 #'             step = 10,
@@ -209,14 +209,57 @@ sits_cube.local_cube <- function(
 #'             ),
 #'             output_dir = tempdir()
 #'      )
+#'      plot(segs_cube)
+#'
 #'      # recover the local segmented cube
-#'      segment_cube <- sits_cube(
+#'      local_segs_cube <- sits_cube(
 #'         source = "BDC",
 #'         collection = "MOD13Q1-6.1",
 #'         raster_cube = modis_cube,
 #'         vector_dir = tempdir(),
 #'         vector_band = "segments"
 #'      )
+#'      # plot the recover model and compare
+#'      plot(local_segs_cube)
+#'
+#'      # classify the segments
+#'      # create a random forest model
+#'      rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
+#'      probs_vector_cube <- sits_classify(
+#'         data = segs_cube,
+#'         ml_model = rfor_model,
+#'         output_dir = tempdir(),
+#'         n_sam_pol = 10
+#'      )
+#'      plot(probs_vector_cube)
+#'
+#'      # recover vector cube
+#'      local_probs_vector_cube <-  sits_cube(
+#'         source = "BDC",
+#'         collection = "MOD13Q1-6.1",
+#'         raster_cube = modis_cube,
+#'         vector_dir = tempdir(),
+#'         vector_band = "probs"
+#'      )
+#'      plot(local_probs_vector_cube)
+#'
+#'      # label the segments
+#'      class_vector_cube <- sits_label_classification(
+#'         cube = probs_vector_cube,
+#'         output_dir = tempdir(),
+#'      )
+#'      plot(class_vector_cube)
+#'
+#'      # recover vector cube
+#'      local_class_vector_cube <-  sits_cube(
+#'         source = "BDC",
+#'         collection = "MOD13Q1-6.1",
+#'         raster_cube = modis_cube,
+#'         vector_dir = tempdir(),
+#'         vector_band = "class"
+#'      )
+#'      plot(local_class_vector_cube)
+#'
 #'}
 #'
 #' @export
@@ -233,6 +276,8 @@ sits_cube.vector_cube <- function(
         multicores = 2L,
         progress = TRUE) {
 
+    # set caller to show in errors
+    .check_set_caller("sits_cube_vector_cube")
     # obtain vector items
     vector_items <- .local_vector_items(
         source = source,
@@ -275,23 +320,22 @@ sits_cube.vector_cube <- function(
 #'                     \code{"CDSE"}, \code{"DEAFRICA"}, \code{"DEAUSTRALIA"},
 #'                     \code{"HLS"}, \code{"PLANETSCOPE"}, \code{"MPC"},
 #'                     \code{"SDC"} or \code{"USGS"}. This is the source
-#'                     from which the data has been downloaded.
-#' @param collection   Image collection in data source.
+#'                     from which the original data has been downloaded.
+#' @param collection   Image collection in data source from which
+#'                     the original data has been downloaded.
 #'                     To find out the supported collections,
 #'                     use \code{\link{sits_list_collections}()}).
 #' @param ...          Other parameters to be passed for specific types.
 #' @param data_dir     Local directory where images are stored
 #' @param tiles        Tiles from the collection to be included in
-#'                     the cube (see details below).
+#'                     the cube.
[…]
