+++ tests/testthat/test-cube-usgs.R
@@ -24,9 +24,9 @@ test_that("Creating LANDSAT cubes from USGS with ROI", {
     bbox_cube_1 <- sits_bbox(.tile(l8_cube_usgs), as_crs = "EPSG:4326")
     expect_true(bbox_cube["xmax"] >= bbox_cube_1["xmax"])
     expect_true(bbox_cube["ymax"] >= bbox_cube_1["ymax"])
-    r_obj <- .raster_open_rast(l8_cube_usgs$file_info[[1]]$path[1])
+    rast <- .raster_open_rast(l8_cube_usgs$file_info[[1]]$path[1])
     tile_nrows <- .tile_nrows(l8_cube_usgs)[[1]]
-    expect_true(.raster_nrows(r_obj) == tile_nrows)
+    expect_true(.raster_nrows(rast) == tile_nrows)
 })
 
 test_that("Creating LANDSAT cubes from USGS with WRS", {
@@ -48,7 +48,7 @@ test_that("Creating LANDSAT cubes from USGS with WRS", {
     testthat::skip_if(purrr::is_null(l8_cube_223067), "USGS is not accessible")
     expect_true(all(sits_bands(l8_cube_223067) %in% c("NIR08")))
     expect_equal(nrow(l8_cube_223067), 1)
-    r_obj <- .raster_open_rast(l8_cube_223067$file_info[[1]]$path[1])
+    rast <- .raster_open_rast(l8_cube_223067$file_info[[1]]$path[1])
     tile_nrows <- .tile_nrows(l8_cube_223067)[[1]]
-    expect_true(.raster_nrows(r_obj) == tile_nrows)
+    expect_true(.raster_nrows(rast) == tile_nrows)
 })
