### 2025-04-07 e-sensing/sits#1310 @Nowosad
@gilbertocamara Thanks. Could you also add an explanation of what is shown by default in the documentation (help file)?

I do not fully understand yet what is shown -- one tile or many tiles merged? Is the outcome composite the result of averaging many dates, or is it the median date in the data? Etc.
