### 2025-04-07 e-sensing/sits#1313 @Nowosad
@gilbertocamara do you have examples of this functions in action? I've tried the code below, but got an error:

```r
# devtools::install_github("e-sensing/sits@dev")
library(sits)
s2_cube <- sits_cube(
  source = "MPC",
  collection = "SENTINEL-2-L2A",
  tiles = "20LKP",
  start_date = "2021-03-01",
  end_date = "2021-09-30"
)
system.time({
  s2_cube_terra = sits_as_terra(s2_cube)
})
#> Error: [rast] number of rows and/or columns do not match
```
