### 2025-04-07 e-sensing/sits#1313 @gilbertocamara
Dear @Nowosad, many thanks for sharing the code. This is a `terra` problem, not a `sits` problem. The `terra` package does not accept spatial rasters with different sizes and/or resultions. The following code works:

```{r}
library(sits)
s2_cube <- sits_cube(
  source = "MPC",
  collection = "SENTINEL-2-L2A",
  tiles = "20LKP",
  bands = c("B04", "B03", "B02"),
  start_date = "2021-03-01",
  end_date = "2021-09-30"
)
  s2_cube_terra <- sits_as_terra(s2_cube)
```
