### 2025-04-08 e-sensing/sits#1313 @gilbertocamara
Dear @Nowosad, to export TIFF files with different resolutions to `terra`, one could use `sits_regularize` to convert all files to the same resolution and bonding box. The resulting data cube can be exported to `terra`, on a date-by-date basis. This arrangement is guaranteed to work. If this does not work, we have a bug in `sits`.
