### 2025-04-10 e-sensing/sits#1310 @Nowosad
Thanks @gilbertocamara. One small (optional, as always) suggestion: maybe it would be good to add a text message when the plot() is used without any other arguments than the cube object. Something like:

`Composite (B11/B08/B02) from 2021-07-02 imagery with minimal cloud cover`
