### 2025-04-10 e-sensing/sits#1313 @Nowosad
@gilbertocamara maybe in the case of the user trying to convert the sits object to spatial rasters with different sizes and/or resolutions, a custom error message (with your suggestion about sits_regularize) could be added
