### 2025-04-12 #596 @gilbertocamara
Dear @mpadge @maelle Please allow @loreabad6 @Nowosad @alexgleith @mdsumner @edzer at least until the end of May to complete their reviews. The comments by reviewers have been as helpful as they are demanding. Their remarks can be broken into two partes: (a) package code and function documentation; (b) important changes in the online book. We have almost completed part (A) and will be starting on part (B) next week.  It will take us about a month to make all the requested changes in the online book. So, please, allow more time by the reviewers to be able to complete their work.
