### 2025-04-12 #596 @mdsumner
I haven't been able to get to the examples that Lorena provide me but will try to very soon, I also wanted to meet to see how the process looked but I couldn't in the end because of timetabling (sorry @loreabad6), we might still be able to do that and I also asked to sign up myself so I may get to have a closer look. I will also ask a colleague who I think has access to Planet.

I wanted to discuss the GDAL topic with the authors, my poorly expressed and interrupted "one thing", but frankly I consider that out of scope here. The efforts of the sit authors within a very challenging software ecosystem are absolutely stellar and while I hope to get to more considered feedback again I don't see any roadblocks at all, the package is an excellent body of work and has strived really hard to make the best of the available software and plumbing required. This scene is also every evolving, STAC being born in "always traverse the tree" json is now starting to come down to tidier tabular principles with GeoParquet, GTI, and sits itself was a strong early iilustration in treating the "cube" as a table of related but as yet uneresolved inputs that are *always* unhidden and available to user at any level. I think it's excellent.

For preparation to the committee meeting, and in subsequent light reading and exploring I spent 8.5 hours reviewing.
