### 2025-04-12 #596 @gilbertocamara
Dear @mdsumner, your support and your help are greatly appreciated. For your reference and that of @Nowosad, @alexgleith, @edzer, @loreabad6, @mpadge, please allow me to try to explain how we use GDAL in `sits`:

- (1) Calls to GDAL utilities (e.g., `gdal_translate`, `gdal_warp`) are handled through `sf` which provides a reliable and efficient interface to them. These calls are part of `api_gdal.R`. Kudos to @edzer for a masterful work on `sf`!
- (2) Using GDAL to create data cubes is done by an interface to `gdalcubes`, an excellent package done by @applemar.
- (3) Reading and writing to raster files, and extracting properties of such files, is done using `terra`.

Every six months, @rolfsimoes, @OldLipe, @M3nin0 and me discuss replacing `terra` with calls to GDAL to reduce the number of dependencies in `sits`. Then we look at the amount of functions that `terra` provides and consider it's not worth it. In any case, there are no direct calls to `terra` in the `sits` code. All codes are handled by the `api_raster.R` which encapsulates `terra` so that eventually we will be able to replace it, should the need arise.

Another key point is that, by relying on `sf` to provide an interface to GDAL, we delegate to @edzer the tough job of making sure the right GDAL libraries are loaded. It is a wonderful work. Rumor has it that @edzer had to use some magical powers to convince the R zealous guardians to allow `sf` and its predecessors to load GDAL on CRAN.

Thus, long story short, all geospatial R developers rely (whether they know it or not) on @ezder to carry the heavy load. I hope this clarifies the views of the `sits` developers regarding GDAL.
