### 2025-04-13 #596 @gilbertocamara
Dear @mdsumner, your generous comments are much appreciated.

There are two complementary viewpoints on STAC. From a conceptual Data Science point of view, STAC is a step in reverse gear. Before STAC came to prominence, @alexgleith and his team at DE Australia had shown there was a better solution based on well-established object-relational database structures. Had the solution originally conceived by @alexgleith and DE Australia prevailed, we would have a much better mechanism for setting up queries to cloud providers. The success of RDBMS in commercial markets shows that the ODC original solution was the best option.

STAC is the VHS of the big EO data world (or should we say Windows-95?),  Despite our grievances, STAC is here to stay; even ODC had to adapt to this reality. From a Computer Science point of view, STAC smells of the 1970s. Nevertheless, there are many 1970s technologies still very much alive (think C, UNIX, and LaTeX). My CS side regrets that we have been left with STAC. At the end of the day, clever people such as the STAC lead developers can do good things even with questionable specifications.  BTW, I am a proud user of LaTeX and have no plans to change anytime soon.

Reality always trumps utopias. So our team of @rolfsimoes, @OldLipe and @M3nin0 also had to develop `rstac` to complement `sits`. Life is short.
