### 2025-04-14 #596 @mpadge
@edzer @alexgleith A repeat of [previous request](https://github.com/ropensci/software-review/issues/596#issuecomment-2760982712): Could you both make brief statements here like [this](https://github.com/ropensci/software-review/issues/596#issuecomment-2772835335) and [this](https://github.com/ropensci/software-review/issues/596#issuecomment-2776130707) that you were part of the review committee, and support the [review posted by @mdsumner](https://github.com/ropensci/software-review/issues/596#issuecomment-2752429783). Thank you
