### 2025-04-14 #596 @edzer
As part of the review committee, I also support the [above review comments](https://github.com/ropensci/software-review/issues/596#issuecomment-2752429783), as well as the summary of the review discussion posted by @mdsumner .
