### 2025-04-16 e-sensing/sits#1310 @gilbertocamara
@Nowosad, all your suggestions have been implemented in the new sits version (1.5.3)
