### 2025-04-16 #596 @gilbertocamara
For information of @edzer @loreabad6 @alexgleith @Nowosad @mdsumner @mpadge, I believe we have completed all issues related to coding of `sits` that have been requested by you and that we could address in a reasonable time. The most relevant changes have been a complete revision of function documentation.

Although not specifically requested by you, we also performed a significant code revision, based on guidance from `lintr`. As those of you who have used `lintr`, this package is notorious for pointing out many significant possible improvements by also some unreliable and undesirable changes. We took what we considered to be important advice and revised the code from top to bottom. The most changes were required by the following linters:

````
- improving function returns to emphasize functional programming features of R;
- removing old code no longer relevant and unused variables;
- avoid implicit integers;
- use lambda functions whereever applicable;
```

We also revised all checking functions involved in pre-conditions. Thanks for making us do a lot of work now that has resulted in much improvement of the software!

for reference, copied to @rolfsimoes @m3nin0 @OldLipe
