### 2025-04-17 e-sensing/sits#1322 @gilbertocamara
Revise `sits` code to harmonize API usage and conform to `best practices` and `lintr`
**Describe the requested improvement**
Review the package code using `lintr` and `best practices` tools. Address all relevant issues and improve the code as required, with emphasis on:

- Strive to make consistent use of the `sits` API, ensuring that all pre- and post-conditions are handled by functions in `api_check.R`.
- Improve function returns to emphasise functional programming features of R;
- Remove old code no longer relevant and unused variables;
- Avoid implicit integers;
- Use lambda functions appropriately;
- Correct poor usages of paste;
