### 2025-04-17 #596 @gilbertocamara
Hi @mpadge, the reviewers made two types of requests: (a) improvements on the package code and function documentation; (b) Improvements on the online book. Thus we created two sets of issues:

As a part of this process, the following issues have been opened:

- Issues related to coding and function documentation: all of these issues have been addressed; either they have let to code improvements or were considered not to be related to the `sits` development itself. See below

[#1306 quoting in 'sits_labels<-' replacement function](https://github.com/e-sensing/sits/issues/1306) - ADDRESSED

[#1307: sits_regularize: "named" issue](https://github.com/e-sensing/sits/issues/1307) - ADDRESSED

[#1308 Improving detection of snow cover](https://github.com/e-sensing/sits/issues/1308) - SOLVED

[#1309 Improve function documentation](https://github.com/e-sensing/sits/issues/1309) - SOLVED

[#1310 Improve and review plot](https://github.com/e-sensing/sits/issues/1310) - SOLVED

[#1311 Improve creation of data cubes from PlanetScope data](https://github.com/e-sensing/sits/issues/1311) - ADDRESSED

[#1313 Provide conversion of sits objects to/from other R packages](https://github.com/e-sensing/sits/issues/1313) - SOLVED

[#1322 Revise sits code to harmonize API usage and conform to best practices and lintr](https://github.com/e-sensing/sits/issues/1322) - SOLVED

- Issues related to the online book which have not been addressed yet

[#1312 Improve documentation of configuration files](https://github.com/e-sensing/sits/issues/1312)

[#1314 Improve the documentation by include an Overview chapter with concise, objective use cases](https://github.com/e-sensing/sits/issues/1314)

Since we consider to have addressed and solved all coding-related issue, we will now concentrate on the online book. I would like to ask @mpadge @loreabad6 @edzer @Nowosad @mdsumner @alexgleith to signal if they disagree with our decision. Otherwise we will concentrate on the book. We will keep you informed on progress on the book.

Best regards
@gilbertocamara @rolfsimoes @OldLipe @m3nin0
