### 2025-04-17 #596 @mpadge
Sounds great @gilbertocamara, please continue then with the book updates, and let us know here when you're done with those. Thanks as always for the very quick and informative response.
