### 2025-04-20 e-sensing/sits#1307 @mdsumner
ok confirmed works, the other problem I mentioned seems to be on one machine only
