### 2025-05-04 #596 @maurolepore
@gilbertocamara just to thank you for sharing your work with rOpenSci. I'm wrapping up my EiC rotation today. I look forward to working with you again.
