### 2025-06-22 #596 @gilbertocamara
Dear @mpadge @edzer @alexgleith @Nowosad @mdsumner,

we are happy to report that all tasks required by the ROpenSci reviewers have been completed. The most laborious one was the documentation review, both in the code and in the book. The sitsbook has been thoroughly revised, as follows:
- the book has been broken in smaller, more concise parts
- there is a new chapter with representative examples that gives a good idea of the functionality of the package
- a new chapter has been added (in the Annex) describing how cloud providers are handled internally in `sits`
- a new chapter on exporting `sits` objects to `R` packages `sf`, `stars` and `terra` is available; it also includes exports of `sits` objects to Python structures such as `pandas` and `xarray`.

Furthermore, `sits` now offers a full Python API. So all `sits` functions can now be called from Python. Python calls are then transformed into calls to R and returned as python objects (e.g., pandas). Apart from the need to install R in the backend, it is a full Python experience.

Dear @mpadge, consider yourself proud. Your effort to coordinate a review of `sits` has made an important difference in the software's quality. Our warm thanks to @edzer @alexgleith @Nowosad @mdsumner, and also "in memoriam" of @abad (Such a sad and early loss!).

My thanks to our `sits` team, @rolfsimoes @oldlipe @M3nin0 for a job well done!
