### 2025-06-23 #596 @mpadge
Thank you very much indeed @gilbertocamara @rolfsimoes @OldLipe @M3nin0 for all of your work. I'm extremely pleased to hear that the review process has helped so much.

@edzer @alexgleith @Nowosad @mdsumner Could you all please have a look through relevant issues in the sits package linked above, and/or at least peruse the now markedly improved [book](https://e-sensing.github.io/sitsbook/) to ensure that you're happy with all changes? Note that the final phases of our review process often involve a few minor back-and-forth questions and comments for final tweaks, so please feel free to make any minor additional suggestions here. Once you're all happy, could each of you please submit [our reviewer approval template](https://devguide.ropensci.org/approval2template.html)?

Thank you all, and thank you @gilbertocamara for expressing your sorrow at the loss of our dear colleague Lorena.
