### 2025-06-24 #596 @mdsumner
## Reviewer Response

I'm extremely impressed with this package and with the thorough and detailed responses of the authors to review.

#### Final approval (post-review)

Estimated hours spent reviewing: 8.5 hours.

Vale Lorena. 😢
