### 2025-06-30 #596 @edzer
## Reviewer Response

The SITS package is an important contribution to the R community, as well as to the wider remote sensing community (also beyond R). In terms of methodology, it is well state of the art.

#### Final approval (post-review)

Estimated hours spent reviewing: 4h
