### 2025-07-01 #596 @mpadge
Thank you so much to everybody involved here - to @edzer @Nowosad @mdsumner for the extremely helpful reviews, to @gilbertocamara @rolfsimoes, @OldLipe, @M3nin0, @lorenalves to all the work you've put in to improving the package. And we will remain grateful in memorium to Lorena Abad. Before describing the next steps, the following represents the most important next step for rOpenSci itself.

### Committee Review follow-up

@edzer @Nowosad @mdsumner Please read this section! The remainder is for package authors, so you may ignore that.

This entire process was devised for this uniquely large and complex submission. We've had similar submissions in the past which may also have been better handled through such a committee review process, and expect more  to come in the future. So we're going to take what we've learnt here and formally offer committee review as an alternative to our standard review. We would greatly appreciate the help of all involved here in documenting the process, particularly through:

- A blog post which we will announce this new review possibility, and which will include a description of the process as it happened here. All involved will be co-authors of that blog post, unless you explicitly state otherwise. We will ask for feedback from the pull request to our website when we get that far.
- Updates to the documentation for our review process, primarily in the [_Editors_ chapter of our _Dev Guide_](https://devguide.ropensci.org/softwarereview_editor.html), for which we will also ask for your feedback there.

This is an exciting development for us as an organization, for which we are grateful to all involved here. Great work!

---

### Next steps

I will approve the package in a moment, but a few notes first, because that will prompt an extended list of instructions from the bot.

The first section of the bot's list will include instructions for transferring the repository to the rOpenSci GitHub org. We haven't discussed this yet, but I presume that you will want the package to remain in the `github.com/e-sensing` organization? That's perfectly okay,  we just need to know to ensure that our system will know where to find your package.

The one relevant bit of the initial bot response under _Repository Transfer_ is an instruction to:

> Delete your current code of conduct file if you had one since rOpenSci's default one will apply, see [devguide.ropensci.org/collaboration.html#coc-file](https://devguide.ropensci.org/collaboration.html#coc-file)

I note that you currently have a CODE_OF_CONDUCT.md file in SITS, which seems to be derived from the [contributor covenant](https://www.contributor-covenant.org/version/2/1/code_of_conduct/). From our side, it is easier and safer for you to remove that and allow the [rOpenSci code of conduct](https://ropensci.org/code-of-conduct/) to automatically apply. (With one immediate benefit for you that we also have [a Portugese version](https://ropensci.org/pt/c%C3%B3digo-de-conduta/)! 😄 ) In that case, instructions are in [this section of our _Dev Guide_](https://devguide.ropensci.org/maintenance_collaboration.html#coc-file). We are nevertheless open to discussing alternatives if you prefer to retain your current Coc, but please note that this can create difficulties in resolving issues when two potentially divergent codes of conduct must be reconciled, and so we do hope that you'll agree to transfer to our CoC. Please let us know.

---

Following that there is also a statement from the bot to remove any `pkgdown` workflows from GitHub as our doc server will generate sites for you. In your case, you'll obviously want to retain your own separate book compilation and URL, which is fine. Please note, however, that docs will still be generated from your internal package documentation. This may be advantageous, as the lack of direct online function references was discussed during the review process. You'd then potentially be able to link to function documentation from within the main book, which may make it even more helpful. For comparison, we also have [the `eDNAjoint` package](https://github.com/ropensci/eDNAjoint) with the main documentation in [a separate book](https://ednajoint.netlify.app/), as well as [our auto-generated rOpenSci docs site](http://docs.ropensci.org/eDNAjoint/). Your set-up will be the same as this, which I hope will be useful for you.

---

The remainder of the bot instructions should be fairly clear. The r-universe comments just mean that SITS will appear both in your current https://e-sensing.r-universe.dev/ universe, as well as in https://ropensci.r-universe.dev. You can of course choose whichever of those r-universe links suits you best.

As always, I'll happily answer any questions you might have along the way. Finally, thank you @gilbertocamara @rolfsimoes, @OldLipe, @M3nin0, @lorenalves for your patience throughout this very extended process, and please accept my apologies for things taking much longer than initially envisioned. From my side, we are very happy with the end result here, and it seems that the review process has led to excellent improvements to your package. I also note that this approval statement will be issued 2 years minus 2 days after the initial submission 😃
