### 2025-07-01 #596 @gilbertocamara
Dear @mpadge many thanks for your patience and guidance. The package has benefited significantly from the review process. As for your points:

1. We plan to keep the package in the e-sensing organization, since `sits` is part of a larger ecosystem, which includes the book, data, and other packages such as `pysits` (which supports `sits` in Python).
2. We will replace the current CODE_OF_CONDUCT statement with the ROpenSci one. It will be included in the upcoming version that will be submitted to CRAN.
 3. The support of ROpenSci in generating additional documentation for the package is most welcome and appreciated. There is no such a thing as too much documentation.
 4. We are delighted to appear in both R-universes.
